Struct gapi_grpc::google::actions::sdk::v2::actions_sdk_client::ActionsSdkClient

pub struct ActionsSdkClient<T> { /* fields omitted */ }

Actions SDK API which allows developers to build projects using the SDK.

Implementations

impl ActionsSdkClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ActionsSdkClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn write_draft<'_>(
&'_ mut self,
request: impl IntoStreamingRequest<Message = WriteDraftRequest>
) -> Result<Response<Draft>, Status>

Updates the project draft based on the model.

pub async fn write_preview<'_>(
&'_ mut self,
request: impl IntoStreamingRequest<Message = WritePreviewRequest>
) -> Result<Response<Preview>, Status>

Updates the user's project preview based on the model.

pub async fn create_version<'_>(
&'_ mut self,
request: impl IntoStreamingRequest<Message = CreateVersionRequest>
) -> Result<Response<Version>, Status>

Creates a project version based on the model and triggers deployment to the specified release channel, if specified.

pub async fn read_draft<'_>(
&'_ mut self,
request: impl IntoRequest<ReadDraftRequest>
) -> Result<Response<Streaming<ReadDraftResponse>>, Status>

Reads the entire content of the project draft.

pub async fn read_version<'_>(
&'_ mut self,
request: impl IntoRequest<ReadVersionRequest>
) -> Result<Response<Streaming<ReadVersionResponse>>, Status>

Reads the entire content of a project version.

pub async fn encrypt_secret<'_>(
&'_ mut self,
request: impl IntoRequest<EncryptSecretRequest>
) -> Result<Response<EncryptSecretResponse>, Status>

Encrypts the OAuth client secret used in account linking flows. This can be used to encrypt the client secret for the first time (e.g. before the first push or after changing the client secret) or to re-encrypt a client secret using the latest primary key version (considering key rotation).

pub async fn decrypt_secret<'_>(
&'_ mut self,
request: impl IntoRequest<DecryptSecretRequest>
) -> Result<Response<DecryptSecretResponse>, Status>

Decrypts the OAuth client secret used in account linking flows. This can be used to view the client secret (e.g. after pulling a project).

pub async fn list_release_channels<'_>(
&'_ mut self,
request: impl IntoRequest<ListReleaseChannelsRequest>
) -> Result<Response<ListReleaseChannelsResponse>, Status>

Lists all release channels and corresponding versions, if any.

pub async fn list_versions<'_>(
&'_ mut self,
request: impl IntoRequest<ListVersionsRequest>
) -> Result<Response<ListVersionsResponse>, Status>

Lists all versions and their current states.