Module gapi_grpc::google::actions::sdk::v2::interactionmodel

Modules

event_handler

intent

prompt

slot

type

Structs

ConditionalEvent
Registers events that trigger as the result of a true condition.

EventHandler
Defines a handler to be executed after an event. Examples of events are intent and condition based events in a scene.

GlobalIntentEvent
Defines a global intent handler. Global intent events are scoped to the entire Actions project and may be overridden by intent handlers in a scene. Intent names must be unique within an Actions project.

Intent
Intents map open-ended user input to structured objects. Spoken phrases are matched to intents with Google's Natural Language Understanding (NLU). Intent matches can trigger events in your conversation design to progress the user's conversation. The intent name is specified in the name of the file.

IntentEvent
Registers Events which trigger as the result of an intent match.

Scene
Scene is the basic unit of control flow when designing a conversation. They can be chained together with other scenes, generate prompts for the end user, and define slots. The scene name is specified in the name of the file.

Slot
Configuration for a slot. Slots are single units of data that can be filled through natural language (ie. intent parameters), session parameters, and other sources.