Module gapi_grpc::google::ads::admob::v1::network_report_spec

Modules

dimension_filter

sort_condition

Structs

DimensionFilter
Describes which report rows to match based on their dimension values.

SortCondition
Sorting direction to be applied on a dimension or a metric.

Enums

Dimension
The dimensions of the network report. Dimensions are data attributes to break down or refine the quantitative measurements (metrics) by certain attributes, such as the ad format or the platform an ad was viewed on.

Metric
The metrics of the network report. Metrics are quantitative measurements indicating how the publisher business is performing. They are aggregated from the individual ad events and grouped by the report dimensions. The metric value is either integer, or decimal (without rounding).