Module gapi_grpc::google::ads::googleads::v1::common::policy_topic_constraint

Structs

CountryConstraint
Indicates that a resource's ability to serve in a particular country is constrained.

CountryConstraintList
A list of countries where a resource's serving is constrained.

ResellerConstraint
Indicates that a policy topic was constrained due to disapproval of the website for reseller purposes.

Enums

Value
Specific information about the constraint.