Module gapi_grpc::google::ads::googleads::v1::services

Modules

account_budget_proposal_operation

account_budget_proposal_service_client
Generated client implementations.

account_budget_service_client
Generated client implementations.

ad_group_ad_label_operation

ad_group_ad_label_service_client
Generated client implementations.

ad_group_ad_operation

ad_group_ad_service_client
Generated client implementations.

ad_group_audience_view_service_client
Generated client implementations.

ad_group_bid_modifier_operation

ad_group_bid_modifier_service_client
Generated client implementations.

ad_group_criterion_label_operation

ad_group_criterion_label_service_client
Generated client implementations.

ad_group_criterion_operation

ad_group_criterion_service_client
Generated client implementations.

ad_group_criterion_simulation_service_client
Generated client implementations.

ad_group_extension_setting_operation

ad_group_extension_setting_service_client
Generated client implementations.

ad_group_feed_operation

ad_group_feed_service_client
Generated client implementations.

ad_group_label_operation

ad_group_label_service_client
Generated client implementations.

ad_group_operation

ad_group_service_client
Generated client implementations.

ad_group_simulation_service_client
Generated client implementations.

ad_parameter_operation

ad_parameter_service_client
Generated client implementations.

ad_schedule_view_service_client
Generated client implementations.

age_range_view_service_client
Generated client implementations.

apply_recommendation_operation

asset_operation

asset_service_client
Generated client implementations.

bidding_strategy_operation

bidding_strategy_service_client
Generated client implementations.

billing_setup_operation

billing_setup_service_client
Generated client implementations.

campaign_audience_view_service_client
Generated client implementations.

campaign_bid_modifier_operation

campaign_bid_modifier_service_client
Generated client implementations.

campaign_budget_operation

campaign_budget_service_client
Generated client implementations.

campaign_criterion_operation

campaign_criterion_service_client
Generated client implementations.

campaign_criterion_simulation_service_client
Generated client implementations.

campaign_draft_operation

campaign_draft_service_client
Generated client implementations.

campaign_experiment_operation

campaign_experiment_service_client
Generated client implementations.

campaign_extension_setting_operation

campaign_extension_setting_service_client
Generated client implementations.

campaign_feed_operation

campaign_feed_service_client
Generated client implementations.

campaign_label_operation

campaign_label_service_client
Generated client implementations.

campaign_operation

campaign_service_client
Generated client implementations.

campaign_shared_set_operation

campaign_shared_set_service_client
Generated client implementations.

carrier_constant_service_client
Generated client implementations.

change_status_service_client
Generated client implementations.

click_view_service_client
Generated client implementations.

conversion_action_operation

conversion_action_service_client
Generated client implementations.

conversion_adjustment

conversion_adjustment_result

conversion_adjustment_upload_service_client
Generated client implementations.

conversion_upload_service_client
Generated client implementations.

custom_interest_operation

custom_interest_service_client
Generated client implementations.

customer_client_link_operation

customer_client_link_service_client
Generated client implementations.

customer_client_service_client
Generated client implementations.

customer_extension_setting_operation

customer_extension_setting_service_client
Generated client implementations.

customer_feed_operation

customer_feed_service_client
Generated client implementations.

customer_label_operation

customer_label_service_client
Generated client implementations.

customer_manager_link_operation

customer_manager_link_service_client
Generated client implementations.

customer_negative_criterion_operation

customer_negative_criterion_service_client
Generated client implementations.

customer_service_client
Generated client implementations.

detail_placement_view_service_client
Generated client implementations.

dismiss_recommendation_request

dismiss_recommendation_response

display_keyword_view_service_client
Generated client implementations.

domain_category_service_client
Generated client implementations.

dynamic_search_ads_search_term_view_service_client
Generated client implementations.

expanded_landing_page_view_service_client
Generated client implementations.

extension_feed_item_operation

extension_feed_item_service_client
Generated client implementations.

feed_item_operation

feed_item_service_client
Generated client implementations.

feed_item_target_operation

feed_item_target_service_client
Generated client implementations.

feed_mapping_operation

feed_mapping_service_client
Generated client implementations.

feed_operation

feed_placeholder_view_service_client
Generated client implementations.

feed_service_client
Generated client implementations.

gender_view_service_client
Generated client implementations.

generate_keyword_ideas_request

geo_target_constant_service_client
Generated client implementations.

geographic_view_service_client
Generated client implementations.

google_ads_field_service_client
Generated client implementations.

google_ads_service_client
Generated client implementations.

group_placement_view_service_client
Generated client implementations.

hotel_group_view_service_client
Generated client implementations.

hotel_performance_view_service_client
Generated client implementations.

keyword_plan_ad_group_operation

keyword_plan_ad_group_service_client
Generated client implementations.

keyword_plan_campaign_operation

keyword_plan_campaign_service_client
Generated client implementations.

keyword_plan_idea_service_client
Generated client implementations.

keyword_plan_keyword_operation

keyword_plan_keyword_service_client
Generated client implementations.

keyword_plan_negative_keyword_operation

keyword_plan_negative_keyword_service_client
Generated client implementations.

keyword_plan_operation

keyword_plan_service_client
Generated client implementations.

keyword_view_service_client
Generated client implementations.

label_operation

label_service_client
Generated client implementations.

landing_page_view_service_client
Generated client implementations.

language_constant_service_client
Generated client implementations.

location_view_service_client
Generated client implementations.

managed_placement_view_service_client
Generated client implementations.

media_file_operation

media_file_service_client
Generated client implementations.

merchant_center_link_operation

merchant_center_link_service_client
Generated client implementations.

mobile_app_category_constant_service_client
Generated client implementations.

mobile_device_constant_service_client
Generated client implementations.

mutate_job_service_client
Generated client implementations.

mutate_operation

mutate_operation_response

operating_system_version_constant_service_client
Generated client implementations.

paid_organic_search_term_view_service_client
Generated client implementations.

parental_status_view_service_client
Generated client implementations.

payments_account_service_client
Generated client implementations.

product_bidding_category_constant_service_client
Generated client implementations.

product_group_view_service_client
Generated client implementations.

recommendation_service_client
Generated client implementations.

remarketing_action_operation

remarketing_action_service_client
Generated client implementations.

search_term_view_service_client
Generated client implementations.

shared_criterion_operation

shared_criterion_service_client
Generated client implementations.

shared_set_operation

shared_set_service_client
Generated client implementations.

shopping_performance_view_service_client
Generated client implementations.

suggest_geo_target_constants_request

topic_constant_service_client
Generated client implementations.

topic_view_service_client
Generated client implementations.

user_interest_service_client
Generated client implementations.

user_list_operation

user_list_service_client
Generated client implementations.

video_service_client
Generated client implementations.

Structs

AccountBudgetProposalOperation
A single operation to propose the creation of a new account-level budget or edit/end/remove an existing one.

AdGroupAdLabelOperation
A single operation (create, remove) on an ad group ad label.

AdGroupAdOperation
A single operation (create, update, remove) on an ad group ad.

AdGroupBidModifierOperation
A single operation (create, remove, update) on an ad group bid modifier.

AdGroupCriterionLabelOperation
A single operation (create, remove) on an ad group criterion label.

AdGroupCriterionOperation
A single operation (create, remove, update) on an ad group criterion.

AdGroupExtensionSettingOperation
A single operation (create, update, remove) on an ad group extension setting.

AdGroupFeedOperation
A single operation (create, update, remove) on an ad group feed.

AdGroupLabelOperation
A single operation (create, remove) on an ad group label.

AdGroupOperation
A single operation (create, update, remove) on an ad group.

AdParameterOperation
A single operation (create, update, remove) on ad parameter.

AddMutateJobOperationsRequest
Request message for [MutateJobService.AddMutateJobOperations][google.ads.googleads.v1.services.MutateJobService.AddMutateJobOperations]

AddMutateJobOperationsResponse
Response message for [MutateJobService.AddMutateJobOperations][google.ads.googleads.v1.services.MutateJobService.AddMutateJobOperations]

ApplyRecommendationOperation
Information about the operation to apply a recommendation and any parameters to customize it.

ApplyRecommendationRequest
Request message for [RecommendationService.ApplyRecommendation][google.ads.googleads.v1.services.RecommendationService.ApplyRecommendation].

ApplyRecommendationResponse
Response message for [RecommendationService.ApplyRecommendation][google.ads.googleads.v1.services.RecommendationService.ApplyRecommendation].

ApplyRecommendationResult
The result of applying a recommendation.

AssetOperation
A single operation to create an asset. Supported asset types are YoutubeVideoAsset, MediaBundleAsset, ImageAsset, and LeadFormAsset. TextAsset should be created with Ad inline.

BiddingStrategyOperation
A single operation (create, update, remove) on a bidding strategy.

BillingSetupOperation
A single operation on a billing setup, which describes the cancellation of an existing billing setup.

CallConversion
A call conversion.

CallConversionResult
Identifying information for a successfully processed CallConversionUpload.

CampaignBidModifierOperation
A single operation (create, remove, update) on a campaign bid modifier.

CampaignBudgetOperation
A single operation (create, update, remove) on a campaign budget.

CampaignCriterionOperation
A single operation (create, update, remove) on a campaign criterion.

CampaignDraftOperation
A single operation (create, update, remove) on a campaign draft.

CampaignExperimentOperation
A single update operation on a campaign experiment.

CampaignExtensionSettingOperation
A single operation (create, update, remove) on a campaign extension setting.

CampaignFeedOperation
A single operation (create, update, remove) on a campaign feed.

CampaignLabelOperation
A single operation (create, remove) on a campaign-label relationship.

CampaignOperation
A single operation (create, update, remove) on a campaign.

CampaignSharedSetOperation
A single operation (create, remove) on an campaign shared set.

ClickConversion
A click conversion.

ClickConversionResult
Identifying information for a successfully processed ClickConversion.

ConversionActionOperation
A single operation (create, update, remove) on a conversion action.

ConversionAdjustment
A conversion adjustment.

ConversionAdjustmentResult
Information identifying a successfully processed ConversionAdjustment.

CreateCampaignExperimentMetadata
Message used as metadata returned in Long Running Operations for CreateCampaignExperimentRequest

CreateCampaignExperimentRequest
Request message for [CampaignExperimentService.CreateCampaignExperiment][google.ads.googleads.v1.services.CampaignExperimentService.CreateCampaignExperiment].

CreateCustomerClientRequest
Request message for [CustomerService.CreateCustomerClient][google.ads.googleads.v1.services.CustomerService.CreateCustomerClient].

CreateCustomerClientResponse
Response message for CreateCustomerClient mutate.

CreateMutateJobRequest
Request message for [MutateJobService.CreateMutateJobRequest][]

CreateMutateJobResponse
Response message for [MutateJobService.CreateMutateJobResponse][]

CustomInterestOperation
A single operation (create, update) on a custom interest.

CustomerClientLinkOperation
A single operation (create, update) on a CustomerClientLink.

CustomerExtensionSettingOperation
A single operation (create, update, remove) on a customer extension setting.

CustomerFeedOperation
A single operation (create, update, remove) on a customer feed.

CustomerLabelOperation
A single operation (create, remove) on a customer-label relationship.

CustomerManagerLinkOperation
Updates the status of a CustomerManagerLink. The following actions are possible:

CustomerNegativeCriterionOperation
A single operation (create or remove) on a customer level negative criterion.

CustomerOperation
A single update on a customer.

DismissRecommendationRequest
Request message for [RecommendationService.DismissRecommendation][google.ads.googleads.v1.services.RecommendationService.DismissRecommendation].

DismissRecommendationResponse
Response message for [RecommendationService.DismissRecommendation][google.ads.googleads.v1.services.RecommendationService.DismissRecommendation].

EndCampaignExperimentRequest
Request message for [CampaignExperimentService.EndCampaignExperiment][google.ads.googleads.v1.services.CampaignExperimentService.EndCampaignExperiment].

ExtensionFeedItemOperation
A single operation (create, update, remove) on an extension feed item.

ExternalAttributionData
Contains additional information about externally attributed conversions.

FeedItemOperation
A single operation (create, update, remove) on an feed item.

FeedItemTargetOperation
A single operation (create, remove) on an feed item target.

FeedMappingOperation
A single operation (create, remove) on a feed mapping.

FeedOperation
A single operation (create, update, remove) on an feed.

ForecastMetrics
Forecast metrics.

GclidDateTimePair
Uniquely identifies a conversion that was reported without an order ID specified.

GenerateForecastMetricsRequest
Request message for [KeywordPlanService.GenerateForecastMetrics][google.ads.googleads.v1.services.KeywordPlanService.GenerateForecastMetrics].

GenerateForecastMetricsResponse
Response message for [KeywordPlanService.GenerateForecastMetrics][google.ads.googleads.v1.services.KeywordPlanService.GenerateForecastMetrics].

GenerateHistoricalMetricsRequest
Request message for [KeywordPlanService.GenerateHistoricalMetrics][google.ads.googleads.v1.services.KeywordPlanService.GenerateHistoricalMetrics].

GenerateHistoricalMetricsResponse
Response message for [KeywordPlanService.GenerateHistoricalMetrics][google.ads.googleads.v1.services.KeywordPlanService.GenerateHistoricalMetrics].

GenerateKeywordIdeaResponse
Response message for [KeywordIdeaService.GenerateKeywordIdeas][].

GenerateKeywordIdeaResult
The result of generating keyword ideas.

GenerateKeywordIdeasRequest
Request message for [KeywordIdeaService.GenerateKeywordIdeas][].

GeoTargetConstantSuggestion
A geo target constant suggestion.

GetAccountBudgetProposalRequest
Request message for [AccountBudgetProposalService.GetAccountBudgetProposal][google.ads.googleads.v1.services.AccountBudgetProposalService.GetAccountBudgetProposal].

GetAccountBudgetRequest
Request message for [AccountBudgetService.GetAccountBudget][google.ads.googleads.v1.services.AccountBudgetService.GetAccountBudget].

GetAdGroupAdLabelRequest
Request message for [AdGroupAdLabelService.GetAdGroupAdLabel][google.ads.googleads.v1.services.AdGroupAdLabelService.GetAdGroupAdLabel].

GetAdGroupAdRequest
Request message for [AdGroupAdService.GetAdGroupAd][google.ads.googleads.v1.services.AdGroupAdService.GetAdGroupAd].

GetAdGroupAudienceViewRequest
Request message for [AdGroupAudienceViewService.GetAdGoupAudienceView][].

GetAdGroupBidModifierRequest
Request message for [AdGroupBidModifierService.GetAdGroupBidModifier][google.ads.googleads.v1.services.AdGroupBidModifierService.GetAdGroupBidModifier].

GetAdGroupCriterionLabelRequest
Request message for [AdGroupCriterionLabelService.GetAdGroupCriterionLabel][google.ads.googleads.v1.services.AdGroupCriterionLabelService.GetAdGroupCriterionLabel].

GetAdGroupCriterionRequest
Request message for [AdGroupCriterionService.GetAdGroupCriterion][google.ads.googleads.v1.services.AdGroupCriterionService.GetAdGroupCriterion].

GetAdGroupCriterionSimulationRequest
Request message for [AdGroupCriterionSimulationService.GetAdGroupCriterionSimulation][google.ads.googleads.v1.services.AdGroupCriterionSimulationService.GetAdGroupCriterionSimulation].

GetAdGroupExtensionSettingRequest
Request message for [AdGroupExtensionSettingService.GetAdGroupExtensionSetting][google.ads.googleads.v1.services.AdGroupExtensionSettingService.GetAdGroupExtensionSetting].

GetAdGroupFeedRequest
Request message for [AdGroupFeedService.GetAdGroupFeed][google.ads.googleads.v1.services.AdGroupFeedService.GetAdGroupFeed].

GetAdGroupLabelRequest
Request message for [AdGroupLabelService.GetAdGroupLabel][google.ads.googleads.v1.services.AdGroupLabelService.GetAdGroupLabel].

GetAdGroupRequest
Request message for [AdGroupService.GetAdGroup][google.ads.googleads.v1.services.AdGroupService.GetAdGroup].

GetAdGroupSimulationRequest
Request message for [AdGroupSimulationService.GetAdGroupSimulation][google.ads.googleads.v1.services.AdGroupSimulationService.GetAdGroupSimulation].

GetAdParameterRequest
Request message for [AdParameterService.GetAdParameter][google.ads.googleads.v1.services.AdParameterService.GetAdParameter]

GetAdScheduleViewRequest
Request message for [AdScheduleViewService.GetAdScheduleView][google.ads.googleads.v1.services.AdScheduleViewService.GetAdScheduleView].

GetAgeRangeViewRequest
Request message for [AgeRangeViewService.GetAgeRangeView][google.ads.googleads.v1.services.AgeRangeViewService.GetAgeRangeView].

GetAssetRequest
Request message for [AssetService.GetAsset][google.ads.googleads.v1.services.AssetService.GetAsset]

GetBiddingStrategyRequest
Request message for [BiddingStrategyService.GetBiddingStrategy][google.ads.googleads.v1.services.BiddingStrategyService.GetBiddingStrategy].

GetBillingSetupRequest
Request message for [BillingSetupService.GetBillingSetup][google.ads.googleads.v1.services.BillingSetupService.GetBillingSetup].

GetCampaignAudienceViewRequest
Request message for [CampaignAudienceViewService.GetCampaignAudienceView][google.ads.googleads.v1.services.CampaignAudienceViewService.GetCampaignAudienceView].

GetCampaignBidModifierRequest
Request message for [CampaignBidModifierService.GetCampaignBidModifier][google.ads.googleads.v1.services.CampaignBidModifierService.GetCampaignBidModifier].

GetCampaignBudgetRequest
Request message for [CampaignBudgetService.GetCampaignBudget][google.ads.googleads.v1.services.CampaignBudgetService.GetCampaignBudget].

GetCampaignCriterionRequest
Request message for [CampaignCriterionService.GetCampaignCriterion][google.ads.googleads.v1.services.CampaignCriterionService.GetCampaignCriterion].

GetCampaignCriterionSimulationRequest
Request message for [CampaignCriterionSimulationService.GetCampaignCriterionSimulation][google.ads.googleads.v1.services.CampaignCriterionSimulationService.GetCampaignCriterionSimulation].

GetCampaignDraftRequest
Request message for [CampaignDraftService.GetCampaignDraft][google.ads.googleads.v1.services.CampaignDraftService.GetCampaignDraft].

GetCampaignExperimentRequest
Request message for [CampaignExperimentService.GetCampaignExperiment][google.ads.googleads.v1.services.CampaignExperimentService.GetCampaignExperiment].

GetCampaignExtensionSettingRequest
Request message for [CampaignExtensionSettingService.GetCampaignExtensionSetting][google.ads.googleads.v1.services.CampaignExtensionSettingService.GetCampaignExtensionSetting].

GetCampaignFeedRequest
Request message for [CampaignFeedService.GetCampaignFeed][google.ads.googleads.v1.services.CampaignFeedService.GetCampaignFeed].

GetCampaignLabelRequest
Request message for [CampaignLabelService.GetCampaignLabel][google.ads.googleads.v1.services.CampaignLabelService.GetCampaignLabel].

GetCampaignRequest
Request message for [CampaignService.GetCampaign][google.ads.googleads.v1.services.CampaignService.GetCampaign].

GetCampaignSharedSetRequest
Request message for [CampaignSharedSetService.GetCampaignSharedSet][google.ads.googleads.v1.services.CampaignSharedSetService.GetCampaignSharedSet].

GetCarrierConstantRequest
Request message for [CarrierConstantService.GetCarrierConstant][google.ads.googleads.v1.services.CarrierConstantService.GetCarrierConstant].

GetChangeStatusRequest
Request message for '[ChangeStatusService.GetChangeStatus][google.ads.googleads.v1.services.ChangeStatusService.GetChangeStatus]'.

GetClickViewRequest
Request message for [ClickViewService.GetClickView][google.ads.googleads.v1.services.ClickViewService.GetClickView].

GetConversionActionRequest
Request message for [ConversionActionService.GetConversionAction][google.ads.googleads.v1.services.ConversionActionService.GetConversionAction].

GetCustomInterestRequest
Request message for [CustomInterestService.GetCustomInterest][google.ads.googleads.v1.services.CustomInterestService.GetCustomInterest].

GetCustomerClientLinkRequest
Request message for [CustomerClientLinkService.GetCustomerClientLink][google.ads.googleads.v1.services.CustomerClientLinkService.GetCustomerClientLink].

GetCustomerClientRequest
Request message for [CustomerClientService.GetCustomerClient][google.ads.googleads.v1.services.CustomerClientService.GetCustomerClient].

GetCustomerExtensionSettingRequest
Request message for [CustomerExtensionSettingService.GetCustomerExtensionSetting][google.ads.googleads.v1.services.CustomerExtensionSettingService.GetCustomerExtensionSetting].

GetCustomerFeedRequest
Request message for [CustomerFeedService.GetCustomerFeed][google.ads.googleads.v1.services.CustomerFeedService.GetCustomerFeed].

GetCustomerLabelRequest
Request message for [CustomerLabelService.GetCustomerLabel][google.ads.googleads.v1.services.CustomerLabelService.GetCustomerLabel].

GetCustomerManagerLinkRequest
Request message for [CustomerManagerLinkService.GetCustomerManagerLink][google.ads.googleads.v1.services.CustomerManagerLinkService.GetCustomerManagerLink].

GetCustomerNegativeCriterionRequest
Request message for [CustomerNegativeCriterionService.GetCustomerNegativeCriterion][google.ads.googleads.v1.services.CustomerNegativeCriterionService.GetCustomerNegativeCriterion].

GetCustomerRequest
Request message for [CustomerService.GetCustomer][google.ads.googleads.v1.services.CustomerService.GetCustomer].

GetDetailPlacementViewRequest
Request message for [DetailPlacementViewService.GetDetailPlacementView][google.ads.googleads.v1.services.DetailPlacementViewService.GetDetailPlacementView].

GetDisplayKeywordViewRequest
Request message for [DisplayKeywordViewService.GetDisplayKeywordView][google.ads.googleads.v1.services.DisplayKeywordViewService.GetDisplayKeywordView].

GetDomainCategoryRequest
Request message for [DomainCategoryService.GetDomainCategory][google.ads.googleads.v1.services.DomainCategoryService.GetDomainCategory].

GetDynamicSearchAdsSearchTermViewRequest
Request message for [DynamicSearchAdsSearchTermViewService.GetDynamicSearchAdsSearchTermView][google.ads.googleads.v1.services.DynamicSearchAdsSearchTermViewService.GetDynamicSearchAdsSearchTermView].

GetExpandedLandingPageViewRequest
Request message for [ExpandedLandingPageViewService.GetExpandedLandingPageView][google.ads.googleads.v1.services.ExpandedLandingPageViewService.GetExpandedLandingPageView].

GetExtensionFeedItemRequest
Request message for [ExtensionFeedItemService.GetExtensionFeedItem][google.ads.googleads.v1.services.ExtensionFeedItemService.GetExtensionFeedItem].

GetFeedItemRequest
Request message for [FeedItemService.GetFeedItem][google.ads.googleads.v1.services.FeedItemService.GetFeedItem].

GetFeedItemTargetRequest
Request message for [FeedItemTargetService.GetFeedItemTarget][google.ads.googleads.v1.services.FeedItemTargetService.GetFeedItemTarget].

GetFeedMappingRequest
Request message for [FeedMappingService.GetFeedMapping][google.ads.googleads.v1.services.FeedMappingService.GetFeedMapping].

GetFeedPlaceholderViewRequest
Request message for [FeedPlaceholderViewService.GetFeedPlaceholderView][google.ads.googleads.v1.services.FeedPlaceholderViewService.GetFeedPlaceholderView].

GetFeedRequest
Request message for [FeedService.GetFeed][google.ads.googleads.v1.services.FeedService.GetFeed].

GetGenderViewRequest
Request message for [GenderViewService.GetGenderView][google.ads.googleads.v1.services.GenderViewService.GetGenderView].

GetGeoTargetConstantRequest
Request message for [GeoTargetConstantService.GetGeoTargetConstant][google.ads.googleads.v1.services.GeoTargetConstantService.GetGeoTargetConstant].

GetGeographicViewRequest
Request message for [GeographicViewService.GetGeographicView][google.ads.googleads.v1.services.GeographicViewService.GetGeographicView].

GetGoogleAdsFieldRequest
Request message for [GoogleAdsFieldService.GetGoogleAdsField][google.ads.googleads.v1.services.GoogleAdsFieldService.GetGoogleAdsField].

GetGroupPlacementViewRequest
Request message for [GroupPlacementViewService.GetGroupPlacementView][google.ads.googleads.v1.services.GroupPlacementViewService.GetGroupPlacementView].

GetHotelGroupViewRequest
Request message for [HotelGroupViewService.GetHotelGroupView][google.ads.googleads.v1.services.HotelGroupViewService.GetHotelGroupView].

GetHotelPerformanceViewRequest
Request message for [HotelPerformanceViewService.GetHotelPerformanceView][google.ads.googleads.v1.services.HotelPerformanceViewService.GetHotelPerformanceView].

GetKeywordPlanAdGroupRequest
Request message for [KeywordPlanAdGroupService.GetKeywordPlanAdGroup][google.ads.googleads.v1.services.KeywordPlanAdGroupService.GetKeywordPlanAdGroup].

GetKeywordPlanCampaignRequest
Request message for [KeywordPlanCampaignService.GetKeywordPlanCampaign][google.ads.googleads.v1.services.KeywordPlanCampaignService.GetKeywordPlanCampaign].

GetKeywordPlanKeywordRequest
Request message for [KeywordPlanKeywordService.GetKeywordPlanKeyword][google.ads.googleads.v1.services.KeywordPlanKeywordService.GetKeywordPlanKeyword].

GetKeywordPlanNegativeKeywordRequest
Request message for [KeywordPlanNegativeKeywordService.GetKeywordPlanNegativeKeyword][google.ads.googleads.v1.services.KeywordPlanNegativeKeywordService.GetKeywordPlanNegativeKeyword].

GetKeywordPlanRequest
Request message for [KeywordPlanService.GetKeywordPlan][google.ads.googleads.v1.services.KeywordPlanService.GetKeywordPlan].

GetKeywordViewRequest
Request message for [KeywordViewService.GetKeywordView][google.ads.googleads.v1.services.KeywordViewService.GetKeywordView].

GetLabelRequest
Request message for [LabelService.GetLabel][google.ads.googleads.v1.services.LabelService.GetLabel].

GetLandingPageViewRequest
Request message for [LandingPageViewService.GetLandingPageView][google.ads.googleads.v1.services.LandingPageViewService.GetLandingPageView].

GetLanguageConstantRequest
Request message for [LanguageConstantService.GetLanguageConstant][google.ads.googleads.v1.services.LanguageConstantService.GetLanguageConstant].

GetLocationViewRequest
Request message for [LocationViewService.GetLocationView][google.ads.googleads.v1.services.LocationViewService.GetLocationView].

GetManagedPlacementViewRequest
Request message for [ManagedPlacementViewService.GetManagedPlacementView][google.ads.googleads.v1.services.ManagedPlacementViewService.GetManagedPlacementView].

GetMediaFileRequest
Request message for [MediaFileService.GetMediaFile][google.ads.googleads.v1.services.MediaFileService.GetMediaFile]

GetMerchantCenterLinkRequest
Request message for [MerchantCenterLinkService.GetMerchantCenterLink][google.ads.googleads.v1.services.MerchantCenterLinkService.GetMerchantCenterLink].

GetMobileAppCategoryConstantRequest
Request message for [MobileAppCategoryConstantService.GetMobileAppCategoryConstant][google.ads.googleads.v1.services.MobileAppCategoryConstantService.GetMobileAppCategoryConstant].

GetMobileDeviceConstantRequest
Request message for [MobileDeviceConstantService.GetMobileDeviceConstant][google.ads.googleads.v1.services.MobileDeviceConstantService.GetMobileDeviceConstant].

GetMutateJobRequest
Request message for [MutateJobService.GetMutateJob][google.ads.googleads.v1.services.MutateJobService.GetMutateJob]

GetOperatingSystemVersionConstantRequest
Request message for [OperatingSystemVersionConstantService.GetOperatingSystemVersionConstant][google.ads.googleads.v1.services.OperatingSystemVersionConstantService.GetOperatingSystemVersionConstant].

GetPaidOrganicSearchTermViewRequest
Request message for [PaidOrganicSearchTermViewService.GetPaidOrganicSearchTermView][google.ads.googleads.v1.services.PaidOrganicSearchTermViewService.GetPaidOrganicSearchTermView].

GetParentalStatusViewRequest
Request message for [ParentalStatusViewService.GetParentalStatusView][google.ads.googleads.v1.services.ParentalStatusViewService.GetParentalStatusView].

GetProductBiddingCategoryConstantRequest
Request message for [ProductBiddingCategoryService.GetProductBiddingCategory][].

GetProductGroupViewRequest
Request message for [ProductGroupViewService.GetProductGroupView][google.ads.googleads.v1.services.ProductGroupViewService.GetProductGroupView].

GetRecommendationRequest
Request message for [RecommendationService.GetRecommendation][google.ads.googleads.v1.services.RecommendationService.GetRecommendation].

GetRemarketingActionRequest
Request message for [RemarketingActionService.GetRemarketingAction][google.ads.googleads.v1.services.RemarketingActionService.GetRemarketingAction].

GetSearchTermViewRequest
Request message for [SearchTermViewService.GetSearchTermView][google.ads.googleads.v1.services.SearchTermViewService.GetSearchTermView].

GetSharedCriterionRequest
Request message for [SharedCriterionService.GetSharedCriterion][google.ads.googleads.v1.services.SharedCriterionService.GetSharedCriterion].

GetSharedSetRequest
Request message for [SharedSetService.GetSharedSet][google.ads.googleads.v1.services.SharedSetService.GetSharedSet].

GetShoppingPerformanceViewRequest
Request message for [ShoppingPerformanceViewService.GetShoppingPerformanceView][google.ads.googleads.v1.services.ShoppingPerformanceViewService.GetShoppingPerformanceView].

GetTopicConstantRequest
Request message for [TopicConstantService.GetTopicConstant][google.ads.googleads.v1.services.TopicConstantService.GetTopicConstant].

GetTopicViewRequest
Request message for [TopicViewService.GetTopicView][google.ads.googleads.v1.services.TopicViewService.GetTopicView].

GetUserInterestRequest
Request message for [UserInterestService.GetUserInterest][google.ads.googleads.v1.services.UserInterestService.GetUserInterest].

GetUserListRequest
Request message for [UserListService.GetUserList][google.ads.googleads.v1.services.UserListService.GetUserList].

GetVideoRequest
Request message for [VideoService.GetVideo][google.ads.googleads.v1.services.VideoService.GetVideo].

GoogleAdsRow
A returned row from the query.

GraduateCampaignExperimentRequest
Request message for [CampaignExperimentService.GraduateCampaignExperiment][google.ads.googleads.v1.services.CampaignExperimentService.GraduateCampaignExperiment].

GraduateCampaignExperimentResponse
Response message for campaign experiment graduate.

KeywordAndUrlSeed
Keyword And Url Seed

KeywordPlanAdGroupForecast
An ad group forecast.

KeywordPlanAdGroupOperation
A single operation (create, update, remove) on a Keyword Plan ad group.

KeywordPlanCampaignForecast
A campaign forecast.

KeywordPlanCampaignOperation
A single operation (create, update, remove) on a Keyword Plan campaign.

KeywordPlanKeywordForecast
A keyword forecast.

KeywordPlanKeywordHistoricalMetrics
A keyword historical metrics.

KeywordPlanKeywordOperation
A single operation (create, update, remove) on a Keyword Plan keyword.

KeywordPlanNegativeKeywordOperation
A single operation (create, update, remove) on a Keyword Plan negative keyword.

KeywordPlanOperation
A single operation (create, update, remove) on a keyword plan.

KeywordSeed
Keyword Seed

LabelOperation
A single operation (create, remove, update) on a label.

ListAccessibleCustomersRequest
Request message for [CustomerService.ListAccessibleCustomers][google.ads.googleads.v1.services.CustomerService.ListAccessibleCustomers].

ListAccessibleCustomersResponse
Response message for [CustomerService.ListAccessibleCustomers][google.ads.googleads.v1.services.CustomerService.ListAccessibleCustomers].

ListCampaignDraftAsyncErrorsRequest
Request message for [CampaignDraftService.ListCampaignDraftAsyncErrors][google.ads.googleads.v1.services.CampaignDraftService.ListCampaignDraftAsyncErrors].

ListCampaignDraftAsyncErrorsResponse
Response message for [CampaignDraftService.ListCampaignDraftAsyncErrors][google.ads.googleads.v1.services.CampaignDraftService.ListCampaignDraftAsyncErrors].

ListCampaignExperimentAsyncErrorsRequest
Request message for [CampaignExperimentService.ListCampaignExperimentAsyncErrors][google.ads.googleads.v1.services.CampaignExperimentService.ListCampaignExperimentAsyncErrors].

ListCampaignExperimentAsyncErrorsResponse
Response message for [CampaignExperimentService.ListCampaignExperimentAsyncErrors][google.ads.googleads.v1.services.CampaignExperimentService.ListCampaignExperimentAsyncErrors].

ListMerchantCenterLinksRequest
Request message for [MerchantCenterLinkService.ListMerchantCenterLinks][google.ads.googleads.v1.services.MerchantCenterLinkService.ListMerchantCenterLinks].

ListMerchantCenterLinksResponse
Response message for [MerchantCenterLinkService.ListMerchantCenterLinks][google.ads.googleads.v1.services.MerchantCenterLinkService.ListMerchantCenterLinks].

ListMutateJobResultsRequest
Request message for [MutateJobService.ListMutateJobResults][google.ads.googleads.v1.services.MutateJobService.ListMutateJobResults].

ListMutateJobResultsResponse
Response message for [MutateJobService.ListMutateJobResults][google.ads.googleads.v1.services.MutateJobService.ListMutateJobResults].

ListPaymentsAccountsRequest
Request message for fetching all accessible payments accounts.

ListPaymentsAccountsResponse
Response message for [PaymentsAccountService.ListPaymentsAccounts][google.ads.googleads.v1.services.PaymentsAccountService.ListPaymentsAccounts].

MediaFileOperation
A single operation to create media file.

MerchantCenterLinkOperation
A single update on a Merchant Center link.

MutateAccountBudgetProposalRequest
Request message for [AccountBudgetProposalService.MutateAccountBudgetProposal][google.ads.googleads.v1.services.AccountBudgetProposalService.MutateAccountBudgetProposal].

MutateAccountBudgetProposalResponse
Response message for account-level budget mutate operations.

MutateAccountBudgetProposalResult
The result for the account budget proposal mutate.

MutateAdGroupAdLabelResult
The result for an ad group ad label mutate.

MutateAdGroupAdLabelsRequest
Request message for [AdGroupAdLabelService.MutateAdGroupAdLabels][google.ads.googleads.v1.services.AdGroupAdLabelService.MutateAdGroupAdLabels].

MutateAdGroupAdLabelsResponse
Response message for an ad group ad labels mutate.

MutateAdGroupAdResult
The result for the ad mutate.

MutateAdGroupAdsRequest
Request message for [AdGroupAdService.MutateAdGroupAds][google.ads.googleads.v1.services.AdGroupAdService.MutateAdGroupAds].

MutateAdGroupAdsResponse
Response message for an ad group ad mutate.

MutateAdGroupBidModifierResult
The result for the criterion mutate.

MutateAdGroupBidModifiersRequest
Request message for [AdGroupBidModifierService.MutateAdGroupBidModifiers][google.ads.googleads.v1.services.AdGroupBidModifierService.MutateAdGroupBidModifiers].

MutateAdGroupBidModifiersResponse
Response message for ad group bid modifiers mutate.

MutateAdGroupCriteriaRequest
Request message for [AdGroupCriterionService.MutateAdGroupCriteria][google.ads.googleads.v1.services.AdGroupCriterionService.MutateAdGroupCriteria].

MutateAdGroupCriteriaResponse
Response message for an ad group criterion mutate.

MutateAdGroupCriterionLabelResult
The result for an ad group criterion label mutate.

MutateAdGroupCriterionLabelsRequest
Request message for [AdGroupCriterionLabelService.MutateAdGroupCriterionLabels][google.ads.googleads.v1.services.AdGroupCriterionLabelService.MutateAdGroupCriterionLabels].

MutateAdGroupCriterionLabelsResponse
Response message for an ad group criterion labels mutate.

MutateAdGroupCriterionResult
The result for the criterion mutate.

MutateAdGroupExtensionSettingResult
The result for the ad group extension setting mutate.

MutateAdGroupExtensionSettingsRequest
Request message for [AdGroupExtensionSettingService.MutateAdGroupExtensionSettings][google.ads.googleads.v1.services.AdGroupExtensionSettingService.MutateAdGroupExtensionSettings].

MutateAdGroupExtensionSettingsResponse
Response message for an ad group extension setting mutate.

MutateAdGroupFeedResult
The result for the ad group feed mutate.

MutateAdGroupFeedsRequest
Request message for [AdGroupFeedService.MutateAdGroupFeeds][google.ads.googleads.v1.services.AdGroupFeedService.MutateAdGroupFeeds].

MutateAdGroupFeedsResponse
Response message for an ad group feed mutate.

MutateAdGroupLabelResult
The result for an ad group label mutate.

MutateAdGroupLabelsRequest
Request message for [AdGroupLabelService.MutateAdGroupLabels][google.ads.googleads.v1.services.AdGroupLabelService.MutateAdGroupLabels].

MutateAdGroupLabelsResponse
Response message for an ad group labels mutate.

MutateAdGroupResult
The result for the ad group mutate.

MutateAdGroupsRequest
Request message for [AdGroupService.MutateAdGroups][google.ads.googleads.v1.services.AdGroupService.MutateAdGroups].

MutateAdGroupsResponse
Response message for an ad group mutate.

MutateAdParameterResult
The result for the ad parameter mutate.

MutateAdParametersRequest
Request message for [AdParameterService.MutateAdParameters][google.ads.googleads.v1.services.AdParameterService.MutateAdParameters]

MutateAdParametersResponse
Response message for an ad parameter mutate.

MutateAssetResult
The result for the asset mutate.

MutateAssetsRequest
Request message for [AssetService.MutateAssets][google.ads.googleads.v1.services.AssetService.MutateAssets]

MutateAssetsResponse
Response message for an asset mutate.

MutateBiddingStrategiesRequest
Request message for [BiddingStrategyService.MutateBiddingStrategies][google.ads.googleads.v1.services.BiddingStrategyService.MutateBiddingStrategies].

MutateBiddingStrategiesResponse
Response message for bidding strategy mutate.

MutateBiddingStrategyResult
The result for the bidding strategy mutate.

MutateBillingSetupRequest
Request message for billing setup mutate operations.

MutateBillingSetupResponse
Response message for a billing setup operation.

MutateBillingSetupResult
Result for a single billing setup mutate.

MutateCampaignBidModifierResult
The result for the criterion mutate.

MutateCampaignBidModifiersRequest
Request message for [CampaignBidModifierService.MutateCampaignBidModifier][].

MutateCampaignBidModifiersResponse
Response message for campaign bid modifiers mutate.

MutateCampaignBudgetResult
The result for the campaign budget mutate.

MutateCampaignBudgetsRequest
Request message for [CampaignBudgetService.MutateCampaignBudgets][google.ads.googleads.v1.services.CampaignBudgetService.MutateCampaignBudgets].

MutateCampaignBudgetsResponse
Response message for campaign budget mutate.

MutateCampaignCriteriaRequest
Request message for [CampaignCriterionService.MutateCampaignCriteria][google.ads.googleads.v1.services.CampaignCriterionService.MutateCampaignCriteria].

MutateCampaignCriteriaResponse
Response message for campaign criterion mutate.

MutateCampaignCriterionResult
The result for the criterion mutate.

MutateCampaignDraftResult
The result for the campaign draft mutate.

MutateCampaignDraftsRequest
Request message for [CampaignDraftService.MutateCampaignDrafts][google.ads.googleads.v1.services.CampaignDraftService.MutateCampaignDrafts].

MutateCampaignDraftsResponse
Response message for campaign draft mutate.

MutateCampaignExperimentResult
The result for the campaign experiment mutate.

MutateCampaignExperimentsRequest
Request message for [CampaignExperimentService.MutateCampaignExperiments][google.ads.googleads.v1.services.CampaignExperimentService.MutateCampaignExperiments].

MutateCampaignExperimentsResponse
Response message for campaign experiment mutate.

MutateCampaignExtensionSettingResult
The result for the campaign extension setting mutate.

MutateCampaignExtensionSettingsRequest
Request message for [CampaignExtensionSettingService.MutateCampaignExtensionSettings][google.ads.googleads.v1.services.CampaignExtensionSettingService.MutateCampaignExtensionSettings].

MutateCampaignExtensionSettingsResponse
Response message for a campaign extension setting mutate.

MutateCampaignFeedResult
The result for the campaign feed mutate.

MutateCampaignFeedsRequest
Request message for [CampaignFeedService.MutateCampaignFeeds][google.ads.googleads.v1.services.CampaignFeedService.MutateCampaignFeeds].

MutateCampaignFeedsResponse
Response message for a campaign feed mutate.

MutateCampaignLabelResult
The result for a campaign label mutate.

MutateCampaignLabelsRequest
Request message for [CampaignLabelService.MutateCampaignLabels][google.ads.googleads.v1.services.CampaignLabelService.MutateCampaignLabels].

MutateCampaignLabelsResponse
Response message for a campaign labels mutate.

MutateCampaignResult
The result for the campaign mutate.

MutateCampaignSharedSetResult
The result for the campaign shared set mutate.

MutateCampaignSharedSetsRequest
Request message for [CampaignSharedSetService.MutateCampaignSharedSets][google.ads.googleads.v1.services.CampaignSharedSetService.MutateCampaignSharedSets].

MutateCampaignSharedSetsResponse
Response message for a campaign shared set mutate.

MutateCampaignsRequest
Request message for [CampaignService.MutateCampaigns][google.ads.googleads.v1.services.CampaignService.MutateCampaigns].

MutateCampaignsResponse
Response message for campaign mutate.

MutateConversionActionResult
The result for the conversion action mutate.

MutateConversionActionsRequest
Request message for [ConversionActionService.MutateConversionActions][google.ads.googleads.v1.services.ConversionActionService.MutateConversionActions].

MutateConversionActionsResponse
Response message for [ConversionActionService.MutateConversionActions][google.ads.googleads.v1.services.ConversionActionService.MutateConversionActions].

MutateCustomInterestResult
The result for the custom interest mutate.

MutateCustomInterestsRequest
Request message for [CustomInterestService.MutateCustomInterests][google.ads.googleads.v1.services.CustomInterestService.MutateCustomInterests].

MutateCustomInterestsResponse
Response message for custom interest mutate.

MutateCustomerClientLinkRequest
Request message for [CustomerClientLinkService.MutateCustomerClientLink][google.ads.googleads.v1.services.CustomerClientLinkService.MutateCustomerClientLink].

MutateCustomerClientLinkResponse
Response message for a CustomerClientLink mutate.

MutateCustomerClientLinkResult
The result for a single customer client link mutate.

MutateCustomerExtensionSettingResult
The result for the customer extension setting mutate.

MutateCustomerExtensionSettingsRequest
Request message for [CustomerExtensionSettingService.MutateCustomerExtensionSettings][google.ads.googleads.v1.services.CustomerExtensionSettingService.MutateCustomerExtensionSettings].

MutateCustomerExtensionSettingsResponse
Response message for a customer extension setting mutate.

MutateCustomerFeedResult
The result for the customer feed mutate.

MutateCustomerFeedsRequest
Request message for [CustomerFeedService.MutateCustomerFeeds][google.ads.googleads.v1.services.CustomerFeedService.MutateCustomerFeeds].

MutateCustomerFeedsResponse
Response message for a customer feed mutate.

MutateCustomerLabelResult
The result for a customer label mutate.

MutateCustomerLabelsRequest
Request message for [CustomerLabelService.MutateCustomerLabels][google.ads.googleads.v1.services.CustomerLabelService.MutateCustomerLabels].

MutateCustomerLabelsResponse
Response message for a customer labels mutate.

MutateCustomerManagerLinkRequest
Request message for [CustomerManagerLinkService.MutateCustomerManagerLink][google.ads.googleads.v1.services.CustomerManagerLinkService.MutateCustomerManagerLink].

MutateCustomerManagerLinkResponse
Response message for a CustomerManagerLink mutate.

MutateCustomerManagerLinkResult
The result for the customer manager link mutate.

MutateCustomerNegativeCriteriaRequest
Request message for [CustomerNegativeCriterionService.MutateCustomerNegativeCriteria][google.ads.googleads.v1.services.CustomerNegativeCriterionService.MutateCustomerNegativeCriteria].

MutateCustomerNegativeCriteriaResponse
Response message for customer negative criterion mutate.

MutateCustomerNegativeCriteriaResult
The result for the criterion mutate.

MutateCustomerRequest
Request message for [CustomerService.MutateCustomer][google.ads.googleads.v1.services.CustomerService.MutateCustomer].

MutateCustomerResponse
Response message for customer mutate.

MutateCustomerResult
The result for the customer mutate.

MutateExtensionFeedItemResult
The result for the extension feed item mutate.

MutateExtensionFeedItemsRequest
Request message for [ExtensionFeedItemService.MutateExtensionFeedItems][google.ads.googleads.v1.services.ExtensionFeedItemService.MutateExtensionFeedItems].

MutateExtensionFeedItemsResponse
Response message for an extension feed item mutate.

MutateFeedItemResult
The result for the feed item mutate.

MutateFeedItemTargetResult
The result for the feed item target mutate.

MutateFeedItemTargetsRequest
Request message for [FeedItemTargetService.MutateFeedItemTargets][google.ads.googleads.v1.services.FeedItemTargetService.MutateFeedItemTargets].

MutateFeedItemTargetsResponse
Response message for an feed item target mutate.

MutateFeedItemsRequest
Request message for [FeedItemService.MutateFeedItems][google.ads.googleads.v1.services.FeedItemService.MutateFeedItems].

MutateFeedItemsResponse
Response message for an feed item mutate.

MutateFeedMappingResult
The result for the feed mapping mutate.

MutateFeedMappingsRequest
Request message for [FeedMappingService.MutateFeedMappings][google.ads.googleads.v1.services.FeedMappingService.MutateFeedMappings].

MutateFeedMappingsResponse
Response message for a feed mapping mutate.

MutateFeedResult
The result for the feed mutate.

MutateFeedsRequest
Request message for [FeedService.MutateFeeds][google.ads.googleads.v1.services.FeedService.MutateFeeds].

MutateFeedsResponse
Response message for an feed mutate.

MutateGoogleAdsRequest
Request message for [GoogleAdsService.Mutate][google.ads.googleads.v1.services.GoogleAdsService.Mutate].

MutateGoogleAdsResponse
Response message for [GoogleAdsService.Mutate][google.ads.googleads.v1.services.GoogleAdsService.Mutate].

MutateJobResult
MutateJob result.

MutateKeywordPlanAdGroupResult
The result for the Keyword Plan ad group mutate.

MutateKeywordPlanAdGroupsRequest
Request message for [KeywordPlanAdGroupService.MutateKeywordPlanAdGroups][google.ads.googleads.v1.services.KeywordPlanAdGroupService.MutateKeywordPlanAdGroups].

MutateKeywordPlanAdGroupsResponse
Response message for a Keyword Plan ad group mutate.

MutateKeywordPlanCampaignResult
The result for the Keyword Plan campaign mutate.

MutateKeywordPlanCampaignsRequest
Request message for [KeywordPlanCampaignService.MutateKeywordPlanCampaigns][google.ads.googleads.v1.services.KeywordPlanCampaignService.MutateKeywordPlanCampaigns].

MutateKeywordPlanCampaignsResponse
Response message for a Keyword Plan campaign mutate.

MutateKeywordPlanKeywordResult
The result for the Keyword Plan keyword mutate.

MutateKeywordPlanKeywordsRequest
Request message for [KeywordPlanKeywordService.MutateKeywordPlanKeywords][google.ads.googleads.v1.services.KeywordPlanKeywordService.MutateKeywordPlanKeywords].

MutateKeywordPlanKeywordsResponse
Response message for a Keyword Plan keyword mutate.

MutateKeywordPlanNegativeKeywordResult
The result for the Keyword Plan negative keyword mutate.

MutateKeywordPlanNegativeKeywordsRequest
Request message for [KeywordPlanNegativeKeywordService.MutateKeywordPlanNegativeKeywords][google.ads.googleads.v1.services.KeywordPlanNegativeKeywordService.MutateKeywordPlanNegativeKeywords].

MutateKeywordPlanNegativeKeywordsResponse
Response message for a Keyword Plan negative keyword mutate.

MutateKeywordPlansRequest
Request message for [KeywordPlanService.MutateKeywordPlans][google.ads.googleads.v1.services.KeywordPlanService.MutateKeywordPlans].

MutateKeywordPlansResponse
Response message for a keyword plan mutate.

MutateKeywordPlansResult
The result for the keyword plan mutate.

MutateLabelResult
The result for a label mutate.

MutateLabelsRequest
Request message for [LabelService.MutateLabels][google.ads.googleads.v1.services.LabelService.MutateLabels].

MutateLabelsResponse
Response message for a labels mutate.

MutateMediaFileResult
The result for the media file mutate.

MutateMediaFilesRequest
Request message for [MediaFileService.MutateMediaFiles][google.ads.googleads.v1.services.MediaFileService.MutateMediaFiles]

MutateMediaFilesResponse
Response message for a media file mutate.

MutateMerchantCenterLinkRequest
Request message for [MerchantCenterLinkService.MutateMerchantCenterLink][google.ads.googleads.v1.services.MerchantCenterLinkService.MutateMerchantCenterLink].

MutateMerchantCenterLinkResponse
Response message for Merchant Center link mutate.

MutateMerchantCenterLinkResult
The result for the Merchant Center link mutate.

MutateOperation
A single operation (create, update, remove) on a resource.

MutateOperationResponse
Response message for the resource mutate.

MutateRemarketingActionResult
The result for the remarketing action mutate.

MutateRemarketingActionsRequest
Request message for [RemarketingActionService.MutateRemarketingActions][google.ads.googleads.v1.services.RemarketingActionService.MutateRemarketingActions].

MutateRemarketingActionsResponse
Response message for remarketing action mutate.

MutateSharedCriteriaRequest
Request message for [SharedCriterionService.MutateSharedCriteria][google.ads.googleads.v1.services.SharedCriterionService.MutateSharedCriteria].

MutateSharedCriteriaResponse
Response message for a shared criterion mutate.

MutateSharedCriterionResult
The result for the shared criterion mutate.

MutateSharedSetResult
The result for the shared set mutate.

MutateSharedSetsRequest
Request message for [SharedSetService.MutateSharedSets][google.ads.googleads.v1.services.SharedSetService.MutateSharedSets].

MutateSharedSetsResponse
Response message for a shared set mutate.

MutateUserListResult
The result for the user list mutate.

MutateUserListsRequest
Request message for [UserListService.MutateUserLists][google.ads.googleads.v1.services.UserListService.MutateUserLists].

MutateUserListsResponse
Response message for user list mutate.

PromoteCampaignDraftRequest
Request message for [CampaignDraftService.PromoteCampaignDraft][google.ads.googleads.v1.services.CampaignDraftService.PromoteCampaignDraft].

PromoteCampaignExperimentRequest
Request message for [CampaignExperimentService.PromoteCampaignExperiment][google.ads.googleads.v1.services.CampaignExperimentService.PromoteCampaignExperiment].

RemarketingActionOperation
A single operation (create, update) on a remarketing action.

RestatementValue
Contains information needed to restate a conversion's value.

RunMutateJobRequest
Request message for [MutateJobService.RunMutateJob][google.ads.googleads.v1.services.MutateJobService.RunMutateJob]

SearchGoogleAdsFieldsRequest
Request message for [GoogleAdsFieldService.SearchGoogleAdsFields][google.ads.googleads.v1.services.GoogleAdsFieldService.SearchGoogleAdsFields].

SearchGoogleAdsFieldsResponse
Response message for [GoogleAdsFieldService.SearchGoogleAdsFields][google.ads.googleads.v1.services.GoogleAdsFieldService.SearchGoogleAdsFields].

SearchGoogleAdsRequest
Request message for [GoogleAdsService.Search][google.ads.googleads.v1.services.GoogleAdsService.Search].

SearchGoogleAdsResponse
Response message for [GoogleAdsService.Search][google.ads.googleads.v1.services.GoogleAdsService.Search].

SharedCriterionOperation
A single operation (create, remove) on an shared criterion.

SharedSetOperation
A single operation (create, update, remove) on an shared set.

SuggestGeoTargetConstantsRequest
Request message for [GeoTargetConstantService.SuggestGeoTargetConstantsRequest][].

SuggestGeoTargetConstantsResponse
Response message for [GeoTargetConstantService.SuggestGeoTargetConstants][google.ads.googleads.v1.services.GeoTargetConstantService.SuggestGeoTargetConstants]

UploadCallConversionsRequest
Request message for [ConversionUploadService.UploadCallConversions][google.ads.googleads.v1.services.ConversionUploadService.UploadCallConversions].

UploadCallConversionsResponse
Response message for [ConversionUploadService.UploadCallConversions][google.ads.googleads.v1.services.ConversionUploadService.UploadCallConversions].

UploadClickConversionsRequest
Request message for [ConversionUploadService.UploadClickConversions][google.ads.googleads.v1.services.ConversionUploadService.UploadClickConversions].

UploadClickConversionsResponse
Response message for [ConversionUploadService.UploadClickConversions][google.ads.googleads.v1.services.ConversionUploadService.UploadClickConversions].

UploadConversionAdjustmentsRequest
Request message for [ConversionAdjustmentUploadService.UploadConversionAdjustments][google.ads.googleads.v1.services.ConversionAdjustmentUploadService.UploadConversionAdjustments].

UploadConversionAdjustmentsResponse
Response message for [ConversionAdjustmentUploadService.UploadConversionAdjustments][google.ads.googleads.v1.services.ConversionAdjustmentUploadService.UploadConversionAdjustments].

UrlSeed
Url Seed

UserListOperation
A single operation (create, update) on a user list.