Enum gapi_grpc::google::ads::googleads::v3::enums::conversion_or_adjustment_lag_bucket_enum::ConversionOrAdjustmentLagBucket

#[repr(i32)]pub enum ConversionOrAdjustmentLagBucket {
    Unspecified,
    Unknown,
    ConversionLessThanOneDay,
    ConversionOneToTwoDays,
    ConversionTwoToThreeDays,
    ConversionThreeToFourDays,
    ConversionFourToFiveDays,
    ConversionFiveToSixDays,
    ConversionSixToSevenDays,
    ConversionSevenToEightDays,
    ConversionEightToNineDays,
    ConversionNineToTenDays,
    ConversionTenToElevenDays,
    ConversionElevenToTwelveDays,
    ConversionTwelveToThirteenDays,
    ConversionThirteenToFourteenDays,
    ConversionFourteenToTwentyOneDays,
    ConversionTwentyOneToThirtyDays,
    ConversionThirtyToFortyFiveDays,
    ConversionFortyFiveToSixtyDays,
    ConversionSixtyToNinetyDays,
    AdjustmentLessThanOneDay,
    AdjustmentOneToTwoDays,
    AdjustmentTwoToThreeDays,
    AdjustmentThreeToFourDays,
    AdjustmentFourToFiveDays,
    AdjustmentFiveToSixDays,
    AdjustmentSixToSevenDays,
    AdjustmentSevenToEightDays,
    AdjustmentEightToNineDays,
    AdjustmentNineToTenDays,
    AdjustmentTenToElevenDays,
    AdjustmentElevenToTwelveDays,
    AdjustmentTwelveToThirteenDays,
    AdjustmentThirteenToFourteenDays,
    AdjustmentFourteenToTwentyOneDays,
    AdjustmentTwentyOneToThirtyDays,
    AdjustmentThirtyToFortyFiveDays,
    AdjustmentFortyFiveToSixtyDays,
    AdjustmentSixtyToNinetyDays,
    AdjustmentNinetyToOneHundredAndFortyFiveDays,
    ConversionUnknown,
    AdjustmentUnknown,
}

Enum representing the number of days between the impression and the conversion or between the impression and adjustments to the conversion.

Variants

Unspecified

Not specified.

Unknown

Used for return value only. Represents value unknown in this version.

ConversionLessThanOneDay

Conversion lag bucket from 0 to 1 day. 0 day is included, 1 day is not.

ConversionOneToTwoDays

Conversion lag bucket from 1 to 2 days. 1 day is included, 2 days is not.

ConversionTwoToThreeDays

Conversion lag bucket from 2 to 3 days. 2 days is included, 3 days is not.

ConversionThreeToFourDays

Conversion lag bucket from 3 to 4 days. 3 days is included, 4 days is not.

ConversionFourToFiveDays

Conversion lag bucket from 4 to 5 days. 4 days is included, 5 days is not.

ConversionFiveToSixDays

Conversion lag bucket from 5 to 6 days. 5 days is included, 6 days is not.

ConversionSixToSevenDays

Conversion lag bucket from 6 to 7 days. 6 days is included, 7 days is not.

ConversionSevenToEightDays

Conversion lag bucket from 7 to 8 days. 7 days is included, 8 days is not.

ConversionEightToNineDays

Conversion lag bucket from 8 to 9 days. 8 days is included, 9 days is not.

ConversionNineToTenDays

Conversion lag bucket from 9 to 10 days. 9 days is included, 10 days is not.

ConversionTenToElevenDays

Conversion lag bucket from 10 to 11 days. 10 days is included, 11 days is not.

ConversionElevenToTwelveDays

Conversion lag bucket from 11 to 12 days. 11 days is included, 12 days is not.

ConversionTwelveToThirteenDays

Conversion lag bucket from 12 to 13 days. 12 days is included, 13 days is not.

ConversionThirteenToFourteenDays

Conversion lag bucket from 13 to 14 days. 13 days is included, 14 days is not.

ConversionFourteenToTwentyOneDays

Conversion lag bucket from 14 to 21 days. 14 days is included, 21 days is not.

ConversionTwentyOneToThirtyDays

Conversion lag bucket from 21 to 30 days. 21 days is included, 30 days is not.

ConversionThirtyToFortyFiveDays

Conversion lag bucket from 30 to 45 days. 30 days is included, 45 days is not.

ConversionFortyFiveToSixtyDays

Conversion lag bucket from 45 to 60 days. 45 days is included, 60 days is not.

ConversionSixtyToNinetyDays

Conversion lag bucket from 60 to 90 days. 60 days is included, 90 days is not.

AdjustmentLessThanOneDay

Conversion adjustment lag bucket from 0 to 1 day. 0 day is included, 1 day is not.

AdjustmentOneToTwoDays

Conversion adjustment lag bucket from 1 to 2 days. 1 day is included, 2 days is not.

AdjustmentTwoToThreeDays

Conversion adjustment lag bucket from 2 to 3 days. 2 days is included, 3 days is not.

AdjustmentThreeToFourDays

Conversion adjustment lag bucket from 3 to 4 days. 3 days is included, 4 days is not.

AdjustmentFourToFiveDays

Conversion adjustment lag bucket from 4 to 5 days. 4 days is included, 5 days is not.

AdjustmentFiveToSixDays

Conversion adjustment lag bucket from 5 to 6 days. 5 days is included, 6 days is not.

AdjustmentSixToSevenDays

Conversion adjustment lag bucket from 6 to 7 days. 6 days is included, 7 days is not.

AdjustmentSevenToEightDays

Conversion adjustment lag bucket from 7 to 8 days. 7 days is included, 8 days is not.

AdjustmentEightToNineDays

Conversion adjustment lag bucket from 8 to 9 days. 8 days is included, 9 days is not.

AdjustmentNineToTenDays

Conversion adjustment lag bucket from 9 to 10 days. 9 days is included, 10 days is not.

AdjustmentTenToElevenDays

Conversion adjustment lag bucket from 10 to 11 days. 10 days is included, 11 days is not.

AdjustmentElevenToTwelveDays

Conversion adjustment lag bucket from 11 to 12 days. 11 days is included, 12 days is not.

AdjustmentTwelveToThirteenDays

Conversion adjustment lag bucket from 12 to 13 days. 12 days is included, 13 days is not.

AdjustmentThirteenToFourteenDays

Conversion adjustment lag bucket from 13 to 14 days. 13 days is included, 14 days is not.

AdjustmentFourteenToTwentyOneDays

Conversion adjustment lag bucket from 14 to 21 days. 14 days is included, 21 days is not.

AdjustmentTwentyOneToThirtyDays

Conversion adjustment lag bucket from 21 to 30 days. 21 days is included, 30 days is not.

AdjustmentThirtyToFortyFiveDays

Conversion adjustment lag bucket from 30 to 45 days. 30 days is included, 45 days is not.

AdjustmentFortyFiveToSixtyDays

Conversion adjustment lag bucket from 45 to 60 days. 45 days is included, 60 days is not.

AdjustmentSixtyToNinetyDays

Conversion adjustment lag bucket from 60 to 90 days. 60 days is included, 90 days is not.

AdjustmentNinetyToOneHundredAndFortyFiveDays

Conversion adjustment lag bucket from 90 to 145 days. 90 days is included, 145 days is not.

ConversionUnknown

Conversion lag bucket UNKNOWN. This is for dates before conversion lag bucket was available in Google Ads.

AdjustmentUnknown

Conversion adjustment lag bucket UNKNOWN. This is for dates before conversion adjustment lag bucket was available in Google Ads.

Implementations

impl ConversionOrAdjustmentLagBucket

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of ConversionOrAdjustmentLagBucket.

pub fn from_i32(value: i32) -> Option<ConversionOrAdjustmentLagBucket>

Converts an i32 to a ConversionOrAdjustmentLagBucket, or None if value is not a valid variant.