Enum gapi_grpc::google::ads::googleads::v4::resources::ad::AdData

pub enum AdData {
    TextAd(TextAdInfo),
    ExpandedTextAd(ExpandedTextAdInfo),
    CallOnlyAd(CallOnlyAdInfo),
    ExpandedDynamicSearchAd(ExpandedDynamicSearchAdInfo),
    HotelAd(HotelAdInfo),
    ShoppingSmartAd(ShoppingSmartAdInfo),
    ShoppingProductAd(ShoppingProductAdInfo),
    GmailAd(GmailAdInfo),
    ImageAd(ImageAdInfo),
    VideoAd(VideoAdInfo),
    ResponsiveSearchAd(ResponsiveSearchAdInfo),
    LegacyResponsiveDisplayAd(LegacyResponsiveDisplayAdInfo),
    AppAd(AppAdInfo),
    LegacyAppInstallAd(LegacyAppInstallAdInfo),
    ResponsiveDisplayAd(ResponsiveDisplayAdInfo),
    LocalAd(LocalAdInfo),
    DisplayUploadAd(DisplayUploadAdInfo),
    AppEngagementAd(AppEngagementAdInfo),
    ShoppingComparisonListingAd(ShoppingComparisonListingAdInfo),
}

Details pertinent to the ad type. Exactly one value must be set.

Variants

TextAd(TextAdInfo)

Immutable. Details pertaining to a text ad.

ExpandedTextAd(ExpandedTextAdInfo)

Details pertaining to an expanded text ad.

CallOnlyAd(CallOnlyAdInfo)

Details pertaining to a call-only ad.

ExpandedDynamicSearchAd(ExpandedDynamicSearchAdInfo)

Immutable. Details pertaining to an Expanded Dynamic Search Ad. This type of ad has its headline, final URLs, and display URL auto-generated at serving time according to domain name specific information provided by dynamic_search_ads_setting linked at the campaign level.

HotelAd(HotelAdInfo)

Details pertaining to a hotel ad.

ShoppingSmartAd(ShoppingSmartAdInfo)

Details pertaining to a Smart Shopping ad.

ShoppingProductAd(ShoppingProductAdInfo)

Details pertaining to a Shopping product ad.

GmailAd(GmailAdInfo)

Immutable. Details pertaining to a Gmail ad.

ImageAd(ImageAdInfo)

Immutable. Details pertaining to an Image ad.

VideoAd(VideoAdInfo)

Details pertaining to a Video ad.

ResponsiveSearchAd(ResponsiveSearchAdInfo)

Details pertaining to a responsive search ad.

LegacyResponsiveDisplayAd(LegacyResponsiveDisplayAdInfo)

Details pertaining to a legacy responsive display ad.

AppAd(AppAdInfo)

Details pertaining to an app ad.

LegacyAppInstallAd(LegacyAppInstallAdInfo)

Immutable. Details pertaining to a legacy app install ad.

ResponsiveDisplayAd(ResponsiveDisplayAdInfo)

Details pertaining to a responsive display ad.

LocalAd(LocalAdInfo)

Details pertaining to a local ad.

DisplayUploadAd(DisplayUploadAdInfo)

Details pertaining to a display upload ad.

AppEngagementAd(AppEngagementAdInfo)

Details pertaining to an app engagement ad.

ShoppingComparisonListingAd(ShoppingComparisonListingAdInfo)

Details pertaining to a Shopping Comparison Listing ad.

Implementations

impl AdData

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<AdData>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize