Module gapi_grpc::google::ads::googleads::v4::resources::recommendation

Modules

campaign_budget_recommendation

target_cpa_opt_in_recommendation

Structs

CallExtensionRecommendation
The Call extension recommendation.

CalloutExtensionRecommendation
The Callout extension recommendation.

CampaignBudgetRecommendation
The budget recommendation for budget constrained campaigns.

EnhancedCpcOptInRecommendation
The Enhanced Cost-Per-Click Opt-In recommendation.

KeywordMatchTypeRecommendation
The keyword match type recommendation.

KeywordRecommendation
The keyword recommendation.

MaximizeClicksOptInRecommendation
The Maximize Clicks opt-in recommendation.

MaximizeConversionsOptInRecommendation
The Maximize Conversions Opt-In recommendation.

MoveUnusedBudgetRecommendation
The move unused budget recommendation.

OptimizeAdRotationRecommendation
The Optimize Ad Rotation recommendation.

RecommendationImpact
The impact of making the change as described in the recommendation. Some types of recommendations may not have impact information.

RecommendationMetrics
Weekly account performance metrics. For some recommendation types, these are averaged over the past 90-day period and hence can be fractional.

SearchPartnersOptInRecommendation
The Search Partners Opt-In recommendation.

SitelinkExtensionRecommendation
The Sitelink extension recommendation.

TargetCpaOptInRecommendation
The Target CPA opt-in recommendation.

TextAdRecommendation
The text ad recommendation.

Enums

Recommendation
The details of recommendation.