Module gapi_grpc::google::ads::googleads::v5::common::operand

Modules

constant_operand

Structs

ConstantOperand
A constant operand in a matching function.

FeedAttributeOperand
A feed attribute operand in a matching function. Used to represent a feed attribute in feed.

FunctionOperand
A function operand in a matching function. Used to represent nested functions.

RequestContextOperand
An operand in a function referring to a value in the request context.

Enums

FunctionArgumentOperand
Different operands that can be used in a matching function. Required.