Struct gapi_grpc::google::ads::googleads::v5::common::Segments

pub struct Segments {
    pub ad_network_type: i32,
    pub budget_campaign_association_status: Option<BudgetCampaignAssociationStatus>,
    pub click_type: i32,
    pub conversion_action: String,
    pub conversion_action_category: i32,
    pub conversion_action_name: String,
    pub conversion_adjustment: bool,
    pub conversion_attribution_event_type: i32,
    pub conversion_lag_bucket: i32,
    pub conversion_or_adjustment_lag_bucket: i32,
    pub date: String,
    pub day_of_week: i32,
    pub device: i32,
    pub external_conversion_source: i32,
    pub geo_target_airport: String,
    pub geo_target_canton: String,
    pub geo_target_city: String,
    pub geo_target_country: String,
    pub geo_target_county: String,
    pub geo_target_district: String,
    pub geo_target_metro: String,
    pub geo_target_most_specific_location: String,
    pub geo_target_postal_code: String,
    pub geo_target_province: String,
    pub geo_target_region: String,
    pub geo_target_state: String,
    pub hotel_booking_window_days: i64,
    pub hotel_center_id: i64,
    pub hotel_check_in_date: String,
    pub hotel_check_in_day_of_week: i32,
    pub hotel_city: String,
    pub hotel_class: i32,
    pub hotel_country: String,
    pub hotel_date_selection_type: i32,
    pub hotel_length_of_stay: i32,
    pub hotel_rate_rule_id: String,
    pub hotel_rate_type: i32,
    pub hotel_price_bucket: i32,
    pub hotel_state: String,
    pub hour: i32,
    pub interaction_on_this_extension: bool,
    pub keyword: Option<Keyword>,
    pub month: String,
    pub month_of_year: i32,
    pub partner_hotel_id: String,
    pub placeholder_type: i32,
    pub product_aggregator_id: i64,
    pub product_bidding_category_level1: String,
    pub product_bidding_category_level2: String,
    pub product_bidding_category_level3: String,
    pub product_bidding_category_level4: String,
    pub product_bidding_category_level5: String,
    pub product_brand: String,
    pub product_channel: i32,
    pub product_channel_exclusivity: i32,
    pub product_condition: i32,
    pub product_country: String,
    pub product_custom_attribute0: String,
    pub product_custom_attribute1: String,
    pub product_custom_attribute2: String,
    pub product_custom_attribute3: String,
    pub product_custom_attribute4: String,
    pub product_item_id: String,
    pub product_language: String,
    pub product_merchant_id: i64,
    pub product_store_id: String,
    pub product_title: String,
    pub product_type_l1: String,
    pub product_type_l2: String,
    pub product_type_l3: String,
    pub product_type_l4: String,
    pub product_type_l5: String,
    pub quarter: String,
    pub search_engine_results_page_type: i32,
    pub search_term_match_type: i32,
    pub slot: i32,
    pub webpage: String,
    pub week: String,
    pub year: i32,
}

Segment only fields.

Fields
ad_network_type: i32

Ad network type.

budget_campaign_association_status: Option<BudgetCampaignAssociationStatus>

Budget campaign association status.

click_type: i32

Click type.

conversion_action: String

Resource name of the conversion action.

conversion_action_category: i32

Conversion action category.

conversion_action_name: String

Conversion action name.

conversion_adjustment: bool

This segments your conversion columns by the original conversion and conversion value vs. the delta if conversions were adjusted. False row has the data as originally stated; While true row has the delta between data now and the data as originally stated. Summing the two together results post-adjustment data.

conversion_attribution_event_type: i32

Conversion attribution event type.

conversion_lag_bucket: i32

An enum value representing the number of days between the impression and the conversion.

conversion_or_adjustment_lag_bucket: i32

An enum value representing the number of days between the impression and the conversion or between the impression and adjustments to the conversion.

date: String

Date to which metrics apply. yyyy-MM-dd format, e.g., 2018-04-17.

day_of_week: i32

Day of the week, e.g., MONDAY.

device: i32

Device to which metrics apply.

external_conversion_source: i32

External conversion source.

geo_target_airport: String

Resource name of the geo target constant that represents an airport.

geo_target_canton: String

Resource name of the geo target constant that represents a canton.

geo_target_city: String

Resource name of the geo target constant that represents a city.

geo_target_country: String

Resource name of the geo target constant that represents a country.

geo_target_county: String

Resource name of the geo target constant that represents a county.

geo_target_district: String

Resource name of the geo target constant that represents a district.

geo_target_metro: String

Resource name of the geo target constant that represents a metro.

geo_target_most_specific_location: String

Resource name of the geo target constant that represents the most specific location.

geo_target_postal_code: String

Resource name of the geo target constant that represents a postal code.

geo_target_province: String

Resource name of the geo target constant that represents a province.

geo_target_region: String

Resource name of the geo target constant that represents a region.

geo_target_state: String

Resource name of the geo target constant that represents a state.

hotel_booking_window_days: i64

Hotel booking window in days.

hotel_center_id: i64

Hotel center ID.

hotel_check_in_date: String

Hotel check-in date. Formatted as yyyy-MM-dd.

hotel_check_in_day_of_week: i32

Hotel check-in day of week.

hotel_city: String

Hotel city.

hotel_class: i32

Hotel class.

hotel_country: String

Hotel country.

hotel_date_selection_type: i32

Hotel date selection type.

hotel_length_of_stay: i32

Hotel length of stay.

hotel_rate_rule_id: String

Hotel rate rule ID.

hotel_rate_type: i32

Hotel rate type.

hotel_price_bucket: i32

Hotel price bucket.

hotel_state: String

Hotel state.

hour: i32

Hour of day as a number between 0 and 23, inclusive.

interaction_on_this_extension: bool

Only used with feed item metrics. Indicates whether the interaction metrics occurred on the feed item itself or a different extension or ad unit.

keyword: Option<Keyword>

Keyword criterion.

month: String

Month as represented by the date of the first day of a month. Formatted as yyyy-MM-dd.

month_of_year: i32

Month of the year, e.g., January.

partner_hotel_id: String

Partner hotel ID.

placeholder_type: i32

Placeholder type. This is only used with feed item metrics.

product_aggregator_id: i64

Aggregator ID of the product.

product_bidding_category_level1: String

Bidding category (level 1) of the product.

product_bidding_category_level2: String

Bidding category (level 2) of the product.

product_bidding_category_level3: String

Bidding category (level 3) of the product.

product_bidding_category_level4: String

Bidding category (level 4) of the product.

product_bidding_category_level5: String

Bidding category (level 5) of the product.

product_brand: String

Brand of the product.

product_channel: i32

Channel of the product.

product_channel_exclusivity: i32

Channel exclusivity of the product.

product_condition: i32

Condition of the product.

product_country: String

Resource name of the geo target constant for the country of sale of the product.

product_custom_attribute0: String

Custom attribute 0 of the product.

product_custom_attribute1: String

Custom attribute 1 of the product.

product_custom_attribute2: String

Custom attribute 2 of the product.

product_custom_attribute3: String

Custom attribute 3 of the product.

product_custom_attribute4: String

Custom attribute 4 of the product.

product_item_id: String

Item ID of the product.

product_language: String

Resource name of the language constant for the language of the product.

product_merchant_id: i64

Merchant ID of the product.

product_store_id: String

Store ID of the product.

product_title: String

Title of the product.

product_type_l1: String

Type (level 1) of the product.

product_type_l2: String

Type (level 2) of the product.

product_type_l3: String

Type (level 3) of the product.

product_type_l4: String

Type (level 4) of the product.

product_type_l5: String

Type (level 5) of the product.

quarter: String

Quarter as represented by the date of the first day of a quarter. Uses the calendar year for quarters, e.g., the second quarter of 2018 starts on 2018-04-01. Formatted as yyyy-MM-dd.

search_engine_results_page_type: i32

Type of the search engine results page.

search_term_match_type: i32

Match type of the keyword that triggered the ad, including variants.

slot: i32

Position of the ad.

webpage: String

Resource name of the ad group criterion that represents webpage criterion.

week: String

Week as defined as Monday through Sunday, and represented by the date of Monday. Formatted as yyyy-MM-dd.

year: i32

Year, formatted as yyyy.

Implementations

impl Segments

pub fn device(&self) -> Device

Returns the enum value of device, or the default if the field is set to an invalid enum value.

pub fn set_device(&mut self, value: Device)

Sets device to the provided enum value.

pub fn conversion_attribution_event_type(
&self
) -> ConversionAttributionEventType

Returns the enum value of conversion_attribution_event_type, or the default if the field is set to an invalid enum value.

pub fn set_conversion_attribution_event_type(
&mut self,
value: ConversionAttributionEventType
)

Sets conversion_attribution_event_type to the provided enum value.

pub fn ad_network_type(&self) -> AdNetworkType

Returns the enum value of ad_network_type, or the default if the field is set to an invalid enum value.

pub fn set_ad_network_type(&mut self, value: AdNetworkType)

Sets ad_network_type to the provided enum value.

pub fn day_of_week(&self) -> DayOfWeek

Returns the enum value of day_of_week, or the default if the field is set to an invalid enum value.

pub fn set_day_of_week(&mut self, value: DayOfWeek)

Sets day_of_week to the provided enum value.

pub fn hotel_check_in_day_of_week(&self) -> DayOfWeek

Returns the enum value of hotel_check_in_day_of_week, or the default if the field is set to an invalid enum value.

pub fn set_hotel_check_in_day_of_week(&mut self, value: DayOfWeek)

Sets hotel_check_in_day_of_week to the provided enum value.

pub fn hotel_date_selection_type(&self) -> HotelDateSelectionType

Returns the enum value of hotel_date_selection_type, or the default if the field is set to an invalid enum value.

pub fn set_hotel_date_selection_type(&mut self, value: HotelDateSelectionType)

Sets hotel_date_selection_type to the provided enum value.

pub fn month_of_year(&self) -> MonthOfYear

Returns the enum value of month_of_year, or the default if the field is set to an invalid enum value.

pub fn set_month_of_year(&mut self, value: MonthOfYear)

Sets month_of_year to the provided enum value.

pub fn placeholder_type(&self) -> PlaceholderType

Returns the enum value of placeholder_type, or the default if the field is set to an invalid enum value.

pub fn set_placeholder_type(&mut self, value: PlaceholderType)

Sets placeholder_type to the provided enum value.

pub fn search_term_match_type(&self) -> SearchTermMatchType

Returns the enum value of search_term_match_type, or the default if the field is set to an invalid enum value.

pub fn set_search_term_match_type(&mut self, value: SearchTermMatchType)

Sets search_term_match_type to the provided enum value.

pub fn slot(&self) -> Slot

Returns the enum value of slot, or the default if the field is set to an invalid enum value.

pub fn set_slot(&mut self, value: Slot)

Sets slot to the provided enum value.

pub fn click_type(&self) -> ClickType

Returns the enum value of click_type, or the default if the field is set to an invalid enum value.

pub fn set_click_type(&mut self, value: ClickType)

Sets click_type to the provided enum value.

pub fn product_channel(&self) -> ProductChannel

Returns the enum value of product_channel, or the default if the field is set to an invalid enum value.

pub fn set_product_channel(&mut self, value: ProductChannel)

Sets product_channel to the provided enum value.

pub fn product_channel_exclusivity(&self) -> ProductChannelExclusivity

Returns the enum value of product_channel_exclusivity, or the default if the field is set to an invalid enum value.

pub fn set_product_channel_exclusivity(
&mut self,
value: ProductChannelExclusivity
)

Sets product_channel_exclusivity to the provided enum value.

pub fn product_condition(&self) -> ProductCondition

Returns the enum value of product_condition, or the default if the field is set to an invalid enum value.

pub fn set_product_condition(&mut self, value: ProductCondition)

Sets product_condition to the provided enum value.

pub fn conversion_lag_bucket(&self) -> ConversionLagBucket

Returns the enum value of conversion_lag_bucket, or the default if the field is set to an invalid enum value.

pub fn set_conversion_lag_bucket(&mut self, value: ConversionLagBucket)

Sets conversion_lag_bucket to the provided enum value.

pub fn conversion_or_adjustment_lag_bucket(
&self
) -> ConversionOrAdjustmentLagBucket

Returns the enum value of conversion_or_adjustment_lag_bucket, or the default if the field is set to an invalid enum value.

pub fn set_conversion_or_adjustment_lag_bucket(
&mut self,
value: ConversionOrAdjustmentLagBucket
)

Sets conversion_or_adjustment_lag_bucket to the provided enum value.

pub fn conversion_action_category(&self) -> ConversionActionCategory

Returns the enum value of conversion_action_category, or the default if the field is set to an invalid enum value.

pub fn set_conversion_action_category(
&mut self,
value: ConversionActionCategory
)

Sets conversion_action_category to the provided enum value.

pub fn external_conversion_source(&self) -> ExternalConversionSource

Returns the enum value of external_conversion_source, or the default if the field is set to an invalid enum value.

pub fn set_external_conversion_source(
&mut self,
value: ExternalConversionSource
)

Sets external_conversion_source to the provided enum value.

pub fn search_engine_results_page_type(&self) -> SearchEngineResultsPageType

Returns the enum value of search_engine_results_page_type, or the default if the field is set to an invalid enum value.

pub fn set_search_engine_results_page_type(
&mut self,
value: SearchEngineResultsPageType
)

Sets search_engine_results_page_type to the provided enum value.

pub fn hotel_rate_type(&self) -> HotelRateType

Returns the enum value of hotel_rate_type, or the default if the field is set to an invalid enum value.

pub fn set_hotel_rate_type(&mut self, value: HotelRateType)

Sets hotel_rate_type to the provided enum value.

pub fn hotel_price_bucket(&self) -> HotelPriceBucket

Returns the enum value of hotel_price_bucket, or the default if the field is set to an invalid enum value.

pub fn set_hotel_price_bucket(&mut self, value: HotelPriceBucket)

Sets hotel_price_bucket to the provided enum value.