Module gapi_grpc::google::ads::googleads::v5::errors

Modules

access_invitation_error_enum

account_budget_proposal_error_enum

account_link_error_enum

ad_customizer_error_enum

ad_error_enum

ad_group_ad_error_enum

ad_group_bid_modifier_error_enum

ad_group_criterion_error_enum

ad_group_error_enum

ad_group_feed_error_enum

ad_parameter_error_enum

ad_sharing_error_enum

adx_error_enum

asset_error_enum

asset_link_error_enum

authentication_error_enum

authorization_error_enum

batch_job_error_enum

bidding_error_enum

bidding_strategy_error_enum

billing_setup_error_enum

campaign_budget_error_enum

campaign_criterion_error_enum

campaign_draft_error_enum

campaign_error_enum

campaign_experiment_error_enum

campaign_feed_error_enum

campaign_shared_set_error_enum

change_status_error_enum

collection_size_error_enum

context_error_enum

conversion_action_error_enum

conversion_adjustment_upload_error_enum

conversion_upload_error_enum

country_code_error_enum

criterion_error_enum

currency_code_error_enum

custom_interest_error_enum

customer_client_link_error_enum

customer_error_enum

customer_feed_error_enum

customer_manager_link_error_enum

database_error_enum

date_error_enum

date_range_error_enum

distinct_error_enum

enum_error_enum

error_code

error_location

extension_feed_item_error_enum

extension_setting_error_enum

feed_attribute_reference_error_enum

feed_error_enum

feed_item_error_enum

feed_item_target_error_enum

feed_item_validation_error_enum

feed_mapping_error_enum

field_error_enum

field_mask_error_enum

function_error_enum

function_parsing_error_enum

geo_target_constant_suggestion_error_enum

header_error_enum

id_error_enum

image_error_enum

internal_error_enum

invoice_error_enum

keyword_plan_ad_group_error_enum

keyword_plan_ad_group_keyword_error_enum

keyword_plan_campaign_error_enum

keyword_plan_campaign_keyword_error_enum

keyword_plan_error_enum

keyword_plan_idea_error_enum

label_error_enum

language_code_error_enum

list_operation_error_enum

manager_link_error_enum

media_bundle_error_enum

media_file_error_enum

media_upload_error_enum

multiplier_error_enum

mutate_error_enum

new_resource_creation_error_enum

not_allowlisted_error_enum

not_empty_error_enum

null_error_enum

offline_user_data_job_error_enum

operation_access_denied_error_enum

operator_error_enum

partial_failure_error_enum

payments_account_error_enum

policy_finding_error_enum

policy_validation_parameter_error_enum

policy_violation_error_enum

query_error_enum

quota_error_details

quota_error_enum

range_error_enum

reach_plan_error_enum

recommendation_error_enum

region_code_error_enum

request_error_enum

resource_access_denied_error_enum

resource_count_limit_exceeded_error_enum

setting_error_enum

shared_criterion_error_enum

shared_set_error_enum

size_limit_error_enum

string_format_error_enum

string_length_error_enum

third_party_app_analytics_link_error_enum

time_zone_error_enum

url_field_error_enum

user_data_error_enum

user_list_error_enum

youtube_video_registration_error_enum

Structs

AccessInvitationErrorEnum
Container for enum describing possible AccessInvitation errors.

AccountBudgetProposalErrorEnum
Container for enum describing possible account budget proposal errors.

AccountLinkErrorEnum
Container for enum describing possible account link errors.

AdCustomizerErrorEnum
Container for enum describing possible ad customizer errors.

AdErrorEnum
Container for enum describing possible ad errors.

AdGroupAdErrorEnum
Container for enum describing possible ad group ad errors.

AdGroupBidModifierErrorEnum
Container for enum describing possible ad group bid modifier errors.

AdGroupCriterionErrorEnum
Container for enum describing possible ad group criterion errors.

AdGroupErrorEnum
Container for enum describing possible ad group errors.

AdGroupFeedErrorEnum
Container for enum describing possible ad group feed errors.

AdParameterErrorEnum
Container for enum describing possible ad parameter errors.

AdSharingErrorEnum
Container for enum describing possible ad sharing errors.

AdxErrorEnum
Container for enum describing possible adx errors.

AssetErrorEnum
Container for enum describing possible asset errors.

AssetLinkErrorEnum
Container for enum describing possible asset link errors.

AuthenticationErrorEnum
Container for enum describing possible authentication errors.

AuthorizationErrorEnum
Container for enum describing possible authorization errors.

BatchJobErrorEnum
Container for enum describing possible batch job errors.

BiddingErrorEnum
Container for enum describing possible bidding errors.

BiddingStrategyErrorEnum
Container for enum describing possible bidding strategy errors.

BillingSetupErrorEnum
Container for enum describing possible billing setup errors.

CampaignBudgetErrorEnum
Container for enum describing possible campaign budget errors.

CampaignCriterionErrorEnum
Container for enum describing possible campaign criterion errors.

CampaignDraftErrorEnum
Container for enum describing possible campaign draft errors.

CampaignErrorEnum
Container for enum describing possible campaign errors.

CampaignExperimentErrorEnum
Container for enum describing possible campaign experiment errors.

CampaignFeedErrorEnum
Container for enum describing possible campaign feed errors.

CampaignSharedSetErrorEnum
Container for enum describing possible campaign shared set errors.

ChangeStatusErrorEnum
Container for enum describing possible change status errors.

CollectionSizeErrorEnum
Container for enum describing possible collection size errors.

ContextErrorEnum
Container for enum describing possible context errors.

ConversionActionErrorEnum
Container for enum describing possible conversion action errors.

ConversionAdjustmentUploadErrorEnum
Container for enum describing possible conversion adjustment upload errors.

ConversionUploadErrorEnum
Container for enum describing possible conversion upload errors.

CountryCodeErrorEnum
Container for enum describing country code errors.

CriterionErrorEnum
Container for enum describing possible criterion errors.

CurrencyCodeErrorEnum
Container for enum describing possible currency code errors.

CustomInterestErrorEnum
Container for enum describing possible custom interest errors.

CustomerClientLinkErrorEnum
Container for enum describing possible CustomeClientLink errors.

CustomerErrorEnum
Container for enum describing possible customer errors.

CustomerFeedErrorEnum
Container for enum describing possible customer feed errors.

CustomerManagerLinkErrorEnum
Container for enum describing possible CustomerManagerLink errors.

DatabaseErrorEnum
Container for enum describing possible database errors.

DateErrorEnum
Container for enum describing possible date errors.

DateRangeErrorEnum
Container for enum describing possible date range errors.

DistinctErrorEnum
Container for enum describing possible distinct errors.

EnumErrorEnum
Container for enum describing possible enum errors.

ErrorCode
The error reason represented by type and enum.

ErrorDetails
Additional error details.

ErrorLocation
Describes the part of the request proto that caused the error.

ExtensionFeedItemErrorEnum
Container for enum describing possible extension feed item error.

ExtensionSettingErrorEnum
Container for enum describing validation errors of extension settings.

FeedAttributeReferenceErrorEnum
Container for enum describing possible feed attribute reference errors.

FeedErrorEnum
Container for enum describing possible feed errors.

FeedItemErrorEnum
Container for enum describing possible feed item errors.

FeedItemTargetErrorEnum
Container for enum describing possible feed item target errors.

FeedItemValidationErrorEnum
Container for enum describing possible validation errors of a feed item.

FeedMappingErrorEnum
Container for enum describing possible feed item errors.

FieldErrorEnum
Container for enum describing possible field errors.

FieldMaskErrorEnum
Container for enum describing possible field mask errors.

FunctionErrorEnum
Container for enum describing possible function errors.

FunctionParsingErrorEnum
Container for enum describing possible function parsing errors.

GeoTargetConstantSuggestionErrorEnum
Container for enum describing possible geo target constant suggestion errors.

GoogleAdsError
GoogleAds-specific error.

GoogleAdsFailure
Describes how a GoogleAds API call failed. It's returned inside google.rpc.Status.details when a call fails.

HeaderErrorEnum
Container for enum describing possible header errors.

IdErrorEnum
Container for enum describing possible id errors.

ImageErrorEnum
Container for enum describing possible image errors.

InternalErrorEnum
Container for enum describing possible internal errors.

InvoiceErrorEnum
Container for enum describing possible invoice errors.

KeywordPlanAdGroupErrorEnum
Container for enum describing possible errors from applying a keyword plan ad group.

KeywordPlanAdGroupKeywordErrorEnum
Container for enum describing possible errors from applying an ad group keyword or a campaign keyword from a keyword plan.

KeywordPlanCampaignErrorEnum
Container for enum describing possible errors from applying a keyword plan campaign.

KeywordPlanCampaignKeywordErrorEnum
Container for enum describing possible errors from applying a keyword plan campaign keyword.

KeywordPlanErrorEnum
Container for enum describing possible errors from applying a keyword plan resource (keyword plan, keyword plan campaign, keyword plan ad group or keyword plan keyword) or KeywordPlanService RPC.

KeywordPlanIdeaErrorEnum
Container for enum describing possible errors from KeywordPlanIdeaService.

LabelErrorEnum
Container for enum describing possible label errors.

LanguageCodeErrorEnum
Container for enum describing language code errors.

ListOperationErrorEnum
Container for enum describing possible list operation errors.

ManagerLinkErrorEnum
Container for enum describing possible ManagerLink errors.

MediaBundleErrorEnum
Container for enum describing possible media bundle errors.

MediaFileErrorEnum
Container for enum describing possible media file errors.

MediaUploadErrorEnum
Container for enum describing possible media uploading errors.

MultiplierErrorEnum
Container for enum describing possible multiplier errors.

MutateErrorEnum
Container for enum describing possible mutate errors.

NewResourceCreationErrorEnum
Container for enum describing possible new resource creation errors.

NotAllowlistedErrorEnum
Container for enum describing possible not allowlisted errors.

NotEmptyErrorEnum
Container for enum describing possible not empty errors.

NullErrorEnum
Container for enum describing possible null errors.

OfflineUserDataJobErrorEnum
Container for enum describing possible offline user data job errors.

OperationAccessDeniedErrorEnum
Container for enum describing possible operation access denied errors.

OperatorErrorEnum
Container for enum describing possible operator errors.

PartialFailureErrorEnum
Container for enum describing possible partial failure errors.

PaymentsAccountErrorEnum
Container for enum describing possible errors in payments account service.

PolicyFindingDetails
Error returned as part of a mutate response. This error indicates one or more policy findings in the fields of a resource.

PolicyFindingErrorEnum
Container for enum describing possible policy finding errors.

PolicyValidationParameterErrorEnum
Container for enum describing possible policy validation parameter errors.

PolicyViolationDetails
Error returned as part of a mutate response. This error indicates single policy violation by some text in one of the fields.

PolicyViolationErrorEnum
Container for enum describing possible policy violation errors.

QueryErrorEnum
Container for enum describing possible query errors.

QuotaErrorDetails
Additional quota error details when there is QuotaError.

QuotaErrorEnum
Container for enum describing possible quota errors.

RangeErrorEnum
Container for enum describing possible range errors.

ReachPlanErrorEnum
Container for enum describing possible errors returned from the ReachPlanService.

RecommendationErrorEnum
Container for enum describing possible errors from applying a recommendation.

RegionCodeErrorEnum
Container for enum describing possible region code errors.

RequestErrorEnum
Container for enum describing possible request errors.

ResourceAccessDeniedErrorEnum
Container for enum describing possible resource access denied errors.

ResourceCountLimitExceededErrorEnum
Container for enum describing possible resource count limit exceeded errors.

SettingErrorEnum
Container for enum describing possible setting errors.

SharedCriterionErrorEnum
Container for enum describing possible shared criterion errors.

SharedSetErrorEnum
Container for enum describing possible shared set errors.

SizeLimitErrorEnum
Container for enum describing possible size limit errors.

StringFormatErrorEnum
Container for enum describing possible string format errors.

StringLengthErrorEnum
Container for enum describing possible string length errors.

ThirdPartyAppAnalyticsLinkErrorEnum
Container for enum describing possible third party app analytics link errors.

TimeZoneErrorEnum
Container for enum describing possible time zone errors.

UrlFieldErrorEnum
Container for enum describing possible url field errors.

UserDataErrorEnum
Container for enum describing possible user data errors.

UserListErrorEnum
Container for enum describing possible user list errors.

YoutubeVideoRegistrationErrorEnum
Container for enum describing YouTube video registration errors.