Enum gapi_grpc::google::ads::googleads::v5::services::mutate_operation_response::Response

pub enum Response {
    AdGroupAdLabelResult(MutateAdGroupAdLabelResult),
    AdGroupAdResult(MutateAdGroupAdResult),
    AdGroupBidModifierResult(MutateAdGroupBidModifierResult),
    AdGroupCriterionLabelResult(MutateAdGroupCriterionLabelResult),
    AdGroupCriterionResult(MutateAdGroupCriterionResult),
    AdGroupExtensionSettingResult(MutateAdGroupExtensionSettingResult),
    AdGroupFeedResult(MutateAdGroupFeedResult),
    AdGroupLabelResult(MutateAdGroupLabelResult),
    AdGroupResult(MutateAdGroupResult),
    AdParameterResult(MutateAdParameterResult),
    AdResult(MutateAdResult),
    AssetResult(MutateAssetResult),
    BiddingStrategyResult(MutateBiddingStrategyResult),
    CampaignAssetResult(MutateCampaignAssetResult),
    CampaignBidModifierResult(MutateCampaignBidModifierResult),
    CampaignBudgetResult(MutateCampaignBudgetResult),
    CampaignCriterionResult(MutateCampaignCriterionResult),
    CampaignDraftResult(MutateCampaignDraftResult),
    CampaignExperimentResult(MutateCampaignExperimentResult),
    CampaignExtensionSettingResult(MutateCampaignExtensionSettingResult),
    CampaignFeedResult(MutateCampaignFeedResult),
    CampaignLabelResult(MutateCampaignLabelResult),
    CampaignResult(MutateCampaignResult),
    CampaignSharedSetResult(MutateCampaignSharedSetResult),
    ConversionActionResult(MutateConversionActionResult),
    CustomerExtensionSettingResult(MutateCustomerExtensionSettingResult),
    CustomerFeedResult(MutateCustomerFeedResult),
    CustomerLabelResult(MutateCustomerLabelResult),
    CustomerNegativeCriterionResult(MutateCustomerNegativeCriteriaResult),
    CustomerResult(MutateCustomerResult),
    ExtensionFeedItemResult(MutateExtensionFeedItemResult),
    FeedItemResult(MutateFeedItemResult),
    FeedItemTargetResult(MutateFeedItemTargetResult),
    FeedMappingResult(MutateFeedMappingResult),
    FeedResult(MutateFeedResult),
    KeywordPlanAdGroupResult(MutateKeywordPlanAdGroupResult),
    KeywordPlanCampaignResult(MutateKeywordPlanCampaignResult),
    KeywordPlanAdGroupKeywordResult(MutateKeywordPlanAdGroupKeywordResult),
    KeywordPlanCampaignKeywordResult(MutateKeywordPlanCampaignKeywordResult),
    KeywordPlanResult(MutateKeywordPlansResult),
    LabelResult(MutateLabelResult),
    MediaFileResult(MutateMediaFileResult),
    RemarketingActionResult(MutateRemarketingActionResult),
    SharedCriterionResult(MutateSharedCriterionResult),
    SharedSetResult(MutateSharedSetResult),
    UserListResult(MutateUserListResult),
}

The mutate response.

Variants

AdGroupAdLabelResult(MutateAdGroupAdLabelResult)

The result for the ad group ad label mutate.

AdGroupAdResult(MutateAdGroupAdResult)

The result for the ad group ad mutate.

AdGroupBidModifierResult(MutateAdGroupBidModifierResult)

The result for the ad group bid modifier mutate.

AdGroupCriterionLabelResult(MutateAdGroupCriterionLabelResult)

The result for the ad group criterion label mutate.

AdGroupCriterionResult(MutateAdGroupCriterionResult)

The result for the ad group criterion mutate.

AdGroupExtensionSettingResult(MutateAdGroupExtensionSettingResult)

The result for the ad group extension setting mutate.

AdGroupFeedResult(MutateAdGroupFeedResult)

The result for the ad group feed mutate.

AdGroupLabelResult(MutateAdGroupLabelResult)

The result for the ad group label mutate.

AdGroupResult(MutateAdGroupResult)

The result for the ad group mutate.

AdParameterResult(MutateAdParameterResult)

The result for the ad parameter mutate.

AdResult(MutateAdResult)

The result for the ad mutate.

AssetResult(MutateAssetResult)

The result for the asset mutate.

BiddingStrategyResult(MutateBiddingStrategyResult)

The result for the bidding strategy mutate.

CampaignAssetResult(MutateCampaignAssetResult)

The result for the campaign asset mutate.

CampaignBidModifierResult(MutateCampaignBidModifierResult)

The result for the campaign bid modifier mutate.

CampaignBudgetResult(MutateCampaignBudgetResult)

The result for the campaign budget mutate.

CampaignCriterionResult(MutateCampaignCriterionResult)

The result for the campaign criterion mutate.

CampaignDraftResult(MutateCampaignDraftResult)

The result for the campaign draft mutate.

CampaignExperimentResult(MutateCampaignExperimentResult)

The result for the campaign experiment mutate.

CampaignExtensionSettingResult(MutateCampaignExtensionSettingResult)

The result for the campaign extension setting mutate.

CampaignFeedResult(MutateCampaignFeedResult)

The result for the campaign feed mutate.

CampaignLabelResult(MutateCampaignLabelResult)

The result for the campaign label mutate.

CampaignResult(MutateCampaignResult)

The result for the campaign mutate.

CampaignSharedSetResult(MutateCampaignSharedSetResult)

The result for the campaign shared set mutate.

ConversionActionResult(MutateConversionActionResult)

The result for the conversion action mutate.

CustomerExtensionSettingResult(MutateCustomerExtensionSettingResult)

The result for the customer extension setting mutate.

CustomerFeedResult(MutateCustomerFeedResult)

The result for the customer feed mutate.

CustomerLabelResult(MutateCustomerLabelResult)

The result for the customer label mutate.

CustomerNegativeCriterionResult(MutateCustomerNegativeCriteriaResult)

The result for the customer negative criterion mutate.

CustomerResult(MutateCustomerResult)

The result for the customer mutate.

ExtensionFeedItemResult(MutateExtensionFeedItemResult)

The result for the extension feed item mutate.

FeedItemResult(MutateFeedItemResult)

The result for the feed item mutate.

FeedItemTargetResult(MutateFeedItemTargetResult)

The result for the feed item target mutate.

FeedMappingResult(MutateFeedMappingResult)

The result for the feed mapping mutate.

FeedResult(MutateFeedResult)

The result for the feed mutate.

KeywordPlanAdGroupResult(MutateKeywordPlanAdGroupResult)

The result for the keyword plan ad group mutate.

KeywordPlanCampaignResult(MutateKeywordPlanCampaignResult)

The result for the keyword plan campaign mutate.

KeywordPlanAdGroupKeywordResult(MutateKeywordPlanAdGroupKeywordResult)

The result for the keyword plan ad group keyword mutate.

KeywordPlanCampaignKeywordResult(MutateKeywordPlanCampaignKeywordResult)

The result for the keyword plan campaign keyword mutate.

KeywordPlanResult(MutateKeywordPlansResult)

The result for the keyword plan mutate.

LabelResult(MutateLabelResult)

The result for the label mutate.

MediaFileResult(MutateMediaFileResult)

The result for the media file mutate.

RemarketingActionResult(MutateRemarketingActionResult)

The result for the remarketing action mutate.

SharedCriterionResult(MutateSharedCriterionResult)

The result for the shared criterion mutate.

SharedSetResult(MutateSharedSetResult)

The result for the shared set mutate.

UserListResult(MutateUserListResult)

The result for the user list mutate.

Implementations

impl Response

pub fn encode<B>(&self, buf: &mut B) where
B: BufMut,

pub fn merge<B>(
field: &mut Option<Response>,
tag: u32,
wire_type: WireType,
buf: &mut B,
ctx: DecodeContext
) -> Result<(), DecodeError> where
B: Buf,

pub fn encoded_len(&self) -> usize