Module gapi_grpc::google::ads::googleads::v6::enums

Modules

access_reason_enum

access_role_enum

account_budget_proposal_status_enum

account_budget_proposal_type_enum

account_budget_status_enum

account_link_status_enum

ad_customizer_placeholder_field_enum

ad_destination_type_enum

ad_group_ad_rotation_mode_enum

ad_group_ad_status_enum

ad_group_criterion_approval_status_enum

ad_group_criterion_status_enum

ad_group_status_enum

ad_group_type_enum

ad_network_type_enum

ad_serving_optimization_status_enum

ad_strength_enum

ad_type_enum

advertising_channel_sub_type_enum

advertising_channel_type_enum

affiliate_location_feed_relationship_type_enum

affiliate_location_placeholder_field_enum

age_range_type_enum

app_campaign_app_store_enum

app_campaign_bidding_strategy_goal_type_enum

app_payment_model_type_enum

app_placeholder_field_enum

app_store_enum

app_url_operating_system_type_enum

asset_field_type_enum

asset_link_status_enum

asset_performance_label_enum

asset_type_enum

attribution_model_enum

batch_job_status_enum

bid_modifier_source_enum

bidding_source_enum

bidding_strategy_status_enum

bidding_strategy_type_enum

billing_setup_status_enum

brand_safety_suitability_enum

budget_campaign_association_status_enum

budget_delivery_method_enum

budget_period_enum

budget_status_enum

budget_type_enum

call_conversion_reporting_state_enum

call_placeholder_field_enum

callout_placeholder_field_enum

campaign_criterion_status_enum

campaign_draft_status_enum

campaign_experiment_status_enum

campaign_experiment_traffic_split_type_enum

campaign_experiment_type_enum

campaign_serving_status_enum

campaign_shared_set_status_enum

campaign_status_enum

change_client_type_enum

change_event_resource_type_enum

change_status_operation_enum

change_status_resource_type_enum

click_type_enum

combined_audience_status_enum

content_label_type_enum

conversion_action_category_enum

conversion_action_counting_type_enum

conversion_action_status_enum

conversion_action_type_enum

conversion_adjustment_type_enum

conversion_attribution_event_type_enum

conversion_lag_bucket_enum

conversion_or_adjustment_lag_bucket_enum

criterion_category_channel_availability_mode_enum

criterion_category_locale_availability_mode_enum

criterion_system_serving_status_enum

criterion_type_enum

custom_audience_member_type_enum

custom_audience_status_enum

custom_audience_type_enum

custom_interest_member_type_enum

custom_interest_status_enum

custom_interest_type_enum

custom_placeholder_field_enum

customer_match_upload_key_type_enum

customer_pay_per_conversion_eligibility_failure_reason_enum

data_driven_model_status_enum

day_of_week_enum

device_enum

display_ad_format_setting_enum

display_upload_product_type_enum

distance_bucket_enum

dsa_page_feed_criterion_field_enum

education_placeholder_field_enum

extension_setting_device_enum

extension_type_enum

external_conversion_source_enum

feed_attribute_type_enum

feed_item_quality_approval_status_enum

feed_item_quality_disapproval_reason_enum

feed_item_set_status_enum

feed_item_set_string_filter_type_enum

feed_item_status_enum

feed_item_target_device_enum

feed_item_target_status_enum

feed_item_target_type_enum

feed_item_validation_status_enum

feed_link_status_enum

feed_mapping_criterion_type_enum

feed_mapping_status_enum

feed_origin_enum

feed_status_enum

flight_placeholder_field_enum

frequency_cap_event_type_enum

frequency_cap_level_enum

frequency_cap_time_unit_enum

gender_type_enum

geo_target_constant_status_enum

geo_targeting_restriction_enum

geo_targeting_type_enum

google_ads_field_category_enum

google_ads_field_data_type_enum

hotel_date_selection_type_enum

hotel_placeholder_field_enum

hotel_price_bucket_enum

hotel_rate_type_enum

income_range_type_enum

interaction_event_type_enum

interaction_type_enum

invoice_type_enum

job_placeholder_field_enum

keyword_match_type_enum

keyword_plan_competition_level_enum

keyword_plan_forecast_interval_enum

keyword_plan_network_enum

label_status_enum

lead_form_call_to_action_type_enum

lead_form_desired_intent_enum

lead_form_field_user_input_type_enum

lead_form_post_submit_call_to_action_type_enum

legacy_app_install_ad_app_store_enum

linked_account_type_enum

listing_group_type_enum

local_placeholder_field_enum

location_extension_targeting_criterion_field_enum

location_group_radius_units_enum

location_placeholder_field_enum

location_source_type_enum

manager_link_status_enum

matching_function_context_type_enum

matching_function_operator_enum

media_type_enum

merchant_center_link_status_enum

message_placeholder_field_enum

mime_type_enum

minute_of_hour_enum

mobile_app_vendor_enum

mobile_device_type_enum

month_of_year_enum

negative_geo_target_type_enum

offline_user_data_job_failure_reason_enum

offline_user_data_job_status_enum

offline_user_data_job_type_enum

operating_system_version_operator_type_enum

optimization_goal_type_enum

parental_status_type_enum

payment_mode_enum

placeholder_type_enum

placement_type_enum

policy_approval_status_enum

policy_review_status_enum

policy_topic_entry_type_enum

policy_topic_evidence_destination_mismatch_url_type_enum

policy_topic_evidence_destination_not_working_device_enum

policy_topic_evidence_destination_not_working_dns_error_type_enum

positive_geo_target_type_enum

preferred_content_type_enum

price_extension_price_qualifier_enum

price_extension_price_unit_enum

price_extension_type_enum

price_placeholder_field_enum

product_bidding_category_level_enum

product_bidding_category_status_enum

product_channel_enum

product_channel_exclusivity_enum

product_condition_enum

product_custom_attribute_index_enum

product_type_level_enum

promotion_extension_discount_modifier_enum

promotion_extension_occasion_enum

promotion_placeholder_field_enum

proximity_radius_units_enum

quality_score_bucket_enum

reach_plan_ad_length_enum

reach_plan_age_range_enum

reach_plan_network_enum

real_estate_placeholder_field_enum

recommendation_type_enum

resource_change_operation_enum

response_content_type_enum

search_engine_results_page_type_enum

search_term_match_type_enum

search_term_targeting_status_enum

served_asset_field_type_enum

shared_set_status_enum

shared_set_type_enum

simulation_modification_method_enum

simulation_type_enum

sitelink_placeholder_field_enum

slot_enum

spending_limit_type_enum

structured_snippet_placeholder_field_enum

summary_row_setting_enum

system_managed_resource_source_enum

target_cpa_opt_in_recommendation_goal_enum

target_impression_share_location_enum

targeting_dimension_enum

time_type_enum

tracking_code_page_format_enum

tracking_code_type_enum

travel_placeholder_field_enum

user_identifier_source_enum

user_interest_taxonomy_type_enum

user_list_access_status_enum

user_list_closing_reason_enum

user_list_combined_rule_operator_enum

user_list_crm_data_source_type_enum

user_list_date_rule_item_operator_enum

user_list_logical_rule_operator_enum

user_list_membership_status_enum

user_list_number_rule_item_operator_enum

user_list_prepopulation_status_enum

user_list_rule_type_enum

user_list_size_range_enum

user_list_string_rule_item_operator_enum

user_list_type_enum

vanity_pharma_display_url_mode_enum

vanity_pharma_text_enum

webpage_condition_operand_enum

webpage_condition_operator_enum

Structs

AccessReasonEnum
Indicates the way the resource such as user list is related to a user.

AccessRoleEnum
Container for enum describing possible access role for user.

AccountBudgetProposalStatusEnum
Message describing AccountBudgetProposal statuses.

AccountBudgetProposalTypeEnum
Message describing AccountBudgetProposal types.

AccountBudgetStatusEnum
Message describing AccountBudget statuses.

AccountLinkStatusEnum
Container for enum describing possible statuses of an account link.

AdCustomizerPlaceholderFieldEnum
Values for Ad Customizer placeholder fields.

AdDestinationTypeEnum
Container for enumeration of Google Ads destination types.

AdGroupAdRotationModeEnum
Container for enum describing possible ad rotation modes of ads within an ad group.

AdGroupAdStatusEnum
Container for enum describing possible statuses of an AdGroupAd.

AdGroupCriterionApprovalStatusEnum
Container for enum describing possible AdGroupCriterion approval statuses.

AdGroupCriterionStatusEnum
Message describing AdGroupCriterion statuses.

AdGroupStatusEnum
Container for enum describing possible statuses of an ad group.

AdGroupTypeEnum
Defines types of an ad group, specific to a particular campaign channel type. This type drives validations that restrict which entities can be added to the ad group.

AdNetworkTypeEnum
Container for enumeration of Google Ads network types.

AdServingOptimizationStatusEnum
Possible ad serving statuses of a campaign.

AdStrengthEnum
Container for enum describing possible ad strengths.

AdTypeEnum
Container for enum describing possible types of an ad.

AdvertisingChannelSubTypeEnum
An immutable specialization of an Advertising Channel.

AdvertisingChannelTypeEnum
The channel type a campaign may target to serve on.

AffiliateLocationFeedRelationshipTypeEnum
Container for enum describing possible values for a relationship type for an affiliate location feed.

AffiliateLocationPlaceholderFieldEnum
Values for Affiliate Location placeholder fields.

AgeRangeTypeEnum
Container for enum describing the type of demographic age ranges.

AppCampaignAppStoreEnum
The application store that distributes mobile applications.

AppCampaignBiddingStrategyGoalTypeEnum
Container for enum describing goal towards which the bidding strategy of an app campaign should optimize for.

AppPaymentModelTypeEnum
Represents a criterion for targeting paid apps.

AppPlaceholderFieldEnum
Values for App placeholder fields.

AppStoreEnum
Container for enum describing app store type in an app extension.

AppUrlOperatingSystemTypeEnum
The possible OS types for a deeplink AppUrl.

AssetFieldTypeEnum
Container for enum describing the possible placements of an asset.

AssetLinkStatusEnum
Container for enum describing possible statuses of an asset link.

AssetPerformanceLabelEnum
Container for enum describing the performance label of an asset.

AssetTypeEnum
Container for enum describing the types of asset.

AttributionModelEnum
Container for enum representing the attribution model that describes how to distribute credit for a particular conversion across potentially many prior interactions.

BatchJobStatusEnum
Container for enum describing possible batch job statuses.

BidModifierSourceEnum
Container for enum describing possible bid modifier sources.

BiddingSourceEnum
Container for enum describing possible bidding sources.

BiddingStrategyStatusEnum
Message describing BiddingStrategy statuses.

BiddingStrategyTypeEnum
Container for enum describing possible bidding strategy types.

BillingSetupStatusEnum
Message describing BillingSetup statuses.

BrandSafetySuitabilityEnum
Container for enum with 3-Tier brand safety suitability control.

BudgetCampaignAssociationStatusEnum
Message describing the status of the association between the Budget and the Campaign.

BudgetDeliveryMethodEnum
Message describing Budget delivery methods. A delivery method determines the rate at which the Budget is spent.

BudgetPeriodEnum
Message describing Budget period.

BudgetStatusEnum
Message describing a Budget status

BudgetTypeEnum
Describes Budget types.

CallConversionReportingStateEnum
Container for enum describing possible data types for call conversion reporting state.

CallPlaceholderFieldEnum
Values for Call placeholder fields.

CalloutPlaceholderFieldEnum
Values for Callout placeholder fields.

CampaignCriterionStatusEnum
Message describing CampaignCriterion statuses.

CampaignDraftStatusEnum
Container for enum describing possible statuses of a campaign draft.

CampaignExperimentStatusEnum
Container for enum describing possible statuses of a campaign experiment.

CampaignExperimentTrafficSplitTypeEnum
Container for enum describing campaign experiment traffic split type.

CampaignExperimentTypeEnum
Container for enum describing campaign experiment type.

CampaignServingStatusEnum
Message describing Campaign serving statuses.

CampaignSharedSetStatusEnum
Container for enum describing types of campaign shared set statuses.

CampaignStatusEnum
Container for enum describing possible statuses of a campaign.

ChangeClientTypeEnum
Container for enum describing the sources that the change event resource was made through.

ChangeEventResourceTypeEnum
Container for enum describing supported resource types for the ChangeEvent resource.

ChangeStatusOperationEnum
Container for enum describing operations for the ChangeStatus resource.

ChangeStatusResourceTypeEnum
Container for enum describing supported resource types for the ChangeStatus resource.

ClickTypeEnum
Container for enumeration of Google Ads click types.

CombinedAudienceStatusEnum
The status of combined audience.

ContentLabelTypeEnum
Container for enum describing content label types in ContentLabel.

ConversionActionCategoryEnum
Container for enum describing the category of conversions that are associated with a ConversionAction.

ConversionActionCountingTypeEnum
Container for enum describing the conversion deduplication mode for conversion optimizer.

ConversionActionStatusEnum
Container for enum describing possible statuses of a conversion action.

ConversionActionTypeEnum
Container for enum describing possible types of a conversion action.

ConversionAdjustmentTypeEnum
Container for enum describing conversion adjustment types.

ConversionAttributionEventTypeEnum
Container for enum indicating the event type the conversion is attributed to.

ConversionLagBucketEnum
Container for enum representing the number of days between impression and conversion.

ConversionOrAdjustmentLagBucketEnum
Container for enum representing the number of days between the impression and the conversion or between the impression and adjustments to the conversion.

CriterionCategoryChannelAvailabilityModeEnum
Describes channel availability mode for a criterion availability - whether the availability is meant to include all advertising channels, or a particular channel with all its channel subtypes, or a channel with a certain subset of channel subtypes.

CriterionCategoryLocaleAvailabilityModeEnum
Describes locale availability mode for a criterion availability - whether it's available globally, or a particular country with all languages, or a particular language with all countries, or a country-language pair.

CriterionSystemServingStatusEnum
Container for enum describing possible criterion system serving statuses.

CriterionTypeEnum
The possible types of a criterion.

CustomAudienceMemberTypeEnum
The type of custom audience member.

CustomAudienceStatusEnum
The status of custom audience.

CustomAudienceTypeEnum
The types of custom audience.

CustomInterestMemberTypeEnum
The types of custom interest member, either KEYWORD or URL.

CustomInterestStatusEnum
The status of custom interest.

CustomInterestTypeEnum
The types of custom interest.

CustomPlaceholderFieldEnum
Values for Custom placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

CustomerMatchUploadKeyTypeEnum
Indicates what type of data are the user list's members matched from.

CustomerPayPerConversionEligibilityFailureReasonEnum
Container for enum describing reasons why a customer is not eligible to use PaymentMode.CONVERSIONS.

DataDrivenModelStatusEnum
Container for enum indicating data driven model status.

DayOfWeekEnum
Container for enumeration of days of the week, e.g., "Monday".

DeviceEnum
Container for enumeration of Google Ads devices available for targeting.

DisplayAdFormatSettingEnum
Container for display ad format settings.

DisplayUploadProductTypeEnum
Container for display upload product types. Product types that have the word "DYNAMIC" in them must be associated with a campaign that has a dynamic remarketing feed. See https://support.google.com/google-ads/answer/6053288 for more info about dynamic remarketing. Other product types are regarded as "static" and do not have this requirement.

DistanceBucketEnum
Container for distance buckets of a user’s distance from an advertiser’s location extension.

DsaPageFeedCriterionFieldEnum
Values for Dynamic Search Ad Page Feed criterion fields.

EducationPlaceholderFieldEnum
Values for Education placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

ExtensionSettingDeviceEnum
Container for enum describing extension setting device types.

ExtensionTypeEnum
Container for enum describing possible data types for an extension in an extension setting.

ExternalConversionSourceEnum
Container for enum describing the external conversion source that is associated with a ConversionAction.

FeedAttributeTypeEnum
Container for enum describing possible data types for a feed attribute.

FeedItemQualityApprovalStatusEnum
Container for enum describing possible quality evaluation approval statuses of a feed item.

FeedItemQualityDisapprovalReasonEnum
Container for enum describing possible quality evaluation disapproval reasons of a feed item.

FeedItemSetStatusEnum
Container for enum describing possible statuses of a feed item set.

FeedItemSetStringFilterTypeEnum
The type of string matching to be used for a dynamic FeedItemSet filter.

FeedItemStatusEnum
Container for enum describing possible statuses of a feed item.

FeedItemTargetDeviceEnum
Container for enum describing possible data types for a feed item target device.

FeedItemTargetStatusEnum
Container for enum describing possible statuses of a feed item target.

FeedItemTargetTypeEnum
Container for enum describing possible types of a feed item target.

FeedItemValidationStatusEnum
Container for enum describing possible validation statuses of a feed item.

FeedLinkStatusEnum
Container for an enum describing possible statuses of a feed link.

FeedMappingCriterionTypeEnum
Container for enum describing possible criterion types for a feed mapping.

FeedMappingStatusEnum
Container for enum describing possible statuses of a feed mapping.

FeedOriginEnum
Container for enum describing possible values for a feed origin.

FeedStatusEnum
Container for enum describing possible statuses of a feed.

FlightPlaceholderFieldEnum
Values for Flight placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

FrequencyCapEventTypeEnum
Container for enum describing the type of event that the cap applies to.

FrequencyCapLevelEnum
Container for enum describing the level on which the cap is to be applied.

FrequencyCapTimeUnitEnum
Container for enum describing the unit of time the cap is defined at.

GenderTypeEnum
Container for enum describing the type of demographic genders.

GeoTargetConstantStatusEnum
Container for describing the status of a geo target constant.

GeoTargetingRestrictionEnum
Message describing feed item geo targeting restriction.

GeoTargetingTypeEnum
Container for enum describing possible geo targeting types.

GoogleAdsFieldCategoryEnum
Container for enum that determines if the described artifact is a resource or a field, and if it is a field, when it segments search queries.

GoogleAdsFieldDataTypeEnum
Container holding the various data types.

HotelDateSelectionTypeEnum
Container for enum describing possible hotel date selection types

HotelPlaceholderFieldEnum
Values for Hotel placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

HotelPriceBucketEnum
Container for enum describing hotel price bucket for a hotel itinerary.

HotelRateTypeEnum
Container for enum describing possible hotel rate types.

IncomeRangeTypeEnum
Container for enum describing the type of demographic income ranges.

InteractionEventTypeEnum
Container for enum describing types of payable and free interactions.

InteractionTypeEnum
Container for enum describing possible interaction types.

InvoiceTypeEnum
Container for enum describing the type of invoices.

JobPlaceholderFieldEnum
Values for Job placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

KeywordMatchTypeEnum
Message describing Keyword match types.

KeywordPlanCompetitionLevelEnum
Container for enumeration of keyword competition levels. The competition level indicates how competitive ad placement is for a keyword and is determined by the number of advertisers bidding on that keyword relative to all keywords across Google. The competition level can depend on the location and Search Network targeting options you've selected.

KeywordPlanForecastIntervalEnum
Container for enumeration of forecast intervals.

KeywordPlanNetworkEnum
Container for enumeration of keyword plan forecastable network types.

LabelStatusEnum
Container for enum describing possible status of a label.

LeadFormCallToActionTypeEnum
Describes the type of call-to-action phrases in a lead form.

LeadFormDesiredIntentEnum
Describes the desired level of intent of generated leads.

LeadFormFieldUserInputTypeEnum
Describes the input type of a lead form field.

LeadFormPostSubmitCallToActionTypeEnum
Describes the type of post-submit call-to-action phrases for a lead form.

LegacyAppInstallAdAppStoreEnum
Container for enum describing app store type in a legacy app install ad.

LinkedAccountTypeEnum
Container for enum describing different types of Linked accounts.

ListingGroupTypeEnum
Container for enum describing the type of the listing group.

LocalPlaceholderFieldEnum
Values for Local placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

LocationExtensionTargetingCriterionFieldEnum
Values for Location Extension Targeting criterion fields.

LocationGroupRadiusUnitsEnum
Container for enum describing unit of radius in location group.

LocationPlaceholderFieldEnum
Values for Location placeholder fields.

LocationSourceTypeEnum
Used to distinguish the location source type.

ManagerLinkStatusEnum
Container for enum describing possible status of a manager and client link.

MatchingFunctionContextTypeEnum
Container for context types for an operand in a matching function.

MatchingFunctionOperatorEnum
Container for enum describing matching function operator.

MediaTypeEnum
Container for enum describing the types of media.

MerchantCenterLinkStatusEnum
Container for enum describing possible statuses of a Google Merchant Center link.

MessagePlaceholderFieldEnum
Values for Message placeholder fields.

MimeTypeEnum
Container for enum describing the mime types.

MinuteOfHourEnum
Container for enumeration of quarter-hours.

MobileAppVendorEnum
Container for enum describing different types of mobile app vendors.

MobileDeviceTypeEnum
Container for enum describing the types of mobile device.

MonthOfYearEnum
Container for enumeration of months of the year, e.g., "January".

NegativeGeoTargetTypeEnum
Container for enum describing possible negative geo target types.

OfflineUserDataJobFailureReasonEnum
Container for enum describing reasons why an offline user data job failed to be processed.

OfflineUserDataJobStatusEnum
Container for enum describing status of an offline user data job.

OfflineUserDataJobTypeEnum
Container for enum describing types of an offline user data job.

OperatingSystemVersionOperatorTypeEnum
Container for enum describing the type of OS operators.

OptimizationGoalTypeEnum
Container for enum describing the type of optimization goal.

ParentalStatusTypeEnum
Container for enum describing the type of demographic parental statuses.

PaymentModeEnum
Container for enum describing possible payment modes.

PlaceholderTypeEnum
Container for enum describing possible placeholder types for a feed mapping.

PlacementTypeEnum
Container for enum describing possible placement types.

PolicyApprovalStatusEnum
Container for enum describing possible policy approval statuses.

PolicyReviewStatusEnum
Container for enum describing possible policy review statuses.

PolicyTopicEntryTypeEnum
Container for enum describing possible policy topic entry types.

PolicyTopicEvidenceDestinationMismatchUrlTypeEnum
Container for enum describing possible policy topic evidence destination mismatch url types.

PolicyTopicEvidenceDestinationNotWorkingDeviceEnum
Container for enum describing possible policy topic evidence destination not working devices.

PolicyTopicEvidenceDestinationNotWorkingDnsErrorTypeEnum
Container for enum describing possible policy topic evidence destination not working DNS error types.

PositiveGeoTargetTypeEnum
Container for enum describing possible positive geo target types.

PreferredContentTypeEnum
Container for enumeration of preferred content criterion type.

PriceExtensionPriceQualifierEnum
Container for enum describing a price extension price qualifier.

PriceExtensionPriceUnitEnum
Container for enum describing price extension price unit.

PriceExtensionTypeEnum
Container for enum describing types for a price extension.

PricePlaceholderFieldEnum
Values for Price placeholder fields.

ProductBiddingCategoryLevelEnum
Level of a product bidding category.

ProductBiddingCategoryStatusEnum
Status of the product bidding category.

ProductChannelEnum
Locality of a product offer.

ProductChannelExclusivityEnum
Availability of a product offer.

ProductConditionEnum
Condition of a product offer.

ProductCustomAttributeIndexEnum
Container for enum describing the index of the product custom attribute.

ProductTypeLevelEnum
Level of the type of a product offer.

PromotionExtensionDiscountModifierEnum
Container for enum describing possible a promotion extension discount modifier.

PromotionExtensionOccasionEnum
Container for enum describing a promotion extension occasion. For more information about the occasions please check: https://support.google.com/google-ads/answer/7367521

PromotionPlaceholderFieldEnum
Values for Promotion placeholder fields.

ProximityRadiusUnitsEnum
Container for enum describing unit of radius in proximity.

QualityScoreBucketEnum
The relative performance compared to other advertisers.

ReachPlanAdLengthEnum
Message describing length of a plannable video ad.

ReachPlanAgeRangeEnum
Message describing plannable age ranges.

ReachPlanNetworkEnum
Container for enum describing plannable networks.

RealEstatePlaceholderFieldEnum
Values for Real Estate placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

RecommendationTypeEnum
Container for enum describing types of recommendations.

ResourceChangeOperationEnum
Container for enum describing resource change operations in the ChangeEvent resource.

ResponseContentTypeEnum
Container for possible response content types.

SearchEngineResultsPageTypeEnum
The type of the search engine results page.

SearchTermMatchTypeEnum
Container for enum describing match types for a keyword triggering an ad.

SearchTermTargetingStatusEnum
Container for enum indicating whether a search term is one of your targeted or excluded keywords.

ServedAssetFieldTypeEnum
Container for enum describing possible asset field types.

SharedSetStatusEnum
Container for enum describing types of shared set statuses.

SharedSetTypeEnum
Container for enum describing types of shared sets.

SimulationModificationMethodEnum
Container for enum describing the method by which a simulation modifies a field.

SimulationTypeEnum
Container for enum describing the field a simulation modifies.

SitelinkPlaceholderFieldEnum
Values for Sitelink placeholder fields.

SlotEnum
Container for enumeration of possible positions of the Ad.

SpendingLimitTypeEnum
Message describing spending limit types.

StructuredSnippetPlaceholderFieldEnum
Values for Structured Snippet placeholder fields.

SummaryRowSettingEnum
Indicates summary row setting in request parameter.

SystemManagedResourceSourceEnum
Container for enum describing possible system managed entity sources.

TargetCpaOptInRecommendationGoalEnum
Container for enum describing goals for TargetCpaOptIn recommendation.

TargetImpressionShareLocationEnum
Container for enum describing where on the first search results page the automated bidding system should target impressions for the TargetImpressionShare bidding strategy.

TargetingDimensionEnum
The dimensions that can be targeted.

TimeTypeEnum
Message describing time types.

TrackingCodePageFormatEnum
Container for enum describing the format of the web page where the tracking tag and snippet will be installed.

TrackingCodeTypeEnum
Container for enum describing the type of the generated tag snippets for tracking conversions.

TravelPlaceholderFieldEnum
Values for Travel placeholder fields. For more information about dynamic remarketing feeds, see https://support.google.com/google-ads/answer/6053288.

UserIdentifierSourceEnum
Container for enum describing the source of the user identifier for offline Store Sales third party uploads.

UserInterestTaxonomyTypeEnum
Message describing a UserInterestTaxonomyType.

UserListAccessStatusEnum
Indicates if this client still has access to the list.

UserListClosingReasonEnum
Indicates the reason why the userlist was closed. This enum is only used when a list is auto-closed by the system.

UserListCombinedRuleOperatorEnum
Logical operator connecting two rules.

UserListCrmDataSourceTypeEnum
Indicates source of Crm upload data.

UserListDateRuleItemOperatorEnum
Supported rule operator for date type.

UserListLogicalRuleOperatorEnum
The logical operator of the rule.

UserListMembershipStatusEnum
Membership status of this user list. Indicates whether a user list is open or active. Only open user lists can accumulate more users and can be used for targeting.

UserListNumberRuleItemOperatorEnum
Supported rule operator for number type.

UserListPrepopulationStatusEnum
Indicates status of prepopulation based on the rule.

UserListRuleTypeEnum
Rule based user list rule type.

UserListSizeRangeEnum
Size range in terms of number of users of a UserList.

UserListStringRuleItemOperatorEnum
Supported rule operator for string type.

UserListTypeEnum
The user list types.

VanityPharmaDisplayUrlModeEnum
The display mode for vanity pharma URLs.

VanityPharmaTextEnum
The text that will be displayed in display URL of the text ad when website description is the selected display mode for vanity pharma URLs.

WebpageConditionOperandEnum
Container for enum describing webpage condition operand in webpage criterion.

WebpageConditionOperatorEnum
Container for enum describing webpage condition operator in webpage criterion.