Struct gapi_grpc::google::ads::googleads::v6::resources::Campaign

pub struct Campaign {
    pub resource_name: String,
    pub id: i64,
    pub name: String,
    pub status: i32,
    pub serving_status: i32,
    pub ad_serving_optimization_status: i32,
    pub advertising_channel_type: i32,
    pub advertising_channel_sub_type: i32,
    pub tracking_url_template: String,
    pub url_custom_parameters: Vec<CustomParameter>,
    pub real_time_bidding_setting: Option<RealTimeBiddingSetting>,
    pub network_settings: Option<NetworkSettings>,
    pub hotel_setting: Option<HotelSettingInfo>,
    pub dynamic_search_ads_setting: Option<DynamicSearchAdsSetting>,
    pub shopping_setting: Option<ShoppingSetting>,
    pub targeting_setting: Option<TargetingSetting>,
    pub geo_target_type_setting: Option<GeoTargetTypeSetting>,
    pub local_campaign_setting: Option<LocalCampaignSetting>,
    pub app_campaign_setting: Option<AppCampaignSetting>,
    pub labels: Vec<String>,
    pub experiment_type: i32,
    pub base_campaign: String,
    pub campaign_budget: String,
    pub bidding_strategy_type: i32,
    pub start_date: String,
    pub end_date: String,
    pub final_url_suffix: String,
    pub frequency_caps: Vec<FrequencyCapEntry>,
    pub video_brand_safety_suitability: i32,
    pub vanity_pharma: Option<VanityPharma>,
    pub selective_optimization: Option<SelectiveOptimization>,
    pub optimization_goal_setting: Option<OptimizationGoalSetting>,
    pub tracking_setting: Option<TrackingSetting>,
    pub payment_mode: i32,
    pub optimization_score: f64,
    pub campaign_bidding_strategy: Option<CampaignBiddingStrategy>,
}

A campaign.

Fields
resource_name: String

Immutable. The resource name of the campaign. Campaign resource names have the form:

customers/{customer_id}/campaigns/{campaign_id}

id: i64

Output only. The ID of the campaign.

name: String

The name of the campaign.

This field is required and should not be empty when creating new campaigns.

It must not contain any null (code point 0x0), NL line feed (code point 0xA) or carriage return (code point 0xD) characters.

status: i32

The status of the campaign.

When a new campaign is added, the status defaults to ENABLED.

serving_status: i32

Output only. The ad serving status of the campaign.

ad_serving_optimization_status: i32

The ad serving optimization status of the campaign.

advertising_channel_type: i32

Immutable. The primary serving target for ads within the campaign. The targeting options can be refined in network_settings.

This field is required and should not be empty when creating new campaigns.

Can be set only when creating campaigns. After the campaign is created, the field can not be changed.

advertising_channel_sub_type: i32

Immutable. Optional refinement to advertising_channel_type. Must be a valid sub-type of the parent channel type.

Can be set only when creating campaigns. After campaign is created, the field can not be changed.

tracking_url_template: String

The URL template for constructing a tracking URL.

url_custom_parameters: Vec<CustomParameter>

The list of mappings used to substitute custom parameter tags in a tracking_url_template, final_urls, or mobile_final_urls.

real_time_bidding_setting: Option<RealTimeBiddingSetting>

Settings for Real-Time Bidding, a feature only available for campaigns targeting the Ad Exchange network.

network_settings: Option<NetworkSettings>

The network settings for the campaign.

hotel_setting: Option<HotelSettingInfo>

Immutable. The hotel setting for the campaign.

dynamic_search_ads_setting: Option<DynamicSearchAdsSetting>

The setting for controlling Dynamic Search Ads (DSA).

shopping_setting: Option<ShoppingSetting>

The setting for controlling Shopping campaigns.

targeting_setting: Option<TargetingSetting>

Setting for targeting related features.

geo_target_type_setting: Option<GeoTargetTypeSetting>

The setting for ads geotargeting.

local_campaign_setting: Option<LocalCampaignSetting>

The setting for local campaign.

app_campaign_setting: Option<AppCampaignSetting>

The setting related to App Campaign.

labels: Vec<String>

Output only. The resource names of labels attached to this campaign.

experiment_type: i32

Output only. The type of campaign: normal, draft, or experiment.

base_campaign: String

Output only. The resource name of the base campaign of a draft or experiment campaign. For base campaigns, this is equal to resource_name.

This field is read-only.

campaign_budget: String

The budget of the campaign.

bidding_strategy_type: i32

Output only. The type of bidding strategy.

A bidding strategy can be created by setting either the bidding scheme to create a standard bidding strategy or the bidding_strategy field to create a portfolio bidding strategy.

This field is read-only.

start_date: String

The date when campaign started.

end_date: String

The date when campaign ended.

final_url_suffix: String

Suffix used to append query parameters to landing pages that are served with parallel tracking.

frequency_caps: Vec<FrequencyCapEntry>

A list that limits how often each user will see this campaign's ads.

video_brand_safety_suitability: i32

Output only. 3-Tier Brand Safety setting for the campaign.

vanity_pharma: Option<VanityPharma>

Describes how unbranded pharma ads will be displayed.

selective_optimization: Option<SelectiveOptimization>

Selective optimization setting for this campaign, which includes a set of conversion actions to optimize this campaign towards.

optimization_goal_setting: Option<OptimizationGoalSetting>

Optimization goal setting for this campaign, which includes a set of optimization goal types.

tracking_setting: Option<TrackingSetting>

Output only. Campaign-level settings for tracking information.

payment_mode: i32

Payment mode for the campaign.

optimization_score: f64

Output only. Optimization score of the campaign.

Optimization score is an estimate of how well a campaign is set to perform. It ranges from 0% (0.0) to 100% (1.0), with 100% indicating that the campaign is performing at full potential. This field is null for unscored campaigns.

See "About optimization score" at https://support.google.com/google-ads/answer/9061546.

This field is read-only.

campaign_bidding_strategy: Option<CampaignBiddingStrategy>

The bidding strategy for the campaign.

Must be either portfolio (created via BiddingStrategy service) or standard, that is embedded into the campaign.

Implementations

impl Campaign

pub fn status(&self) -> CampaignStatus

Returns the enum value of status, or the default if the field is set to an invalid enum value.

pub fn set_status(&mut self, value: CampaignStatus)

Sets status to the provided enum value.

pub fn ad_serving_optimization_status(&self) -> AdServingOptimizationStatus

Returns the enum value of ad_serving_optimization_status, or the default if the field is set to an invalid enum value.

pub fn set_ad_serving_optimization_status(
&mut self,
value: AdServingOptimizationStatus
)

Sets ad_serving_optimization_status to the provided enum value.

pub fn advertising_channel_type(&self) -> AdvertisingChannelType

Returns the enum value of advertising_channel_type, or the default if the field is set to an invalid enum value.

pub fn set_advertising_channel_type(&mut self, value: AdvertisingChannelType)

Sets advertising_channel_type to the provided enum value.

pub fn advertising_channel_sub_type(&self) -> AdvertisingChannelSubType

Returns the enum value of advertising_channel_sub_type, or the default if the field is set to an invalid enum value.

pub fn set_advertising_channel_sub_type(
&mut self,
value: AdvertisingChannelSubType
)

Sets advertising_channel_sub_type to the provided enum value.

pub fn experiment_type(&self) -> CampaignExperimentType

Returns the enum value of experiment_type, or the default if the field is set to an invalid enum value.

pub fn set_experiment_type(&mut self, value: CampaignExperimentType)

Sets experiment_type to the provided enum value.

pub fn serving_status(&self) -> CampaignServingStatus

Returns the enum value of serving_status, or the default if the field is set to an invalid enum value.

pub fn set_serving_status(&mut self, value: CampaignServingStatus)

Sets serving_status to the provided enum value.

pub fn bidding_strategy_type(&self) -> BiddingStrategyType

Returns the enum value of bidding_strategy_type, or the default if the field is set to an invalid enum value.

pub fn set_bidding_strategy_type(&mut self, value: BiddingStrategyType)

Sets bidding_strategy_type to the provided enum value.

pub fn video_brand_safety_suitability(&self) -> BrandSafetySuitability

Returns the enum value of video_brand_safety_suitability, or the default if the field is set to an invalid enum value.

pub fn set_video_brand_safety_suitability(
&mut self,
value: BrandSafetySuitability
)

Sets video_brand_safety_suitability to the provided enum value.

pub fn payment_mode(&self) -> PaymentMode

Returns the enum value of payment_mode, or the default if the field is set to an invalid enum value.

pub fn set_payment_mode(&mut self, value: PaymentMode)

Sets payment_mode to the provided enum value.