Module gapi_grpc::google::analytics::admin::v1alpha

Modules

analytics_admin_service_client
Generated client implementations.

Structs

Account
A resource message representing a Google Analytics account.

AccountSummary
A virtual resource representing an overview of an account and all its child GA4 properties.

AndroidAppDataStream
A resource message representing a Google Analytics Android app stream.

AuditUserLink
Read-only resource used to summarize a principal's effective roles.

AuditUserLinksRequest
Request message for AuditUserLinks RPC.

AuditUserLinksResponse
Response message for AuditUserLinks RPC.

BatchCreateUserLinksRequest
Request message for BatchCreateUserLinks RPC.

BatchCreateUserLinksResponse
Response message for BatchCreateUserLinks RPC.

BatchDeleteUserLinksRequest
Request message for BatchDeleteUserLinks RPC.

BatchGetUserLinksRequest
Request message for BatchGetUserLinks RPC.

BatchGetUserLinksResponse
Response message for BatchGetUserLinks RPC.

BatchUpdateUserLinksRequest
Request message for BatchUpdateUserLinks RPC.

BatchUpdateUserLinksResponse
Response message for BatchUpdateUserLinks RPC.

CreateAndroidAppDataStreamRequest
Request message for CreateAndroidAppDataStream RPC.

CreateFirebaseLinkRequest
Request message for CreateFirebaseLink RPC

CreateGoogleAdsLinkRequest
Request message for CreateGoogleAdsLink RPC

CreateIosAppDataStreamRequest
Request message for CreateIosAppDataStream RPC.

CreatePropertyRequest
Request message for CreateProperty RPC.

CreateUserLinkRequest
Request message for CreateUserLink RPC.

CreateWebDataStreamRequest
Request message for CreateWebDataStream RPC.

DataSharingSettings
A resource message representing data sharing settings of a Google Analytics account.

DeleteAccountRequest
Request message for DeleteAccount RPC.

DeleteAndroidAppDataStreamRequest
Request message for DeleteAndroidAppDataStream RPC.

DeleteFirebaseLinkRequest
Request message for DeleteFirebaseLink RPC

DeleteGoogleAdsLinkRequest
Request message for DeleteGoogleAdsLink RPC.

DeleteIosAppDataStreamRequest
Request message for DeleteIosAppDataStream RPC.

DeletePropertyRequest
Request message for DeleteProperty RPC.

DeleteUserLinkRequest
Request message for DeleteUserLink RPC.

DeleteWebDataStreamRequest
Request message for DeleteWebDataStream RPC.

EnhancedMeasurementSettings
Singleton resource under a WebDataStream, configuring measurement of additional site interactions and content.

FirebaseLink
A link between an GA4 property and a Firebase project.

GetAccountRequest
Request message for GetAccount RPC.

GetAndroidAppDataStreamRequest
Request message for GetAndroidAppDataStream RPC.

GetDataSharingSettingsRequest
Request message for GetDataSharingSettings RPC.

GetEnhancedMeasurementSettingsRequest
Request message for GetEnhancedMeasurementSettings RPC.

GetGlobalSiteTagRequest
Request message for GetGlobalSiteTag RPC.

GetIosAppDataStreamRequest
Request message for GetIosAppDataStream RPC.

GetPropertyRequest
Request message for GetProperty RPC.

GetUserLinkRequest
Request message for GetUserLink RPC.

GetWebDataStreamRequest
Request message for GetWebDataStream RPC.

GlobalSiteTag
Read-only resource with the tag for sending data from a website to a WebDataStream.

GoogleAdsLink
A link between an GA4 property and a Google Ads account.

IosAppDataStream
A resource message representing a Google Analytics IOS app stream.

ListAccountSummariesRequest
Request message for ListAccountSummaries RPC.

ListAccountSummariesResponse
Response message for ListAccountSummaries RPC.

ListAccountsRequest
Request message for ListAccounts RPC.

ListAccountsResponse
Request message for ListAccounts RPC.

ListAndroidAppDataStreamsRequest
Request message for ListAndroidAppDataStreams RPC.

ListAndroidAppDataStreamsResponse
Request message for ListAndroidDataStreams RPC.

ListFirebaseLinksRequest
Request message for ListFirebaseLinks RPC

ListFirebaseLinksResponse
Response message for ListFirebaseLinks RPC

ListGoogleAdsLinksRequest
Request message for ListGoogleAdsLinks RPC.

ListGoogleAdsLinksResponse
Response message for ListGoogleAdsLinks RPC.

ListIosAppDataStreamsRequest
Request message for ListIosAppDataStreams RPC.

ListIosAppDataStreamsResponse
Request message for ListIosAppDataStreams RPC.

ListPropertiesRequest
Request message for ListProperties RPC.

ListPropertiesResponse
Response message for ListProperties RPC.

ListUserLinksRequest
Request message for ListUserLinks RPC.

ListUserLinksResponse
Response message for ListUserLinks RPC.

ListWebDataStreamsRequest
Request message for ListWebDataStreams RPC.

ListWebDataStreamsResponse
Request message for ListWebDataStreams RPC.

Property
A resource message representing a Google Analytics GA4 property.

PropertySummary
A virtual resource representing metadata for an GA4 property.

ProvisionAccountTicketRequest
Request message for ProvisionAccountTicket RPC.

ProvisionAccountTicketResponse
Response message for ProvisionAccountTicket RPC.

UpdateAccountRequest
Request message for UpdateAccount RPC.

UpdateAndroidAppDataStreamRequest
Request message for UpdateAndroidAppDataStream RPC.

UpdateEnhancedMeasurementSettingsRequest
Request message for UpdateEnhancedMeasurementSettings RPC.

UpdateFirebaseLinkRequest
Request message for UpdateFirebaseLink RPC

UpdateGoogleAdsLinkRequest
Request message for UpdateGoogleAdsLink RPC

UpdateIosAppDataStreamRequest
Request message for UpdateIosAppDataStream RPC.

UpdatePropertyRequest
Request message for UpdateProperty RPC.

UpdateUserLinkRequest
Request message for UpdateUserLink RPC.

UpdateWebDataStreamRequest
Request message for UpdateWebDataStream RPC.

UserLink
A resource message representing a user's permissions on an Account or Property resource.

WebDataStream
A resource message representing a Google Analytics web stream.

Enums

IndustryCategory
The category selected for this property, used for industry benchmarking.

MaximumUserAccess
Maximum access settings that Firebase user receive on the linked Analytics property.