Module gapi_grpc::google::analytics::data::v1alpha

Modules

alpha_analytics_data_client
Generated client implementations.

cohorts_range

dimension_expression

dimension_value

filter

filter_expression

metric_value

numeric_value

order_by

Structs

BatchRunPivotReportsRequest
The batch request containing multiple pivot report requests.

BatchRunPivotReportsResponse
The batch response containing multiple pivot reports.

BatchRunReportsRequest
The batch request containing multiple report requests.

BatchRunReportsResponse
The batch response containing multiple reports.

Cohort
Defines a cohort. A cohort is a group of users who share a common characteristic. For example, all users with the same acquisition date belong to the same cohort.

CohortReportSettings
Settings of a cohort report.

CohortSpec
Specification for a cohort report.

CohortsRange
Describes date range for a cohort report.

DateRange
A contiguous set of days: startDate, startDate + 1, ..., endDate. Requests are allowed up to 4 date ranges, and the union of the ranges can cover up to 1 year.

Dimension
Dimensions are attributes of your data. For example, the dimension City indicates the city, for example, "Paris" or "New York", from which an event originates. Requests are allowed up to 8 dimensions.

DimensionExpression
Used to express a dimension which is the result of a formula of multiple dimensions. Example usages:

DimensionHeader
The header for the dimensions.

DimensionValue
The value of a dimension.

Entity
The unique identifier of the property whose events are tracked.

Filter
An expression to filter dimension or metric values.

FilterExpression
To express dimension or metric filters. The fields in the same FilterExpression need to be either all dimensions or all metrics.

FilterExpressionList
A list of filter expressions.

Metric
The quantitative measurements of a report. For example, the metric eventCount is the total number of events. Requests are allowed up to 10 metrics.

MetricHeader
Describes the metric column in the report.

MetricValue
The value of a metric.

NumericValue
To represent a number.

OrderBy
The sort options.

Pivot
Describes the visible dimension columns and rows in the report response.

PivotHeader
Dimensions' values in a pivot.

PropertyQuota
Current state of all quotas for this Analytics Property. If any quota for a property is exhausted, all requests to that property will return Resource Exhausted errors.

QuotaStatus
Current state for a particular quota group.

ResponseMetaData
Response's metadata carrying additional information about the report content.

Row
Report data for each row. For example if RunReportRequest contains:

RunPivotReportRequest
The request to generate a pivot report.

RunPivotReportResponse
The response pivot report table corresponding to a pivot request.

RunReportRequest
The request to generate a report.

RunReportResponse
The response report table corresponding to a request.

Enums

MetricAggregation
Represents aggregation of metrics.

MetricType
Type of a metric value.