Module gapi_grpc::google::analytics::data::v1alpha::filter

Modules

numeric_filter

string_filter

Structs

BetweenFilter
To express that the result needs to be between two numbers (inclusive).

InListFilter
The result needs to be in a list of string values.

NumericFilter
Filters for numeric or date values.

StringFilter
The filter for string

Enums

OneFilter
Specify one type of filter for Filter.