Struct gapi_grpc::google::analytics::management::v1alpha::management_client::ManagementClient

pub struct ManagementClient<T> { /* fields omitted */ }

Service Interface for the GA Management API (App+Web).

Methods

impl ManagementClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ManagementClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn get_account<'_>(
&'_ mut self,
request: impl IntoRequest<GetAccountRequest>
) -> Result<Response<Account>, Status>

Lookup for a single Account. Throws "Target not found" if no such account found, or if caller does not have permissions to access it.

pub async fn list_accounts<'_>(
&'_ mut self,
request: impl IntoRequest<ListAccountsRequest>
) -> Result<Response<ListAccountsResponse>, Status>

Returns all accounts accessible by the caller.

Note that these accounts might not currently have App+Web properties. Soft-deleted (ie: "trashed") accounts are excluded by default. Returns an empty list if no relevant accounts are found.

pub async fn delete_account<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteAccountRequest>
) -> Result<Response<()>, Status>

Marks target Account as soft-deleted (ie: "trashed") and returns it.

This API does not have a method to restore soft-deleted accounts. However, they can be restored using the Trash Can UI.

If the accounts are not restored before the expiration time, the account and all child resources (eg: Properties, GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found.

pub async fn update_account<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateAccountRequest>
) -> Result<Response<Account>, Status>

Updates an account.

pub async fn provision_account_ticket<'_>(
&'_ mut self,
request: impl IntoRequest<ProvisionAccountTicketRequest>
) -> Result<Response<ProvisionAccountTicketResponse>, Status>

Requests a ticket for creating an account.

pub async fn get_property<'_>(
&'_ mut self,
request: impl IntoRequest<GetPropertyRequest>
) -> Result<Response<Property>, Status>

Lookup for a single "App+Web" Property.

Throws "Target not found" if no such property found, if property is not of the type "App+Web", or if caller does not have permissions to access it.

pub async fn list_properties<'_>(
&'_ mut self,
request: impl IntoRequest<ListPropertiesRequest>
) -> Result<Response<ListPropertiesResponse>, Status>

Returns child Properties under the specified parent Account.

Only "App+Web" properties will be returned. Properties will be excluded if the caller does not have access. Soft-deleted (ie: "trashed") properties are excluded by default. Returns an empty list if no relevant properties are found.

pub async fn create_property<'_>(
&'_ mut self,
request: impl IntoRequest<CreatePropertyRequest>
) -> Result<Response<Property>, Status>

Creates an "App+Web" property with the specified location and attributes.

pub async fn delete_property<'_>(
&'_ mut self,
request: impl IntoRequest<DeletePropertyRequest>
) -> Result<Response<()>, Status>

Marks target Property as soft-deleted (ie: "trashed") and returns it.

This API does not have a method to restore soft-deleted properties. However, they can be restored using the Trash Can UI.

If the properties are not restored before the expiration time, the Property and all child resources (eg: GoogleAdsLinks, Streams, UserLinks) will be permanently purged. https://support.google.com/analytics/answer/6154772

Returns an error if the target is not found, or is not an App+Web Property.

pub async fn update_property<'_>(
&'_ mut self,
request: impl IntoRequest<UpdatePropertyRequest>
) -> Result<Response<Property>, Status>

Updates a property.

pub async fn get_user_link<'_>(
&'_ mut self,
request: impl IntoRequest<GetUserLinkRequest>
) -> Result<Response<UserLink>, Status>

Gets information about a user's link to an account or property.

pub async fn batch_get_user_links<'_>(
&'_ mut self,
request: impl IntoRequest<BatchGetUserLinksRequest>
) -> Result<Response<BatchGetUserLinksResponse>, Status>

Gets information about multiple users' links to an account or property.

pub async fn list_user_links<'_>(
&'_ mut self,
request: impl IntoRequest<ListUserLinksRequest>
) -> Result<Response<ListUserLinksResponse>, Status>

Lists all user links on an account or property.

pub async fn audit_user_links<'_>(
&'_ mut self,
request: impl IntoRequest<AuditUserLinksRequest>
) -> Result<Response<AuditUserLinksResponse>, Status>

Lists all user links on an account or property, including implicit ones that come from effective permissions granted by groups or organization admin roles.

If a returned user link does not have direct permissions, they cannot be removed from the account or property directly with the DeleteUserLink command. They have to be removed from the group/etc that gives them permissions, which is currently only usable/discoverable in the GA or GMP UIs.

pub async fn create_user_link<'_>(
&'_ mut self,
request: impl IntoRequest<CreateUserLinkRequest>
) -> Result<Response<UserLink>, Status>

Creates a user link on an account or property.

If the user with the specified email already has permissions on the account or property, then the user's existing permissions will be unioned with the permissions specified in the new UserLink.

pub async fn batch_create_user_links<'_>(
&'_ mut self,
request: impl IntoRequest<BatchCreateUserLinksRequest>
) -> Result<Response<BatchCreateUserLinksResponse>, Status>

Creates information about multiple users' links to an account or property.

This method is transactional. If any UserLink cannot be created, none of the UserLinks will be created.

pub async fn update_user_link<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateUserLinkRequest>
) -> Result<Response<UserLink>, Status>

Updates a user link on an account or property.

pub async fn batch_update_user_links<'_>(
&'_ mut self,
request: impl IntoRequest<BatchUpdateUserLinksRequest>
) -> Result<Response<BatchUpdateUserLinksResponse>, Status>

Updates information about multiple users' links to an account or property.

pub async fn delete_user_link<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteUserLinkRequest>
) -> Result<Response<()>, Status>

Deletes a user link on an account or property.

pub async fn batch_delete_user_links<'_>(
&'_ mut self,
request: impl IntoRequest<BatchDeleteUserLinksRequest>
) -> Result<Response<()>, Status>

Deletes information about multiple users' links to an account or property.

pub async fn get_web_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<GetWebDataStreamRequest>
) -> Result<Response<WebDataStream>, Status>

Lookup for a single WebDataStream

Throws "Target not found" if no such web data stream found, or if the caller does not have permissions to access it.

pub async fn delete_web_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteWebDataStreamRequest>
) -> Result<Response<()>, Status>

Deletes a web stream on a property.

pub async fn update_web_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateWebDataStreamRequest>
) -> Result<Response<WebDataStream>, Status>

Updates a web stream on a property.

pub async fn create_web_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<CreateWebDataStreamRequest>
) -> Result<Response<WebDataStream>, Status>

Creates a web stream with the specified location and attributes.

pub async fn list_web_data_streams<'_>(
&'_ mut self,
request: impl IntoRequest<ListWebDataStreamsRequest>
) -> Result<Response<ListWebDataStreamsResponse>, Status>

Returns child web data streams under the specified parent property.

Web data streams will be excluded if the caller does not have access. Returns an empty list if no relevant web data streams are found.

pub async fn get_ios_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<GetIosAppDataStreamRequest>
) -> Result<Response<IosAppDataStream>, Status>

Lookup for a single IosAppDataStream

Throws "Target not found" if no such iOS app data stream found, or if the caller does not have permissions to access it.

pub async fn delete_ios_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteIosAppDataStreamRequest>
) -> Result<Response<()>, Status>

Deletes an iOS app stream on a property.

pub async fn update_ios_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateIosAppDataStreamRequest>
) -> Result<Response<IosAppDataStream>, Status>

Updates an iOS app stream on a property.

pub async fn create_ios_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<CreateIosAppDataStreamRequest>
) -> Result<Response<IosAppDataStream>, Status>

Creates an iOS app data stream with the specified location and attributes.

pub async fn list_ios_app_data_streams<'_>(
&'_ mut self,
request: impl IntoRequest<ListIosAppDataStreamsRequest>
) -> Result<Response<ListIosAppDataStreamsResponse>, Status>

Returns child iOS app data streams under the specified parent property.

iOS app data streams will be excluded if the caller does not have access. Returns an empty list if no relevant iOS app data streams are found.

pub async fn get_android_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<GetAndroidAppDataStreamRequest>
) -> Result<Response<AndroidAppDataStream>, Status>

Lookup for a single AndroidAppDataStream

Throws "Target not found" if no such android app data stream found, or if the caller does not have permissions to access it.

pub async fn delete_android_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteAndroidAppDataStreamRequest>
) -> Result<Response<()>, Status>

Deletes an android app stream on a property.

pub async fn update_android_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateAndroidAppDataStreamRequest>
) -> Result<Response<AndroidAppDataStream>, Status>

Updates an android app stream on a property.

pub async fn create_android_app_data_stream<'_>(
&'_ mut self,
request: impl IntoRequest<CreateAndroidAppDataStreamRequest>
) -> Result<Response<AndroidAppDataStream>, Status>

Creates an android app stream with the specified location and attributes.

pub async fn list_android_app_data_streams<'_>(
&'_ mut self,
request: impl IntoRequest<ListAndroidAppDataStreamsRequest>
) -> Result<Response<ListAndroidAppDataStreamsResponse>, Status>

Returns child android app streams under the specified parent property.

Android app streams will be excluded if the caller does not have access. Returns an empty list if no relevant android app streams are found.

pub async fn get_enhanced_measurement_settings<'_>(
&'_ mut self,
request: impl IntoRequest<GetEnhancedMeasurementSettingsRequest>
) -> Result<Response<EnhancedMeasurementSettings>, Status>

Returns the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

pub async fn update_enhanced_measurement_settings<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateEnhancedMeasurementSettingsRequest>
) -> Result<Response<EnhancedMeasurementSettings>, Status>

Updates the singleton enhanced measurement settings for this web stream. Note that the stream must enable enhanced measurement for these settings to take effect.

pub async fn create_firebase_link<'_>(
&'_ mut self,
request: impl IntoRequest<CreateFirebaseLinkRequest>
) -> Result<Response<FirebaseLink>, Status>

Creates a FirebaseLink.

Properties can have at most one FirebaseLink.

pub async fn update_firebase_link<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateFirebaseLinkRequest>
) -> Result<Response<FirebaseLink>, Status>

Updates a FirebaseLink on a property

pub async fn delete_firebase_link<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteFirebaseLinkRequest>
) -> Result<Response<()>, Status>

Deletes a FirebaseLink on a property

pub async fn list_firebase_links<'_>(
&'_ mut self,
request: impl IntoRequest<ListFirebaseLinksRequest>
) -> Result<Response<ListFirebaseLinksResponse>, Status>

Lists FirebaseLinks on a property. Properties can have at most one FirebaseLink.

pub async fn get_global_site_tag<'_>(
&'_ mut self,
request: impl IntoRequest<GetGlobalSiteTagRequest>
) -> Result<Response<GlobalSiteTag>, Status>

Returns the Site Tag for the specified web stream. Site Tags are immutable singletons.

pub async fn create_google_ads_link<'_>(
&'_ mut self,
request: impl IntoRequest<CreateGoogleAdsLinkRequest>
) -> Result<Response<GoogleAdsLink>, Status>

Creates a GoogleAdsLink.

pub async fn update_google_ads_link<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateGoogleAdsLinkRequest>
) -> Result<Response<GoogleAdsLink>, Status>

Updates a GoogleAdsLink on a property

pub async fn delete_google_ads_link<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteGoogleAdsLinkRequest>
) -> Result<Response<()>, Status>

Deletes a GoogleAdsLink on a property

pub async fn list_google_ads_links<'_>(
&'_ mut self,
request: impl IntoRequest<ListGoogleAdsLinksRequest>
) -> Result<Response<ListGoogleAdsLinksResponse>, Status>

Lists GoogleAdsLinks on a property.