Module gapi_grpc::google::api::expr::v1alpha1

Modules

conformance_service_client
Generated client implementations.

constant

decl

eval_request

eval_state

explain

expr

expr_value

issue_details

map_value

type

value

Structs

CheckRequest
Request message for the Check method.

CheckResponse
Response message for the Check method.

CheckedExpr
A CEL expression which has been successfully type checked.

Constant
Represents a primitive literal.

Decl
Represents a declaration of a named value or function.

EnumValue
An enum value.

ErrorSet
A set of errors.

EvalRequest
Request message for the Eval method.

EvalResponse
Response message for the Eval method.

EvalState
The state of an evaluation.

Explain
Values of intermediate expressions produced when evaluating expression. Deprecated, use EvalState instead.

Expr
An abstract representation of a common expression.

ExprValue
The value of an evaluated expression.

IssueDetails
Warnings or errors in service execution are represented by [google.rpc.Status][google.rpc.Status] messages, with the following message in the details field.

ListValue
A list.

MapValue
A map.

ParseRequest
Request message for the Parse method.

ParseResponse
Response message for the Parse method.

ParsedExpr
An expression together with source information as returned by the parser.

Reference
Describes a resolved reference to a declaration.

SourceInfo
Source information collected at parse time.

SourcePosition
A specific position in source.

Type
Represents a CEL type.

UnknownSet
A set of expressions for which the value is unknown.

Value
Represents a CEL value.