Struct gapi_grpc::google::api::serviceusage::v1beta1::service_usage_client::ServiceUsageClient

pub struct ServiceUsageClient<T> { /* fields omitted */ }

Service Usage API

Implementations

impl ServiceUsageClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ServiceUsageClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn enable_service(
&mut self,
request: impl IntoRequest<EnableServiceRequest>
) -> Result<Response<Operation>, Status>

Enables a service so that it can be used with a project.

Operation response type: google.protobuf.Empty

pub async fn disable_service(
&mut self,
request: impl IntoRequest<DisableServiceRequest>
) -> Result<Response<Operation>, Status>

Disables a service so that it can no longer be used with a project. This prevents unintended usage that may cause unexpected billing charges or security leaks.

It is not valid to call the disable method on a service that is not currently enabled. Callers will receive a FAILED_PRECONDITION status if the target service is not currently enabled.

Operation response type: google.protobuf.Empty

pub async fn get_service(
&mut self,
request: impl IntoRequest<GetServiceRequest>
) -> Result<Response<Service>, Status>

Returns the service configuration and enabled state for a given service.

pub async fn list_services(
&mut self,
request: impl IntoRequest<ListServicesRequest>
) -> Result<Response<ListServicesResponse>, Status>

Lists all services available to the specified project, and the current state of those services with respect to the project. The list includes all public services, all services for which the calling user has the servicemanagement.services.bind permission, and all services that have already been enabled on the project. The list can be filtered to only include services in a specific state, for example to only include services enabled on the project.

pub async fn batch_enable_services(
&mut self,
request: impl IntoRequest<BatchEnableServicesRequest>
) -> Result<Response<Operation>, Status>

Enables multiple services on a project. The operation is atomic: if enabling any service fails, then the entire batch fails, and no state changes occur.

Operation response type: google.protobuf.Empty

pub async fn list_consumer_quota_metrics(
&mut self,
request: impl IntoRequest<ListConsumerQuotaMetricsRequest>
) -> Result<Response<ListConsumerQuotaMetricsResponse>, Status>

Retrieves a summary of all quota information visible to the service consumer, organized by service metric. Each metric includes information about all of its defined limits. Each limit includes the limit configuration (quota unit, preciseness, default value), the current effective limit value, and all of the overrides applied to the limit.

pub async fn get_consumer_quota_metric(
&mut self,
request: impl IntoRequest<GetConsumerQuotaMetricRequest>
) -> Result<Response<ConsumerQuotaMetric>, Status>

Retrieves a summary of quota information for a specific quota metric

pub async fn get_consumer_quota_limit(
&mut self,
request: impl IntoRequest<GetConsumerQuotaLimitRequest>
) -> Result<Response<ConsumerQuotaLimit>, Status>

Retrieves a summary of quota information for a specific quota limit.

pub async fn create_admin_override(
&mut self,
request: impl IntoRequest<CreateAdminOverrideRequest>
) -> Result<Response<Operation>, Status>

Creates an admin override. An admin override is applied by an administrator of a parent folder or parent organization of the consumer receiving the override. An admin override is intended to limit the amount of quota the consumer can use out of the total quota pool allocated to all children of the folder or organization.

pub async fn update_admin_override(
&mut self,
request: impl IntoRequest<UpdateAdminOverrideRequest>
) -> Result<Response<Operation>, Status>

Updates an admin override.

pub async fn delete_admin_override(
&mut self,
request: impl IntoRequest<DeleteAdminOverrideRequest>
) -> Result<Response<Operation>, Status>

Deletes an admin override.

pub async fn list_admin_overrides(
&mut self,
request: impl IntoRequest<ListAdminOverridesRequest>
) -> Result<Response<ListAdminOverridesResponse>, Status>

Lists all admin overrides on this limit.

pub async fn import_admin_overrides(
&mut self,
request: impl IntoRequest<ImportAdminOverridesRequest>
) -> Result<Response<Operation>, Status>

Creates or updates multiple admin overrides atomically, all on the same consumer, but on many different metrics or limits. The name field in the quota override message should not be set.

pub async fn create_consumer_override(
&mut self,
request: impl IntoRequest<CreateConsumerOverrideRequest>
) -> Result<Response<Operation>, Status>

Creates a consumer override. A consumer override is applied to the consumer on its own authority to limit its own quota usage. Consumer overrides cannot be used to grant more quota than would be allowed by admin overrides, producer overrides, or the default limit of the service.

pub async fn update_consumer_override(
&mut self,
request: impl IntoRequest<UpdateConsumerOverrideRequest>
) -> Result<Response<Operation>, Status>

Updates a consumer override.

pub async fn delete_consumer_override(
&mut self,
request: impl IntoRequest<DeleteConsumerOverrideRequest>
) -> Result<Response<Operation>, Status>

Deletes a consumer override.

pub async fn list_consumer_overrides(
&mut self,
request: impl IntoRequest<ListConsumerOverridesRequest>
) -> Result<Response<ListConsumerOverridesResponse>, Status>

Lists all consumer overrides on this limit.

pub async fn import_consumer_overrides(
&mut self,
request: impl IntoRequest<ImportConsumerOverridesRequest>
) -> Result<Response<Operation>, Status>

Creates or updates multiple consumer overrides atomically, all on the same consumer, but on many different metrics or limits. The name field in the quota override message should not be set.

pub async fn generate_service_identity(
&mut self,
request: impl IntoRequest<GenerateServiceIdentityRequest>
) -> Result<Response<Operation>, Status>

Generates service identity for service.