Struct gapi_grpc::google::appengine::logging::v1::RequestLog

pub struct RequestLog {
    pub app_id: String,
    pub module_id: String,
    pub version_id: String,
    pub request_id: String,
    pub ip: String,
    pub start_time: Option<Timestamp>,
    pub end_time: Option<Timestamp>,
    pub latency: Option<Duration>,
    pub mega_cycles: i64,
    pub method: String,
    pub resource: String,
    pub http_version: String,
    pub status: i32,
    pub response_size: i64,
    pub referrer: String,
    pub user_agent: String,
    pub nickname: String,
    pub url_map_entry: String,
    pub host: String,
    pub cost: f64,
    pub task_queue_name: String,
    pub task_name: String,
    pub was_loading_request: bool,
    pub pending_time: Option<Duration>,
    pub instance_index: i32,
    pub finished: bool,
    pub first: bool,
    pub instance_id: String,
    pub line: Vec<LogLine>,
    pub app_engine_release: String,
    pub trace_id: String,
    pub trace_sampled: bool,
    pub source_reference: Vec<SourceReference>,
}

Complete log information about a single HTTP request to an App Engine application.

Fields
app_id: String

Application that handled this request.

module_id: String

Module of the application that handled this request.

version_id: String

Version of the application that handled this request.

request_id: String

Globally unique identifier for a request, which is based on the request start time. Request IDs for requests which started later will compare greater as strings than those for requests which started earlier.

ip: String

Origin IP address.

start_time: Option<Timestamp>

Time when the request started.

end_time: Option<Timestamp>

Time when the request finished.

latency: Option<Duration>

Latency of the request.

mega_cycles: i64

Number of CPU megacycles used to process request.

method: String

Request method. Example: "GET", "HEAD", "PUT", "POST", "DELETE".

resource: String

Contains the path and query portion of the URL that was requested. For example, if the URL was "http://example.com/app?name=val", the resource would be "/app?name=val". The fragment identifier, which is identified by the # character, is not included.

http_version: String

HTTP version of request. Example: "HTTP/1.1".

status: i32

HTTP response status code. Example: 200, 404.

response_size: i64

Size in bytes sent back to client by request.

referrer: String

Referrer URL of request.

user_agent: String

User agent that made the request.

nickname: String

The logged-in user who made the request.

Most likely, this is the part of the user's email before the @ sign. The field value is the same for different requests from the same user, but different users can have similar names. This information is also available to the application via the App Engine Users API.

This field will be populated starting with App Engine 1.9.21.

url_map_entry: String

File or class that handled the request.

host: String

Internet host and port number of the resource being requested.

cost: f64

An indication of the relative cost of serving this request.

task_queue_name: String

Queue name of the request, in the case of an offline request.

task_name: String

Task name of the request, in the case of an offline request.

was_loading_request: bool

Whether this was a loading request for the instance.

pending_time: Option<Duration>

Time this request spent in the pending request queue.

instance_index: i32

If the instance processing this request belongs to a manually scaled module, then this is the 0-based index of the instance. Otherwise, this value is -1.

finished: bool

Whether this request is finished or active.

first: bool

Whether this is the first RequestLog entry for this request. If an active request has several RequestLog entries written to Stackdriver Logging, then this field will be set for one of them.

instance_id: String

An identifier for the instance that handled the request.

line: Vec<LogLine>

A list of log lines emitted by the application while serving this request.

app_engine_release: String

App Engine release version.

trace_id: String

Stackdriver Trace identifier for this request.

trace_sampled: bool

If true, the value in the 'trace_id' field was sampled for storage in a trace backend.

source_reference: Vec<SourceReference>

Source code for the application that handled this request. There can be more than one source reference per deployed application if source code is distributed among multiple repositories.