Module gapi_grpc::google::assistant::embedded::v1alpha1

Modules

audio_in_config

audio_out_config

converse_request

converse_response

converse_result

embedded_assistant_client
Generated client implementations.

Structs

AudioInConfig
Specifies how to process the audio_in data that will be provided in subsequent requests. For recommended settings, see the Google Assistant SDK best practices.

AudioOut
The audio containing the assistant's response to the query. Sequential chunks of audio data are received in sequential ConverseResponse messages.

AudioOutConfig
Specifies the desired format for the server to use when it returns audio_out messages.

ConverseConfig
Specifies how to process the ConverseRequest messages.

ConverseRequest
The top-level message sent by the client. Clients must send at least two, and typically numerous ConverseRequest messages. The first message must contain a config message and must not contain audio_in data. All subsequent messages must contain audio_in data and must not contain a config message.

ConverseResponse
The top-level message received by the client. A series of one or more ConverseResponse messages are streamed back to the client.

ConverseResult
The semantic result for the user's spoken query.

ConverseState
Provides information about the current dialog state.