Module gapi_grpc::google::bigtable::admin::table::v1

Modules

bigtable_table_service_client
Generated client implementations.

bulk_delete_rows_request

gc_rule

table

Structs

BulkDeleteRowsRequest

ColumnFamily
A set of columns within a table which share a common configuration.

CreateColumnFamilyRequest

CreateTableRequest

DeleteColumnFamilyRequest

DeleteTableRequest

GcRule
Rule for determining which cells to delete during garbage collection.

GetTableRequest

ListTablesRequest

ListTablesResponse

RenameTableRequest

Table
A collection of user data indexed by row, column, and timestamp. Each table is served using the resources of its parent cluster.