Struct gapi_grpc::google::bigtable::admin::v2::bigtable_table_admin_client::BigtableTableAdminClient

pub struct BigtableTableAdminClient<T> { /* fields omitted */ }

Service for creating, configuring, and deleting Cloud Bigtable tables.

Provides access to the table schemas only, not the data stored within the tables.

Implementations

impl BigtableTableAdminClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> BigtableTableAdminClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_table<'_>(
&'_ mut self,
request: impl IntoRequest<CreateTableRequest>
) -> Result<Response<Table>, Status>

Creates a new table in the specified instance. The table can be created with a full set of initial column families, specified in the request.

pub async fn create_table_from_snapshot<'_>(
&'_ mut self,
request: impl IntoRequest<CreateTableFromSnapshotRequest>
) -> Result<Response<Operation>, Status>

Creates a new table from the specified snapshot. The target table must not exist. The snapshot and the table must be in the same instance.

Note: This is a private alpha release of Cloud Bigtable snapshots. This feature is not currently available to most Cloud Bigtable customers. This feature might be changed in backward-incompatible ways and is not recommended for production use. It is not subject to any SLA or deprecation policy.

pub async fn list_tables<'_>(
&'_ mut self,
request: impl IntoRequest<ListTablesRequest>
) -> Result<Response<ListTablesResponse>, Status>

Lists all tables served from a specified instance.

pub async fn get_table<'_>(
&'_ mut self,
request: impl IntoRequest<GetTableRequest>
) -> Result<Response<Table>, Status>

Gets metadata information about the specified table.

pub async fn delete_table<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteTableRequest>
) -> Result<Response<()>, Status>

Permanently deletes a specified table and all of its data.

pub async fn modify_column_families<'_>(
&'_ mut self,
request: impl IntoRequest<ModifyColumnFamiliesRequest>
) -> Result<Response<Table>, Status>

Performs a series of column family modifications on the specified table. Either all or none of the modifications will occur before this method returns, but data requests received prior to that point may see a table where only some modifications have taken effect.

pub async fn drop_row_range<'_>(
&'_ mut self,
request: impl IntoRequest<DropRowRangeRequest>
) -> Result<Response<()>, Status>

Permanently drop/delete a row range from a specified table. The request can specify whether to delete all rows in a table, or only those that match a particular prefix.

pub async fn generate_consistency_token<'_>(
&'_ mut self,
request: impl IntoRequest<GenerateConsistencyTokenRequest>
) -> Result<Response<GenerateConsistencyTokenResponse>, Status>

Generates a consistency token for a Table, which can be used in CheckConsistency to check whether mutations to the table that finished before this call started have been replicated. The tokens will be available for 90 days.

pub async fn check_consistency<'_>(
&'_ mut self,
request: impl IntoRequest<CheckConsistencyRequest>
) -> Result<Response<CheckConsistencyResponse>, Status>

Checks replication consistency based on a consistency token, that is, if replication has caught up based on the conditions specified in the token and the check request.

pub async fn snapshot_table<'_>(
&'_ mut self,
request: impl IntoRequest<SnapshotTableRequest>
) -> Result<Response<Operation>, Status>

Creates a new snapshot in the specified cluster from the specified source table. The cluster and the table must be in the same instance.

Note: This is a private alpha release of Cloud Bigtable snapshots. This feature is not currently available to most Cloud Bigtable customers. This feature might be changed in backward-incompatible ways and is not recommended for production use. It is not subject to any SLA or deprecation policy.

pub async fn get_snapshot<'_>(
&'_ mut self,
request: impl IntoRequest<GetSnapshotRequest>
) -> Result<Response<Snapshot>, Status>

Gets metadata information about the specified snapshot.

Note: This is a private alpha release of Cloud Bigtable snapshots. This feature is not currently available to most Cloud Bigtable customers. This feature might be changed in backward-incompatible ways and is not recommended for production use. It is not subject to any SLA or deprecation policy.

pub async fn list_snapshots<'_>(
&'_ mut self,
request: impl IntoRequest<ListSnapshotsRequest>
) -> Result<Response<ListSnapshotsResponse>, Status>

Lists all snapshots associated with the specified cluster.

Note: This is a private alpha release of Cloud Bigtable snapshots. This feature is not currently available to most Cloud Bigtable customers. This feature might be changed in backward-incompatible ways and is not recommended for production use. It is not subject to any SLA or deprecation policy.

pub async fn delete_snapshot<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteSnapshotRequest>
) -> Result<Response<()>, Status>

Permanently deletes the specified snapshot.

Note: This is a private alpha release of Cloud Bigtable snapshots. This feature is not currently available to most Cloud Bigtable customers. This feature might be changed in backward-incompatible ways and is not recommended for production use. It is not subject to any SLA or deprecation policy.

pub async fn create_backup<'_>(
&'_ mut self,
request: impl IntoRequest<CreateBackupRequest>
) -> Result<Response<Operation>, Status>

Starts creating a new Cloud Bigtable Backup. The returned backup [long-running operation][google.longrunning.Operation] can be used to track creation of the backup. The [metadata][google.longrunning.Operation.metadata] field type is [CreateBackupMetadata][google.bigtable.admin.v2.CreateBackupMetadata]. The [response][google.longrunning.Operation.response] field type is [Backup][google.bigtable.admin.v2.Backup], if successful. Cancelling the returned operation will stop the creation and delete the backup.

pub async fn get_backup<'_>(
&'_ mut self,
request: impl IntoRequest<GetBackupRequest>
) -> Result<Response<Backup>, Status>

Gets metadata on a pending or completed Cloud Bigtable Backup.

pub async fn update_backup<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateBackupRequest>
) -> Result<Response<Backup>, Status>

Updates a pending or completed Cloud Bigtable Backup.

pub async fn delete_backup<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteBackupRequest>
) -> Result<Response<()>, Status>

Deletes a pending or completed Cloud Bigtable backup.

pub async fn list_backups<'_>(
&'_ mut self,
request: impl IntoRequest<ListBackupsRequest>
) -> Result<Response<ListBackupsResponse>, Status>

Lists Cloud Bigtable backups. Returns both completed and pending backups.

pub async fn restore_table<'_>(
&'_ mut self,
request: impl IntoRequest<RestoreTableRequest>
) -> Result<Response<Operation>, Status>

Create a new table by restoring from a completed backup. The new table must be in the same instance as the instance containing the backup. The returned table [long-running operation][google.longrunning.Operation] can be used to track the progress of the operation, and to cancel it. The [metadata][google.longrunning.Operation.metadata] field type is [RestoreTableMetadata][google.bigtable.admin.RestoreTableMetadata]. The [response][google.longrunning.Operation.response] type is [Table][google.bigtable.admin.v2.Table], if successful.

pub async fn get_iam_policy<'_>(
&'_ mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the access control policy for a resource. Returns an empty policy if the resource exists but does not have a policy set.

pub async fn set_iam_policy<'_>(
&'_ mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Sets the access control policy on a Table or Backup resource. Replaces any existing policy.

pub async fn test_iam_permissions<'_>(
&'_ mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns permissions that the caller has on the specified table resource.