Module gapi_grpc::google::bigtable::v1

Modules

bigtable_service_client
Generated client implementations.

column_range

mutate_rows_request

mutation

read_modify_write_rule

read_rows_request

read_rows_response

row_filter

value_range

Structs

Cell
Specifies (some of) the contents of a single row/column/timestamp of a table.

CheckAndMutateRowRequest
Request message for BigtableService.CheckAndMutateRowRequest

CheckAndMutateRowResponse
Response message for BigtableService.CheckAndMutateRowRequest.

Column
Specifies (some of) the contents of a single row/column of a table.

ColumnRange
Specifies a contiguous range of columns within a single column family. The range spans from <column_family>:<start_qualifier> to <column_family>:<end_qualifier>, where both bounds can be either inclusive or exclusive.

Family
Specifies (some of) the contents of a single row/column family of a table.

MutateRowRequest
Request message for BigtableService.MutateRow.

MutateRowsRequest
Request message for BigtableService.MutateRows.

MutateRowsResponse
Response message for BigtableService.MutateRows.

Mutation
Specifies a particular change to be made to the contents of a row.

ReadModifyWriteRowRequest
Request message for BigtableService.ReadModifyWriteRowRequest.

ReadModifyWriteRule
Specifies an atomic read/modify/write operation on the latest value of the specified column.

ReadRowsRequest
Request message for BigtableServer.ReadRows.

ReadRowsResponse
Response message for BigtableService.ReadRows.

Row
Specifies the complete (requested) contents of a single row of a table. Rows which exceed 256MiB in size cannot be read in full.

RowFilter
Takes a row as input and produces an alternate view of the row based on specified rules. For example, a RowFilter might trim down a row to include just the cells from columns matching a given regular expression, or might return all the cells of a row but not their values. More complicated filters can be composed out of these components to express requests such as, "within every column of a particular family, give just the two most recent cells which are older than timestamp X."

RowRange
Specifies a contiguous range of rows.

RowSet
Specifies a non-contiguous set of rows.

SampleRowKeysRequest
Request message for BigtableService.SampleRowKeys.

SampleRowKeysResponse
Response message for BigtableService.SampleRowKeys.

TimestampRange
Specified a contiguous range of microsecond timestamps.

ValueRange
Specifies a contiguous range of raw byte values.