Module gapi_grpc::google::chromeos::moblab::v1beta1

Modules

build_service_client
Generated client implementations.

Structs

Build
Resource that represents a build for the given build target, model, milestone and build version. NEXT_TAG: 4

BuildArtifact
Resource that represents a build artifact stored in Google Cloud Storage for the given build target, model, build version and bucket. NEXT_TAG: 6

BuildTarget
Resource that represents a build target.

CheckBuildStageStatusRequest
Request message for checking if the build artifact is staged.

CheckBuildStageStatusResponse
Response message for checking the stage status of a build artifact. NEXT_TAG: 4

ListBuildsRequest
Request message for listing builds. NEXT_TAG: 7

ListBuildsResponse
Response message for listing builds. NEXT_TAG: 4

Milestone
Resource that represents a chrome OS milestone.

Model
Resource that represents a model. Each model belongs to a build target. For non-unified build, the model name is the same as its build target name.

StageBuildMetadata
Metadata message for staging a build artifact. NEXT_TAG: 4

StageBuildRequest
Request message for staging a build artifact.

StageBuildResponse
Response message for staging a build artifact.