Module gapi_grpc::google::cloud::aiplatform::v1::study_spec::parameter_spec

Modules

conditional_parameter_spec

Structs

CategoricalValueSpec
Value specification for a parameter in CATEGORICAL type.

ConditionalParameterSpec
Represents a parameter spec with condition from its parent parameter.

DiscreteValueSpec
Value specification for a parameter in DISCRETE type.

DoubleValueSpec
Value specification for a parameter in DOUBLE type.

IntegerValueSpec
Value specification for a parameter in INTEGER type.

Enums

ParameterValueSpec

ScaleType
The type of scaling that should be applied to this parameter.