Module gapi_grpc::google::cloud::aiplatform::v1beta1::explanation_metadata::input_metadata

Modules

visualization

Structs

FeatureValueDomain
Domain details of the input feature value. Provides numeric information about the feature, such as its range (min, max). If the feature has been pre-processed, for example with z-scoring, then it provides information about how to recover the original feature. For example, if the input feature is an image and it has been pre-processed to obtain 0-mean and stddev = 1 values, then original_mean, and original_stddev refer to the mean and stddev of the original feature (e.g. image tensor) from which input feature (with mean = 0 and stddev = 1) was obtained.

Visualization
Visualization configurations for image explanation.

Enums

Encoding
Defines how the feature is encoded to [encoded_tensor][]. Defaults to IDENTITY.