Module gapi_grpc::google::cloud::aiplatform::v1beta1::schema

Modules

image_segmentation_annotation

predict

prediction_result

tables_dataset_metadata

text_sentiment_prediction_result

time_series_dataset_metadata

trainingjob

video_object_tracking_prediction_result

Structs

AnnotationSpecColor
An entry of mapping between color and AnnotationSpec. The mapping is used in segmentation mask.

ClassificationPredictionResult
Prediction output format for Image Classification.

ImageBoundingBoxAnnotation
Annotation details specific to image object detection.

ImageClassificationAnnotation
Annotation details specific to image classification.

ImageClassificationPredictionInstance
Prediction input format for Image Classification.

ImageClassificationPredictionParams
Prediction model parameters for Image Classification.

ImageDataItem
Payload of Image DataItem.

ImageDatasetMetadata
The metadata of Datasets that contain Image DataItems.

ImageObjectDetectionPredictionInstance
Prediction input format for Image Object Detection.

ImageObjectDetectionPredictionParams
Prediction model parameters for Image Object Detection.

ImageObjectDetectionPredictionResult
Prediction output format for Image Object Detection.

ImageSegmentationAnnotation
Annotation details specific to image segmentation.

ImageSegmentationPredictionInstance
Prediction input format for Image Segmentation.

ImageSegmentationPredictionParams
Prediction model parameters for Image Segmentation.

PredictionResult
Represents a line of JSONL in the batch prediction output file.

TablesDatasetMetadata
The metadata of Datasets that contain tables data.

TextClassificationAnnotation
Annotation details specific to text classification.

TextClassificationPredictionInstance
Prediction input format for Text Classification.

TextDataItem
Payload of Text DataItem.

TextDatasetMetadata
The metadata of Datasets that contain Text DataItems.

TextExtractionAnnotation
Annotation details specific to text extraction.

TextExtractionPredictionInstance
Prediction input format for Text Extraction.

TextExtractionPredictionResult
Prediction output format for Text Extraction.

TextSegment
The text segment inside of DataItem.

TextSentimentAnnotation
Annotation details specific to text sentiment.

TextSentimentPredictionInstance
Prediction input format for Text Sentiment.

TextSentimentPredictionResult
Represents a line of JSONL in the text sentiment batch prediction output file. This is a hack to allow printing of integer values.

TextSentimentSavedQueryMetadata
The metadata of SavedQuery contains TextSentiment Annotations.

TimeSegment
A time period inside of a DataItem that has a time dimension (e.g. video).

TimeSeriesDatasetMetadata
The metadata of Datasets that contain time series data.

Vertex
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

VideoActionRecognitionAnnotation
Annotation details specific to video action recognition.

VideoActionRecognitionPredictionInstance
Prediction input format for Video Action Recognition.

VideoActionRecognitionPredictionParams
Prediction model parameters for Video Action Recognition.

VideoClassificationAnnotation
Annotation details specific to video classification.

VideoClassificationPredictionInstance
Prediction input format for Video Classification.

VideoClassificationPredictionParams
Prediction model parameters for Video Classification.

VideoClassificationPredictionResult
Prediction output format for Video Classification.

VideoDataItem
Payload of Video DataItem.

VideoDatasetMetadata
The metadata of Datasets that contain Video DataItems.

VideoObjectTrackingAnnotation
Annotation details specific to video object tracking.

VideoObjectTrackingPredictionInstance
Prediction input format for Video Classification.

VideoObjectTrackingPredictionParams
Prediction model parameters for Video Object Tracking.

VideoObjectTrackingPredictionResult
Prediction output format for Video Object Tracking.

VisualInspectionClassificationLabelSavedQueryMetadata

VisualInspectionMaskSavedQueryMetadata