Module gapi_grpc::google::cloud::aiplatform::v1beta1::schema::predict::prediction::video_object_tracking_prediction_result

Structs

Frame
The fields xMin, xMax, yMin, and yMax refer to a bounding box, i.e. the rectangle over the video frame pinpointing the found AnnotationSpec. The coordinates are relative to the frame size, and the point 0,0 is in the top left of the frame.