Module gapi_grpc::google::cloud::asset::v1beta1

Modules

asset_service_client
Generated client implementations.

gcs_destination

output_config

Structs

Asset
Cloud asset. This includes all Google Cloud Platform resources, Cloud IAM policies, and other non-GCP assets.

BatchGetAssetsHistoryRequest
Batch get assets history request.

BatchGetAssetsHistoryResponse
Batch get assets history response.

ExportAssetsRequest
Export asset request.

ExportAssetsResponse
The export asset response. This message is returned by the [google.longrunning.Operations.GetOperation][google.longrunning.Operations.GetOperation] method in the returned [google.longrunning.Operation.response][google.longrunning.Operation.response] field.

GcsDestination
A Cloud Storage location.

OutputConfig
Output configuration for export assets destination.

Resource
Representation of a cloud resource.

TemporalAsset
Temporal asset. In addition to the asset, the temporal asset includes the status of the asset and valid from and to time of it.

TimeWindow
A time window of (start_time, end_time].

Enums

ContentType
Asset content type.