Module gapi_grpc::google::cloud::asset::v1p4beta1::iam_policy_analysis_result

Modules

access

Structs

Access
A role or permission that appears in an access control list.

AccessControlList
An access control list, derived from the above IAM policy binding, which contains a set of resources and accesses. May include one item from each set to compose an access control entry.

AnalysisState
Represents analysis state of each node in the result graph or non-critical errors in the response.

Edge
A directional edge.

Identity
An identity that appears in an access control list.

IdentityList

Resource
A GCP resource that appears in an access control list.