Module gapi_grpc::google::cloud::asset::v1p7beta1

Modules

asset

asset_service_client
Generated client implementations.

gcs_destination

output_config

output_result

partition_spec

Structs

Asset
An asset in Google Cloud. An asset can be any resource in the Google Cloud resource hierarchy, a resource outside the Google Cloud resource hierarchy (such as Google Kubernetes Engine clusters and objects), or a policy (e.g. Cloud IAM policy). See Supported asset types for more information.

BigQueryDestination
A BigQuery destination for exporting assets to.

ExportAssetsRequest
Export asset request.

ExportAssetsResponse
The export asset response. This message is returned by the [google.longrunning.Operations.GetOperation][google.longrunning.Operations.GetOperation] method in the returned [google.longrunning.Operation.response][google.longrunning.Operation.response] field.

GcsDestination
A Cloud Storage location.

GcsOutputResult
A Cloud Storage output result.

OutputConfig
Output configuration for export assets destination.

OutputResult
Output result of export assets.

PartitionSpec
Specifications of BigQuery partitioned table as export destination.

RelatedAsset
An asset identify in Google Cloud which contains its name, type and ancestors. An asset can be any resource in the Google Cloud resource hierarchy, a resource outside the Google Cloud resource hierarchy (such as Google Kubernetes Engine clusters and objects), or a policy (e.g. Cloud IAM policy). See Supported asset types for more information.

RelatedAssets
The detailed related assets with the relationship_type.

RelationshipAttributes
The relationship attributes which include type, source_resource_type, target_resource_type and action.

Resource
A representation of a Google Cloud resource.

Enums

ContentType
Asset content type.