Struct gapi_grpc::google::cloud::automl::v1::auto_ml_client::AutoMlClient

pub struct AutoMlClient<T> { /* fields omitted */ }

AutoML Server API.

The resource names are assigned by the server. The server never reuses names that it has created after the resources with those names are deleted.

An ID of a resource is the last element of the item's resource name. For projects/{project_id}/locations/{location_id}/datasets/{dataset_id}, then the id for the item is {dataset_id}.

Currently the only supported location_id is "us-central1".

On any input that is documented to expect a string parameter in snake_case or kebab-case, either of those cases is accepted.

Implementations

impl AutoMlClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> AutoMlClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_dataset<'_>(
&'_ mut self,
request: impl IntoRequest<CreateDatasetRequest>
) -> Result<Response<Operation>, Status>

Creates a dataset.

pub async fn get_dataset<'_>(
&'_ mut self,
request: impl IntoRequest<GetDatasetRequest>
) -> Result<Response<Dataset>, Status>

Gets a dataset.

pub async fn list_datasets<'_>(
&'_ mut self,
request: impl IntoRequest<ListDatasetsRequest>
) -> Result<Response<ListDatasetsResponse>, Status>

Lists datasets in a project.

pub async fn update_dataset<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateDatasetRequest>
) -> Result<Response<Dataset>, Status>

Updates a dataset.

pub async fn delete_dataset<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteDatasetRequest>
) -> Result<Response<Operation>, Status>

Deletes a dataset and all of its contents. Returns empty response in the [response][google.longrunning.Operation.response] field when it completes, and delete_details in the [metadata][google.longrunning.Operation.metadata] field.

pub async fn import_data<'_>(
&'_ mut self,
request: impl IntoRequest<ImportDataRequest>
) -> Result<Response<Operation>, Status>

Imports data into a dataset. For Tables this method can only be called on an empty Dataset.

For Tables:

A [schema_inference_version][google.cloud.automl.v1.InputConfig.params] parameter must be explicitly set. Returns an empty response in the [response][google.longrunning.Operation.response] field when it completes.

pub async fn export_data<'_>(
&'_ mut self,
request: impl IntoRequest<ExportDataRequest>
) -> Result<Response<Operation>, Status>

Exports dataset's data to the provided output location. Returns an empty response in the [response][google.longrunning.Operation.response] field when it completes.

pub async fn get_annotation_spec<'_>(
&'_ mut self,
request: impl IntoRequest<GetAnnotationSpecRequest>
) -> Result<Response<AnnotationSpec>, Status>

Gets an annotation spec.

pub async fn create_model<'_>(
&'_ mut self,
request: impl IntoRequest<CreateModelRequest>
) -> Result<Response<Operation>, Status>

Creates a model. Returns a Model in the [response][google.longrunning.Operation.response] field when it completes. When you create a model, several model evaluations are created for it: a global evaluation, and one evaluation for each annotation spec.

pub async fn get_model<'_>(
&'_ mut self,
request: impl IntoRequest<GetModelRequest>
) -> Result<Response<Model>, Status>

Gets a model.

pub async fn list_models<'_>(
&'_ mut self,
request: impl IntoRequest<ListModelsRequest>
) -> Result<Response<ListModelsResponse>, Status>

Lists models.

pub async fn delete_model<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteModelRequest>
) -> Result<Response<Operation>, Status>

Deletes a model. Returns google.protobuf.Empty in the [response][google.longrunning.Operation.response] field when it completes, and delete_details in the [metadata][google.longrunning.Operation.metadata] field.

pub async fn update_model<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateModelRequest>
) -> Result<Response<Model>, Status>

Updates a model.

pub async fn deploy_model<'_>(
&'_ mut self,
request: impl IntoRequest<DeployModelRequest>
) -> Result<Response<Operation>, Status>

Deploys a model. If a model is already deployed, deploying it with the same parameters has no effect. Deploying with different parametrs (as e.g. changing

[node_number][google.cloud.automl.v1p1beta.ImageObjectDetectionModelDeploymentMetadata.node_number]) will reset the deployment state without pausing the model's availability.

Only applicable for Text Classification, Image Object Detection , Tables, and Image Segmentation; all other domains manage deployment automatically.

Returns an empty response in the [response][google.longrunning.Operation.response] field when it completes.

pub async fn undeploy_model<'_>(
&'_ mut self,
request: impl IntoRequest<UndeployModelRequest>
) -> Result<Response<Operation>, Status>

Undeploys a model. If the model is not deployed this method has no effect.

Only applicable for Text Classification, Image Object Detection and Tables; all other domains manage deployment automatically.

Returns an empty response in the [response][google.longrunning.Operation.response] field when it completes.

pub async fn export_model<'_>(
&'_ mut self,
request: impl IntoRequest<ExportModelRequest>
) -> Result<Response<Operation>, Status>

Exports a trained, "export-able", model to a user specified Google Cloud Storage location. A model is considered export-able if and only if it has an export format defined for it in [ModelExportOutputConfig][google.cloud.automl.v1.ModelExportOutputConfig].

Returns an empty response in the [response][google.longrunning.Operation.response] field when it completes.

pub async fn get_model_evaluation<'_>(
&'_ mut self,
request: impl IntoRequest<GetModelEvaluationRequest>
) -> Result<Response<ModelEvaluation>, Status>

Gets a model evaluation.

pub async fn list_model_evaluations<'_>(
&'_ mut self,
request: impl IntoRequest<ListModelEvaluationsRequest>
) -> Result<Response<ListModelEvaluationsResponse>, Status>

Lists model evaluations.