Module gapi_grpc::google::cloud::bigquery::storage::v1beta1

Modules

big_query_storage_client
Generated client implementations.

read_rows_response

read_session

Structs

ArrowRecordBatch
Arrow RecordBatch.

ArrowSchema
Arrow schema.

AvroRows
Avro rows.

AvroSchema
Avro schema.

BatchCreateReadSessionStreamsRequest
Information needed to request additional streams for an established read session.

BatchCreateReadSessionStreamsResponse
The response from BatchCreateReadSessionStreams returns the stream identifiers for the newly created streams.

CreateReadSessionRequest
Creates a new read session, which may include additional options such as requested parallelism, projection filters and constraints.

FinalizeStreamRequest
Request information for invoking FinalizeStream.

Progress

ReadRowsRequest
Requesting row data via ReadRows must provide Stream position information.

ReadRowsResponse
Response from calling ReadRows may include row data, progress and throttling information.

ReadSession
Information returned from a CreateReadSession request.

SplitReadStreamRequest
Request information for SplitReadStream.

SplitReadStreamResponse
Response from SplitReadStream.

Stream
Information about a single data stream within a read session.

StreamPosition
Expresses a point within a given stream using an offset position.

StreamStatus
Progress information for a given Stream.

TableModifiers
All fields in this message optional.

TableReadOptions
Options dictating how we read a table.

TableReference
Table reference that includes just the 3 strings needed to identify a table.

ThrottleStatus
Information on if the current connection is being throttled.

Enums

DataFormat
Data format for input or output data.

ShardingStrategy
Strategy for distributing data among multiple streams in a read session.