Enum gapi_grpc::google::cloud::bigquery::v2::model::HolidayRegion

#[repr(i32)]pub enum HolidayRegion {
    Unspecified,
    Global,
    Na,
    Japac,
    Emea,
    Lac,
    Ae,
    Ar,
    At,
    Au,
    Be,
    Br,
    Ca,
    Ch,
    Cl,
    Cn,
    Co,
    Cs,
    Cz,
    De,
    Dk,
    Dz,
    Ec,
    Ee,
    Eg,
    Es,
    Fi,
    Fr,
    Gb,
    Gr,
    Hk,
    Hu,
    Id,
    Ie,
    Il,
    In,
    Ir,
    It,
    Jp,
    Kr,
    Lv,
    Ma,
    Mx,
    My,
    Ng,
    Nl,
    No,
    Nz,
    Pe,
    Ph,
    Pk,
    Pl,
    Pt,
    Ro,
    Rs,
    Ru,
    Sa,
    Se,
    Sg,
    Si,
    Sk,
    Th,
    Tr,
    Tw,
    Ua,
    Us,
    Ve,
    Vn,
    Za,
}

Type of supported holiday regions for time series forecasting models.

Variants

Unspecified

Holiday region unspecified.

Global

Global.

Na

North America.

Japac

Japan and Asia Pacific: Korea, Greater China, India, Australia, and New Zealand.

Emea

Europe, the Middle East and Africa.

Lac

Latin America and the Caribbean.

Ae

United Arab Emirates

Ar

Argentina

At

Austria

Au

Australia

Be

Belgium

Br

Brazil

Ca

Canada

Ch

Switzerland

Cl

Chile

Cn

China

Co

Colombia

Cs

Czechoslovakia

Cz

Czech Republic

De

Germany

Dk

Denmark

Dz

Algeria

Ec

Ecuador

Ee

Estonia

Eg

Egypt

Es

Spain

Fi

Finland

Fr

France

Gb

Great Britain (United Kingdom)

Gr

Greece

Hk

Hong Kong

Hu

Hungary

Id

Indonesia

Ie

Ireland

Il

Israel

In

India

Ir

Iran

It

Italy

Jp

Japan

Kr

Korea (South)

Lv

Latvia

Ma

Morocco

Mx

Mexico

My

Malaysia

Ng

Nigeria

Nl

Netherlands

No

Norway

Nz

New Zealand

Pe

Peru

Ph

Philippines

Pk

Pakistan

Pl

Poland

Pt

Portugal

Ro

Romania

Rs

Serbia

Ru

Russian Federation

Sa

Saudi Arabia

Se

Sweden

Sg

Singapore

Si

Slovenia

Sk

Slovakia

Th

Thailand

Tr

Turkey

Tw

Taiwan

Ua

Ukraine

Us

United States

Ve

Venezuela

Vn

Viet Nam

Za

South Africa

Implementations

impl HolidayRegion

pub fn is_valid(value: i32) -> bool

Returns true if value is a variant of HolidayRegion.

pub fn from_i32(value: i32) -> Option<HolidayRegion>

Converts an i32 to a HolidayRegion, or None if value is not a valid variant.