Module gapi_grpc::google::cloud::billing::v1

Modules

aggregation_info

cloud_billing_client
Generated client implementations.

cloud_catalog_client
Generated client implementations.

pricing_expression

Structs

AggregationInfo
Represents the aggregation level and interval for pricing of a single SKU.

BillingAccount
A billing account in GCP Console. You can assign a billing account to one or more projects.

Category
Represents the category hierarchy of a SKU.

CreateBillingAccountRequest
Request message for CreateBillingAccount.

GetBillingAccountRequest
Request message for GetBillingAccount.

GetProjectBillingInfoRequest
Request message for GetProjectBillingInfo.

ListBillingAccountsRequest
Request message for ListBillingAccounts.

ListBillingAccountsResponse
Response message for ListBillingAccounts.

ListProjectBillingInfoRequest
Request message for ListProjectBillingInfo.

ListProjectBillingInfoResponse
Request message for ListProjectBillingInfoResponse.

ListServicesRequest
Request message for ListServices.

ListServicesResponse
Response message for ListServices.

ListSkusRequest
Request message for ListSkus.

ListSkusResponse
Response message for ListSkus.

PricingExpression
Expresses a mathematical pricing formula. For Example:-

PricingInfo
Represents the pricing information for a SKU at a single point of time.

ProjectBillingInfo
Encapsulation of billing information for a GCP Console project. A project has at most one associated billing account at a time (but a billing account can be assigned to multiple projects).

Service
Encapsulates a single service in Google Cloud Platform.

Sku
Encapsulates a single SKU in Google Cloud Platform

UpdateBillingAccountRequest
Request message for UpdateBillingAccount.

UpdateProjectBillingInfoRequest
Request message for UpdateProjectBillingInfo.