Struct gapi_grpc::google::cloud::channel::v1::cloud_channel_service_client::CloudChannelServiceClient

pub struct CloudChannelServiceClient<T> { /* fields omitted */ }

CloudChannelService lets Google cloud resellers and distributors manage their customers, channel partners, entitlements, and reports.

Using this service:

Resellers and distributors can manage a customer entity.

Distributors can register an authorized reseller in their channel and provide them with delegated admin access.

Resellers and distributors can manage customer entitlements.

CloudChannelService exposes the following resources:

[Customer][google.cloud.channel.v1.Customer]s: An entity—usually an enterprise—managed by a reseller or distributor.

[Entitlement][google.cloud.channel.v1.Entitlement]s: An entity that provides a customer with the means to use a service. Entitlements are created or updated as a result of a successful fulfillment.

[ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink]s: An entity that identifies links between distributors and their indirect resellers in a channel.

Implementations

impl CloudChannelServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> CloudChannelServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn list_customers(
&mut self,
request: impl IntoRequest<ListCustomersRequest>
) -> Result<Response<ListCustomersResponse>, Status>

List [Customer][google.cloud.channel.v1.Customer]s.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

Return value: List of [Customer][google.cloud.channel.v1.Customer]s, or an empty list if there are no customers.

pub async fn get_customer(
&mut self,
request: impl IntoRequest<GetCustomerRequest>
) -> Result<Response<Customer>, Status>

Returns a requested [Customer][google.cloud.channel.v1.Customer] resource.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The customer resource doesn’t exist. Usually the result of an invalid name parameter.

Return value: The [Customer][google.cloud.channel.v1.Customer] resource.

pub async fn check_cloud_identity_accounts_exist(
&mut self,
request: impl IntoRequest<CheckCloudIdentityAccountsExistRequest>
) -> Result<Response<CheckCloudIdentityAccountsExistResponse>, Status>

Confirms the existence of Cloud Identity accounts based on the domain and if the Cloud Identity accounts are owned by the reseller.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

INVALID_VALUE: Invalid domain value in the request.

Return value: A list of [CloudIdentityCustomerAccount][google.cloud.channel.v1.CloudIdentityCustomerAccount] resources for the domain (may be empty)

Note: in the v1alpha1 version of the API, a NOT_FOUND error returns if no [CloudIdentityCustomerAccount][google.cloud.channel.v1.CloudIdentityCustomerAccount] resources match the domain.

pub async fn create_customer(
&mut self,
request: impl IntoRequest<CreateCustomerRequest>
) -> Result<Response<Customer>, Status>

Creates a new [Customer][google.cloud.channel.v1.Customer] resource under the reseller or distributor account.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT:

Required request parameters are missing or invalid.

Domain field value doesn’t match the primary email domain.

Return value: The newly created [Customer][google.cloud.channel.v1.Customer] resource.

pub async fn update_customer(
&mut self,
request: impl IntoRequest<UpdateCustomerRequest>
) -> Result<Response<Customer>, Status>

Updates an existing [Customer][google.cloud.channel.v1.Customer] resource for the reseller or distributor.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: No [Customer][google.cloud.channel.v1.Customer] resource found for the name in the request.

Return value: The updated [Customer][google.cloud.channel.v1.Customer] resource.

pub async fn delete_customer(
&mut self,
request: impl IntoRequest<DeleteCustomerRequest>
) -> Result<Response<()>, Status>

Deletes the given [Customer][google.cloud.channel.v1.Customer] permanently and irreversibly.

Possible error codes:

PERMISSION_DENIED: The account making the request does not own this customer.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

FAILED_PRECONDITION: The customer has existing entitlements.

NOT_FOUND: No [Customer][google.cloud.channel.v1.Customer] resource found for the name in the request.

pub async fn provision_cloud_identity(
&mut self,
request: impl IntoRequest<ProvisionCloudIdentityRequest>
) -> Result<Response<Operation>, Status>

Creates a Cloud Identity for the given customer using the customer’s information, or the information provided here.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The customer was not found.

ALREADY_EXISTS: The customer’s primary email already exists. Retry after changing the customer’s primary contact email.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata contains an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn list_entitlements(
&mut self,
request: impl IntoRequest<ListEntitlementsRequest>
) -> Result<Response<ListEntitlementsResponse>, Status>

Lists [Entitlement][google.cloud.channel.v1.Entitlement]s belonging to a customer.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

Return value: A list of the customer’s [Entitlement][google.cloud.channel.v1.Entitlement]s.

pub async fn list_transferable_skus(
&mut self,
request: impl IntoRequest<ListTransferableSkusRequest>
) -> Result<Response<ListTransferableSkusResponse>, Status>

List [TransferableSku][google.cloud.channel.v1.TransferableSku]s of a customer based on the Cloud Identity ID or Customer Name in the request.

Use this method to list the entitlements information of an unowned customer. You should provide the customer’s Cloud Identity ID or Customer Name.

Possible error codes:

PERMISSION_DENIED:

The customer doesn’t belong to the reseller and has no auth token.

The supplied auth token is invalid.

The reseller account making the request is different from the reseller account in the query.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

Return value: A list of the customer’s [TransferableSku][google.cloud.channel.v1.TransferableSku].

pub async fn list_transferable_offers(
&mut self,
request: impl IntoRequest<ListTransferableOffersRequest>
) -> Result<Response<ListTransferableOffersResponse>, Status>

List [TransferableOffer][google.cloud.channel.v1.TransferableOffer]s of a customer based on Cloud Identity ID or Customer Name in the request.

Use this method when a reseller gets the entitlement information of an unowned customer. The reseller should provide the customer’s Cloud Identity ID or Customer Name.

Possible error codes:

PERMISSION_DENIED:

The customer doesn’t belong to the reseller and has no auth token.

The supplied auth token is invalid.

The reseller account making the request is different from the reseller account in the query.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

Return value: List of [TransferableOffer][google.cloud.channel.v1.TransferableOffer] for the given customer and SKU.

pub async fn get_entitlement(
&mut self,
request: impl IntoRequest<GetEntitlementRequest>
) -> Result<Response<Entitlement>, Status>

Returns a requested [Entitlement][google.cloud.channel.v1.Entitlement] resource.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The customer entitlement was not found.

Return value: The requested [Entitlement][google.cloud.channel.v1.Entitlement] resource.

pub async fn create_entitlement(
&mut self,
request: impl IntoRequest<CreateEntitlementRequest>
) -> Result<Response<Operation>, Status>

Creates an entitlement for a customer.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT:

Required request parameters are missing or invalid.

There is already a customer entitlement for a SKU from the same product family.

INVALID_VALUE: Make sure the OfferId is valid. If it is, contact Google Channel support for further troubleshooting.

NOT_FOUND: The customer or offer resource was not found.

ALREADY_EXISTS:

The SKU was already purchased for the customer.

The customer’s primary email already exists. Retry after changing the customer’s primary contact email.

CONDITION_NOT_MET or FAILED_PRECONDITION:

The domain required for purchasing a SKU has not been verified.

A pre-requisite SKU required to purchase an Add-On SKU is missing. For example, Google Workspace Business Starter is required to purchase Vault or Drive.

(Developer accounts only) Reseller and resold domain must meet the following naming requirements:

Domain names must start with goog-test.

Domain names must include the reseller domain.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn change_parameters(
&mut self,
request: impl IntoRequest<ChangeParametersRequest>
) -> Result<Response<Operation>, Status>

Change parameters of the entitlement.

An entitlement update is a long-running operation and it updates the entitlement as a result of fulfillment.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid. For example, the number of seats being changed is greater than the allowed number of max seats, or decreasing seats for a commitment based plan.

NOT_FOUND: Entitlement resource not found.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn change_renewal_settings(
&mut self,
request: impl IntoRequest<ChangeRenewalSettingsRequest>
) -> Result<Response<Operation>, Status>

Updates the renewal settings for an existing customer entitlement.

An entitlement update is a long-running operation and it updates the entitlement as a result of fulfillment.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: Entitlement resource not found.

NOT_COMMITMENT_PLAN: Renewal Settings are only applicable for a commitment plan. Can’t enable or disable renewals for non-commitment plans.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn change_offer(
&mut self,
request: impl IntoRequest<ChangeOfferRequest>
) -> Result<Response<Operation>, Status>

Updates the Offer for an existing customer entitlement.

An entitlement update is a long-running operation and it updates the entitlement as a result of fulfillment.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: Offer or Entitlement resource not found.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn start_paid_service(
&mut self,
request: impl IntoRequest<StartPaidServiceRequest>
) -> Result<Response<Operation>, Status>

Starts paid service for a trial entitlement.

Starts paid service for a trial entitlement immediately. This method is only applicable if a plan is set up for a trial entitlement but has some trial days remaining.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: Entitlement resource not found.

FAILED_PRECONDITION/NOT_IN_TRIAL: This method only works for entitlement on trial plans.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn suspend_entitlement(
&mut self,
request: impl IntoRequest<SuspendEntitlementRequest>
) -> Result<Response<Operation>, Status>

Suspends a previously fulfilled entitlement.

An entitlement suspension is a long-running operation.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: Entitlement resource not found.

NOT_ACTIVE: Entitlement is not active.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn cancel_entitlement(
&mut self,
request: impl IntoRequest<CancelEntitlementRequest>
) -> Result<Response<Operation>, Status>

Cancels a previously fulfilled entitlement.

An entitlement cancellation is a long-running operation.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

FAILED_PRECONDITION: There are Google Cloud projects linked to the Google Cloud entitlement’s Cloud Billing subaccount.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: Entitlement resource not found.

DELETION_TYPE_NOT_ALLOWED: Cancel is only allowed for Google Workspace add-ons, or entitlements for Google Cloud’s development platform.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The response will contain google.protobuf.Empty on success. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn activate_entitlement(
&mut self,
request: impl IntoRequest<ActivateEntitlementRequest>
) -> Result<Response<Operation>, Status>

Activates a previously suspended entitlement. Entitlements suspended for pending ToS acceptance can’t be activated using this method.

An entitlement activation is a long-running operation and it updates the state of the customer entitlement.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: Entitlement resource not found.

SUSPENSION_NOT_RESELLER_INITIATED: Can only activate reseller-initiated suspensions and entitlements that have accepted the TOS.

NOT_SUSPENDED: Can only activate suspended entitlements not in an ACTIVE state.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn transfer_entitlements(
&mut self,
request: impl IntoRequest<TransferEntitlementsRequest>
) -> Result<Response<Operation>, Status>

Transfers customer entitlements to new reseller.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The customer or offer resource was not found.

ALREADY_EXISTS: The SKU was already transferred for the customer.

CONDITION_NOT_MET or FAILED_PRECONDITION:

The SKU requires domain verification to transfer, but the domain is not verified.

An Add-On SKU (example, Vault or Drive) is missing the pre-requisite SKU (example, G Suite Basic).

(Developer accounts only) Reseller and resold domain must meet the following naming requirements:

Domain names must start with goog-test.

Domain names must include the reseller domain.

Specify all transferring entitlements.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn transfer_entitlements_to_google(
&mut self,
request: impl IntoRequest<TransferEntitlementsToGoogleRequest>
) -> Result<Response<Operation>, Status>

Transfers customer entitlements from their current reseller to Google.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The customer or offer resource was not found.

ALREADY_EXISTS: The SKU was already transferred for the customer.

CONDITION_NOT_MET or FAILED_PRECONDITION:

The SKU requires domain verification to transfer, but the domain is not verified.

An Add-On SKU (example, Vault or Drive) is missing the pre-requisite SKU (example, G Suite Basic).

(Developer accounts only) Reseller and resold domain must meet the following naming requirements:

Domain names must start with goog-test.

Domain names must include the reseller domain.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The ID of a long-running operation.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The response will contain google.protobuf.Empty on success. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn list_channel_partner_links(
&mut self,
request: impl IntoRequest<ListChannelPartnerLinksRequest>
) -> Result<Response<ListChannelPartnerLinksResponse>, Status>

List [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink]s belonging to a distributor. You must be a distributor to call this method.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

Return value: The list of the distributor account’s [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resources.

pub async fn get_channel_partner_link(
&mut self,
request: impl IntoRequest<GetChannelPartnerLinkRequest>
) -> Result<Response<ChannelPartnerLink>, Status>

Returns a requested [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource. You must be a distributor to call this method.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: ChannelPartnerLink resource not found because of an invalid channel partner link name.

Return value: The [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource.

pub async fn create_channel_partner_link(
&mut self,
request: impl IntoRequest<CreateChannelPartnerLinkRequest>
) -> Result<Response<ChannelPartnerLink>, Status>

Initiates a channel partner link between a distributor and a reseller, or between resellers in an n-tier reseller channel. Invited partners need to follow the invite_link_uri provided in the response to accept. After accepting the invitation, a link is set up between the two parties. You must be a distributor to call this method.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

ALREADY_EXISTS: The ChannelPartnerLink sent in the request already exists.

NOT_FOUND: No Cloud Identity customer exists for provided domain.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The new [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource.

pub async fn update_channel_partner_link(
&mut self,
request: impl IntoRequest<UpdateChannelPartnerLinkRequest>
) -> Result<Response<ChannelPartnerLink>, Status>

Updates a channel partner link. Distributors call this method to change a link’s status. For example, to suspend a partner link. You must be a distributor to call this method.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request is different from the reseller account in the API request.

INVALID_ARGUMENT:

Required request parameters are missing or invalid.

Link state cannot change from invited to active or suspended.

Cannot send reseller_cloud_identity_id, invite_url, or name in update mask.

NOT_FOUND: ChannelPartnerLink resource not found.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The updated [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource.

pub async fn list_products(
&mut self,
request: impl IntoRequest<ListProductsRequest>
) -> Result<Response<ListProductsResponse>, Status>

Lists the Products the reseller is authorized to sell.

Possible error codes:

INVALID_ARGUMENT: Required request parameters are missing or invalid.

pub async fn list_skus(
&mut self,
request: impl IntoRequest<ListSkusRequest>
) -> Result<Response<ListSkusResponse>, Status>

Lists the SKUs for a product the reseller is authorized to sell.

Possible error codes:

INVALID_ARGUMENT: Required request parameters are missing or invalid.

pub async fn list_offers(
&mut self,
request: impl IntoRequest<ListOffersRequest>
) -> Result<Response<ListOffersResponse>, Status>

Lists the Offers the reseller can sell.

Possible error codes:

INVALID_ARGUMENT: Required request parameters are missing or invalid.

pub async fn list_purchasable_skus(
&mut self,
request: impl IntoRequest<ListPurchasableSkusRequest>
) -> Result<Response<ListPurchasableSkusResponse>, Status>

Lists the following:

SKUs that you can purchase for a customer

SKUs that you can upgrade or downgrade for an entitlement.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

pub async fn list_purchasable_offers(
&mut self,
request: impl IntoRequest<ListPurchasableOffersRequest>
) -> Result<Response<ListPurchasableOffersResponse>, Status>

Lists the following:

Offers that you can purchase for a customer.

Offers that you can change for an entitlement.

Possible error codes:

PERMISSION_DENIED: The customer doesn’t belong to the reseller

INVALID_ARGUMENT: Required request parameters are missing or invalid.

pub async fn register_subscriber(
&mut self,
request: impl IntoRequest<RegisterSubscriberRequest>
) -> Result<Response<RegisterSubscriberResponse>, Status>

Registers a service account with subscriber privileges on the Cloud Pub/Sub topic for this Channel Services account. After you create a subscriber, you get the events through [SubscriberEvent][google.cloud.channel.v1.SubscriberEvent]

Possible error codes:

PERMISSION_DENIED: The reseller account making the request and the provided reseller account are different, or the impersonated user is not a super admin.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The topic name with the registered service email address.

pub async fn unregister_subscriber(
&mut self,
request: impl IntoRequest<UnregisterSubscriberRequest>
) -> Result<Response<UnregisterSubscriberResponse>, Status>

Unregisters a service account with subscriber privileges on the Cloud Pub/Sub topic created for this Channel Services account. If there are no service accounts left with subscriber privileges, this deletes the topic. You can call ListSubscribers to check for these accounts.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request and the provided reseller account are different, or the impersonated user is not a super admin.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The topic resource doesn’t exist.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: The topic name that unregistered the service email address. Returns a success response if the service email address wasn’t registered with the topic.

pub async fn list_subscribers(
&mut self,
request: impl IntoRequest<ListSubscribersRequest>
) -> Result<Response<ListSubscribersResponse>, Status>

Lists service accounts with subscriber privileges on the Cloud Pub/Sub topic created for this Channel Services account.

Possible error codes:

PERMISSION_DENIED: The reseller account making the request and the provided reseller account are different, or the impersonated user is not a super admin.

INVALID_ARGUMENT: Required request parameters are missing or invalid.

NOT_FOUND: The topic resource doesn’t exist.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support.

Return value: A list of service email addresses.