Module gapi_grpc::google::cloud::contactcenterinsights::v1

Modules

analysis_result

annotation_boundary

answer_feedback

call_annotation

contact_center_insights_client
Generated client implementations.

conversation

conversation_data_source

entity

entity_mention_data

export_insights_data_request

issue_model

issue_model_label_stats

phrase_match_rule_config

phrase_match_rule_group

phrase_matcher

runtime_annotation

settings

Structs

Analysis
The analysis resource.

AnalysisResult
The result of an analysis.

AnnotationBoundary
A point in a conversation that marks the start or the end of an annotation.

AnswerFeedback
The feedback that the customer has about a certain answer in the conversation.

ArticleSuggestionData
Agent Assist Article Suggestion data.

CalculateIssueModelStatsRequest
Request to get statistics of an issue model.

CalculateIssueModelStatsResponse
Response of querying an issue model’s statistics.

CalculateStatsRequest
The request for calculating conversation statistics.

CalculateStatsResponse
The response for calculating conversation statistics.

CallAnnotation
A piece of metadata that applies to a window of a call.

Conversation
The conversation resource.

ConversationDataSource
The conversation source, which is a combination of transcript and audio.

ConversationLevelSentiment
One channel of conversation-level sentiment data.

CreateAnalysisOperationMetadata
Metadata for a create analysis operation.

CreateAnalysisRequest
The request to create an analysis.

CreateConversationRequest
Request to create a conversation.

CreatePhraseMatcherRequest
Request to create a phrase matcher.

DeleteAnalysisRequest
The request to delete an analysis.

DeleteConversationRequest
The request to delete a conversation.

DeletePhraseMatcherRequest
The request to delete a phrase matcher.

DialogflowIntent
The data for a Dialogflow intent. Represents a detected intent in the conversation, e.g. MAKES_PROMISE.

DialogflowInteractionData
Dialogflow interaction data.

DialogflowSource
A Dialogflow source of conversation data.

Entity
The data for an entity annotation. Represents a phrase in the conversation that is a known entity, such as a person, an organization, or location.

EntityMentionData
The data for an entity mention annotation. This represents a mention of an Entity in the conversation.

ExactMatchConfig
Exact match configuration.

ExportInsightsDataMetadata
Metadata for an export insights operation.

ExportInsightsDataRequest
The request to export insights.

ExportInsightsDataResponse
Response for an export insights operation.

FaqAnswerData
Agent Assist frequently-asked-question answer data.

GcsSource
A Cloud Storage source of conversation data.

GetAnalysisRequest
The request to get an analysis.

GetConversationRequest
The request to get a conversation.

GetIssueModelRequest
The request to get an issue model.

GetIssueRequest
The request to get an issue.

GetPhraseMatcherRequest
The request to get a a phrase matcher.

GetSettingsRequest
The request to get project-level settings.

HoldData
The data for a hold annotation.

Intent
The data for an intent. Represents a detected intent in the conversation, for example MAKES_PROMISE.

IntentMatchData
The data for an intent match. Represents an intent match for a text segment in the conversation. A text segment can be part of a sentence, a complete sentence, or an utterance with multiple sentences.

InterruptionData
The data for an interruption annotation.

Issue
The issue resource.

IssueAssignment
Information about the issue.

IssueModel
The issue model resource.

IssueModelLabelStats
Aggregated statistics about an issue model.

IssueModelResult
Issue Modeling result on a conversation.

ListAnalysesRequest
The request to list analyses.

ListAnalysesResponse
The response to list analyses.

ListConversationsRequest
Request to list conversations.

ListConversationsResponse
The response of listing conversations.

ListIssueModelsRequest
Request to list issue models.

ListIssueModelsResponse
The response of listing issue models.

ListIssuesRequest
Request to list issues.

ListIssuesResponse
The response of listing issues.

ListPhraseMatchersRequest
Request to list phrase matchers.

ListPhraseMatchersResponse
The response of listing phrase matchers.

PhraseMatchData
The data for a matched phrase matcher. Represents information identifying a phrase matcher for a given match.

PhraseMatchRule
The data for a phrase match rule.

PhraseMatchRuleConfig
Configuration information of a phrase match rule.

PhraseMatchRuleGroup
A message representing a rule in the phrase matcher.

PhraseMatcher
The phrase matcher resource.

RuntimeAnnotation
An annotation that was generated during the customer and agent interaction.

SentimentData
The data for a sentiment annotation.

Settings
The settings resource.

SilenceData
The data for a silence annotation.

SmartComposeSuggestionData
Agent Assist Smart Compose suggestion data.

SmartReplyData
Agent Assist Smart Reply data.

UpdateConversationRequest
The request to update a conversation.

UpdateSettingsRequest
The request to update project-level settings.

Enums

ConversationView
Represents the options for views of a conversation.