Module gapi_grpc::google::cloud::datacatalog::v1

Modules

big_query_table_spec

data_catalog_client
Generated client implementations.

entry

field_type

lookup_entry_request

search_catalog_request

search_catalog_result

tag

tag_field

Structs

BigQueryDateShardedSpec
Spec for a group of BigQuery tables with name pattern [prefix]YYYYMMDD. Context: https://cloud.google.com/bigquery/docs/partitioned-tables#partitioning_versus_sharding

BigQueryTableSpec
Describes a BigQuery table.

ColumnSchema
Representation of a column within a schema. Columns could be nested inside other columns.

CreateEntryGroupRequest
Request message for [CreateEntryGroup][google.cloud.datacatalog.v1.DataCatalog.CreateEntryGroup].

CreateEntryRequest
Request message for [CreateEntry][google.cloud.datacatalog.v1.DataCatalog.CreateEntry].

CreateTagRequest
Request message for [CreateTag][google.cloud.datacatalog.v1.DataCatalog.CreateTag].

CreateTagTemplateFieldRequest
Request message for [CreateTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.CreateTagTemplateField].

CreateTagTemplateRequest
Request message for [CreateTagTemplate][google.cloud.datacatalog.v1.DataCatalog.CreateTagTemplate].

DeleteEntryGroupRequest
Request message for [DeleteEntryGroup][google.cloud.datacatalog.v1.DataCatalog.DeleteEntryGroup].

DeleteEntryRequest
Request message for [DeleteEntry][google.cloud.datacatalog.v1.DataCatalog.DeleteEntry].

DeleteTagRequest
Request message for [DeleteTag][google.cloud.datacatalog.v1.DataCatalog.DeleteTag].

DeleteTagTemplateFieldRequest
Request message for [DeleteTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.DeleteTagTemplateField].

DeleteTagTemplateRequest
Request message for [DeleteTagTemplate][google.cloud.datacatalog.v1.DataCatalog.DeleteTagTemplate].

Entry
Entry Metadata. A Data Catalog Entry resource represents another resource in Google Cloud Platform (such as a BigQuery dataset or a Pub/Sub topic) or outside of Google Cloud Platform. Clients can use the linked_resource field in the Entry resource to refer to the original resource ID of the source system.

EntryGroup
EntryGroup Metadata. An EntryGroup resource represents a logical grouping of zero or more Data Catalog [Entry][google.cloud.datacatalog.v1.Entry] resources.

FieldType

GcsFileSpec
Specifications of a single file in Cloud Storage.

GcsFilesetSpec
Describes a Cloud Storage fileset entry.

GetEntryGroupRequest
Request message for [GetEntryGroup][google.cloud.datacatalog.v1.DataCatalog.GetEntryGroup].

GetEntryRequest
Request message for [GetEntry][google.cloud.datacatalog.v1.DataCatalog.GetEntry].

GetTagTemplateRequest
Request message for [GetTagTemplate][google.cloud.datacatalog.v1.DataCatalog.GetTagTemplate].

ListEntriesRequest
Request message for [ListEntries][google.cloud.datacatalog.v1.DataCatalog.ListEntries].

ListEntriesResponse
Response message for [ListEntries][google.cloud.datacatalog.v1.DataCatalog.ListEntries].

ListEntryGroupsRequest
Request message for [ListEntryGroups][google.cloud.datacatalog.v1.DataCatalog.ListEntryGroups].

ListEntryGroupsResponse
Response message for [ListEntryGroups][google.cloud.datacatalog.v1.DataCatalog.ListEntryGroups].

ListTagsRequest
Request message for [ListTags][google.cloud.datacatalog.v1.DataCatalog.ListTags].

ListTagsResponse
Response message for [ListTags][google.cloud.datacatalog.v1.DataCatalog.ListTags].

LookupEntryRequest
Request message for [LookupEntry][google.cloud.datacatalog.v1.DataCatalog.LookupEntry].

RenameTagTemplateFieldRequest
Request message for [RenameTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.RenameTagTemplateField].

Schema
Represents a schema (e.g. BigQuery, GoogleSQL, Avro schema).

SearchCatalogRequest
Request message for [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

SearchCatalogResponse
Response message for [SearchCatalog][google.cloud.datacatalog.v1.DataCatalog.SearchCatalog].

SearchCatalogResult
A result that appears in the response of a search request. Each result captures details of one entry that matches the search.

SystemTimestamps
Timestamps about this resource according to a particular system.

TableSpec
Normal BigQuery table spec.

Tag
Tags are used to attach custom metadata to Data Catalog resources. Tags conform to the specifications within their tag template.

TagField
Contains the value and supporting information for a field within a [Tag][google.cloud.datacatalog.v1.Tag].

TagTemplate
A tag template defines a tag, which can have one or more typed fields. The template is used to create and attach the tag to GCP resources. Tag template roles provide permissions to create, edit, and use the template. See, for example, the TagTemplate User role, which includes permission to use the tag template to tag resources.

TagTemplateField
The template for an individual field within a tag template.

UpdateEntryGroupRequest
Request message for [UpdateEntryGroup][google.cloud.datacatalog.v1.DataCatalog.UpdateEntryGroup].

UpdateEntryRequest
Request message for [UpdateEntry][google.cloud.datacatalog.v1.DataCatalog.UpdateEntry].

UpdateTagRequest
Request message for [UpdateTag][google.cloud.datacatalog.v1.DataCatalog.UpdateTag].

UpdateTagTemplateFieldRequest
Request message for [UpdateTagTemplateField][google.cloud.datacatalog.v1.DataCatalog.UpdateTagTemplateField].

UpdateTagTemplateRequest
Request message for [UpdateTagTemplate][google.cloud.datacatalog.v1.DataCatalog.UpdateTagTemplate].

ViewSpec
Table view specification.

Enums

EntryType
Entry resources in Data Catalog can be of different types e.g. a BigQuery Table entry is of type TABLE. This enum describes all the possible types Data Catalog contains.

IntegratedSystem
This enum describes all the possible systems that Data Catalog integrates with.

SearchResultType
The different types of resources that can be returned in search.

TableSourceType
Table source type.