Struct gapi_grpc::google::cloud::datacatalog::v1::data_catalog_client::DataCatalogClient

pub struct DataCatalogClient<T> { /* fields omitted */ }

Data Catalog API service allows clients to discover, understand, and manage their data.

Implementations

impl DataCatalogClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> DataCatalogClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn search_catalog<'_>(
&'_ mut self,
request: impl IntoRequest<SearchCatalogRequest>
) -> Result<Response<SearchCatalogResponse>, Status>

Searches Data Catalog for multiple resources like entries, tags that match a query.

This is a custom method (https://cloud.google.com/apis/design/custom_methods) and does not return the complete resource, only the resource identifier and high level fields. Clients can subsequentally call Get methods.

Note that Data Catalog search queries do not guarantee full recall. Query results that match your query may not be returned, even in subsequent result pages. Also note that results returned (and not returned) can vary across repeated search queries.

See Data Catalog Search Syntax for more information.

pub async fn create_entry_group<'_>(
&'_ mut self,
request: impl IntoRequest<CreateEntryGroupRequest>
) -> Result<Response<EntryGroup>, Status>

Creates an EntryGroup.

An entry group contains logically related entries together with Cloud Identity and Access Management policies that specify the users who can create, edit, and view entries within the entry group.

Data Catalog automatically creates an entry group for BigQuery entries ("@bigquery") and Pub/Sub topics ("@pubsub"). Users create their own entry group to contain Cloud Storage fileset entries or custom type entries, and the IAM policies associated with those entries. Entry groups, like entries, can be searched.

A maximum of 10,000 entry groups may be created per organization across all locations.

Users should enable the Data Catalog API in the project identified by the parent parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn get_entry_group<'_>(
&'_ mut self,
request: impl IntoRequest<GetEntryGroupRequest>
) -> Result<Response<EntryGroup>, Status>

Gets an EntryGroup.

pub async fn update_entry_group<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateEntryGroupRequest>
) -> Result<Response<EntryGroup>, Status>

Updates an EntryGroup. The user should enable the Data Catalog API in the project identified by the entry_group.name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn delete_entry_group<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteEntryGroupRequest>
) -> Result<Response<()>, Status>

Deletes an EntryGroup. Only entry groups that do not contain entries can be deleted. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn list_entry_groups<'_>(
&'_ mut self,
request: impl IntoRequest<ListEntryGroupsRequest>
) -> Result<Response<ListEntryGroupsResponse>, Status>

Lists entry groups.

pub async fn create_entry<'_>(
&'_ mut self,
request: impl IntoRequest<CreateEntryRequest>
) -> Result<Response<Entry>, Status>

Creates an entry. Only entries of 'FILESET' type or user-specified type can be created.

Users should enable the Data Catalog API in the project identified by the parent parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

A maximum of 100,000 entries may be created per entry group.

pub async fn update_entry<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateEntryRequest>
) -> Result<Response<Entry>, Status>

Updates an existing entry. Users should enable the Data Catalog API in the project identified by the entry.name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn delete_entry<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteEntryRequest>
) -> Result<Response<()>, Status>

Deletes an existing entry. Only entries created through [CreateEntry][google.cloud.datacatalog.v1.DataCatalog.CreateEntry] method can be deleted. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn get_entry<'_>(
&'_ mut self,
request: impl IntoRequest<GetEntryRequest>
) -> Result<Response<Entry>, Status>

Gets an entry.

pub async fn lookup_entry<'_>(
&'_ mut self,
request: impl IntoRequest<LookupEntryRequest>
) -> Result<Response<Entry>, Status>

Get an entry by target resource name. This method allows clients to use the resource name from the source Google Cloud Platform service to get the Data Catalog Entry.

pub async fn list_entries<'_>(
&'_ mut self,
request: impl IntoRequest<ListEntriesRequest>
) -> Result<Response<ListEntriesResponse>, Status>

Lists entries.

pub async fn create_tag_template<'_>(
&'_ mut self,
request: impl IntoRequest<CreateTagTemplateRequest>
) -> Result<Response<TagTemplate>, Status>

Creates a tag template. The user should enable the Data Catalog API in the project identified by the parent parameter (see Data Catalog Resource Project for more information).

pub async fn get_tag_template<'_>(
&'_ mut self,
request: impl IntoRequest<GetTagTemplateRequest>
) -> Result<Response<TagTemplate>, Status>

Gets a tag template.

pub async fn update_tag_template<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateTagTemplateRequest>
) -> Result<Response<TagTemplate>, Status>

Updates a tag template. This method cannot be used to update the fields of a template. The tag template fields are represented as separate resources and should be updated using their own create/update/delete methods. Users should enable the Data Catalog API in the project identified by the tag_template.name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn delete_tag_template<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteTagTemplateRequest>
) -> Result<Response<()>, Status>

Deletes a tag template and all tags using the template. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn create_tag_template_field<'_>(
&'_ mut self,
request: impl IntoRequest<CreateTagTemplateFieldRequest>
) -> Result<Response<TagTemplateField>, Status>

Creates a field in a tag template. The user should enable the Data Catalog API in the project identified by the parent parameter (see Data Catalog Resource Project for more information).

pub async fn update_tag_template_field<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateTagTemplateFieldRequest>
) -> Result<Response<TagTemplateField>, Status>

Updates a field in a tag template. This method cannot be used to update the field type. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn rename_tag_template_field<'_>(
&'_ mut self,
request: impl IntoRequest<RenameTagTemplateFieldRequest>
) -> Result<Response<TagTemplateField>, Status>

Renames a field in a tag template. The user should enable the Data Catalog API in the project identified by the name parameter (see Data Catalog Resource Project for more information).

pub async fn delete_tag_template_field<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteTagTemplateFieldRequest>
) -> Result<Response<()>, Status>

Deletes a field in a tag template and all uses of that field. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

pub async fn create_tag<'_>(
&'_ mut self,
request: impl IntoRequest<CreateTagRequest>
) -> Result<Response<Tag>, Status>

Creates a tag on an [Entry][google.cloud.datacatalog.v1.Entry]. Note: The project identified by the parent parameter for the tag and the tag template used to create the tag must be from the same organization.

pub async fn update_tag<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateTagRequest>
) -> Result<Response<Tag>, Status>

Updates an existing tag.

pub async fn delete_tag<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteTagRequest>
) -> Result<Response<()>, Status>

Deletes a tag.

pub async fn list_tags<'_>(
&'_ mut self,
request: impl IntoRequest<ListTagsRequest>
) -> Result<Response<ListTagsResponse>, Status>

Lists the tags on an [Entry][google.cloud.datacatalog.v1.Entry].

pub async fn set_iam_policy<'_>(
&'_ mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Sets the access control policy for a resource. Replaces any existing policy. Supported resources are:

Tag templates.

Entries.

Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog.

Callers must have following Google IAM permission

datacatalog.tagTemplates.setIamPolicy to set policies on tag templates.

datacatalog.entries.setIamPolicy to set policies on entries.

datacatalog.entryGroups.setIamPolicy to set policies on entry groups.

pub async fn get_iam_policy<'_>(
&'_ mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the access control policy for a resource. A NOT_FOUND error is returned if the resource does not exist. An empty policy is returned if the resource exists but does not have a policy set on it.

Supported resources are:

Tag templates.

Entries.

Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog.

Callers must have following Google IAM permission

datacatalog.tagTemplates.getIamPolicy to get policies on tag templates.

datacatalog.entries.getIamPolicy to get policies on entries.

datacatalog.entryGroups.getIamPolicy to get policies on entry groups.

pub async fn test_iam_permissions<'_>(
&'_ mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Returns the caller's permissions on a resource. If the resource does not exist, an empty set of permissions is returned (We don't return a NOT_FOUND error).

Supported resources are:

Tag templates.

Entries.

Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog.

A caller is not required to have Google IAM permission to make this request.