Module gapi_grpc::google::cloud::dialogflow::cx::v3beta1

Modules

agents_client
Generated client implementations.

entity_type

entity_types_client
Generated client implementations.

environment

environments_client
Generated client implementations.

event_handler

experiment

experiments_client
Generated client implementations.

export_agent_response

flows_client
Generated client implementations.

form

fulfillment

intent

intents_client
Generated client implementations.

match

match_intent_response

nlu_settings

page_info

pages_client
Generated client implementations.

query_input

query_result

response_message

restore_agent_request

security_settings

security_settings_service_client
Generated client implementations.

session_entity_type

session_entity_types_client
Generated client implementations.

sessions_client
Generated client implementations.

streaming_detect_intent_response

streaming_recognition_result

transition_route

transition_route_groups_client
Generated client implementations.

variants_history

version

version_variants

versions_client
Generated client implementations.

webhook

webhook_request

webhook_response

webhooks_client
Generated client implementations.

Structs

Agent
Agents are best described as Natural Language Understanding (NLU) modules that transform user requests into actionable data. You can include agents in your app, product, or service to determine user intent and respond to the user in a natural way.

AudioInput
Represents the natural speech audio to be processed.

CreateAgentRequest
The request message for [Agents.CreateAgent][google.cloud.dialogflow.cx.v3beta1.Agents.CreateAgent].

CreateEntityTypeRequest
The request message for [EntityTypes.CreateEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.CreateEntityType].

CreateEnvironmentRequest
The request message for [Environments.CreateEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.CreateEnvironment].

CreateExperimentRequest
The request message for [Experiments.CreateExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.CreateExperiment].

CreateFlowRequest
The request message for [Flows.CreateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.CreateFlow].

CreateIntentRequest
The request message for [Intents.CreateIntent][google.cloud.dialogflow.cx.v3beta1.Intents.CreateIntent].

CreatePageRequest
The request message for [Pages.CreatePage][google.cloud.dialogflow.cx.v3beta1.Pages.CreatePage].

CreateSecuritySettingsRequest
The request message for [SecuritySettings.CreateSecuritySettings][].

CreateSessionEntityTypeRequest
The request message for [SessionEntityTypes.CreateSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.CreateSessionEntityType].

CreateTransitionRouteGroupRequest
The request message for [TransitionRouteGroups.CreateTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.CreateTransitionRouteGroup].

CreateVersionOperationMetadata
Metadata associated with the long running operation for [Versions.CreateVersion][google.cloud.dialogflow.cx.v3beta1.Versions.CreateVersion].

CreateVersionRequest
The request message for [Versions.CreateVersion][google.cloud.dialogflow.cx.v3beta1.Versions.CreateVersion].

CreateWebhookRequest
The request message for [Webhooks.CreateWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.CreateWebhook].

DeleteAgentRequest
The request message for [Agents.DeleteAgent][google.cloud.dialogflow.cx.v3beta1.Agents.DeleteAgent].

DeleteEntityTypeRequest
The request message for [EntityTypes.DeleteEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.DeleteEntityType].

DeleteEnvironmentRequest
The request message for [Environments.DeleteEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.DeleteEnvironment].

DeleteExperimentRequest
The request message for [Experiments.DeleteExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.DeleteExperiment].

DeleteFlowRequest
The request message for [Flows.DeleteFlow][google.cloud.dialogflow.cx.v3beta1.Flows.DeleteFlow].

DeleteIntentRequest
The request message for [Intents.DeleteIntent][google.cloud.dialogflow.cx.v3beta1.Intents.DeleteIntent].

DeletePageRequest
The request message for [Pages.DeletePage][google.cloud.dialogflow.cx.v3beta1.Pages.DeletePage].

DeleteSecuritySettingsRequest
The request message for [SecuritySettings.DeleteSecuritySettings][].

DeleteSessionEntityTypeRequest
The request message for [SessionEntityTypes.DeleteSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.DeleteSessionEntityType].

DeleteTransitionRouteGroupRequest
The request message for [TransitionRouteGroups.DeleteTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.DeleteTransitionRouteGroup].

DeleteVersionRequest
The request message for [Versions.DeleteVersion][google.cloud.dialogflow.cx.v3beta1.Versions.DeleteVersion].

DeleteWebhookRequest
The request message for [Webhooks.DeleteWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.DeleteWebhook].

DetectIntentRequest
The request to detect user's intent.

DetectIntentResponse
The message returned from the DetectIntent method.

DtmfInput
Represents the input for dtmf event.

EntityType
Entities are extracted from user input and represent parameters that are meaningful to your application. For example, a date range, a proper name such as a geographic location or landmark, and so on. Entities represent actionable data for your application.

Environment
Represents an environment for an agent. You can create multiple versions of your agent and publish them to separate environments. When you edit an agent, you are editing the draft agent. At any point, you can save the draft agent as an agent version, which is an immutable snapshot of your agent. When you save the draft agent, it is published to the default environment. When you create agent versions, you can publish them to custom environments. You can create a variety of custom environments for testing, development, production, etc.

EventHandler
An event handler specifies an [event][google.cloud.dialogflow.cx.v3beta1.EventHandler.event] that can be handled during a session. When the specified event happens, the following actions are taken in order:

EventInput
Represents the event to trigger.

Experiment
Represents an experiment in an environment. Next ID: 13

ExportAgentRequest
The request message for [Agents.ExportAgent][google.cloud.dialogflow.cx.v3beta1.Agents.ExportAgent].

ExportAgentResponse
The response message for [Agents.ExportAgent][google.cloud.dialogflow.cx.v3beta1.Agents.ExportAgent].

Flow
Flows represents the conversation flows when you build your chatbot agent.

Form
A form is a data model that groups related parameters that can be collected from the user. The process in which the agent prompts the user and collects parameter values from the user is called form filling. A form can be added to a [page][google.cloud.dialogflow.cx.v3beta1.Page]. When form filling is done, the filled parameters will be written to the [session][google.cloud.dialogflow.cx.v3beta1.SessionInfo.parameters].

FulfillIntentRequest
Request of [FulfillIntent][]

FulfillIntentResponse
Response of [FulfillIntent][]

Fulfillment
A fulfillment can do one or more of the following actions at the same time:

GetAgentRequest
The request message for [Agents.GetAgent][google.cloud.dialogflow.cx.v3beta1.Agents.GetAgent].

GetEntityTypeRequest
The request message for [EntityTypes.GetEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.GetEntityType].

GetEnvironmentRequest
The request message for [Environments.GetEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.GetEnvironment].

GetExperimentRequest
The request message for [Experiments.GetExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.GetExperiment].

GetFlowRequest
The response message for [Flows.GetFlow][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlow].

GetIntentRequest
The request message for [Intents.GetIntent][google.cloud.dialogflow.cx.v3beta1.Intents.GetIntent].

GetPageRequest
The request message for [Pages.GetPage][google.cloud.dialogflow.cx.v3beta1.Pages.GetPage].

GetSecuritySettingsRequest
The request message for [SecuritySettingsService.GetSecuritySettings][google.cloud.dialogflow.cx.v3beta1.SecuritySettingsService.GetSecuritySettings].

GetSessionEntityTypeRequest
The request message for [SessionEntityTypes.GetSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.GetSessionEntityType].

GetTransitionRouteGroupRequest
The request message for [TransitionRouteGroups.GetTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.GetTransitionRouteGroup].

GetVersionRequest
The request message for [Versions.GetVersion][google.cloud.dialogflow.cx.v3beta1.Versions.GetVersion].

GetWebhookRequest
The request message for [Webhooks.GetWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.GetWebhook].

InputAudioConfig
Instructs the speech recognizer on how to process the audio content.

Intent
An intent represents a user's intent to interact with a conversational agent.

IntentInput
Represents the intent to trigger programmatically rather than as a result of natural language processing.

ListAgentsRequest
The request message for [Agents.ListAgents][google.cloud.dialogflow.cx.v3beta1.Agents.ListAgents].

ListAgentsResponse
The response message for [Agents.ListAgents][google.cloud.dialogflow.cx.v3beta1.Agents.ListAgents].

ListEntityTypesRequest
The request message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ListEntityTypes].

ListEntityTypesResponse
The response message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ListEntityTypes].

ListEnvironmentsRequest
The request message for [Environments.ListEnvironments][google.cloud.dialogflow.cx.v3beta1.Environments.ListEnvironments].

ListEnvironmentsResponse
The response message for [Environments.ListEnvironments][google.cloud.dialogflow.cx.v3beta1.Environments.ListEnvironments].

ListExperimentsRequest
The request message for [Experiments.ListExperiments][google.cloud.dialogflow.cx.v3beta1.Experiments.ListExperiments].

ListExperimentsResponse
The response message for [Experiments.ListExperiments][google.cloud.dialogflow.cx.v3beta1.Experiments.ListExperiments].

ListFlowsRequest
The request message for [Flows.ListFlows][google.cloud.dialogflow.cx.v3beta1.Flows.ListFlows].

ListFlowsResponse
The response message for [Flows.ListFlows][google.cloud.dialogflow.cx.v3beta1.Flows.ListFlows].

ListIntentsRequest
The request message for [Intents.ListIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ListIntents].

ListIntentsResponse
The response message for [Intents.ListIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ListIntents].

ListPagesRequest
The request message for [Pages.ListPages][google.cloud.dialogflow.cx.v3beta1.Pages.ListPages].

ListPagesResponse
The response message for [Pages.ListPages][google.cloud.dialogflow.cx.v3beta1.Pages.ListPages].

ListSecuritySettingsRequest
The request message for [SecuritySettings.ListSecuritySettings][].

ListSecuritySettingsResponse
The response message for [SecuritySettings.ListSecuritySettings][].

ListSessionEntityTypesRequest
The request message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.ListSessionEntityTypes].

ListSessionEntityTypesResponse
The response message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.ListSessionEntityTypes].

ListTransitionRouteGroupsRequest
The request message for [TransitionRouteGroups.ListTransitionRouteGroups][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.ListTransitionRouteGroups].

ListTransitionRouteGroupsResponse
The response message for [TransitionRouteGroups.ListTransitionRouteGroups][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.ListTransitionRouteGroups].

ListVersionsRequest
The request message for [Versions.ListVersions][google.cloud.dialogflow.cx.v3beta1.Versions.ListVersions].

ListVersionsResponse
The response message for [Versions.ListVersions][google.cloud.dialogflow.cx.v3beta1.Versions.ListVersions].

ListWebhooksRequest
The request message for [Webhooks.ListWebhooks][google.cloud.dialogflow.cx.v3beta1.Webhooks.ListWebhooks].

ListWebhooksResponse
The response message for [Webhooks.ListWebhooks][google.cloud.dialogflow.cx.v3beta1.Webhooks.ListWebhooks].

LoadVersionRequest
The request message for [Versions.LoadVersion][google.cloud.dialogflow.cx.v3beta1.Versions.LoadVersion].

LookupEnvironmentHistoryRequest
The request message for [Environments.LookupEnvironmentHistory][google.cloud.dialogflow.cx.v3beta1.Environments.LookupEnvironmentHistory].

LookupEnvironmentHistoryResponse
The response message for [Environments.LookupEnvironmentHistory][google.cloud.dialogflow.cx.v3beta1.Environments.LookupEnvironmentHistory].

Match
Represents one match result of [MatchIntent][].

MatchIntentRequest
Request of [MatchIntent][].

MatchIntentResponse
Response of [MatchIntent][].

NluSettings
Settings related to NLU.

OutputAudioConfig
Instructs the speech synthesizer how to generate the output audio content.

Page
A Dialogflow CX conversation (session) can be described and visualized as a state machine. The states of a CX session are represented by pages.

PageInfo
Represents page information communicated to and from the webhook.

QueryInput
Represents the query input. It can contain one of:

QueryParameters
Represents the parameters of a conversational query.

QueryResult
Represents the result of a conversational query.

ResponseMessage
Represents a response message that can be returned by a conversational agent.

RestoreAgentRequest
The request message for [Agents.RestoreAgent][google.cloud.dialogflow.cx.v3beta1.Agents.RestoreAgent].

SecuritySettings
Represents the settings related to security issues, such as data redaction and data retention. It may take hours for updates on the settings to propagate to all the related components and take effect.

SentimentAnalysisResult
The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user's attitude as positive, negative, or neutral.

SessionEntityType
Session entity types are referred to as User entity types and are entities that are built for an individual user such as favorites, preferences, playlists, and so on.

SessionInfo
Represents session information communicated to and from the webhook.

SpeechToTextSettings
Settings related to speech recognition.

SpeechWordInfo
Information for a word recognized by the speech recognizer.

StartExperimentRequest
The request message for [Experiments.StartExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.StartExperiment].

StopExperimentRequest
The request message for [Experiments.StopExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.StopExperiment].

StreamingDetectIntentRequest
The top-level message sent by the client to the [Sessions.StreamingDetectIntent][google.cloud.dialogflow.cx.v3beta1.Sessions.StreamingDetectIntent] method.

StreamingDetectIntentResponse
The top-level message returned from the StreamingDetectIntent method.

StreamingRecognitionResult
Contains a speech recognition result corresponding to a portion of the audio that is currently being processed or an indication that this is the end of the single requested utterance.

SynthesizeSpeechConfig
Configuration of how speech should be synthesized.

TextInput
Represents the natural language text to be processed.

TrainFlowRequest
The request message for [Flows.TrainFlow][google.cloud.dialogflow.cx.v3beta1.Flows.TrainFlow].

TransitionRoute
A transition route specifies a [intent][google.cloud.dialogflow.cx.v3beta1.Intent] that can be matched and/or a data condition that can be evaluated during a session. When a specified transition is matched, the following actions are taken in order:

TransitionRouteGroup
An TransitionRouteGroup represents a group of [TransitionRoutes][google.cloud.dialogflow.cx.v3beta1.TransitionRoute] to be used by a [Page][google.cloud.dialogflow.cx.v3beta1.Page].

UpdateAgentRequest
The request message for [Agents.UpdateAgent][google.cloud.dialogflow.cx.v3beta1.Agents.UpdateAgent].

UpdateEntityTypeRequest
The request message for [EntityTypes.UpdateEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.UpdateEntityType].

UpdateEnvironmentRequest
The request message for [Environments.UpdateEnvironment][google.cloud.dialogflow.cx.v3beta1.Environments.UpdateEnvironment].

UpdateExperimentRequest
The request message for [Experiments.UpdateExperiment][google.cloud.dialogflow.cx.v3beta1.Experiments.UpdateExperiment].

UpdateFlowRequest
The request message for [Flows.UpdateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.UpdateFlow].

UpdateIntentRequest
The request message for [Intents.UpdateIntent][google.cloud.dialogflow.cx.v3beta1.Intents.UpdateIntent].

UpdatePageRequest
The request message for [Pages.UpdatePage][google.cloud.dialogflow.cx.v3beta1.Pages.UpdatePage].

UpdateSecuritySettingsRequest
The request message for [SecuritySettingsService.UpdateSecuritySettings][google.cloud.dialogflow.cx.v3beta1.SecuritySettingsService.UpdateSecuritySettings].

UpdateSessionEntityTypeRequest
The request message for [SessionEntityTypes.UpdateSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.UpdateSessionEntityType].

UpdateTransitionRouteGroupRequest
The request message for [TransitionRouteGroups.UpdateTransitionRouteGroup][google.cloud.dialogflow.cx.v3beta1.TransitionRouteGroups.UpdateTransitionRouteGroup].

UpdateVersionRequest
The request message for [Versions.UpdateVersion][google.cloud.dialogflow.cx.v3beta1.Versions.UpdateVersion].

UpdateWebhookRequest
The request message for [Webhooks.UpdateWebhook][google.cloud.dialogflow.cx.v3beta1.Webhooks.UpdateWebhook].

VariantsHistory
The history of variants update.

Version
Represents a version of a flow.

VersionVariants
A list of flow version variants.

VoiceSelectionParams
Description of which voice to use for speech synthesis.

Webhook
Webhooks host the developer's business logic. During a session, webhooks allow the developer to use the data extracted by Dialogflow's natural language processing to generate dynamic responses, validate collected data, or trigger actions on the backend.

WebhookRequest
The request message for a webhook call.

WebhookResponse
The response message for a webhook call.

Enums

AudioEncoding
Audio encoding of the audio content sent in the conversational query request. Refer to the Cloud Speech API documentation for more details.

IntentView
Represents the options for views of an intent. An intent can be a sizable object. Therefore, we provide a resource view that does not return training phrases in the response.

OutputAudioEncoding
Audio encoding of the output audio format in Text-To-Speech.

SpeechModelVariant
Variant of the specified [Speech model][google.cloud.dialogflow.cx.v3beta1.InputAudioConfig.model] to use.

SsmlVoiceGender
Gender of the voice as described in SSML voice element.