Module gapi_grpc::google::cloud::dialogflow::cx::v3beta1::webhook_response

Modules

fulfillment_response

Structs

FulfillmentResponse
Represents a fulfillment response to the user.

Enums

Transition
The target to transition to. This can be set optionally to indicate an immediate transition to a different page in the same host flow, or a different flow in the same agent.