Module gapi_grpc::google::cloud::dialogflow::v2::intent

Modules

message

training_phrase

Structs

FollowupIntentInfo
Represents a single followup intent in the chain.

Message
A rich response message. Corresponds to the intent Response field in the Dialogflow console. For more information, see Rich response messages.

Parameter
Represents intent parameters.

TrainingPhrase
Represents an example that the agent is trained on.

Enums

WebhookState
Represents the different states that webhooks can be in.