Module gapi_grpc::google::cloud::dialogflow::v2beta1

Modules

agent

agents_client
Generated client implementations.

batch_update_entity_types_request

batch_update_intents_request

contexts_client
Generated client implementations.

document

documents_client
Generated client implementations.

entity_type

entity_types_client
Generated client implementations.

environment

environments_client
Generated client implementations.

export_agent_response

import_agent_request

intent

intents_client
Generated client implementations.

knowledge_answers

knowledge_bases_client
Generated client implementations.

knowledge_operation_metadata

query_input

reload_document_request

restore_agent_request

session_entity_type

session_entity_types_client
Generated client implementations.

sessions_client
Generated client implementations.

streaming_recognition_result

validation_error

Structs

Agent
A Dialogflow agent is a virtual agent that handles conversations with your end-users. It is a natural language understanding module that understands the nuances of human language. Dialogflow translates end-user text or audio during a conversation to structured data that your apps and services can understand. You design and build a Dialogflow agent to handle the types of conversations required for your system.

AutoApproveSmartMessagingEntriesResponse
Response message for [Documents.AutoApproveSmartMessagingEntries].

BatchCreateEntitiesRequest
The request message for [EntityTypes.BatchCreateEntities][google.cloud.dialogflow.v2beta1.EntityTypes.BatchCreateEntities].

BatchDeleteEntitiesRequest
The request message for [EntityTypes.BatchDeleteEntities][google.cloud.dialogflow.v2beta1.EntityTypes.BatchDeleteEntities].

BatchDeleteEntityTypesRequest
The request message for [EntityTypes.BatchDeleteEntityTypes][google.cloud.dialogflow.v2beta1.EntityTypes.BatchDeleteEntityTypes].

BatchDeleteIntentsRequest
The request message for [Intents.BatchDeleteIntents][google.cloud.dialogflow.v2beta1.Intents.BatchDeleteIntents].

BatchUpdateEntitiesRequest
The request message for [EntityTypes.BatchUpdateEntities][google.cloud.dialogflow.v2beta1.EntityTypes.BatchUpdateEntities].

BatchUpdateEntityTypesRequest
The request message for [EntityTypes.BatchUpdateEntityTypes][google.cloud.dialogflow.v2beta1.EntityTypes.BatchUpdateEntityTypes].

BatchUpdateEntityTypesResponse
The response message for [EntityTypes.BatchUpdateEntityTypes][google.cloud.dialogflow.v2beta1.EntityTypes.BatchUpdateEntityTypes].

BatchUpdateIntentsRequest
The request message for [Intents.BatchUpdateIntents][google.cloud.dialogflow.v2beta1.Intents.BatchUpdateIntents].

BatchUpdateIntentsResponse
The response message for [Intents.BatchUpdateIntents][google.cloud.dialogflow.v2beta1.Intents.BatchUpdateIntents].

Context
Dialogflow contexts are similar to natural language context. If a person says to you "they are orange", you need context in order to understand what "they" is referring to. Similarly, for Dialogflow to handle an end-user expression like that, it needs to be provided with context in order to correctly match an intent.

CreateContextRequest
The request message for [Contexts.CreateContext][google.cloud.dialogflow.v2beta1.Contexts.CreateContext].

CreateDocumentRequest
Request message for [Documents.CreateDocument][google.cloud.dialogflow.v2beta1.Documents.CreateDocument].

CreateEntityTypeRequest
The request message for [EntityTypes.CreateEntityType][google.cloud.dialogflow.v2beta1.EntityTypes.CreateEntityType].

CreateIntentRequest
The request message for [Intents.CreateIntent][google.cloud.dialogflow.v2beta1.Intents.CreateIntent].

CreateKnowledgeBaseRequest
Request message for [KnowledgeBases.CreateKnowledgeBase][google.cloud.dialogflow.v2beta1.KnowledgeBases.CreateKnowledgeBase].

CreateSessionEntityTypeRequest
The request message for [SessionEntityTypes.CreateSessionEntityType][google.cloud.dialogflow.v2beta1.SessionEntityTypes.CreateSessionEntityType].

DeleteAgentRequest
The request message for [Agents.DeleteAgent][google.cloud.dialogflow.v2beta1.Agents.DeleteAgent].

DeleteAllContextsRequest
The request message for [Contexts.DeleteAllContexts][google.cloud.dialogflow.v2beta1.Contexts.DeleteAllContexts].

DeleteContextRequest
The request message for [Contexts.DeleteContext][google.cloud.dialogflow.v2beta1.Contexts.DeleteContext].

DeleteDocumentRequest
Request message for [Documents.DeleteDocument][google.cloud.dialogflow.v2beta1.Documents.DeleteDocument].

DeleteEntityTypeRequest
The request message for [EntityTypes.DeleteEntityType][google.cloud.dialogflow.v2beta1.EntityTypes.DeleteEntityType].

DeleteIntentRequest
The request message for [Intents.DeleteIntent][google.cloud.dialogflow.v2beta1.Intents.DeleteIntent].

DeleteKnowledgeBaseRequest
Request message for [KnowledgeBases.DeleteKnowledgeBase][google.cloud.dialogflow.v2beta1.KnowledgeBases.DeleteKnowledgeBase].

DeleteSessionEntityTypeRequest
The request message for [SessionEntityTypes.DeleteSessionEntityType][google.cloud.dialogflow.v2beta1.SessionEntityTypes.DeleteSessionEntityType].

DetectIntentRequest
The request to detect user's intent.

DetectIntentResponse
The message returned from the DetectIntent method.

Document
A knowledge document to be used by a [KnowledgeBase][google.cloud.dialogflow.v2beta1.KnowledgeBase].

EntityType
Each intent parameter has a type, called the entity type, which dictates exactly how data from an end-user expression is extracted.

EntityTypeBatch
This message is a wrapper around a collection of entity types.

Environment
You can create multiple versions of your agent and publish them to separate environments.

EventInput
Events allow for matching intents by event name instead of the natural language input. For instance, input <event: { name: "welcome_event", parameters: { name: "Sam" } }> can trigger a personalized welcome response. The parameter name may be used by the agent in the response: "Hello #welcome_event.name! What can I do for you today?".

ExportAgentRequest
The request message for [Agents.ExportAgent][google.cloud.dialogflow.v2beta1.Agents.ExportAgent].

ExportAgentResponse
The response message for [Agents.ExportAgent][google.cloud.dialogflow.v2beta1.Agents.ExportAgent].

GcsSource
Google Cloud Storage location for single input.

GetAgentRequest
The request message for [Agents.GetAgent][google.cloud.dialogflow.v2beta1.Agents.GetAgent].

GetContextRequest
The request message for [Contexts.GetContext][google.cloud.dialogflow.v2beta1.Contexts.GetContext].

GetDocumentRequest
Request message for [Documents.GetDocument][google.cloud.dialogflow.v2beta1.Documents.GetDocument].

GetEntityTypeRequest
The request message for [EntityTypes.GetEntityType][google.cloud.dialogflow.v2beta1.EntityTypes.GetEntityType].

GetIntentRequest
The request message for [Intents.GetIntent][google.cloud.dialogflow.v2beta1.Intents.GetIntent].

GetKnowledgeBaseRequest
Request message for [KnowledgeBases.GetKnowledgeBase][google.cloud.dialogflow.v2beta1.KnowledgeBases.GetKnowledgeBase].

GetSessionEntityTypeRequest
The request message for [SessionEntityTypes.GetSessionEntityType][google.cloud.dialogflow.v2beta1.SessionEntityTypes.GetSessionEntityType].

GetValidationResultRequest
The request message for [Agents.GetValidationResult][google.cloud.dialogflow.v2beta1.Agents.GetValidationResult].

ImportAgentRequest
The request message for [Agents.ImportAgent][google.cloud.dialogflow.v2beta1.Agents.ImportAgent].

InputAudioConfig
Instructs the speech recognizer on how to process the audio content.

Intent
An intent categorizes an end-user's intention for one conversation turn. For each agent, you define many intents, where your combined intents can handle a complete conversation. When an end-user writes or says something, referred to as an end-user expression or end-user input, Dialogflow matches the end-user input to the best intent in your agent. Matching an intent is also known as intent classification.

IntentBatch
This message is a wrapper around a collection of intents.

KnowledgeAnswers
Represents the result of querying a Knowledge base.

KnowledgeBase
A knowledge base represents a collection of knowledge documents that you provide to Dialogflow. Your knowledge documents contain information that may be useful during conversations with end-users. Some Dialogflow features use knowledge bases when looking for a response to an end-user input.

KnowledgeOperationMetadata
Metadata in google::longrunning::Operation for Knowledge operations.

ListContextsRequest
The request message for [Contexts.ListContexts][google.cloud.dialogflow.v2beta1.Contexts.ListContexts].

ListContextsResponse
The response message for [Contexts.ListContexts][google.cloud.dialogflow.v2beta1.Contexts.ListContexts].

ListDocumentsRequest
Request message for [Documents.ListDocuments][google.cloud.dialogflow.v2beta1.Documents.ListDocuments].

ListDocumentsResponse
Response message for [Documents.ListDocuments][google.cloud.dialogflow.v2beta1.Documents.ListDocuments].

ListEntityTypesRequest
The request message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.v2beta1.EntityTypes.ListEntityTypes].

ListEntityTypesResponse
The response message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.v2beta1.EntityTypes.ListEntityTypes].

ListEnvironmentsRequest
The request message for [Environments.ListEnvironments][google.cloud.dialogflow.v2beta1.Environments.ListEnvironments].

ListEnvironmentsResponse
The response message for [Environments.ListEnvironments][google.cloud.dialogflow.v2beta1.Environments.ListEnvironments].

ListIntentsRequest
The request message for [Intents.ListIntents][google.cloud.dialogflow.v2beta1.Intents.ListIntents].

ListIntentsResponse
The response message for [Intents.ListIntents][google.cloud.dialogflow.v2beta1.Intents.ListIntents].

ListKnowledgeBasesRequest
Request message for [KnowledgeBases.ListKnowledgeBases][google.cloud.dialogflow.v2beta1.KnowledgeBases.ListKnowledgeBases].

ListKnowledgeBasesResponse
Response message for [KnowledgeBases.ListKnowledgeBases][google.cloud.dialogflow.v2beta1.KnowledgeBases.ListKnowledgeBases].

ListSessionEntityTypesRequest
The request message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.v2beta1.SessionEntityTypes.ListSessionEntityTypes].

ListSessionEntityTypesResponse
The response message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.v2beta1.SessionEntityTypes.ListSessionEntityTypes].

OriginalDetectIntentRequest
Represents the contents of the original request that was passed to the [Streaming]DetectIntent call.

OutputAudioConfig
Instructs the speech synthesizer how to generate the output audio content. If this audio config is supplied in a request, it overrides all existing text-to-speech settings applied to the agent.

QueryInput
Represents the query input. It can contain either:

QueryParameters
Represents the parameters of the conversational query.

QueryResult
Represents the result of conversational query or event processing.

ReloadDocumentRequest
Request message for [Documents.ReloadDocument][google.cloud.dialogflow.v2beta1.Documents.ReloadDocument].

RestoreAgentRequest
The request message for [Agents.RestoreAgent][google.cloud.dialogflow.v2beta1.Agents.RestoreAgent].

SearchAgentsRequest
The request message for [Agents.SearchAgents][google.cloud.dialogflow.v2beta1.Agents.SearchAgents].

SearchAgentsResponse
The response message for [Agents.SearchAgents][google.cloud.dialogflow.v2beta1.Agents.SearchAgents].

Sentiment
The sentiment, such as positive/negative feeling or association, for a unit of analysis, such as the query text.

SentimentAnalysisRequestConfig
Configures the types of sentiment analysis to perform.

SentimentAnalysisResult
The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user's attitude as positive, negative, or neutral. For [Participants.AnalyzeContent][google.cloud.dialogflow.v2beta1.Participants.AnalyzeContent], it needs to be configured in [DetectIntentRequest.query_params][google.cloud.dialogflow.v2beta1.DetectIntentRequest.query_params]. For [Participants.StreamingAnalyzeContent][google.cloud.dialogflow.v2beta1.Participants.StreamingAnalyzeContent], it needs to be configured in [StreamingDetectIntentRequest.query_params][google.cloud.dialogflow.v2beta1.StreamingDetectIntentRequest.query_params]. And for [Participants.AnalyzeContent][google.cloud.dialogflow.v2beta1.Participants.AnalyzeContent] and [Participants.StreamingAnalyzeContent][google.cloud.dialogflow.v2beta1.Participants.StreamingAnalyzeContent], it needs to be configured in [ConversationProfile.human_agent_assistant_config][google.cloud.dialogflow.v2beta1.ConversationProfile.human_agent_assistant_config]

SessionEntityType
A session represents a conversation between a Dialogflow agent and an end-user. You can create special entities, called session entities, during a session. Session entities can extend or replace custom entity types and only exist during the session that they were created for. All session data, including session entities, is stored by Dialogflow for 20 minutes.

SetAgentRequest
The request message for [Agents.SetAgent][google.cloud.dialogflow.v2beta1.Agents.SetAgent].

SpeechContext
Hints for the speech recognizer to help with recognition in a specific conversation state.

SpeechWordInfo
Information for a word recognized by the speech recognizer.

StreamingDetectIntentRequest
The top-level message sent by the client to the [Sessions.StreamingDetectIntent][google.cloud.dialogflow.v2beta1.Sessions.StreamingDetectIntent] method.

StreamingDetectIntentResponse
The top-level message returned from the StreamingDetectIntent method.

StreamingRecognitionResult
Contains a speech recognition result corresponding to a portion of the audio that is currently being processed or an indication that this is the end of the single requested utterance.

SubAgent
Contains basic configuration for a sub-agent.

SynthesizeSpeechConfig
Configuration of how speech should be synthesized.

TextInput
Represents the natural language text to be processed.

TrainAgentRequest
The request message for [Agents.TrainAgent][google.cloud.dialogflow.v2beta1.Agents.TrainAgent].

UpdateContextRequest
The request message for [Contexts.UpdateContext][google.cloud.dialogflow.v2beta1.Contexts.UpdateContext].

UpdateDocumentRequest
Request message for [Documents.UpdateDocument][google.cloud.dialogflow.v2beta1.Documents.UpdateDocument].

UpdateEntityTypeRequest
The request message for [EntityTypes.UpdateEntityType][google.cloud.dialogflow.v2beta1.EntityTypes.UpdateEntityType].

UpdateIntentRequest
The request message for [Intents.UpdateIntent][google.cloud.dialogflow.v2beta1.Intents.UpdateIntent].

UpdateKnowledgeBaseRequest
Request message for [KnowledgeBases.UpdateKnowledgeBase][google.cloud.dialogflow.v2beta1.KnowledgeBases.UpdateKnowledgeBase].

UpdateSessionEntityTypeRequest
The request message for [SessionEntityTypes.UpdateSessionEntityType][google.cloud.dialogflow.v2beta1.SessionEntityTypes.UpdateSessionEntityType].

ValidationError
Represents a single validation error.

ValidationResult
Represents the output of agent validation.

VoiceSelectionParams
Description of which voice to use for speech synthesis.

WebhookRequest
The request message for a webhook call.

WebhookResponse
The response message for a webhook call.

Enums

AudioEncoding
Audio encoding of the audio content sent in the conversational query request. Refer to the Cloud Speech API documentation for more details.

IntentView
Represents the options for views of an intent. An intent can be a sizable object. Therefore, we provide a resource view that does not return training phrases in the response by default.

OutputAudioEncoding
Audio encoding of the output audio format in Text-To-Speech.

SpeechModelVariant
Variant of the specified [Speech model][google.cloud.dialogflow.v2beta1.InputAudioConfig.model] to use.

SsmlVoiceGender
Gender of the voice as described in SSML voice element.