Module gapi_grpc::google::cloud::dialogflow::v2beta1::entity_type

Structs

Entity
An entity entry for an associated entity type.

Enums

AutoExpansionMode
Represents different entity type expansion modes. Automated expansion allows an agent to recognize values that have not been explicitly listed in the entity (for example, new kinds of shopping list items).

Kind
Represents kinds of entities.