Module gapi_grpc::google::cloud::dialogflow::v2beta1::intent::message

Modules

basic_card

browse_carousel_card

card

carousel_select

column_properties

list_select

media_content

rbm_card_content

rbm_carousel_card

rbm_standalone_card

rbm_suggested_action

rbm_suggestion

telephony_synthesize_speech

Structs

BasicCard
The basic card message. Useful for displaying information.

BrowseCarouselCard
Browse Carousel Card for Actions on Google. https://developers.google.com/actions/assistant/responses#browsing_carousel

Card
The card response message.

CarouselSelect
The card for presenting a carousel of options to select from.

ColumnProperties
Column properties for [TableCard][google.cloud.dialogflow.v2beta1.Intent.Message.TableCard].

Image
The image response message.

LinkOutSuggestion
The suggestion chip message that allows the user to jump out to the app or website associated with this agent.

ListSelect
The card for presenting a list of options to select from.

MediaContent
The media content card for Actions on Google.

QuickReplies
The quick replies response message.

RbmCardContent
Rich Business Messaging (RBM) Card content

RbmCarouselCard
Carousel Rich Business Messaging (RBM) rich card.

RbmStandaloneCard
Standalone Rich Business Messaging (RBM) rich card.

RbmSuggestedAction
Rich Business Messaging (RBM) suggested client-side action that the user can choose from the card.

RbmSuggestedReply
Rich Business Messaging (RBM) suggested reply that the user can click instead of typing in their own response.

RbmSuggestion
Rich Business Messaging (RBM) suggestion. Suggestions allow user to easily select/click a predefined response or perform an action (like opening a web uri).

RbmText
Rich Business Messaging (RBM) text response with suggestions.

SelectItemInfo
Additional info about the select item for when it is triggered in a dialog.

SimpleResponse
The simple response message containing speech or text.

SimpleResponses
The collection of simple response candidates. This message in QueryResult.fulfillment_messages and WebhookResponse.fulfillment_messages should contain only one SimpleResponse.

Suggestion
The suggestion chip message that the user can tap to quickly post a reply to the conversation.

Suggestions
The collection of suggestions.

TableCard
Table card for Actions on Google.

TableCardCell
Cell of [TableCardRow][google.cloud.dialogflow.v2beta1.Intent.Message.TableCardRow].

TableCardRow
Row of [TableCard][google.cloud.dialogflow.v2beta1.Intent.Message.TableCard].

TelephonyPlayAudio
Plays audio from a file in Telephony Gateway.

TelephonySynthesizeSpeech
Synthesizes speech and plays back the synthesized audio to the caller in Telephony Gateway.

TelephonyTransferCall
Transfers the call in Telephony Gateway.

Text
The text response message.

Enums

Message
Required. The rich response message.

Platform
Represents different platforms that a rich message can be intended for.