Module gapi_grpc::google::cloud::documentai::v1::document

Modules

entity

page

page_anchor

provenance

revision

style

text_anchor

Structs

Entity
A phrase in the text that is a known entity type, such as a person, an organization, or location.

EntityRelation
Relationship between [Entities][google.cloud.documentai.v1.Document.Entity].

Page
A page in a [Document][google.cloud.documentai.v1.Document].

PageAnchor
Referencing the visual context of the entity in the [Document.pages][google.cloud.documentai.v1.Document.pages]. Page anchors can be cross-page, consist of multiple bounding polygons and optionally reference specific layout element types.

Provenance
Structure to identify provenance relationships between annotations in different revisions.

Revision
Contains past or forward revisions of this document.

ShardInfo
For a large document, sharding may be performed to produce several document shards. Each document shard contains this field to detail which shard it is.

Style
Annotation for common text style attributes. This adheres to CSS conventions as much as possible.

TextAnchor
Text reference indexing into the [Document.text][google.cloud.documentai.v1.Document.text].

TextChange
This message is used for text changes aka. OCR corrections.

Enums

Source
Original source document from the user.