Module gapi_grpc::google::cloud::documentai::v1beta2::document::page

Modules

layout

table

token

Structs

Block
A block has a set of lines (collected into paragraphs) that have a common line-spacing and orientation.

DetectedLanguage
Detected language for a structural component.

Dimension
Dimension for the page.

FormField
A form field detected on the page.

Layout
Visual element describing a layout unit on a page.

Line
A collection of tokens that a human would perceive as a line. Does not cross column boundaries, can be horizontal, vertical, etc.

Paragraph
A collection of lines that a human would perceive as a paragraph.

Table
A table representation similar to HTML table structure.

Token
A detected token.

VisualElement
Detected non-text visual elements e.g. checkbox, signature etc. on the page.