Module gapi_grpc::google::cloud::documentai::v1beta3

Modules

batch_process_metadata

batch_process_request

document

document_processor_service_client
Generated client implementations.

review_document_operation_metadata

Structs

BatchProcessMetadata
The long running operation metadata for batch process method.

BatchProcessRequest
Request message for batch process document method.

BatchProcessResponse
Response message for batch process document method.

BoundingPoly
A bounding polygon for the detected image annotation.

Document
Document represents the canonical document resource in Document Understanding AI. It is an interchange format that provides insights into documents and allows for collaboration between users and Document Understanding AI to iterate and optimize for quality.

NormalizedVertex
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ProcessRequest
Request message for the process document method.

ProcessResponse
Response message for the process document method.

ReviewDocumentOperationMetadata
The long running operation metadata for review document method.

ReviewDocumentRequest
Request message for review document method.

ReviewDocumentResponse
Response message for review document method.

Vertex
A vertex represents a 2D point in the image. NOTE: the vertex coordinates are in the same scale as the original image.