Module gapi_grpc::google::cloud::essentialcontacts::v1

Modules

essential_contacts_service_client
Generated client implementations.

Structs

ComputeContactsRequest
Request message for the ComputeContacts method.

ComputeContactsResponse
Response message for the ComputeContacts method.

Contact
A contact that will receive notifications from Google Cloud.

CreateContactRequest
Request message for the CreateContact method.

DeleteContactRequest
Request message for the DeleteContact method.

GetContactRequest
Request message for the GetContact method.

ListContactsRequest
Request message for the ListContacts method.

ListContactsResponse
Response message for the ListContacts method.

SendTestMessageRequest
Request message for the SendTestMessage method.

UpdateContactRequest
Request message for the UpdateContact method.

Enums

NotificationCategory
The notification categories that an essential contact can be subscribed to. Each notification will be categorized by the sender into one of the following categories. All contacts that are subscribed to that category will receive the notification.

ValidationState
A contact’s validation state indicates whether or not it is the correct contact to be receiving notifications for a particular resource.