Module gapi_grpc::google::cloud::functions::v1

Modules

cloud_function

cloud_functions_service_client
Generated client implementations.

failure_policy

Structs

CallFunctionRequest
Request for the CallFunction method.

CallFunctionResponse
Response of CallFunction method.

CloudFunction
Describes a Cloud Function that contains user computation executed in response to an event. It encapsulate function and triggers configurations.

CreateFunctionRequest
Request for the CreateFunction method.

DeleteFunctionRequest
Request for the DeleteFunction method.

EventTrigger
Describes EventTrigger, used to request events be sent from another service.

FailurePolicy
Describes the policy in case of function's execution failure. If empty, then defaults to ignoring failures (i.e. not retrying them).

GenerateDownloadUrlRequest
Request of GenerateDownloadUrl method.

GenerateDownloadUrlResponse
Response of GenerateDownloadUrl method.

GenerateUploadUrlRequest
Request of GenerateSourceUploadUrl method.

GenerateUploadUrlResponse
Response of GenerateSourceUploadUrl method.

GetFunctionRequest
Request for the GetFunction method.

HttpsTrigger
Describes HttpsTrigger, could be used to connect web hooks to function.

ListFunctionsRequest
Request for the ListFunctions method.

ListFunctionsResponse
Response for the ListFunctions method.

OperationMetadataV1
Metadata describing an [Operation][google.longrunning.Operation]

SourceRepository
Describes SourceRepository, used to represent parameters related to source repository where a function is hosted.

UpdateFunctionRequest
Request for the UpdateFunction method.

Enums

CloudFunctionStatus
Describes the current stage of a deployment.

OperationType
A type of an operation.