Module gapi_grpc::google::cloud::functions::v1::failure_policy

Structs

Retry
Describes the retry policy in case of function's execution failure. A function execution will be retried on any failure. A failed execution will be retried up to 7 days with an exponential backoff (capped at 10 seconds). Retried execution is charged as any other execution.

Enums

Action
Defines the action taken in case of a function execution failure.