Struct gapi_grpc::google::cloud::functions::v1beta2::cloud_functions_service_client::CloudFunctionsServiceClient

pub struct CloudFunctionsServiceClient<T> { /* fields omitted */ }

A service that application uses to manipulate triggers and functions.

Methods

impl CloudFunctionsServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> CloudFunctionsServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn list_functions<'_>(
&'_ mut self,
request: impl IntoRequest<ListFunctionsRequest>
) -> Result<Response<ListFunctionsResponse>, Status>

Returns a list of functions that belong to the requested project.

pub async fn get_function<'_>(
&'_ mut self,
request: impl IntoRequest<GetFunctionRequest>
) -> Result<Response<CloudFunction>, Status>

Returns a function with the given name from the requested project.

pub async fn create_function<'_>(
&'_ mut self,
request: impl IntoRequest<CreateFunctionRequest>
) -> Result<Response<Operation>, Status>

Creates a new function. If a function with the given name already exists in the specified project, the long running operation will return ALREADY_EXISTS error.

pub async fn update_function<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateFunctionRequest>
) -> Result<Response<Operation>, Status>

Updates existing function.

pub async fn delete_function<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteFunctionRequest>
) -> Result<Response<Operation>, Status>

Deletes a function with the given name from the specified project. If the given function is used by some trigger, the trigger will be updated to remove this function.

pub async fn call_function<'_>(
&'_ mut self,
request: impl IntoRequest<CallFunctionRequest>
) -> Result<Response<CallFunctionResponse>, Status>

Synchronously invokes a deployed Cloud Function. To be used for testing purposes as very limited traffic is allowed. For more information on the actual limits refer to API Calls.

pub async fn generate_upload_url<'_>(
&'_ mut self,
request: impl IntoRequest<GenerateUploadUrlRequest>
) -> Result<Response<GenerateUploadUrlResponse>, Status>

Returns a signed URL for uploading a function source code. For more information about the signed URL usage see: https://cloud.google.com/storage/docs/access-control/signed-urls Once the function source code upload is complete, the used signed URL should be provided in CreateFunction or UpdateFunction request as a reference to the function source code.

When uploading source code to the generated signed URL, please follow these restrictions:

Source file type should be a zip file.

Source file size should not exceed 100MB limit.

No credentials should be attached - the signed URLs provide access to the target bucket using internal service identity; if credentials were attached, the identity from the credentials would be used, but that identity does not have permissions to upload files to the URL.

When making a HTTP PUT request, these two headers need to be specified:

content-type: application/zip

x-goog-content-length-range: 0,104857600

And this header SHOULD NOT be specified:

Authorization: Bearer YOUR_TOKEN

pub async fn generate_download_url<'_>(
&'_ mut self,
request: impl IntoRequest<GenerateDownloadUrlRequest>
) -> Result<Response<GenerateDownloadUrlResponse>, Status>

Returns a signed URL for downloading deployed function source code. The URL is only valid for a limited period and should be used within minutes after generation. For more information about the signed URL usage see: https://cloud.google.com/storage/docs/access-control/signed-urls