Module gapi_grpc::google::cloud::gaming::v1

Modules

deployed_fleet_details

fetch_deployment_state_response

game_server_cluster_connection_info

game_server_clusters_service_client
Generated client implementations.

game_server_config_override

game_server_configs_service_client
Generated client implementations.

game_server_deployments_service_client
Generated client implementations.

operation_status

realms_service_client
Generated client implementations.

target_details

Structs

CreateGameServerClusterRequest
Request message for GameServerClustersService.CreateGameServerCluster.

CreateGameServerConfigRequest
Request message for GameServerConfigsService.CreateGameServerConfig.

CreateGameServerDeploymentRequest
Request message for GameServerDeploymentsService.CreateGameServerDeployment.

CreateRealmRequest
Request message for RealmsService.CreateRealm.

DeleteGameServerClusterRequest
Request message for GameServerClustersService.DeleteGameServerCluster.

DeleteGameServerConfigRequest
Request message for GameServerConfigsService.DeleteGameServerConfig.

DeleteGameServerDeploymentRequest
Request message for GameServerDeploymentsService.DeleteGameServerDeployment.

DeleteRealmRequest
Request message for RealmsService.DeleteRealm.

DeployedFleetDetails
Details of the deployed Agones fleet.

FetchDeploymentStateRequest
Request message for GameServerDeploymentsService.FetchDeploymentState.

FetchDeploymentStateResponse
Response message for GameServerDeploymentsService.FetchDeploymentState.

FleetConfig
Fleet configs for Agones.

GameServerCluster
A game server cluster resource.

GameServerClusterConnectionInfo
The game server cluster connection information.

GameServerConfig
A game server config resource.

GameServerConfigOverride
A game server config override.

GameServerDeployment
A game server deployment resource.

GameServerDeploymentRollout
The game server deployment rollout which represents the desired rollout state.

GetGameServerClusterRequest
Request message for GameServerClustersService.GetGameServerCluster.

GetGameServerConfigRequest
Request message for GameServerConfigsService.GetGameServerConfig.

GetGameServerDeploymentRequest
Request message for GameServerDeploymentsService.GetGameServerDeployment.

GetGameServerDeploymentRolloutRequest
Request message for GameServerDeploymentsService.GetGameServerDeploymentRollout.

GetRealmRequest
Request message for RealmsService.GetRealm.

GkeClusterReference
A reference to a GKE cluster.

LabelSelector
The label selector, used to group labels on the resources.

ListGameServerClustersRequest
Request message for GameServerClustersService.ListGameServerClusters.

ListGameServerClustersResponse
Response message for GameServerClustersService.ListGameServerClusters.

ListGameServerConfigsRequest
Request message for GameServerConfigsService.ListGameServerConfigs.

ListGameServerConfigsResponse
Response message for GameServerConfigsService.ListGameServerConfigs.

ListGameServerDeploymentsRequest
Request message for GameServerDeploymentsService.ListGameServerDeployments.

ListGameServerDeploymentsResponse
Response message for GameServerDeploymentsService.ListGameServerDeployments.

ListRealmsRequest
Request message for RealmsService.ListRealms.

ListRealmsResponse
Response message for RealmsService.ListRealms.

OperationMetadata
Represents the metadata of the long-running operation.

OperationStatus

PreviewCreateGameServerClusterRequest
Request message for GameServerClustersService.PreviewCreateGameServerCluster.

PreviewCreateGameServerClusterResponse
Response message for GameServerClustersService.PreviewCreateGameServerCluster.

PreviewDeleteGameServerClusterRequest
Request message for GameServerClustersService.PreviewDeleteGameServerCluster.

PreviewDeleteGameServerClusterResponse
Response message for GameServerClustersService.PreviewDeleteGameServerCluster.

PreviewGameServerDeploymentRolloutRequest
Request message for PreviewGameServerDeploymentRollout.

PreviewGameServerDeploymentRolloutResponse
Response message for PreviewGameServerDeploymentRollout. This has details about the Agones fleet and autoscaler to be actuated.

PreviewRealmUpdateRequest
Request message for RealmsService.PreviewRealmUpdate.

PreviewRealmUpdateResponse
Response message for RealmsService.PreviewRealmUpdate.

PreviewUpdateGameServerClusterRequest
Request message for GameServerClustersService.UpdateGameServerCluster.

PreviewUpdateGameServerClusterResponse
Response message for GameServerClustersService.PreviewUpdateGameServerCluster

Realm
A realm resource.

RealmSelector
The realm selector, used to match realm resources.

ScalingConfig
Autoscaling config for an Agones fleet.

Schedule
The schedule of a recurring or one time event. The event's time span is specified by start_time and end_time. If the scheduled event's timespan is larger than the cron_spec + cron_job_duration, the event will be recurring. If only cron_spec + cron_job_duration are specified, the event is effective starting at the local time specified by cron_spec, and is recurring.

SpecSource
Encapsulates Agones fleet spec and Agones autoscaler spec sources.

TargetDetails
Details about the Agones resources.

TargetState
Encapsulates the Target state.

UpdateGameServerClusterRequest
Request message for GameServerClustersService.UpdateGameServerCluster.

UpdateGameServerDeploymentRequest
Request message for GameServerDeploymentsService.UpdateGameServerDeployment. Only allows updates for labels.

UpdateGameServerDeploymentRolloutRequest
Request message for GameServerDeploymentsService.UpdateGameServerRolloutDeployment.

UpdateRealmRequest
Request message for RealmsService.UpdateRealm.