Module gapi_grpc::google::cloud::gkehub::v1alpha2

Modules

gke_hub_client
Generated client implementations.

membership

membership_state

Structs

Authority
Authority encodes how Google will recognize identities from this Membership. See the workload identity documentation for more details: https://cloud.google.com/kubernetes-engine/docs/how-to/workload-identity

ConnectAgentResource
ConnectAgentResource represents a Kubernetes resource manifest for Connect Agent deployment.

CreateMembershipRequest
Request message for the GkeHub.CreateMembership method.

DeleteMembershipRequest
Request message for GkeHub.DeleteMembership method.

GenerateConnectManifestRequest
Request message for GkeHub.GenerateConnectManifest method. .

GenerateConnectManifestResponse
GenerateConnectManifestResponse contains manifest information for installing/upgrading a Connect agent.

GetMembershipRequest
Request message for GkeHub.GetMembership method.

GkeCluster
GkeCluster contains information specific to GKE clusters.

InitializeHubRequest
Request message for the InitializeHub method.

InitializeHubResponse
Response message for the InitializeHub method.

KubernetesMetadata
KubernetesMetadata provides informational metadata for Memberships that are created from Kubernetes Endpoints (currently, these are equivalent to Kubernetes clusters).

KubernetesResource
KubernetesResource contains the YAML manifests and configuration for Membership Kubernetes resources in the cluster. After CreateMembership or UpdateMembership, these resources should be re-applied in the cluster.

ListMembershipsRequest
Request message for GkeHub.ListMemberships method.

ListMembershipsResponse
Response message for the GkeHub.ListMemberships method.

Membership
Membership contains information about a member cluster.

MembershipEndpoint
MembershipEndpoint contains information needed to contact a Kubernetes API, endpoint and any additional Kubernetes metadata.

MembershipState
MembershipState describes the state of a Membership resource.

OperationMetadata
Represents the metadata of the long-running operation.

ResourceManifest
ResourceManifest represents a single Kubernetes resource to be applied to the cluster.

ResourceOptions
ResourceOptions represent options for Kubernetes resource generation.

TypeMeta
TypeMeta is the type information needed for content unmarshalling of Kubernetes resources in the manifest.

UpdateMembershipRequest
Request message for GkeHub.UpdateMembership method.