Module gapi_grpc::google::cloud::irm::v1alpha2

Modules

artifact

communication_channel

incident

incident_role

incident_service_client
Generated client implementations.

lookup_signal_request

search_similar_incidents_response

send_shift_handoff_request

signal

subscription

user

Structs

Annotation
A text annotation by a user.

Artifact
External artifact associated to an incident.

CancelIncidentRoleHandoverRequest
Request to cancel a role handover.

CommunicationChannel
Communication Channels are mechanisms used to receive notifications about changes to incidents.

ConfirmIncidentRoleHandoverRequest
Request to confirm a role handover.

CreateAnnotationRequest
Request for the CreateAnnotation method.

CreateArtifactRequest
Request for the CreateArtifact method.

CreateIncidentRequest
Request for the CreateIncident method.

CreateIncidentRoleAssignmentRequest
Request for creating a role assignment.

CreateSignalRequest
Request for the CreateSignal method.

CreateSubscriptionRequest
Request for the CreateSubscription method.

CreateTagRequest
Request for the CreateTag method.

DeleteArtifactRequest
Request for deleting an artifact.

DeleteIncidentRoleAssignmentRequest
Request for deleting a role assignment.

DeleteSubscriptionRequest
Request for deleting a subscription.

DeleteTagRequest
Request for the DeleteTag method.

EscalateIncidentRequest
Request to escalate an incident.

EscalateIncidentResponse
Response for EscalateIncident.

ForceIncidentRoleHandoverRequest
Request to force a role handover.

GetIncidentRequest
Request for the GetIncident method.

GetSignalRequest
Request for the GetSignal method.

Incident
Representation of an incident.

IncidentRole
Describes a role that can be assigned to an incident.

IncidentRoleAssignment
Stores the assignee of a role as well as the proposed next assignee.

ListAnnotationsRequest
Request for the ListAnnotations method.

ListAnnotationsResponse
Response for the ListAnnotations method.

ListArtifactsRequest
Request for the ListArtifacts method.

ListArtifactsResponse
Response for the ListArtifacts method.

ListIncidentRoleAssignmentsRequest
Request to list role assignments of an incident.

ListIncidentRoleAssignmentsResponse
Response for the ListIncidentRoleAssignments method.

ListSubscriptionsRequest
Request for the ListSubscriptions method.

ListSubscriptionsResponse
Response for the ListSubscriptions method.

ListTagsRequest
Request for the ListTagsForIncident method.

ListTagsResponse
Response for the ListTagsForIncident method.

LookupSignalRequest
Request for the LookupSignal method.

RequestIncidentRoleHandoverRequest
Request to start a role handover.

SearchIncidentsRequest
Request for the SearchIncidents method.

SearchIncidentsResponse
Response for the SearchIncidents method.

SearchSignalsRequest
Request for the SearchSignals method.

SearchSignalsResponse
Response for the SearchSignals method.

SearchSimilarIncidentsRequest
Request for the SearchSimilarIncidents method.

SearchSimilarIncidentsResponse
Response for the SearchSimilarIncidents method.

SendShiftHandoffRequest
SendShiftHandoff and PreviewShiftHandoff RPC request.

SendShiftHandoffResponse
SendShiftHandoff and PreviewShiftHandoff RPC response.

Signal
A signal is a message calling attention to a (potential) incident. An example is a page based on a Stackdriver Alerting policy.

Subscription
A subscription allows users to get notifications about changes to an incident.

Synopsis
Synopsis is a summary of an incident and it contains a textual content, an author and a last updated timestamp.

Tag
A tag by a user.

UpdateArtifactRequest
Request for the UpdateArtifact method.

UpdateIncidentRequest
Request for the UpdateIncident method.

UpdateSignalRequest
Request for the UpdateSignal method.

UpdateSubscriptionRequest
Request for the UpdateSubscription method.

User
A user of the IRM app.