Struct gapi_grpc::google::cloud::irm::v1alpha2::incident_service_client::IncidentServiceClient

pub struct IncidentServiceClient<T> { /* fields omitted */ }

The Incident API for Incident Response & Management.

Implementations

impl IncidentServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> IncidentServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_incident<'_>(
&'_ mut self,
request: impl IntoRequest<CreateIncidentRequest>
) -> Result<Response<Incident>, Status>

Creates a new incident.

pub async fn get_incident<'_>(
&'_ mut self,
request: impl IntoRequest<GetIncidentRequest>
) -> Result<Response<Incident>, Status>

Returns an incident by name.

pub async fn search_incidents<'_>(
&'_ mut self,
request: impl IntoRequest<SearchIncidentsRequest>
) -> Result<Response<SearchIncidentsResponse>, Status>

Returns a list of incidents. Incidents are ordered by start time, with the most recent incidents first.

pub async fn update_incident<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateIncidentRequest>
) -> Result<Response<Incident>, Status>

Updates an existing incident.

pub async fn search_similar_incidents<'_>(
&'_ mut self,
request: impl IntoRequest<SearchSimilarIncidentsRequest>
) -> Result<Response<SearchSimilarIncidentsResponse>, Status>

Returns a list of incidents that are "similar" to the specified incident or signal. This functionality is provided on a best-effort basis and the definition of "similar" is subject to change.

pub async fn create_annotation<'_>(
&'_ mut self,
request: impl IntoRequest<CreateAnnotationRequest>
) -> Result<Response<Annotation>, Status>

Creates an annotation on an existing incident. Only 'text/plain' and 'text/markdown' annotations can be created via this method.

pub async fn list_annotations<'_>(
&'_ mut self,
request: impl IntoRequest<ListAnnotationsRequest>
) -> Result<Response<ListAnnotationsResponse>, Status>

Lists annotations that are part of an incident. No assumptions should be made on the content-type of the annotation returned.

pub async fn create_tag<'_>(
&'_ mut self,
request: impl IntoRequest<CreateTagRequest>
) -> Result<Response<Tag>, Status>

Creates a tag on an existing incident.

pub async fn delete_tag<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteTagRequest>
) -> Result<Response<()>, Status>

Deletes an existing tag.

pub async fn list_tags<'_>(
&'_ mut self,
request: impl IntoRequest<ListTagsRequest>
) -> Result<Response<ListTagsResponse>, Status>

Lists tags that are part of an incident.

pub async fn create_signal<'_>(
&'_ mut self,
request: impl IntoRequest<CreateSignalRequest>
) -> Result<Response<Signal>, Status>

Creates a new signal.

pub async fn search_signals<'_>(
&'_ mut self,
request: impl IntoRequest<SearchSignalsRequest>
) -> Result<Response<SearchSignalsResponse>, Status>

Lists signals that are part of an incident. Signals are returned in reverse chronological order. Note that search should not be relied on for critical functionality. It has lower availability guarantees and might fail to return valid results. Returned results might include stale or extraneous entries.

pub async fn lookup_signal<'_>(
&'_ mut self,
request: impl IntoRequest<LookupSignalRequest>
) -> Result<Response<Signal>, Status>

Finds a signal by other unique IDs.

pub async fn get_signal<'_>(
&'_ mut self,
request: impl IntoRequest<GetSignalRequest>
) -> Result<Response<Signal>, Status>

Returns a signal by name.

pub async fn update_signal<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateSignalRequest>
) -> Result<Response<Signal>, Status>

Updates an existing signal (for example, to assign/unassign it to an incident).

pub async fn escalate_incident<'_>(
&'_ mut self,
request: impl IntoRequest<EscalateIncidentRequest>
) -> Result<Response<EscalateIncidentResponse>, Status>

Escalates an incident.

pub async fn create_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<CreateArtifactRequest>
) -> Result<Response<Artifact>, Status>

Creates a new artifact.

pub async fn list_artifacts<'_>(
&'_ mut self,
request: impl IntoRequest<ListArtifactsRequest>
) -> Result<Response<ListArtifactsResponse>, Status>

Returns a list of artifacts for an incident.

pub async fn update_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateArtifactRequest>
) -> Result<Response<Artifact>, Status>

Updates an existing artifact.

pub async fn delete_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteArtifactRequest>
) -> Result<Response<()>, Status>

Deletes an existing artifact.

pub async fn send_shift_handoff<'_>(
&'_ mut self,
request: impl IntoRequest<SendShiftHandoffRequest>
) -> Result<Response<SendShiftHandoffResponse>, Status>

Sends a summary of the shift for oncall handoff.

pub async fn create_subscription<'_>(
&'_ mut self,
request: impl IntoRequest<CreateSubscriptionRequest>
) -> Result<Response<Subscription>, Status>

Creates a new subscription. This will fail if: a. there are too many (50) subscriptions in the incident already b. a subscription using the given channel already exists

pub async fn update_subscription<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateSubscriptionRequest>
) -> Result<Response<Subscription>, Status>

Updates a subscription.

pub async fn list_subscriptions<'_>(
&'_ mut self,
request: impl IntoRequest<ListSubscriptionsRequest>
) -> Result<Response<ListSubscriptionsResponse>, Status>

Returns a list of subscriptions for an incident.

pub async fn delete_subscription<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteSubscriptionRequest>
) -> Result<Response<()>, Status>

Deletes an existing subscription.

pub async fn create_incident_role_assignment<'_>(
&'_ mut self,
request: impl IntoRequest<CreateIncidentRoleAssignmentRequest>
) -> Result<Response<IncidentRoleAssignment>, Status>

Creates a role assignment on an existing incident. Normally, the user field will be set when assigning a role to oneself, and the next field will be set when proposing another user as the assignee. Setting the next field directly to a user other than oneself is equivalent to proposing and force-assigning the role to the user.

pub async fn delete_incident_role_assignment<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteIncidentRoleAssignmentRequest>
) -> Result<Response<()>, Status>

Deletes an existing role assignment.

pub async fn list_incident_role_assignments<'_>(
&'_ mut self,
request: impl IntoRequest<ListIncidentRoleAssignmentsRequest>
) -> Result<Response<ListIncidentRoleAssignmentsResponse>, Status>

Lists role assignments that are part of an incident.

pub async fn request_incident_role_handover<'_>(
&'_ mut self,
request: impl IntoRequest<RequestIncidentRoleHandoverRequest>
) -> Result<Response<IncidentRoleAssignment>, Status>

Starts a role handover. The proposed assignee will receive an email notifying them of the assignment. This will fail if a role handover is already pending. Handover to an oncall ladder is not permitted. Use CreateIncidentRoleAssignment instead.

pub async fn confirm_incident_role_handover<'_>(
&'_ mut self,
request: impl IntoRequest<ConfirmIncidentRoleHandoverRequest>
) -> Result<Response<IncidentRoleAssignment>, Status>

Confirms a role handover. This will fail if the 'proposed_assignee' field of the IncidentRoleAssignment is not equal to the 'new_assignee' field of the request. If the caller is not the new_assignee, ForceIncidentRoleHandover should be used instead.

pub async fn force_incident_role_handover<'_>(
&'_ mut self,
request: impl IntoRequest<ForceIncidentRoleHandoverRequest>
) -> Result<Response<IncidentRoleAssignment>, Status>

Forces a role handover. This will fail if the 'proposed_assignee' field of the IncidentRoleAssignment is not equal to the 'new_assignee' field of the request. If the caller is the new_assignee, ConfirmIncidentRoleHandover should be used instead.

pub async fn cancel_incident_role_handover<'_>(
&'_ mut self,
request: impl IntoRequest<CancelIncidentRoleHandoverRequest>
) -> Result<Response<IncidentRoleAssignment>, Status>

Cancels a role handover. This will fail if the 'proposed_assignee' field of the IncidentRoleAssignment is not equal to the 'new_assignee' field of the request.