Module gapi_grpc::google::cloud::metastore::v1

Modules

database_dump_spec

dataproc_metastore_client
Generated client implementations.

export_metadata_request

location_metadata

metadata_export

metadata_import

secret

service

Structs

CreateMetadataImportRequest
Request message for [DataprocMetastore.CreateMetadataImport][google.cloud.metastore.v1.DataprocMetastore.CreateMetadataImport].

CreateServiceRequest
Request message for [DataprocMetastore.CreateService][google.cloud.metastore.v1.DataprocMetastore.CreateService].

DatabaseDumpSpec
The specification of database dump to import from or export to.

DeleteServiceRequest
Request message for [DataprocMetastore.DeleteService][google.cloud.metastore.v1.DataprocMetastore.DeleteService].

ExportMetadataRequest
Request message for [DataprocMetastore.ExportMetadata][google.cloud.metastore.v1.DataprocMetastore.ExportMetadata].

GetMetadataImportRequest
Request message for [DataprocMetastore.GetMetadataImport][google.cloud.metastore.v1.DataprocMetastore.GetMetadataImport].

GetServiceRequest
Request message for [DataprocMetastore.GetService][google.cloud.metastore.v1.DataprocMetastore.GetService].

HiveMetastoreConfig
Specifies configuration information specific to running Hive metastore software as the metastore service.

KerberosConfig
Configuration information for a Kerberos principal.

ListMetadataImportsRequest
Request message for [DataprocMetastore.ListMetadataImports][google.cloud.metastore.v1.DataprocMetastore.ListMetadataImports].

ListMetadataImportsResponse
Response message for [DataprocMetastore.ListMetadataImports][google.cloud.metastore.v1.DataprocMetastore.ListMetadataImports].

ListServicesRequest
Request message for [DataprocMetastore.ListServices][google.cloud.metastore.v1.DataprocMetastore.ListServices].

ListServicesResponse
Response message for [DataprocMetastore.ListServices][google.cloud.metastore.v1.DataprocMetastore.ListServices].

LocationMetadata
Metadata about the service in a location.

MaintenanceWindow
Maintenance window. This specifies when Dataproc Metastore may perform system maintenance operation to the service.

MetadataExport
The details of a metadata export operation.

MetadataImport
A metastore resource that imports metadata.

MetadataManagementActivity
The metadata management activities of the metastore service.

OperationMetadata
Represents the metadata of a long-running operation.

Secret
A securely stored value.

Service
A managed metastore service that serves metadata queries.

UpdateMetadataImportRequest
Request message for [DataprocMetastore.UpdateMetadataImport][google.cloud.metastore.v1.DataprocMetastore.UpdateMetadataImport].

UpdateServiceRequest
Request message for [DataprocMetastore.UpdateService][google.cloud.metastore.v1.DataprocMetastore.UpdateService].