Struct gapi_grpc::google::cloud::metastore::v1::dataproc_metastore_client::DataprocMetastoreClient

pub struct DataprocMetastoreClient<T> { /* fields omitted */ }

Configures and manages metastore services. Metastore services are fully managed, highly available, auto-scaled, auto-healing, OSS-native deployments of technical metadata management software. Each metastore service exposes a network endpoint through which metadata queries are served. Metadata queries can originate from a variety of sources, including Apache Hive, Apache Presto, and Apache Spark.

The Dataproc Metastore API defines the following resource model:

The service works with a collection of Google Cloud projects, named: /projects/*

Each project has a collection of available locations, named: /locations/* (a location must refer to a Google Cloud region)

Each location has a collection of services, named: /services/*

Dataproc Metastore services are resources with names of the form:

/projects/{project_number}/locations/{location_id}/services/{service_id}.

Implementations

impl DataprocMetastoreClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> DataprocMetastoreClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn list_services(
&mut self,
request: impl IntoRequest<ListServicesRequest>
) -> Result<Response<ListServicesResponse>, Status>

Lists services in a project and location.

pub async fn get_service(
&mut self,
request: impl IntoRequest<GetServiceRequest>
) -> Result<Response<Service>, Status>

Gets the details of a single service.

pub async fn create_service(
&mut self,
request: impl IntoRequest<CreateServiceRequest>
) -> Result<Response<Operation>, Status>

Creates a metastore service in a project and location.

pub async fn update_service(
&mut self,
request: impl IntoRequest<UpdateServiceRequest>
) -> Result<Response<Operation>, Status>

Updates the parameters of a single service.

pub async fn delete_service(
&mut self,
request: impl IntoRequest<DeleteServiceRequest>
) -> Result<Response<Operation>, Status>

Deletes a single service.

pub async fn list_metadata_imports(
&mut self,
request: impl IntoRequest<ListMetadataImportsRequest>
) -> Result<Response<ListMetadataImportsResponse>, Status>

Lists imports in a service.

pub async fn get_metadata_import(
&mut self,
request: impl IntoRequest<GetMetadataImportRequest>
) -> Result<Response<MetadataImport>, Status>

Gets details of a single import.

pub async fn create_metadata_import(
&mut self,
request: impl IntoRequest<CreateMetadataImportRequest>
) -> Result<Response<Operation>, Status>

Creates a new MetadataImport in a given project and location.

pub async fn update_metadata_import(
&mut self,
request: impl IntoRequest<UpdateMetadataImportRequest>
) -> Result<Response<Operation>, Status>

Updates a single import. Only the description field of MetadataImport is supported to be updated.

pub async fn export_metadata(
&mut self,
request: impl IntoRequest<ExportMetadataRequest>
) -> Result<Response<Operation>, Status>

Exports metadata from a service.