Module gapi_grpc::google::cloud::ml::v1

Modules

hyperparameter_output

hyperparameter_spec

job

job_service_client
Generated client implementations.

model_service_client
Generated client implementations.

online_prediction_service_client
Generated client implementations.

operation_metadata

parameter_spec

prediction_input

project_management_service_client
Generated client implementations.

training_input

Structs

CancelJobRequest
Request message for the CancelJob method.

CreateJobRequest
Request message for the CreateJob method.

CreateModelRequest
Request message for the CreateModel method.

CreateVersionRequest
Uploads the provided trained model version to Cloud Machine Learning.

DeleteModelRequest
Request message for the DeleteModel method.

DeleteVersionRequest
Request message for the DeleteVerionRequest method.

GetConfigRequest
Requests service account information associated with a project.

GetConfigResponse
Returns service account information associated with a project.

GetJobRequest
Request message for the GetJob method.

GetModelRequest
Request message for the GetModel method.

GetVersionRequest
Request message for the GetVersion method.

HyperparameterOutput
Represents the result of a single hyperparameter tuning trial from a training job. The TrainingOutput object that is returned on successful completion of a training job with hyperparameter tuning includes a list of HyperparameterOutput objects, one for each successful trial.

HyperparameterSpec
Represents a set of hyperparameters to optimize.

Job
Represents a training or prediction job.

ListJobsRequest
Request message for the ListJobs method.

ListJobsResponse
Response message for the ListJobs method.

ListModelsRequest
Request message for the ListModels method.

ListModelsResponse
Response message for the ListModels method.

ListVersionsRequest
Request message for the ListVersions method.

ListVersionsResponse
Response message for the ListVersions method.

ManualScaling
Options for manually scaling a model.

Model
Represents a machine learning solution.

OperationMetadata
Represents the metadata of the long-running operation.

ParameterSpec
Represents a single hyperparameter to optimize.

PredictRequest
Request for predictions to be issued against a trained model.

PredictionInput
Represents input parameters for a prediction job.

PredictionOutput
Represents results of a prediction job.

SetDefaultVersionRequest
Request message for the SetDefaultVersion request.

TrainingInput
Represents input parameters for a training job.

TrainingOutput
Represents results of a training job. Output only.

Version
Represents a version of the model.