Module gapi_grpc::google::cloud::ml::v1::prediction_input

Enums

DataFormat
The format used to separate data instances in the source files.

ModelVersion
Required. The model or the version to use for prediction.