Module gapi_grpc::google::cloud::networkconnectivity::v1alpha1

Modules

hub_service_client
Generated client implementations.

Structs

CreateHubRequest
Request for [HubService.CreateHub][google.cloud.networkconnectivity.v1alpha1.HubService.CreateHub] method.

CreateSpokeRequest
The request for [HubService.CreateSpoke][google.cloud.networkconnectivity.v1alpha1.HubService.CreateSpoke].

DeleteHubRequest
The request for [HubService.DeleteHub][google.cloud.networkconnectivity.v1alpha1.HubService.DeleteHub].

DeleteSpokeRequest
The request for [HubService.DeleteSpoke][google.cloud.networkconnectivity.v1alpha1.HubService.DeleteSpoke].

GetHubRequest
Request for [HubService.GetHub][google.cloud.networkconnectivity.v1alpha1.HubService.GetHub] method.

GetSpokeRequest
The request for [HubService.GetSpoke][google.cloud.networkconnectivity.v1alpha1.HubService.GetSpoke].

Hub
The Connectivity Hub is a hub-and-spoke abstraction for network connectivity management in Google Cloud. It aims to reduce operational complexity through a simple, centralized connectivity management model. Here is the resource message of a Hub.

ListHubsRequest
Request for [HubService.ListHubs][google.cloud.networkconnectivity.v1alpha1.HubService.ListHubs] method.

ListHubsResponse
Response for [HubService.ListHubs][google.cloud.networkconnectivity.v1alpha1.HubService.ListHubs] method.

ListSpokesRequest
The request for [HubService.ListSpokes][google.cloud.networkconnectivity.v1alpha1.HubService.ListSpokes].

ListSpokesResponse
The response for [HubService.ListSpokes][google.cloud.networkconnectivity.v1alpha1.HubService.ListSpokes].

OperationMetadata
Represents the metadata of the long-running operation.

RouterApplianceInstance
RouterAppliance represents a Router appliance which is specified by a VM URI and a NIC address.

Spoke
A Spoke is an abstraction of a network attachment being attached to a Hub. A Spoke can be underlying a VPN tunnel, a VLAN (interconnect) attachment, a Router appliance, etc.

UpdateHubRequest
Request for [HubService.UpdateHub][google.cloud.networkconnectivity.v1alpha1.HubService.UpdateHub] method.

UpdateSpokeRequest
Request for [HubService.UpdateSpoke][google.cloud.networkconnectivity.v1alpha1.HubService.UpdateSpoke] method.