Module gapi_grpc::google::cloud::orgpolicy::v2

Modules

constraint

org_policy_client
Generated client implementations.

policy_spec

Structs

AlternatePolicySpec
Similar to PolicySpec but with an extra 'launch' field for launch reference. The PolicySpec here is specific for dry-run/darklaunch.

Constraint
A constraint describes a way to restrict resource's configuration. For example, you could enforce a constraint that controls which cloud services can be activated across an organization, or whether a Compute Engine instance can have serial port connections established. Constraints can be configured by the organization's policy adminstrator to fit the needs of the organzation by setting a policy that includes constraints at different locations in the organization's resource hierarchy. Policies are inherited down the resource hierarchy from higher levels, but can also be overridden. For details about the inheritance rules please read about [policies][google.cloud.OrgPolicy.v2.Policy].

CreatePolicyRequest
The request sent to the CreatePolicyRequest [google.cloud.orgpolicy.v2.OrgPolicy.CreatePolicy] method.

DeletePolicyRequest
The request sent to the [DeletePolicy] [google.cloud.orgpolicy.v2.OrgPolicy.DeletePolicy] method.

GetEffectivePolicyRequest
The request sent to the [GetEffectivePolicy] [google.cloud.orgpolicy.v2.OrgPolicy.GetEffectivePolicy] method.

GetPolicyRequest
The request sent to the [GetPolicy] [google.cloud.orgpolicy.v2.OrgPolicy.GetPolicy] method.

ListConstraintsRequest
The request sent to the [ListConstraints] [google.cloud.orgpolicy.v2.OrgPolicy.ListConstraints] method.

ListConstraintsResponse
The response returned from the [ListConstraints] [google.cloud.orgpolicy.v2.OrgPolicy.ListConstraints] method.

ListPoliciesRequest
The request sent to the [ListPolicies] [google.cloud.orgpolicy.v2.OrgPolicy.ListPolicies] method.

ListPoliciesResponse
The response returned from the [ListPolicies] [google.cloud.orgpolicy.v2.OrgPolicy.ListPolicies] method. It will be empty if no Policies are set on the resource.

Policy
Defines a Cloud Organization Policy which is used to specify Constraints for configurations of Cloud Platform resources.

PolicySpec
Defines a Cloud Organization PolicySpec which is used to specify Constraints for configurations of Cloud Platform resources.

UpdatePolicyRequest
The request sent to the UpdatePolicyRequest [google.cloud.orgpolicy.v2.OrgPolicy.UpdatePolicy] method.