Module gapi_grpc::google::cloud::osconfig::agentendpoint::v1::os_policy::resource::repository_resource

Modules

apt_repository

Structs

AptRepository
Represents a single apt package repository. These will be added to a repo file that will be managed at /etc/apt/sources.list.d/google_osconfig.list.

GooRepository
Represents a Goo package repository. These are added to a repo file that is managed at C:/ProgramData/GooGet/repos/google_osconfig.repo.

YumRepository
Represents a single yum package repository. These are added to a repo file that is managed at /etc/yum.repos.d/google_osconfig.repo.

ZypperRepository
Represents a single zypper package repository. These are added to a repo file that is managed at /etc/zypp/repos.d/google_osconfig.repo.

Enums

Repository
A specific type of repository.