Module gapi_grpc::google::cloud::osconfig::v1::inventory

Modules

item

software_package

windows_update_package

Structs

Item
A single piece of inventory on a VM.

OsInfo
Operating system information for the VM.

SoftwarePackage
Software package information of the operating system.

VersionedPackage
Information related to the a standard versioned package. This includes package info for APT, Yum, Zypper, and Googet package managers.

WindowsQuickFixEngineeringPackage
Information related to a Quick Fix Engineering package. Fields are taken from Windows QuickFixEngineering Interface and match the source names: https://docs.microsoft.com/en-us/windows/win32/cimwin32prov/win32-quickfixengineering

WindowsUpdatePackage
Details related to a Windows Update package. Field data and names are taken from Windows Update API IUpdate Interface: https://docs.microsoft.com/en-us/windows/win32/api/_wua/ Descriptive fields like title, and description are localized based on the locale of the VM being updated.

ZypperPatch
Details related to a Zypper Patch.