Module gapi_grpc::google::cloud::osconfig::v1beta::software_recipe::step

Modules

exec_file

extract_archive

run_script

Structs

CopyFile
Copies the artifact to the specified path on the instance.

ExecFile
Executes an artifact or local file.

ExtractArchive
Extracts an archive of the type specified in the specified directory.

InstallDpkg
Installs a deb via dpkg.

InstallMsi
Installs an MSI file.

InstallRpm
Installs an rpm file via the rpm utility.

RunScript
Runs a script through an interpreter.

Enums

Step
A specific type of step.