Module gapi_grpc::google::cloud::policytroubleshooter::v1

Modules

binding_explanation

iam_checker_client
Generated client implementations.

Structs

AccessTuple
Information about the member, resource, and permission to check.

BindingExplanation
Details about how a binding in a policy affects a member's ability to use a permission.

ExplainedPolicy
Details about how a specific IAM [Policy][google.iam.v1.Policy] contributed to the access check.

TroubleshootIamPolicyRequest
Request for [TroubleshootIamPolicy][google.cloud.policytroubleshooter.v1.IamChecker.TroubleshootIamPolicy].

TroubleshootIamPolicyResponse
Response for [TroubleshootIamPolicy][google.cloud.policytroubleshooter.v1.IamChecker.TroubleshootIamPolicy].

Enums

AccessState
Whether a member has a permission for a resource.

HeuristicRelevance
The extent to which a single data point contributes to an overall determination.