Module gapi_grpc::google::cloud::recaptchaenterprise::v1beta1

Modules

annotate_assessment_request

assessment

key

recaptcha_enterprise_service_v1_beta1_client
Generated client implementations.

token_properties

web_key_settings

Structs

AndroidKeySettings
Settings specific to keys that can be used by Android apps.

AnnotateAssessmentRequest
The request message to annotate an Assessment.

AnnotateAssessmentResponse
Empty response for AnnotateAssessment.

Assessment
A recaptcha assessment resource.

CreateAssessmentRequest
The create assessment request message.

CreateKeyRequest
The create key request message.

DeleteKeyRequest
The delete key request message.

Event

GetKeyRequest
The get key request message.

IosKeySettings
Settings specific to keys that can be used by iOS apps.

Key
A key used to identify and configure applications (web and/or mobile) that use reCAPTCHA Enterprise.

ListKeysRequest
The list keys request message.

ListKeysResponse
Response to request to list keys in a project.

TokenProperties

UpdateKeyRequest
The update key request message.

WebKeySettings
Settings specific to keys that can be used by websites.