Module gapi_grpc::google::cloud::recommendationengine::v1beta1

Modules

catalog_item

catalog_service_client
Generated client implementations.

feature_map

import_errors_config

input_config

predict_response

prediction_api_key_registry_client
Generated client implementations.

prediction_service_client
Generated client implementations.

product_catalog_item

user_event

user_event_service_client
Generated client implementations.

Structs

CatalogInlineSource
The inline source for the input config for ImportCatalogItems method.

CatalogItem
CatalogItem captures all metadata information of items to be recommended.

CollectUserEventRequest
Request message for CollectUserEvent method.

CreateCatalogItemRequest
Request message for CreateCatalogItem method.

CreatePredictionApiKeyRegistrationRequest
Request message for the CreatePredictionApiKeyRegistration method.

DeleteCatalogItemRequest
Request message for DeleteCatalogItem method.

DeletePredictionApiKeyRegistrationRequest
Request message for DeletePredictionApiKeyRegistration method.

EventDetail
User event details shared by all recommendation types.

FeatureMap
FeatureMap represents extra features that customers want to include in the recommendation model for catalogs/user events as categorical/numerical features.

GcsSource
Google Cloud Storage location for input content. format.

GetCatalogItemRequest
Request message for GetCatalogItem method.

Image
Catalog item thumbnail/detail image.

ImportCatalogItemsRequest
Request message for Import methods.

ImportCatalogItemsResponse
Response of the ImportCatalogItemsRequest. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ImportErrorsConfig
Configuration of destination for Import related errors.

ImportMetadata
Metadata related to the progress of the Import operation. This will be returned by the google.longrunning.Operation.metadata field.

ImportUserEventsRequest
Request message for the ImportUserEvents request.

ImportUserEventsResponse
Response of the ImportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

InputConfig
The input config source.

ListCatalogItemsRequest
Request message for ListCatalogItems method.

ListCatalogItemsResponse
Response message for ListCatalogItems method.

ListPredictionApiKeyRegistrationsRequest
Request message for the ListPredictionApiKeyRegistrations.

ListPredictionApiKeyRegistrationsResponse
Response message for the ListPredictionApiKeyRegistrations.

ListUserEventsRequest
Request message for ListUserEvents method.

ListUserEventsResponse
Response message for ListUserEvents method.

PredictRequest
Request message for Predict method.

PredictResponse
Response message for predict method.

PredictionApiKeyRegistration
Registered Api Key.

ProductCatalogItem
ProductCatalogItem captures item metadata specific to retail products.

ProductDetail
Detailed product information associated with a user event.

ProductEventDetail
ProductEventDetail captures user event information specific to retail products.

PurchaseTransaction
A transaction represents the entire purchase transaction.

PurgeUserEventsMetadata
Metadata related to the progress of the PurgeUserEvents operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeUserEventsRequest
Request message for PurgeUserEvents method.

PurgeUserEventsResponse
Response of the PurgeUserEventsRequest. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

UpdateCatalogItemRequest
Request message for UpdateCatalogItem method.

UserEvent
UserEvent captures all metadata information recommendation engine needs to know about how end users interact with customers' website.

UserEventImportSummary
A summary of import result. The UserEventImportSummary summarizes the import status for user events.

UserEventInlineSource
The inline source for the input config for ImportUserEvents method.

UserInfo
Information of end users.

WriteUserEventRequest
Request message for WriteUserEvent method.