Module gapi_grpc::google::cloud::resourcesettings::v1

Modules

resource_settings_service_client
Generated client implementations.

setting

value

Structs

CreateSettingValueRequest
The request for CreateSettingValue.

DeleteSettingValueRequest
The request for DeleteSettingValue.

GetSettingValueRequest
The request for GetSettingValue.

ListSettingsRequest
The request for ListSettings.

ListSettingsResponse
The response from ListSettings.

LookupEffectiveSettingValueRequest
The request for LookupEffectiveSettingValue.

SearchSettingValuesRequest
The request for SearchSettingValues.

SearchSettingValuesResponse
The response from SearchSettingValues.

Setting
The schema for setting values. At a given Cloud resource, a setting can parent at most one setting value.

SettingValue
The instantiation of a setting. Every setting value is parented by its corresponding setting.

UpdateSettingValueRequest
The request for UpdateSettingValue.

Value
The data in a setting value.