Module gapi_grpc::google::cloud::retail::v2

Modules

catalog_service_client
Generated client implementations.

import_errors_config

predict_response

prediction_service_client
Generated client implementations.

product

product_input_config

product_service_client
Generated client implementations.

rejoin_user_events_request

user_event_input_config

user_event_service_client
Generated client implementations.

Structs

BigQuerySource
BigQuery source import data from.

Catalog
The catalog configuration.

CollectUserEventRequest
Request message for CollectUserEvent method.

CreateProductRequest
Request message for [CreateProduct][] method.

CustomAttribute
A custom attribute that is not explicitly modeled in [Product][google.cloud.retail.v2.Product]].

DeleteProductRequest
Request message for [DeleteProduct][] method.

GcsSource
Google Cloud Storage location for input content. format.

GetProductRequest
Request message for [GetProduct][] method.

Image
[Product][google.cloud.retail.v2.Product] thumbnail/detail image.

ImportErrorsConfig
Configuration of destination for Import related errors.

ImportMetadata
Metadata related to the progress of the Import operation. This will be returned by the google.longrunning.Operation.metadata field.

ImportProductsRequest
Request message for Import methods.

ImportProductsResponse
Response of the [ImportProductsRequest][google.cloud.retail.v2.ImportProductsRequest]. If the long running operation is done, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ImportUserEventsRequest
Request message for the ImportUserEvents request.

ImportUserEventsResponse
Response of the ImportUserEventsRequest. If the long running operation was successful, then this message is returned by the google.longrunning.Operations.response field if the operation was successful.

ListCatalogsRequest
Request for [CatalogService.ListCatalogs][google.cloud.retail.v2.CatalogService.ListCatalogs] method.

ListCatalogsResponse
Response for [CatalogService.ListCatalogs][google.cloud.retail.v2.CatalogService.ListCatalogs] method.

PredictRequest
Request message for Predict method.

PredictResponse
Response message for predict method.

PriceInfo
The price information of a [Product][google.cloud.retail.v2.Product].

Product
Product captures all metadata information of items to be recommended or searched.

ProductDetail
Detailed product information associated with a user event.

ProductInlineSource
The inline source for the input config for ImportProducts method.

ProductInputConfig
The input config source for products.

ProductLevelConfig
Configures what level the product should be uploaded with regards to how users will be send events and how predictions will be made.

PurchaseTransaction
A transaction represents the entire purchase transaction.

PurgeMetadata
Metadata related to the progress of the Purge operation. This will be returned by the google.longrunning.Operation.metadata field.

PurgeUserEventsRequest
Request message for PurgeUserEvents method.

PurgeUserEventsResponse
Response of the PurgeUserEventsRequest. If the long running operation is successfully done, then this message is returned by the google.longrunning.Operations.response field.

RejoinUserEventsMetadata
Metadata for RejoinUserEvents method.

RejoinUserEventsRequest
Request message for RejoinUserEvents method.

RejoinUserEventsResponse
Response message for RejoinUserEvents method.

UpdateCatalogRequest
Request for [CatalogService.UpdateCatalog][google.cloud.retail.v2.CatalogService.UpdateCatalog] method.

UpdateProductRequest
Request message for [UpdateProduct][] method.

UserEvent
UserEvent captures all metadata information Retail API needs to know about how end users interact with customers' website.

UserEventImportSummary
A summary of import result. The UserEventImportSummary summarizes the import status for user events.

UserEventInlineSource
The inline source for the input config for ImportUserEvents method.

UserEventInputConfig
The input config source for user events.

UserInfo
Information of an end user.

WriteUserEventRequest
Request message for WriteUserEvent method.