Struct gapi_grpc::google::cloud::servicedirectory::v1::registration_service_client::RegistrationServiceClient

pub struct RegistrationServiceClient<T> { /* fields omitted */ }

Service Directory API for registering services. It defines the following resource model:

The API has a collection of [Namespace][google.cloud.servicedirectory.v1.Namespace] resources, named projects/*/locations/*/namespaces/*.

Each Namespace has a collection of [Service][google.cloud.servicedirectory.v1.Service] resources, named projects/*/locations/*/namespaces/*/services/*.

Each Service has a collection of [Endpoint][google.cloud.servicedirectory.v1.Endpoint] resources, named projects/*/locations/*/namespaces/*/services/*/endpoints/*.

Implementations

impl RegistrationServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> RegistrationServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_namespace(
&mut self,
request: impl IntoRequest<CreateNamespaceRequest>
) -> Result<Response<Namespace>, Status>

Creates a namespace, and returns the new Namespace.

pub async fn list_namespaces(
&mut self,
request: impl IntoRequest<ListNamespacesRequest>
) -> Result<Response<ListNamespacesResponse>, Status>

Lists all namespaces.

pub async fn get_namespace(
&mut self,
request: impl IntoRequest<GetNamespaceRequest>
) -> Result<Response<Namespace>, Status>

Gets a namespace.

pub async fn update_namespace(
&mut self,
request: impl IntoRequest<UpdateNamespaceRequest>
) -> Result<Response<Namespace>, Status>

Updates a namespace.

pub async fn delete_namespace(
&mut self,
request: impl IntoRequest<DeleteNamespaceRequest>
) -> Result<Response<()>, Status>

Deletes a namespace. This also deletes all services and endpoints in the namespace.

pub async fn create_service(
&mut self,
request: impl IntoRequest<CreateServiceRequest>
) -> Result<Response<Service>, Status>

Creates a service, and returns the new Service.

pub async fn list_services(
&mut self,
request: impl IntoRequest<ListServicesRequest>
) -> Result<Response<ListServicesResponse>, Status>

Lists all services belonging to a namespace.

pub async fn get_service(
&mut self,
request: impl IntoRequest<GetServiceRequest>
) -> Result<Response<Service>, Status>

Gets a service.

pub async fn update_service(
&mut self,
request: impl IntoRequest<UpdateServiceRequest>
) -> Result<Response<Service>, Status>

Updates a service.

pub async fn delete_service(
&mut self,
request: impl IntoRequest<DeleteServiceRequest>
) -> Result<Response<()>, Status>

Deletes a service. This also deletes all endpoints associated with the service.

pub async fn create_endpoint(
&mut self,
request: impl IntoRequest<CreateEndpointRequest>
) -> Result<Response<Endpoint>, Status>

Creates a endpoint, and returns the new Endpoint.

pub async fn list_endpoints(
&mut self,
request: impl IntoRequest<ListEndpointsRequest>
) -> Result<Response<ListEndpointsResponse>, Status>

Lists all endpoints.

pub async fn get_endpoint(
&mut self,
request: impl IntoRequest<GetEndpointRequest>
) -> Result<Response<Endpoint>, Status>

Gets a endpoint.

pub async fn update_endpoint(
&mut self,
request: impl IntoRequest<UpdateEndpointRequest>
) -> Result<Response<Endpoint>, Status>

Updates a endpoint.

pub async fn delete_endpoint(
&mut self,
request: impl IntoRequest<DeleteEndpointRequest>
) -> Result<Response<()>, Status>

Deletes a endpoint.

pub async fn get_iam_policy(
&mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Gets the IAM Policy for a resource (namespace or service only).

pub async fn set_iam_policy(
&mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Sets the IAM Policy for a resource (namespace or service only).

pub async fn test_iam_permissions(
&mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Tests IAM permissions for a resource (namespace or service only).