Struct gapi_grpc::google::cloud::sql::v1beta4::sql_instances_service_client::SqlInstancesServiceClient

pub struct SqlInstancesServiceClient<T> { /* fields omitted */ }

Implementations

impl SqlInstancesServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> SqlInstancesServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn add_server_ca(
&mut self,
request: impl IntoRequest<SqlInstancesAddServerCaRequest>
) -> Result<Response<Operation>, Status>

Add a new trusted Certificate Authority (CA) version for the specified instance. Required to prepare for a certificate rotation. If a CA version was previously added but never used in a certificate rotation, this operation replaces that version. There cannot be more than one CA version waiting to be rotated in.

pub async fn clone(
&mut self,
request: impl IntoRequest<SqlInstancesCloneRequest>
) -> Result<Response<Operation>, Status>

Creates a Cloud SQL instance as a clone of the source instance. Using this operation might cause your instance to restart.

pub async fn delete(
&mut self,
request: impl IntoRequest<SqlInstancesDeleteRequest>
) -> Result<Response<Operation>, Status>

Deletes a Cloud SQL instance.

pub async fn demote_master(
&mut self,
request: impl IntoRequest<SqlInstancesDemoteMasterRequest>
) -> Result<Response<Operation>, Status>

Demotes the stand-alone instance to be a Cloud SQL read replica for an external database server.

pub async fn export(
&mut self,
request: impl IntoRequest<SqlInstancesExportRequest>
) -> Result<Response<Operation>, Status>

Exports data from a Cloud SQL instance to a Cloud Storage bucket as a SQL dump or CSV file.

pub async fn failover(
&mut self,
request: impl IntoRequest<SqlInstancesFailoverRequest>
) -> Result<Response<Operation>, Status>

Failover the instance to its failover replica instance. Using this operation might cause your instance to restart.

pub async fn get(
&mut self,
request: impl IntoRequest<SqlInstancesGetRequest>
) -> Result<Response<DatabaseInstance>, Status>

Retrieves a resource containing information about a Cloud SQL instance.

pub async fn import(
&mut self,
request: impl IntoRequest<SqlInstancesImportRequest>
) -> Result<Response<Operation>, Status>

Imports data into a Cloud SQL instance from a SQL dump or CSV file in Cloud Storage.

pub async fn insert(
&mut self,
request: impl IntoRequest<SqlInstancesInsertRequest>
) -> Result<Response<Operation>, Status>

Creates a new Cloud SQL instance.

pub async fn list(
&mut self,
request: impl IntoRequest<SqlInstancesListRequest>
) -> Result<Response<InstancesListResponse>, Status>

Lists instances under a given project.

pub async fn list_server_cas(
&mut self,
request: impl IntoRequest<SqlInstancesListServerCasRequest>
) -> Result<Response<InstancesListServerCasResponse>, Status>

Lists all of the trusted Certificate Authorities (CAs) for the specified instance. There can be up to three CAs listed: the CA that was used to sign the certificate that is currently in use, a CA that has been added but not yet used to sign a certificate, and a CA used to sign a certificate that has previously rotated out.

pub async fn patch(
&mut self,
request: impl IntoRequest<SqlInstancesPatchRequest>
) -> Result<Response<Operation>, Status>

Updates settings of a Cloud SQL instance. This method supports patch semantics.

pub async fn promote_replica(
&mut self,
request: impl IntoRequest<SqlInstancesPromoteReplicaRequest>
) -> Result<Response<Operation>, Status>

Promotes the read replica instance to be a stand-alone Cloud SQL instance. Using this operation might cause your instance to restart.

pub async fn reset_ssl_config(
&mut self,
request: impl IntoRequest<SqlInstancesResetSslConfigRequest>
) -> Result<Response<Operation>, Status>

Deletes all client certificates and generates a new server SSL certificate for the instance.

pub async fn restart(
&mut self,
request: impl IntoRequest<SqlInstancesRestartRequest>
) -> Result<Response<Operation>, Status>

Restarts a Cloud SQL instance.

pub async fn restore_backup(
&mut self,
request: impl IntoRequest<SqlInstancesRestoreBackupRequest>
) -> Result<Response<Operation>, Status>

Restores a backup of a Cloud SQL instance. Using this operation might cause your instance to restart.

pub async fn rotate_server_ca(
&mut self,
request: impl IntoRequest<SqlInstancesRotateServerCaRequest>
) -> Result<Response<Operation>, Status>

Rotates the server certificate to one signed by the Certificate Authority (CA) version previously added with the addServerCA method.

pub async fn start_replica(
&mut self,
request: impl IntoRequest<SqlInstancesStartReplicaRequest>
) -> Result<Response<Operation>, Status>

Starts the replication in the read replica instance.

pub async fn stop_replica(
&mut self,
request: impl IntoRequest<SqlInstancesStopReplicaRequest>
) -> Result<Response<Operation>, Status>

Stops the replication in the read replica instance.

pub async fn truncate_log(
&mut self,
request: impl IntoRequest<SqlInstancesTruncateLogRequest>
) -> Result<Response<Operation>, Status>

Truncate MySQL general and slow query log tables MySQL only.

pub async fn update(
&mut self,
request: impl IntoRequest<SqlInstancesUpdateRequest>
) -> Result<Response<Operation>, Status>

Updates settings of a Cloud SQL instance. Using this operation might cause your instance to restart.

pub async fn create_ephemeral(
&mut self,
request: impl IntoRequest<SqlInstancesCreateEphemeralCertRequest>
) -> Result<Response<SslCert>, Status>

Generates a short-lived X509 certificate containing the provided public key and signed by a private key specific to the target instance. Users may use the certificate to authenticate as themselves when connecting to the database.

pub async fn reschedule_maintenance(
&mut self,
request: impl IntoRequest<SqlInstancesRescheduleMaintenanceRequest>
) -> Result<Response<Operation>, Status>

Reschedules the maintenance on the given instance.

pub async fn verify_external_sync_settings(
&mut self,
request: impl IntoRequest<SqlInstancesVerifyExternalSyncSettingsRequest>
) -> Result<Response<SqlInstancesVerifyExternalSyncSettingsResponse>, Status>

Verify External primary instance external sync settings.

pub async fn start_external_sync(
&mut self,
request: impl IntoRequest<SqlInstancesStartExternalSyncRequest>
) -> Result<Response<Operation>, Status>

Start External primary instance migration.