Module gapi_grpc::google::cloud::tasks::v2beta2

Modules

cloud_tasks_client
Generated client implementations.

queue

retry_config

task

Structs

AcknowledgeTaskRequest
Request message for acknowledging a task using [AcknowledgeTask][google.cloud.tasks.v2beta2.CloudTasks.AcknowledgeTask].

AppEngineHttpRequest
App Engine HTTP request.

AppEngineHttpTarget
App Engine HTTP target.

AppEngineRouting
App Engine Routing.

AttemptStatus
The status of a task attempt.

CancelLeaseRequest
Request message for canceling a lease using [CancelLease][google.cloud.tasks.v2beta2.CloudTasks.CancelLease].

CreateQueueRequest
Request message for [CreateQueue][google.cloud.tasks.v2beta2.CloudTasks.CreateQueue].

CreateTaskRequest
Request message for [CreateTask][google.cloud.tasks.v2beta2.CloudTasks.CreateTask].

DeleteQueueRequest
Request message for [DeleteQueue][google.cloud.tasks.v2beta2.CloudTasks.DeleteQueue].

DeleteTaskRequest
Request message for deleting a task using [DeleteTask][google.cloud.tasks.v2beta2.CloudTasks.DeleteTask].

GetQueueRequest
Request message for [GetQueue][google.cloud.tasks.v2beta2.CloudTasks.GetQueue].

GetTaskRequest
Request message for getting a task using [GetTask][google.cloud.tasks.v2beta2.CloudTasks.GetTask].

LeaseTasksRequest
Request message for leasing tasks using [LeaseTasks][google.cloud.tasks.v2beta2.CloudTasks.LeaseTasks].

LeaseTasksResponse
Response message for leasing tasks using [LeaseTasks][google.cloud.tasks.v2beta2.CloudTasks.LeaseTasks].

ListQueuesRequest
Request message for [ListQueues][google.cloud.tasks.v2beta2.CloudTasks.ListQueues].

ListQueuesResponse
Response message for [ListQueues][google.cloud.tasks.v2beta2.CloudTasks.ListQueues].

ListTasksRequest
Request message for listing tasks using [ListTasks][google.cloud.tasks.v2beta2.CloudTasks.ListTasks].

ListTasksResponse
Response message for listing tasks using [ListTasks][google.cloud.tasks.v2beta2.CloudTasks.ListTasks].

PauseQueueRequest
Request message for [PauseQueue][google.cloud.tasks.v2beta2.CloudTasks.PauseQueue].

PullMessage
The pull message contains data that can be used by the caller of [LeaseTasks][google.cloud.tasks.v2beta2.CloudTasks.LeaseTasks] to process the task.

PullTarget
Pull target.

PurgeQueueRequest
Request message for [PurgeQueue][google.cloud.tasks.v2beta2.CloudTasks.PurgeQueue].

Queue
A queue is a container of related tasks. Queues are configured to manage how those tasks are dispatched. Configurable properties include rate limits, retry options, target types, and others.

RateLimits
Rate limits.

RenewLeaseRequest
Request message for renewing a lease using [RenewLease][google.cloud.tasks.v2beta2.CloudTasks.RenewLease].

ResumeQueueRequest
Request message for [ResumeQueue][google.cloud.tasks.v2beta2.CloudTasks.ResumeQueue].

RetryConfig
Retry config.

RunTaskRequest
Request message for forcing a task to run now using [RunTask][google.cloud.tasks.v2beta2.CloudTasks.RunTask].

Task
A unit of scheduled work.

TaskStatus
Status of the task.

UpdateQueueRequest
Request message for [UpdateQueue][google.cloud.tasks.v2beta2.CloudTasks.UpdateQueue].

Enums

HttpMethod
The HTTP method used to execute the task.