Module gapi_grpc::google::cloud::translation::v3beta1

Modules

batch_translate_metadata

create_glossary_metadata

delete_glossary_metadata

detect_language_request

glossary

glossary_input_config

input_config

output_config

translation_service_client
Generated client implementations.

Structs

BatchTranslateMetadata
State metadata for the batch translation operation.

BatchTranslateResponse
Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by BatchTranslateText if at least one sentence is translated successfully.

BatchTranslateTextRequest
The batch translation request.

CreateGlossaryMetadata
Stored in the [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata] field returned by CreateGlossary.

CreateGlossaryRequest
Request message for CreateGlossary.

DeleteGlossaryMetadata
Stored in the [google.longrunning.Operation.metadata][google.longrunning.Operation.metadata] field returned by DeleteGlossary.

DeleteGlossaryRequest
Request message for DeleteGlossary.

DeleteGlossaryResponse
Stored in the [google.longrunning.Operation.response][google.longrunning.Operation.response] field returned by DeleteGlossary.

DetectLanguageRequest
The request message for language detection.

DetectLanguageResponse
The response message for language detection.

DetectedLanguage
The response message for language detection.

GcsDestination
The Google Cloud Storage location for the output content.

GcsSource
The Google Cloud Storage location for the input content.

GetGlossaryRequest
Request message for GetGlossary.

GetSupportedLanguagesRequest
The request message for discovering supported languages.

Glossary
Represents a glossary built from user provided data.

GlossaryInputConfig
Input configuration for glossaries.

InputConfig
Input configuration for BatchTranslateText request.

ListGlossariesRequest
Request message for ListGlossaries.

ListGlossariesResponse
Response message for ListGlossaries.

OutputConfig
Output configuration for BatchTranslateText request.

SupportedLanguage
A single supported language response corresponds to information related to one supported language.

SupportedLanguages
The response message for discovering supported languages.

TranslateTextGlossaryConfig
Configures which glossary should be used for a specific target language, and defines options for applying that glossary.

TranslateTextRequest
The request message for synchronous translation.

TranslateTextResponse

Translation
A single translation response.