Module gapi_grpc::google::cloud::visualinspection::v1beta1

Modules

annotation

annotation_set

confusion_matrix

import_data_config

labeling_job

metrics_slice

module

solution

solution_artifact

visual_inspection_client
Generated client implementations.

Structs

ActiveLearningConfig
Configuration of active learning.

Annotation
Annotation is used to assign specific AnnotationSpec to a particular Image.

AnnotationInput
Specifies the source of Annotations for a Module input.

AnnotationOutput
Specifies the output for Annotations produced by a Module.

AnnotationSet
Resource of an AnnotationSet.

AnnotationSpec
Resource of an AnnotationSpec.

AnnotationSpecColor
Mapping of color to AnnotationSpec

AnomalyDetectionLabelingConfig
Labeling configuration of anomaly detection.

AnomalyDetectionModule
Detailed input and output information of an anomaly detection Module.

BatchPredictOperationMetadata
Operation metadata for BatchPreidct.

BatchPredictRequest
Request message for BatchPredict.

BatchPredictResponse

BoundingBoxAnnotation
Detailed information for BoundingBox Annotation.

BoundingBoxAnnotationSet
BoundingBox Annotation type.

BoundingPolyAnnotation
Detailed information for BoundingPoly Annotation.

CancelLabelingJobRequest
Request message for cancelling a labeling job.

ClassificationLabelAnnotation
Detailed information for ClassificationLabel Annotation.

ClassificationLabelAnnotationSet
ClassificationLabel Annotation type.

ClassificationModule
Detailed input and output information of an image classification module.

CommonOperationMetadata
Common operation metadata across Visual Inspection AI operations.

ConfusionMatrix
Confusion matrix across different class labels (if present).

CreateAnnotationSetRequest
Request for creating a AnnotationSet.

CreateAnnotationSpecRequest
Request for creating a AnnotationSpec.

CreateDatasetOperationMetadata
CreateDataset operation metadata.

CreateLabelingJobRequest
Request for creating a labeling job.

CreateModelOperationMetadata
Operation metadata for CreateModel.

CreateModelRequest
Request for creating a Model. This invokes the training process for the parent Module.

CreateModuleRequest
Request for creating a Module.

CreateSolutionArtifactOperationMetadata
CreateSolutionArtifact operation metadata.

CreateSolutionArtifactRequest
Request for creating a SolutionArtifact.

CreateSolutionRequest
Request for creating a solution.

CreateSpecialistPoolOperationMetadata
Operation metadata for CreateSpecialistPool.

CreateSpecialistPoolRequest
Request for creating a SpecialistPool.

Dataset
A Dataset is a collection of Images and Annotations, which could be used to solve a specifc visual inspection problem.

DatasetSource
Describes the source Dataset from where we import data.

DeleteAnnotationOperationMetadata
Operation metadata for DeleteAnnotation.

DeleteAnnotationSetOperationMetadata
Operation metadata for DeleteAnnotationSet.

DeleteAnnotationSpecOperationMetadata
Operation Metadata for DeleteAnnotationSpec.

DeleteDatasetOperationMetadata
DeleteDataset operation metadata.

DeleteDatasetRequest
Request for deleting a Dataset.

DeleteLabelingJobOperationMetadata
Operation metadata for deleting a labeling job.

DeleteLabelingJobRequest
Request for deleting a labeling job.

DeleteModelOperationMetadata
Operation metadata for DeleteModel.

DeleteModelRequest
Request for deleting a Model.

DeleteModuleOperationMetadata
Operation metadata for DeleteModule.

DeleteModuleRequest
Request for deleting a Module.

DeleteSolutionArtifactOperationMetadata
Operation metadata for DeleteSolutionArtifact.

DeleteSolutionArtifactRequest
Request for deleting a SolutionArtifact.

DeleteSpecialistPoolOperationMetadata
Operation metadata for DeleteSpecialistPool.

DeleteSpecialistPoolRequest
Request for deleting a specialist pool.

DeploySolutionOperationMetadata
Deploy solution metadata to distinguish that the operation is an export solution operation.

ExportDataConfig
Describes the location to where we export data from a Dataset.

ExportDataOperationMetadata
ExportData operation metadata.

ExportDataRequest
Request for exporting a Dataset.

ExportDataResponse

ExportSolutionOperationMetadata
Export solution metadata to distinguish that the operation is an export solution operation.

GcrDestination
The GCR location where the container will be pushed.

GcsDestination
The Google Cloud Storage location where the output is to be written to.

GcsSource
The Google Cloud Storage location for the input content.

GetDatasetRequest
Request for getting a Dataset.

GetImageRequest
Request for getting an Image.

GetLabelingJobRequest
Request for getting a labeling job.

GetModelEvaluationRequest
Request for getting a ModelEvaluation.

GetModelRequest
Request for getting a Model.

GetModuleRequest
Request for getting a Module.

GetSolutionArtifactRequest
Request for getting a SolutionArtifact.

GetSolutionRequest
For requesting details about a particular solution.

GetSpecialistPoolRequest
Request for getting a SpecialistPool.

Image
A piece of data in a Dataset.

ImportDataConfig
Describes the location from where we import data into a Dataset.

ImportDataOperationMetadata
ImportData operation metadata.

ImportDataRequest
Request for importing a Dataset.

ImportDataResponse

LabelingJob
Labeling job to trigger human labeling for adding annotations to images.

ListAnnotationSetsRequest
Request for listing AnnotationSets.

ListAnnotationSetsResponse
Response for ListAnnotationSets.

ListAnnotationSpecsRequest
Request for listing AnnotationSpecs.

ListAnnotationSpecsResponse
Response for ListAnnotationSpecs.

ListAnnotationsRequest
Request for listing Annotations.

ListAnnotationsResponse
Response for ListAnnotations.

ListDatasetsRequest
Request for listing Datasets.

ListDatasetsResponse
Response for ListDatasets.

ListImagesRequest
Request for listing Images.

ListImagesResponse
Response for ListImages.

ListLabelingJobsRequest
Request message for listing labeling jobs.

ListLabelingJobsResponse
Response message for listing labeling jobs.

ListModelEvaluationsRequest
Request for listing ModelEvaluations.

ListModelEvaluationsResponse
Response for ListModelEvaluations.

ListModelsRequest
Request for listing Models.

ListModelsResponse
Response for ListModels.

ListModulesRequest
Request for listing Modules.

ListModulesResponse
Response for ListModules.

ListSolutionArtifactsRequest
Request for listing SolutionArtifacts.

ListSolutionArtifactsResponse
Response for ListSolutionArtifacts.

ListSolutionsRequest
Request method for listing solutions.

ListSolutionsResponse
Response method for listing solutions.

ListSpecialistPoolsRequest
Request message for listing SpecialistPools.

ListSpecialistPoolsResponse
Response message for ListSpecialistPools.

LocalizationRefinementModule
Detailed input and output information of a localization refinment module.

MaskAnnotation
Detailed information for Mask Annotation.

MaskAnnotationSet
Mask Annotation type.

Metrics
All metrics for the solution.

MetricsSlice
Metrics at a particular slice. A slice can be "pixel" level, "image" level, "component" level, "unit" level, and/or "product" level depending on the solution and the input data.

Model
A Model is the trained artifact of a Module which makes predictions.

ModelEvaluation
ModelEvaluation for visual inspection solutions.

ModelMetrics
Model metrics which can vary depending on the model. These are not assumed to refer to defect states in particular.

Module
A Module applies a computer vision algorithm to a subproblem of a visual inspection task.

NormalizedBoundingPoly
Normalized Bounding Polygon.

NormalizedPolyline
Normalized Line, which represents a curve consisting of connected straight-line segments.

NormalizedVertex
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

ObjectDetectionModule
Detailed input and output information of an object detection module.

PolygonAnnotationSet
Polygon Annotation type.

PolylineAnnotation
Detailed information for Polyline Annotation.

PolylineAnnotationSet
Polyline Annotation type.

SegmentationModule
Detailed input and output information of an image segmentation module.

Solution
A solution is a set of models and algorithms combined together that will solve a specific visual inspection problem.

SolutionArtifact
A SolutionArtifact is a fixed set of Models which can be exported or deployed.

SolutionOperationMetadata
Solution operation metadata.

SparseAlignmentLocalizationModule
Detailed input and output information of a sparse alignment localization module.

SpecialistPool
SpecialistPool represents customers' own workforce to work on their data labeling jobs. It includes a group of specialist managers who are responsible for managing the labelers in this pool as well as customers' data labeling jobs associated with this pool. Customers create specialist pool as well as start data labeling jobs on Cloud, managers and labelers work with the jobs using CrowdCompute console.

UndeploySolutionOperationMetadata
Undeploy solution metadata to distinguish that the operation is an export solution operation.

UpdateModelRequest
Request for updating a Model.

UpdateModuleRequest
Request for updating a Module.

UpdateSolutionArtifactRequest
Request for updating a SolutionArtifact.

UpdateSpecialistPoolOperationMetadata
Operation metadata for UpdateSpecialistPool.

UpdateSpecialistPoolRequest
Request for updating a SpecialistPool.

WritePredictionsOperationMetadata
Operation metadata for WritePredictions.

WritePredictionsRequest
Request for enriching the Dataset with predicted Annotations from a Model.

WritePredictionsResponse