Struct gapi_grpc::google::cloud::visualinspection::v1beta1::visual_inspection_client::VisualInspectionClient

pub struct VisualInspectionClient<T> { /* fields omitted */ }

Visual Inspection service enables users to train and manage solutions for visually inspecting parts.

Implementations

impl VisualInspectionClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> VisualInspectionClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_solution<'_>(
&'_ mut self,
request: impl IntoRequest<CreateSolutionRequest>
) -> Result<Response<Operation>, Status>

Creates a solution.

pub async fn list_solutions<'_>(
&'_ mut self,
request: impl IntoRequest<ListSolutionsRequest>
) -> Result<Response<ListSolutionsResponse>, Status>

Lists solutions.

pub async fn get_solution<'_>(
&'_ mut self,
request: impl IntoRequest<GetSolutionRequest>
) -> Result<Response<Solution>, Status>

Gets a solution.

pub async fn create_solution_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<CreateSolutionArtifactRequest>
) -> Result<Response<Operation>, Status>

Creates a SolutionArtifact.

pub async fn list_solution_artifacts<'_>(
&'_ mut self,
request: impl IntoRequest<ListSolutionArtifactsRequest>
) -> Result<Response<ListSolutionArtifactsResponse>, Status>

Lists SolutionArtifacts in a Solution.

pub async fn get_solution_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<GetSolutionArtifactRequest>
) -> Result<Response<SolutionArtifact>, Status>

Gets a SolutionArtifact.

pub async fn delete_solution_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteSolutionArtifactRequest>
) -> Result<Response<Operation>, Status>

Deletes a SolutionArtifact. Only the subject SolutionArtifact is affected by this operation. The SolutionArtifact's Models remain unchanged.

pub async fn update_solution_artifact<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateSolutionArtifactRequest>
) -> Result<Response<SolutionArtifact>, Status>

Updates a SolutionArtifact.

pub async fn batch_predict<'_>(
&'_ mut self,
request: impl IntoRequest<BatchPredictRequest>
) -> Result<Response<Operation>, Status>

Perform batch prediction on a given SolutionArtifact.

pub async fn get_dataset<'_>(
&'_ mut self,
request: impl IntoRequest<GetDatasetRequest>
) -> Result<Response<Dataset>, Status>

Gets a Dataset.

pub async fn list_datasets<'_>(
&'_ mut self,
request: impl IntoRequest<ListDatasetsRequest>
) -> Result<Response<ListDatasetsResponse>, Status>

Lists Datasets in a Project.

pub async fn delete_dataset<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteDatasetRequest>
) -> Result<Response<Operation>, Status>

Deletes a Dataset.

pub async fn import_data<'_>(
&'_ mut self,
request: impl IntoRequest<ImportDataRequest>
) -> Result<Response<Operation>, Status>

Imports data to a Dataset.

pub async fn export_data<'_>(
&'_ mut self,
request: impl IntoRequest<ExportDataRequest>
) -> Result<Response<Operation>, Status>

Exports data from a Dataset.

pub async fn get_image<'_>(
&'_ mut self,
request: impl IntoRequest<GetImageRequest>
) -> Result<Response<Image>, Status>

Gets an Image. ImportData will create images. Therefore CreateImage is omitted.

pub async fn list_images<'_>(
&'_ mut self,
request: impl IntoRequest<ListImagesRequest>
) -> Result<Response<ListImagesResponse>, Status>

Lists Images in a Dataset.

pub async fn create_annotation_set<'_>(
&'_ mut self,
request: impl IntoRequest<CreateAnnotationSetRequest>
) -> Result<Response<AnnotationSet>, Status>

Creates an AnnotationSet.

pub async fn list_annotation_sets<'_>(
&'_ mut self,
request: impl IntoRequest<ListAnnotationSetsRequest>
) -> Result<Response<ListAnnotationSetsResponse>, Status>

Lists AnnotationSets in a Dataset.

pub async fn create_annotation_spec<'_>(
&'_ mut self,
request: impl IntoRequest<CreateAnnotationSpecRequest>
) -> Result<Response<AnnotationSpec>, Status>

Creates an AnnotationSpec.

pub async fn list_annotation_specs<'_>(
&'_ mut self,
request: impl IntoRequest<ListAnnotationSpecsRequest>
) -> Result<Response<ListAnnotationSpecsResponse>, Status>

Lists AnnotationSpecs in a dataset.

pub async fn list_annotations<'_>(
&'_ mut self,
request: impl IntoRequest<ListAnnotationsRequest>
) -> Result<Response<ListAnnotationsResponse>, Status>

Lists Annotations on an Image.

pub async fn create_module<'_>(
&'_ mut self,
request: impl IntoRequest<CreateModuleRequest>
) -> Result<Response<Module>, Status>

Creates a Module.

pub async fn list_modules<'_>(
&'_ mut self,
request: impl IntoRequest<ListModulesRequest>
) -> Result<Response<ListModulesResponse>, Status>

Lists Modules in a Solution.

pub async fn get_module<'_>(
&'_ mut self,
request: impl IntoRequest<GetModuleRequest>
) -> Result<Response<Module>, Status>

Gets a Module.

pub async fn delete_module<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteModuleRequest>
) -> Result<Response<Operation>, Status>

Deletes a Module. Only the subject Module is affected by this operation. The Module's input and output AnnotationSets and any connected Modules remain unchanged.

pub async fn update_module<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateModuleRequest>
) -> Result<Response<Module>, Status>

Updates a Module.

pub async fn create_model<'_>(
&'_ mut self,
request: impl IntoRequest<CreateModelRequest>
) -> Result<Response<Operation>, Status>

Creates a Model.

pub async fn list_models<'_>(
&'_ mut self,
request: impl IntoRequest<ListModelsRequest>
) -> Result<Response<ListModelsResponse>, Status>

Lists Models for a Module.

pub async fn get_model<'_>(
&'_ mut self,
request: impl IntoRequest<GetModelRequest>
) -> Result<Response<Model>, Status>

Gets a Model.

pub async fn delete_model<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteModelRequest>
) -> Result<Response<Operation>, Status>

Deletes a Model.

pub async fn update_model<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateModelRequest>
) -> Result<Response<Model>, Status>

Updates a Model.

pub async fn list_model_evaluations<'_>(
&'_ mut self,
request: impl IntoRequest<ListModelEvaluationsRequest>
) -> Result<Response<ListModelEvaluationsResponse>, Status>

Lists ModelEvaluations for a Model. For Beta, we will have only one evaluation produced when the Model is created.

pub async fn get_model_evaluation<'_>(
&'_ mut self,
request: impl IntoRequest<GetModelEvaluationRequest>
) -> Result<Response<ModelEvaluation>, Status>

Gets a ModelEvaluation.

pub async fn write_predictions<'_>(
&'_ mut self,
request: impl IntoRequest<WritePredictionsRequest>
) -> Result<Response<Operation>, Status>

Writes a Model's predictions to the parent Module's output AnnotationSets.

pub async fn create_specialist_pool<'_>(
&'_ mut self,
request: impl IntoRequest<CreateSpecialistPoolRequest>
) -> Result<Response<Operation>, Status>

Creates a SpecialistPool.

pub async fn get_specialist_pool<'_>(
&'_ mut self,
request: impl IntoRequest<GetSpecialistPoolRequest>
) -> Result<Response<SpecialistPool>, Status>

Gets a SpecialistPool.

pub async fn update_specialist_pool<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateSpecialistPoolRequest>
) -> Result<Response<Operation>, Status>

Updates a SpecialistPool.

pub async fn list_specialist_pools<'_>(
&'_ mut self,
request: impl IntoRequest<ListSpecialistPoolsRequest>
) -> Result<Response<ListSpecialistPoolsResponse>, Status>

Lists SpecialistPools.

pub async fn delete_specialist_pool<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteSpecialistPoolRequest>
) -> Result<Response<Operation>, Status>

Deletes a SpecialistPool.

pub async fn create_labeling_job<'_>(
&'_ mut self,
request: impl IntoRequest<CreateLabelingJobRequest>
) -> Result<Response<LabelingJob>, Status>

Creates a LabelingJob.

pub async fn get_labeling_job<'_>(
&'_ mut self,
request: impl IntoRequest<GetLabelingJobRequest>
) -> Result<Response<LabelingJob>, Status>

Gets a LabelingJob.

pub async fn list_labeling_jobs<'_>(
&'_ mut self,
request: impl IntoRequest<ListLabelingJobsRequest>
) -> Result<Response<ListLabelingJobsResponse>, Status>

Lists LabelingJobs.

pub async fn delete_labeling_job<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteLabelingJobRequest>
) -> Result<Response<Operation>, Status>

Deletes a LabelingJob.

pub async fn cancel_labeling_job<'_>(
&'_ mut self,
request: impl IntoRequest<CancelLabelingJobRequest>
) -> Result<Response<()>, Status>

Cancels a LabelingJob. Success of cancellation is not guaranteed.