Struct gapi_grpc::google::container::v1alpha1::cluster_manager_client::ClusterManagerClient

pub struct ClusterManagerClient<T> { /* fields omitted */ }

Google Container Engine Cluster Manager v1alpha1

Implementations

impl ClusterManagerClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> ClusterManagerClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn list_clusters(
&mut self,
request: impl IntoRequest<ListClustersRequest>
) -> Result<Response<ListClustersResponse>, Status>

Lists all clusters owned by a project in either the specified zone or all zones.

pub async fn get_cluster(
&mut self,
request: impl IntoRequest<GetClusterRequest>
) -> Result<Response<Cluster>, Status>

Gets the details of a specific cluster.

pub async fn create_cluster(
&mut self,
request: impl IntoRequest<CreateClusterRequest>
) -> Result<Response<Operation>, Status>

Creates a cluster, consisting of the specified number and type of Google Compute Engine instances.

By default, the cluster is created in the project's default network.

One firewall is added for the cluster. After cluster creation, the cluster creates routes for each node to allow the containers on that node to communicate with all other instances in the cluster.

Finally, an entry is added to the project's global metadata indicating which CIDR range is being used by the cluster.

pub async fn update_cluster(
&mut self,
request: impl IntoRequest<UpdateClusterRequest>
) -> Result<Response<Operation>, Status>

Updates the settings of a specific cluster.

pub async fn update_node_pool(
&mut self,
request: impl IntoRequest<UpdateNodePoolRequest>
) -> Result<Response<Operation>, Status>

Updates the version and/or iamge type of a specific node pool.

pub async fn set_node_pool_autoscaling(
&mut self,
request: impl IntoRequest<SetNodePoolAutoscalingRequest>
) -> Result<Response<Operation>, Status>

Sets the autoscaling settings of a specific node pool.

pub async fn set_logging_service(
&mut self,
request: impl IntoRequest<SetLoggingServiceRequest>
) -> Result<Response<Operation>, Status>

Sets the logging service of a specific cluster.

pub async fn set_monitoring_service(
&mut self,
request: impl IntoRequest<SetMonitoringServiceRequest>
) -> Result<Response<Operation>, Status>

Sets the monitoring service of a specific cluster.

pub async fn set_addons_config(
&mut self,
request: impl IntoRequest<SetAddonsConfigRequest>
) -> Result<Response<Operation>, Status>

Sets the addons of a specific cluster.

pub async fn set_locations(
&mut self,
request: impl IntoRequest<SetLocationsRequest>
) -> Result<Response<Operation>, Status>

Sets the locations of a specific cluster.

pub async fn update_master(
&mut self,
request: impl IntoRequest<UpdateMasterRequest>
) -> Result<Response<Operation>, Status>

Updates the master of a specific cluster.

pub async fn set_master_auth(
&mut self,
request: impl IntoRequest<SetMasterAuthRequest>
) -> Result<Response<Operation>, Status>

Used to set master auth materials. Currently supports :- Changing the admin password of a specific cluster. This can be either via password generation or explicitly set. Modify basic_auth.csv and reset the K8S API server.

pub async fn delete_cluster(
&mut self,
request: impl IntoRequest<DeleteClusterRequest>
) -> Result<Response<Operation>, Status>

Deletes the cluster, including the Kubernetes endpoint and all worker nodes.

Firewalls and routes that were configured during cluster creation are also deleted.

Other Google Compute Engine resources that might be in use by the cluster (e.g. load balancer resources) will not be deleted if they weren't present at the initial create time.

pub async fn list_operations(
&mut self,
request: impl IntoRequest<ListOperationsRequest>
) -> Result<Response<ListOperationsResponse>, Status>

Lists all operations in a project in a specific zone or all zones.

pub async fn get_operation(
&mut self,
request: impl IntoRequest<GetOperationRequest>
) -> Result<Response<Operation>, Status>

Gets the specified operation.

pub async fn cancel_operation(
&mut self,
request: impl IntoRequest<CancelOperationRequest>
) -> Result<Response<()>, Status>

Cancels the specified operation.

pub async fn get_server_config(
&mut self,
request: impl IntoRequest<GetServerConfigRequest>
) -> Result<Response<ServerConfig>, Status>

Returns configuration info about the Container Engine service.

pub async fn list_node_pools(
&mut self,
request: impl IntoRequest<ListNodePoolsRequest>
) -> Result<Response<ListNodePoolsResponse>, Status>

Lists the node pools for a cluster.

pub async fn get_node_pool(
&mut self,
request: impl IntoRequest<GetNodePoolRequest>
) -> Result<Response<NodePool>, Status>

Retrieves the node pool requested.

pub async fn create_node_pool(
&mut self,
request: impl IntoRequest<CreateNodePoolRequest>
) -> Result<Response<Operation>, Status>

Creates a node pool for a cluster.

pub async fn delete_node_pool(
&mut self,
request: impl IntoRequest<DeleteNodePoolRequest>
) -> Result<Response<Operation>, Status>

Deletes a node pool from a cluster.

pub async fn rollback_node_pool_upgrade(
&mut self,
request: impl IntoRequest<RollbackNodePoolUpgradeRequest>
) -> Result<Response<Operation>, Status>

Roll back the previously Aborted or Failed NodePool upgrade. This will be an no-op if the last upgrade successfully completed.

pub async fn set_node_pool_management(
&mut self,
request: impl IntoRequest<SetNodePoolManagementRequest>
) -> Result<Response<Operation>, Status>

Sets the NodeManagement options for a node pool.

pub async fn set_labels(
&mut self,
request: impl IntoRequest<SetLabelsRequest>
) -> Result<Response<Operation>, Status>

Sets labels on a cluster.

pub async fn set_legacy_abac(
&mut self,
request: impl IntoRequest<SetLegacyAbacRequest>
) -> Result<Response<Operation>, Status>

Enables or disables the ABAC authorization mechanism on a cluster.

pub async fn start_ip_rotation(
&mut self,
request: impl IntoRequest<StartIpRotationRequest>
) -> Result<Response<Operation>, Status>

Start master IP rotation.

pub async fn complete_ip_rotation(
&mut self,
request: impl IntoRequest<CompleteIpRotationRequest>
) -> Result<Response<Operation>, Status>

Completes master IP rotation.

pub async fn set_node_pool_size(
&mut self,
request: impl IntoRequest<SetNodePoolSizeRequest>
) -> Result<Response<Operation>, Status>

Sets the size of a specific node pool.

pub async fn set_network_policy(
&mut self,
request: impl IntoRequest<SetNetworkPolicyRequest>
) -> Result<Response<Operation>, Status>

Enables/Disables Network Policy for a cluster.

pub async fn set_maintenance_policy(
&mut self,
request: impl IntoRequest<SetMaintenancePolicyRequest>
) -> Result<Response<Operation>, Status>

Sets the maintenance policy for a cluster.