Module gapi_grpc::google::devtools::clouddebugger::v2::controller2_client

Generated client implementations.

Structs

Controller2Client
The Controller service provides the API for orchestrating a collection of debugger agents to perform debugging tasks. These agents are each attached to a process of an application which may include one or more replicas.