Module gapi_grpc::google::devtools::cloudprofiler::v2

Modules

profiler_service_client
Generated client implementations.

Structs

CreateOfflineProfileRequest
CreateOfflineProfileRequest describes a profile resource offline creation request. Profile field must be set.

CreateProfileRequest
CreateProfileRequest describes a profile resource online creation request. The deployment field must be populated. The profile_type specifies the list of profile types supported by the agent. The creation call will hang until a profile of one of these types needs to be collected.

Deployment
Deployment contains the deployment identification information.

Profile
Profile resource.

UpdateProfileRequest
UpdateProfileRequest contains the profile to update.

Enums

ProfileType
ProfileType is type of profiling data. NOTE: the enumeration member names are used (in lowercase) as unique string identifiers of profile types, so they must not be renamed.