Module gapi_grpc::google::devtools::containeranalysis::v1beta1

Modules

container_analysis_v1_beta1_client
Generated client implementations.

Structs

GetScanConfigRequest
Request to get a scan configuration.

ListScanConfigsRequest
Request to list scan configurations.

ListScanConfigsResponse
Response for listing scan configurations.

ScanConfig
A scan configuration specifies whether Cloud components in a project have a particular type of analysis being run. For example, it can configure whether vulnerability scanning is being done on Docker images or not.

UpdateScanConfigRequest
A request to update a scan configuration.