Module gapi_grpc::google::devtools::sourcerepo::v1

Modules

source_repo_client
Generated client implementations.

Structs

CreateRepoRequest
Request for CreateRepo

DeleteRepoRequest
Request for DeleteRepo.

GetRepoRequest
Request for GetRepo.

ListReposRequest
Request for ListRepos.

ListReposResponse
Response for ListRepos. The size is not set in the returned repositories.

MirrorConfig
Configuration to automatically mirror a repository from another hosting service, for example GitHub or BitBucket.

Repo
A repository (or repo) is a Git repository storing versioned source content.