Struct gapi_grpc::google::firestore::admin::v1beta1::firestore_admin_client::FirestoreAdminClient

pub struct FirestoreAdminClient<T> { /* fields omitted */ }

The Cloud Firestore Admin API.

This API provides several administrative services for Cloud Firestore.

Concepts

Project, Database, Namespace, Collection, and Document are used as defined in the Google Cloud Firestore API.

Operation: An Operation represents work being performed in the background.

Services

Index

The index service manages Cloud Firestore indexes.

Index creation is performed asynchronously. An Operation resource is created for each such asynchronous operation. The state of the operation (including any errors encountered) may be queried via the Operation resource.

Metadata

Provides metadata and statistical information about data in Cloud Firestore. The data provided as part of this API may be stale.

Operation

The Operations collection provides a record of actions performed for the specified Project (including any Operations in progress). Operations are not created directly but through calls on other collections or resources.

An Operation that is not yet done may be cancelled. The request to cancel is asynchronous and the Operation may continue to run for some time after the request to cancel is made.

An Operation that is done may be deleted so that it is no longer listed as part of the Operation collection.

Operations are created by service FirestoreAdmin, but are accessed via service google.longrunning.Operations.

Implementations

impl FirestoreAdminClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> FirestoreAdminClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn create_index(
&mut self,
request: impl IntoRequest<CreateIndexRequest>
) -> Result<Response<Operation>, Status>

Creates the specified index. A newly created index's initial state is CREATING. On completion of the returned [google.longrunning.Operation][google.longrunning.Operation], the state will be READY. If the index already exists, the call will return an ALREADY_EXISTS status.

During creation, the process could result in an error, in which case the index will move to the ERROR state. The process can be recovered by fixing the data that caused the error, removing the index with [delete][google.firestore.admin.v1beta1.FirestoreAdmin.DeleteIndex], then re-creating the index with [create][google.firestore.admin.v1beta1.FirestoreAdmin.CreateIndex].

Indexes with a single field cannot be created.

pub async fn list_indexes(
&mut self,
request: impl IntoRequest<ListIndexesRequest>
) -> Result<Response<ListIndexesResponse>, Status>

Lists the indexes that match the specified filters.

pub async fn get_index(
&mut self,
request: impl IntoRequest<GetIndexRequest>
) -> Result<Response<Index>, Status>

Gets an index.

pub async fn delete_index(
&mut self,
request: impl IntoRequest<DeleteIndexRequest>
) -> Result<Response<()>, Status>

Deletes an index.

pub async fn export_documents(
&mut self,
request: impl IntoRequest<ExportDocumentsRequest>
) -> Result<Response<Operation>, Status>

Exports a copy of all or a subset of documents from Google Cloud Firestore to another storage system, such as Google Cloud Storage. Recent updates to documents may not be reflected in the export. The export occurs in the background and its progress can be monitored and managed via the Operation resource that is created. The output of an export may only be used once the associated operation is done. If an export operation is cancelled before completion it may leave partial data behind in Google Cloud Storage.

pub async fn import_documents(
&mut self,
request: impl IntoRequest<ImportDocumentsRequest>
) -> Result<Response<Operation>, Status>

Imports documents into Google Cloud Firestore. Existing documents with the same name are overwritten. The import occurs in the background and its progress can be monitored and managed via the Operation resource that is created. If an ImportDocuments operation is cancelled, it is possible that a subset of the data has already been imported to Cloud Firestore.