Module gapi_grpc::google::firestore::admin::v1beta2

Modules

field

field_operation_metadata

firestore_admin_client
Generated client implementations.

index

Structs

CreateIndexRequest
The request for [FirestoreAdmin.CreateIndex][google.firestore.admin.v1beta2.FirestoreAdmin.CreateIndex].

DeleteIndexRequest
The request for [FirestoreAdmin.DeleteIndex][google.firestore.admin.v1beta2.FirestoreAdmin.DeleteIndex].

ExportDocumentsMetadata
Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.ExportDocuments][google.firestore.admin.v1beta2.FirestoreAdmin.ExportDocuments].

ExportDocumentsRequest
The request for [FirestoreAdmin.ExportDocuments][google.firestore.admin.v1beta2.FirestoreAdmin.ExportDocuments].

ExportDocumentsResponse
Returned in the [google.longrunning.Operation][google.longrunning.Operation] response field.

Field
Represents a single field in the database.

FieldOperationMetadata
Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.UpdateField][google.firestore.admin.v1beta2.FirestoreAdmin.UpdateField].

GetFieldRequest
The request for [FirestoreAdmin.GetField][google.firestore.admin.v1beta2.FirestoreAdmin.GetField].

GetIndexRequest
The request for [FirestoreAdmin.GetIndex][google.firestore.admin.v1beta2.FirestoreAdmin.GetIndex].

ImportDocumentsMetadata
Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.ImportDocuments][google.firestore.admin.v1beta2.FirestoreAdmin.ImportDocuments].

ImportDocumentsRequest
The request for [FirestoreAdmin.ImportDocuments][google.firestore.admin.v1beta2.FirestoreAdmin.ImportDocuments].

Index
Cloud Firestore indexes enable simple and complex queries against documents in a database.

IndexOperationMetadata
Metadata for [google.longrunning.Operation][google.longrunning.Operation] results from [FirestoreAdmin.CreateIndex][google.firestore.admin.v1beta2.FirestoreAdmin.CreateIndex].

ListFieldsRequest
The request for [FirestoreAdmin.ListFields][google.firestore.admin.v1beta2.FirestoreAdmin.ListFields].

ListFieldsResponse
The response for [FirestoreAdmin.ListFields][google.firestore.admin.v1beta2.FirestoreAdmin.ListFields].

ListIndexesRequest
The request for [FirestoreAdmin.ListIndexes][google.firestore.admin.v1beta2.FirestoreAdmin.ListIndexes].

ListIndexesResponse
The response for [FirestoreAdmin.ListIndexes][google.firestore.admin.v1beta2.FirestoreAdmin.ListIndexes].

Progress
Describes the progress of the operation. Unit of work is generic and must be interpreted based on where [Progress][google.firestore.admin.v1beta2.Progress] is used.

UpdateFieldRequest
The request for [FirestoreAdmin.UpdateField][google.firestore.admin.v1beta2.FirestoreAdmin.UpdateField].

Enums

OperationState
Describes the state of the operation.