Module gapi_grpc::google::iam::admin::v1

Modules

audit_data

iam_client
Generated client implementations.

list_service_account_keys_request

permission

role

Structs

AuditData
Audit log information specific to Cloud IAM admin APIs. This message is serialized as an Any type in the ServiceData message of an AuditLog message.

CreateRoleRequest
The request to create a new role.

CreateServiceAccountKeyRequest
The service account key create request.

CreateServiceAccountRequest
The service account create request.

DeleteRoleRequest
The request to delete an existing role.

DeleteServiceAccountKeyRequest
The service account key delete request.

DeleteServiceAccountRequest
The service account delete request.

GetRoleRequest
The request to get the definition of an existing role.

GetServiceAccountKeyRequest
The service account key get by id request.

GetServiceAccountRequest
The service account get request.

ListRolesRequest
The request to get all roles defined under a resource.

ListRolesResponse
The response containing the roles defined under a resource.

ListServiceAccountKeysRequest
The service account keys list request.

ListServiceAccountKeysResponse
The service account keys list response.

ListServiceAccountsRequest
The service account list request.

ListServiceAccountsResponse
The service account list response.

Permission
A permission which can be included by a role.

QueryGrantableRolesRequest
The grantable role query request.

QueryGrantableRolesResponse
The grantable role query response.

QueryTestablePermissionsRequest
A request to get permissions which can be tested on a resource.

QueryTestablePermissionsResponse
The response containing permissions which can be tested on a resource.

Role
A role in the Identity and Access Management API.

ServiceAccount
A service account in the Identity and Access Management API.

ServiceAccountKey
Represents a service account key.

SignBlobRequest
The service account sign blob request.

SignBlobResponse
The service account sign blob response.

SignJwtRequest
The service account sign JWT request.

SignJwtResponse
The service account sign JWT response.

UndeleteRoleRequest
The request to undelete an existing role.

UpdateRoleRequest
The request to update a role.

Enums

RoleView
A view for Role objects.

ServiceAccountKeyAlgorithm
Supported key algorithms.

ServiceAccountPrivateKeyType
Supported private key output formats.

ServiceAccountPublicKeyType
Supported public key output formats.