Module gapi_grpc::google::monitoring::dashboard::v1

Modules

aggregation

chart_options

column_layout

dashboard

dashboards_service_client
Generated client implementations.

pick_time_series_filter

row_layout

scorecard

statistical_time_series_filter

text

threshold

time_series_filter

time_series_filter_ratio

time_series_query

widget

xy_chart

Structs

Aggregation
Describes how to combine multiple time series to provide a different view of the data. Aggregation of time series is done in two steps. First, each time series in the set is aligned to the same time interval boundaries, then the set of time series is optionally reduced in number.

ChartOptions
Options to control visual rendering of a chart.

ColumnLayout
A simplified layout that divides the available space into vertical columns and arranges a set of widgets vertically in each column.

CreateDashboardRequest
The CreateDashboard request.

Dashboard
A Google Stackdriver dashboard. Dashboards define the content and layout of pages in the Stackdriver web application.

DeleteDashboardRequest
The DeleteDashboard request.

GetDashboardRequest
The GetDashboard request.

GridLayout
A basic layout divides the available space into vertical columns of equal width and arranges a list of widgets using a row-first strategy.

ListDashboardsRequest
The ListDashboards request.

ListDashboardsResponse
The ListDashboards request.

PickTimeSeriesFilter
Describes a ranking-based time series filter. Each input time series is ranked with an aligner. The filter will allow up to num_time_series time series to pass through it, selecting them based on the relative ranking.

RowLayout
A simplified layout that divides the available space into rows and arranges a set of widgets horizontally in each row.

Scorecard
A widget showing the latest value of a metric, and how this value relates to one or more thresholds.

StatisticalTimeSeriesFilter
A filter that ranks streams based on their statistical relation to other streams in a request. Note: This field is deprecated and completely ignored by the API.

Text
A widget that displays textual content.

Threshold
Defines a threshold for categorizing time series values.

TimeSeriesFilter
A filter that defines a subset of time series data that is displayed in a widget. Time series data is fetched using the ListTimeSeries method.

TimeSeriesFilterRatio
A pair of time series filters that define a ratio computation. The output time series is the pair-wise division of each aligned element from the numerator and denominator time series.

TimeSeriesQuery
TimeSeriesQuery collects the set of supported methods for querying time series data from the Stackdriver metrics API.

UpdateDashboardRequest
The UpdateDashboard request.

Widget
Widget contains a single dashboard component and configuration of how to present the component in the dashboard.

XyChart
A chart that displays data on a 2D (X and Y axes) plane.

Enums

SparkChartType
Defines the possible types of spark chart supported by the Scorecard.