Module gapi_grpc::google::monitoring::dashboard::v1::threshold

Enums

Color
The color suggests an interpretation to the viewer when actual values cross the threshold. Comments on each color provide UX guidance on how users can be expected to interpret a given state color.

Direction
Whether the threshold is considered crossed by an actual value above or below its threshold value.