Module gapi_grpc::google::monitoring::v3::alert_policy::condition

Modules

trigger

Structs

MetricAbsence
A condition type that checks that monitored resources are reporting data. The configuration defines a metric and a set of monitored resources. The predicate is considered in violation when a time series for the specified metric of a monitored resource does not include any data in the specified duration.

MetricThreshold
A condition type that compares a collection of time series against a threshold.

Trigger
Specifies how many time series must fail a predicate to trigger a condition. If not specified, then a {count: 1} trigger is used.

Enums

Condition
Only one of the following condition types will be specified.