Module gapi_grpc::google::privacy::dlp::v2::analyze_data_source_risk_details

Modules

categorical_stats_result

delta_presence_estimation_result

k_anonymity_result

k_map_estimation_result

l_diversity_result

Structs

CategoricalStatsResult
Result of the categorical stats computation.

DeltaPresenceEstimationResult
Result of the δ-presence computation. Note that these results are an estimation, not exact values.

KAnonymityResult
Result of the k-anonymity computation.

KMapEstimationResult
Result of the reidentifiability analysis. Note that these results are an estimation, not exact values.

LDiversityResult
Result of the l-diversity computation.

NumericalStatsResult
Result of the numerical stats computation.

RequestedRiskAnalysisOptions
Risk analysis options.

Enums

Result
Values associated with this metric.