Module gapi_grpc::google::privacy::dlp::v2::job_trigger

Modules

trigger

Structs

Trigger
What event needs to occur for a new job to be started.

Enums

Job
The configuration details for the specific type of job to run.

Status
Whether the trigger is currently active. If PAUSED or CANCELLED, no jobs will be created with this configuration. The service may automatically pause triggers experiencing frequent errors. To restart a job, set the status to HEALTHY after correcting user errors.