Module gapi_grpc::google::spanner::admin::instance::v1

Modules

instance

instance_admin_client
Generated client implementations.

replica_info

Structs

CreateInstanceMetadata
Metadata type for the operation returned by [CreateInstance][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstance].

CreateInstanceRequest
The request for [CreateInstance][google.spanner.admin.instance.v1.InstanceAdmin.CreateInstance].

DeleteInstanceRequest
The request for [DeleteInstance][google.spanner.admin.instance.v1.InstanceAdmin.DeleteInstance].

GetInstanceConfigRequest
The request for [GetInstanceConfigRequest][google.spanner.admin.instance.v1.InstanceAdmin.GetInstanceConfig].

GetInstanceRequest
The request for [GetInstance][google.spanner.admin.instance.v1.InstanceAdmin.GetInstance].

Instance
An isolated set of Cloud Spanner resources on which databases can be hosted.

InstanceConfig
A possible configuration for a Cloud Spanner instance. Configurations define the geographic placement of nodes and their replication.

ListInstanceConfigsRequest
The request for [ListInstanceConfigs][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigs].

ListInstanceConfigsResponse
The response for [ListInstanceConfigs][google.spanner.admin.instance.v1.InstanceAdmin.ListInstanceConfigs].

ListInstancesRequest
The request for [ListInstances][google.spanner.admin.instance.v1.InstanceAdmin.ListInstances].

ListInstancesResponse
The response for [ListInstances][google.spanner.admin.instance.v1.InstanceAdmin.ListInstances].

ReplicaInfo

UpdateInstanceMetadata
Metadata type for the operation returned by [UpdateInstance][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstance].

UpdateInstanceRequest
The request for [UpdateInstance][google.spanner.admin.instance.v1.InstanceAdmin.UpdateInstance].