Module gapi_grpc::google::storagetransfer::v1

Modules

storage_transfer_service_client
Generated client implementations.

transfer_job

transfer_operation

transfer_spec

Structs

AwsAccessKey
AWS access key (see AWS Security Credentials).

AwsS3Data
An AwsS3Data can be a data source, but not a data sink. In an AwsS3Data, an object's name is the S3 object's key name.

CreateTransferJobRequest
Request passed to CreateTransferJob.

ErrorLogEntry
An entry describing an error that has occurred.

ErrorSummary
A summary of errors by error code, plus a count and sample error log entries.

GcsData
In a GcsData, an object's name is the Google Cloud Storage object's name and its lastModificationTime refers to the object's updated time, which changes when the content or the metadata of the object is updated.

GetGoogleServiceAccountRequest
Request passed to GetGoogleServiceAccount.

GetTransferJobRequest
Request passed to GetTransferJob.

GoogleServiceAccount
Google service account

HttpData
An HttpData specifies a list of objects on the web to be transferred over HTTP. The information of the objects to be transferred is contained in a file referenced by a URL. The first line in the file must be "TsvHttpData-1.0", which specifies the format of the file. Subsequent lines specify the information of the list of objects, one object per list entry. Each entry has the following tab-delimited fields:

ListTransferJobsRequest
project_id, job_names, and job_statuses are query parameters that can be specified when listing transfer jobs.

ListTransferJobsResponse
Response from ListTransferJobs.

ObjectConditions
Conditions that determine which objects will be transferred.

PauseTransferOperationRequest
Request passed to PauseTransferOperation.

ResumeTransferOperationRequest
Request passed to ResumeTransferOperation.

Schedule
Transfers can be scheduled to recur or to run just once.

TransferCounters
A collection of counters that report the progress of a transfer operation.

TransferJob
This resource represents the configuration of a transfer job that runs periodically.

TransferOperation
A description of the execution of a transfer.

TransferOptions
TransferOptions uses three boolean parameters to define the actions to be performed on objects in a transfer.

TransferSpec
Configuration for running a transfer.

UpdateTransferJobRequest
Request passed to UpdateTransferJob.