Module gapi_grpc::grafeas::v1beta1

Modules

attestation

build

deployment

discovery

grafeas_v1_beta1_client
Generated client implementations.

image

note

occurrence

package

provenance

source

vulnerability

vulnerability_occurrences_summary

Structs

BatchCreateNotesRequest
Request to create notes in batch.

BatchCreateNotesResponse
Response for creating notes in batch.

BatchCreateOccurrencesRequest
Request to create occurrences in batch.

BatchCreateOccurrencesResponse
Response for creating occurrences in batch.

CreateNoteRequest
Request to create a new note.

CreateOccurrenceRequest
Request to create a new occurrence.

DeleteNoteRequest
Request to delete a note.

DeleteOccurrenceRequest
Request to delete a occurrence.

GetNoteRequest
Request to get a note.

GetOccurrenceNoteRequest
Request to get the note to which the specified occurrence is attached.

GetOccurrenceRequest
Request to get an occurrence.

GetVulnerabilityOccurrencesSummaryRequest
Request to get a vulnerability summary for some set of occurrences.

ListNoteOccurrencesRequest
Request to list occurrences for a note.

ListNoteOccurrencesResponse
Response for listing occurrences for a note.

ListNotesRequest
Request to list notes.

ListNotesResponse
Response for listing notes.

ListOccurrencesRequest
Request to list occurrences.

ListOccurrencesResponse
Response for listing occurrences.

Note
A type of analysis that can be done for a resource.

Occurrence
An instance of an analysis type that has been found on a resource.

RelatedUrl
Metadata for any related URL information.

Resource
An entity that can have metadata. For example, a Docker image.

Signature
Verifiers (e.g. Kritis implementations) MUST verify signatures with respect to the trust anchors defined in policy (e.g. a Kritis policy). Typically this means that the verifier has been configured with a map from public_key_id to public key material (and any required parameters, e.g. signing algorithm).

UpdateNoteRequest
Request to update a note.

UpdateOccurrenceRequest
Request to update an occurrence.

VulnerabilityOccurrencesSummary
A summary of how many vulnerability occurrences there are per resource and severity type.

Enums

NoteKind
Kind represents the kinds of notes supported.