Struct gapi_grpc::grafeas::v1beta1::grafeas_v1_beta1_client::GrafeasV1Beta1Client

pub struct GrafeasV1Beta1Client<T> { /* fields omitted */ }

Grafeas API.

Retrieves analysis results of Cloud components such as Docker container images.

Analysis results are stored as a series of occurrences. An Occurrence contains information about a specific analysis instance on a resource. An occurrence refers to a Note. A note contains details describing the analysis and is generally stored in a separate project, called a Provider. Multiple occurrences can refer to the same note.

For example, an SSL vulnerability could affect multiple images. In this case, there would be one note for the vulnerability and an occurrence for each image with the vulnerability referring to that note.

Methods

impl GrafeasV1Beta1Client<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> GrafeasV1Beta1Client<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn get_occurrence<'_>(
&'_ mut self,
request: impl IntoRequest<GetOccurrenceRequest>
) -> Result<Response<Occurrence>, Status>

Gets the specified occurrence.

pub async fn list_occurrences<'_>(
&'_ mut self,
request: impl IntoRequest<ListOccurrencesRequest>
) -> Result<Response<ListOccurrencesResponse>, Status>

Lists occurrences for the specified project.

pub async fn delete_occurrence<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteOccurrenceRequest>
) -> Result<Response<()>, Status>

Deletes the specified occurrence. For example, use this method to delete an occurrence when the occurrence is no longer applicable for the given resource.

pub async fn create_occurrence<'_>(
&'_ mut self,
request: impl IntoRequest<CreateOccurrenceRequest>
) -> Result<Response<Occurrence>, Status>

Creates a new occurrence.

pub async fn batch_create_occurrences<'_>(
&'_ mut self,
request: impl IntoRequest<BatchCreateOccurrencesRequest>
) -> Result<Response<BatchCreateOccurrencesResponse>, Status>

Creates new occurrences in batch.

pub async fn update_occurrence<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateOccurrenceRequest>
) -> Result<Response<Occurrence>, Status>

Updates the specified occurrence.

pub async fn get_occurrence_note<'_>(
&'_ mut self,
request: impl IntoRequest<GetOccurrenceNoteRequest>
) -> Result<Response<Note>, Status>

Gets the note attached to the specified occurrence. Consumer projects can use this method to get a note that belongs to a provider project.

pub async fn get_note<'_>(
&'_ mut self,
request: impl IntoRequest<GetNoteRequest>
) -> Result<Response<Note>, Status>

Gets the specified note.

pub async fn list_notes<'_>(
&'_ mut self,
request: impl IntoRequest<ListNotesRequest>
) -> Result<Response<ListNotesResponse>, Status>

Lists notes for the specified project.

pub async fn delete_note<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteNoteRequest>
) -> Result<Response<()>, Status>

Deletes the specified note.

pub async fn create_note<'_>(
&'_ mut self,
request: impl IntoRequest<CreateNoteRequest>
) -> Result<Response<Note>, Status>

Creates a new note.

pub async fn batch_create_notes<'_>(
&'_ mut self,
request: impl IntoRequest<BatchCreateNotesRequest>
) -> Result<Response<BatchCreateNotesResponse>, Status>

Creates new notes in batch.

pub async fn update_note<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateNoteRequest>
) -> Result<Response<Note>, Status>

Updates the specified note.

pub async fn list_note_occurrences<'_>(
&'_ mut self,
request: impl IntoRequest<ListNoteOccurrencesRequest>
) -> Result<Response<ListNoteOccurrencesResponse>, Status>

Lists occurrences referencing the specified note. Provider projects can use this method to get all occurrences across consumer projects referencing the specified note.

pub async fn get_vulnerability_occurrences_summary<'_>(
&'_ mut self,
request: impl IntoRequest<GetVulnerabilityOccurrencesSummaryRequest>
) -> Result<Response<VulnerabilityOccurrencesSummary>, Status>

Gets a summary of the number and severity of occurrences.