Module gapi_grpc::grafeas::v1beta1::image

Modules

layer

Structs

Basis
Basis describes the base image portion (Note) of the DockerImage relationship. Linked occurrences are derived from this or an equivalent image via: FROM <Basis.resource_url> Or an equivalent reference, e.g. a tag of the resource_url.

Derived
Derived describes the derived image portion (Occurrence) of the DockerImage relationship. This image would be produced from a Dockerfile with FROM <DockerImage.Basis in attached Note>.

Details
Details of an image occurrence.

Fingerprint
A set of properties that uniquely identify a given Docker image.

Layer
Layer holds metadata specific to a layer of a Docker image.