Module gapi_grpc::grafeas::v1beta1::package

Modules

version

Structs

Details
Details of a package occurrence.

Distribution
This represents a particular channel of distribution for a given package. E.g., Debian's jessie-backports dpkg mirror.

Installation
This represents how a particular software package may be installed on a system.

Location
An occurrence of a particular package installation found within a system's filesystem. E.g., glibc was found in /var/lib/dpkg/status.

Package
This represents a particular package that is distributed over various channels. E.g., glibc (aka libc6) is distributed by many, at various versions.

Version
Version contains structured information about the version of a package.

Enums

Architecture
Instruction set architectures supported by various package managers.