Module gapi_grpc::grafeas::v1beta1::source

Modules

alias_context

cloud_repo_source_context

gerrit_source_context

repo_id

source_context

Structs

AliasContext
An alias to a repo revision.

CloudRepoSourceContext
A CloudRepoSourceContext denotes a particular revision in a Google Cloud Source Repo.

GerritSourceContext
A SourceContext referring to a Gerrit project.

GitSourceContext
A GitSourceContext denotes a particular revision in a third party Git repository (e.g., GitHub).

ProjectRepoId
Selects a repo using a Google Cloud Platform project ID (e.g., winged-cargo-31) and a repo name within that project.

RepoId
A unique identifier for a Cloud Repo.

SourceContext
A SourceContext is a reference to a tree of files. A SourceContext together with a path point to a unique revision of a single file or directory.