Module gapi_grpc::google::cloud::aiplatform::v1beta1::schema

Modules

image_segmentation_annotation

predict

tables_dataset_metadata

time_series_dataset_metadata

trainingjob

Structs

AnnotationSpecColor
An entry of mapping between color and AnnotationSpec. The mapping is used in segmentation mask.

ImageBoundingBoxAnnotation
Annotation details specific to image object detection.

ImageClassificationAnnotation
Annotation details specific to image classification.

ImageDataItem
Payload of Image DataItem.

ImageDatasetMetadata
The metadata of Datasets that contain Image DataItems.

ImageSegmentationAnnotation
Annotation details specific to image segmentation.

TablesDatasetMetadata
The metadata of Datasets that contain tables data.

TextClassificationAnnotation
Annotation details specific to text classification.

TextDataItem
Payload of Text DataItem.

TextDatasetMetadata
The metadata of Datasets that contain Text DataItems.

TextExtractionAnnotation
Annotation details specific to text extraction.

TextSegment
The text segment inside of DataItem.

TextSentimentAnnotation
Annotation details specific to text sentiment.

TimeSegment
A time period inside of a DataItem that has a time dimension (e.g. video).

TimeSeriesDatasetMetadata
The metadata of Datasets that contain time series data.

Vertex
A vertex represents a 2D point in the image. NOTE: the normalized vertex coordinates are relative to the original image and range from 0 to 1.

VideoActionRecognitionAnnotation
Annotation details specific to video action recognition.

VideoClassificationAnnotation
Annotation details specific to video classification.

VideoDataItem
Payload of Video DataItem.

VideoDatasetMetadata
The metadata of Datasets that contain Video DataItems.

VideoObjectTrackingAnnotation
Annotation details specific to video object tracking.