Struct gapi_grpc::google::cloud::channel::v1::cloud_channel_service_client::CloudChannelServiceClient

pub struct CloudChannelServiceClient<T> { /* fields omitted */ }

CloudChannelService enables Google cloud resellers and distributors to manage their customers, channel partners, entitlements and reports.

Using this service:

Resellers or distributors can manage a customer entity.

Distributors can register an authorized reseller in their channel and then enable delegated admin access for the reseller.

Resellers or distributors can manage entitlements for their customers.

The service primarily exposes the following resources:

[Customer][google.cloud.channel.v1.Customer]s: A Customer represents an entity managed by a reseller or distributor. A customer typically represents an enterprise. In an n-tier resale channel hierarchy, customers are generally represented as leaf nodes. Customers primarily have an Entitlement sub-resource discussed below.

[Entitlement][google.cloud.channel.v1.Entitlement]s: An Entitlement represents an entity which provides a customer means to start using a service. Entitlements are created or updated as a result of a successful fulfillment.

[ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink]s: A ChannelPartnerLink is an entity that identifies links between distributors and their indirect resellers in a channel.

Implementations

impl CloudChannelServiceClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> CloudChannelServiceClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn list_customers<'_>(
&'_ mut self,
request: impl IntoRequest<ListCustomersRequest>
) -> Result<Response<ListCustomersResponse>, Status>

List downstream [Customer][google.cloud.channel.v1.Customer]s.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

Return Value:
List of [Customer][google.cloud.channel.v1.Customer]s pertaining to the reseller or empty list if there are none.

pub async fn get_customer<'_>(
&'_ mut self,
request: impl IntoRequest<GetCustomerRequest>
) -> Result<Response<Customer>, Status>

Returns a requested [Customer][google.cloud.channel.v1.Customer] resource.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: If the customer resource doesn't exist. Usually the result of an invalid name parameter.

Return Value:
[Customer][google.cloud.channel.v1.Customer] resource if found, error otherwise.

pub async fn check_cloud_identity_accounts_exist<'_>(
&'_ mut self,
request: impl IntoRequest<CheckCloudIdentityAccountsExistRequest>
) -> Result<Response<CheckCloudIdentityAccountsExistResponse>, Status>

Confirms the existence of Cloud Identity accounts, based on the domain and whether the Cloud Identity accounts are owned by the reseller.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

INVALID_VALUE: Invalid domain value in the request.

NOT_FOUND: If there is no [CloudIdentityCustomerAccount][google.cloud.channel.v1.CloudIdentityCustomerAccount] customer for the domain specified in the request.

Return Value:
List of [CloudIdentityCustomerAccount][google.cloud.channel.v1.CloudIdentityCustomerAccount] resources if any exist for the domain, otherwise an error is returned.

pub async fn create_customer<'_>(
&'_ mut self,
request: impl IntoRequest<CreateCustomerRequest>
) -> Result<Response<Customer>, Status>

Creates a new [Customer][google.cloud.channel.v1.Customer] resource under the reseller or distributor account.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT:

Missing or invalid required parameters in the request.

Domain field value doesn't match the domain specified in primary email.

Return Value:
If successful, the newly created [Customer][google.cloud.channel.v1.Customer] resource, otherwise returns an error.

pub async fn update_customer<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateCustomerRequest>
) -> Result<Response<Customer>, Status>

Updates an existing [Customer][google.cloud.channel.v1.Customer] resource belonging to the reseller or distributor.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: No [Customer][google.cloud.channel.v1.Customer] resource found for the name specified in the request.

Return Value:
If successful, the updated [Customer][google.cloud.channel.v1.Customer] resource, otherwise returns an error.

pub async fn delete_customer<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteCustomerRequest>
) -> Result<Response<()>, Status>

Deletes the given [Customer][google.cloud.channel.v1.Customer] permanently and irreversibly.

Possible Error Codes:

PERMISSION_DENIED: If the account making the request does not own this customer.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

FAILED_PRECONDITION: If the customer has existing entitlements.

NOT_FOUND: No [Customer][google.cloud.channel.v1.Customer] resource found for the name specified in the request.

pub async fn provision_cloud_identity<'_>(
&'_ mut self,
request: impl IntoRequest<ProvisionCloudIdentityRequest>
) -> Result<Response<Operation>, Status>

Creates a Cloud Identity for the given customer using the customer's information or the information provided here, if present.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: If the customer is not found for the reseller.

ALREADY_EXISTS: If the customer's primary email already exists. In this case, retry after changing the customer's primary contact email.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support in this case.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel support in this case.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn list_entitlements<'_>(
&'_ mut self,
request: impl IntoRequest<ListEntitlementsRequest>
) -> Result<Response<ListEntitlementsResponse>, Status>

List [Entitlement][google.cloud.channel.v1.Entitlement]s belonging to a customer.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

Return Value:
List of [Entitlement][google.cloud.channel.v1.Entitlement]s belonging to the customer, or empty list if there are none.

pub async fn list_transferable_skus<'_>(
&'_ mut self,
request: impl IntoRequest<ListTransferableSkusRequest>
) -> Result<Response<ListTransferableSkusResponse>, Status>

List [TransferableSku][google.cloud.channel.v1.TransferableSku]s of a customer based on Cloud Identity ID or Customer Name in the request.

This method is used when a reseller lists the entitlements information of a customer that is not owned. The reseller should provide the customer's Cloud Identity ID or Customer Name.

Possible Error Codes:

PERMISSION_DENIED, due to one of the following reasons:

If the customer doesn't belong to the reseller and no auth token, or an invalid auth token is supplied.

If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

Return Value:
List of [TransferableSku][google.cloud.channel.v1.TransferableSku] for the given customer.

pub async fn list_transferable_offers<'_>(
&'_ mut self,
request: impl IntoRequest<ListTransferableOffersRequest>
) -> Result<Response<ListTransferableOffersResponse>, Status>

List [TransferableOffer][google.cloud.channel.v1.TransferableOffer]s of a customer based on Cloud Identity ID or Customer Name in the request.

This method is used when a reseller gets the entitlement information of a customer that is not owned. The reseller should provide the customer's Cloud Identity ID or Customer Name.

Possible Error Codes:

PERMISSION_DENIED, due to one of the following reasons: (a) If the customer doesn't belong to the reseller and no auth token or invalid auth token is supplied. (b) If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

Return Value: List of [TransferableOffer][google.cloud.channel.v1.TransferableOffer] for the given customer and SKU.

pub async fn get_entitlement<'_>(
&'_ mut self,
request: impl IntoRequest<GetEntitlementRequest>
) -> Result<Response<Entitlement>, Status>

Returns a requested [Entitlement][google.cloud.channel.v1.Entitlement] resource.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: If the entitlement is not found for the customer.

Return Value:
If found, the requested [Entitlement][google.cloud.channel.v1.Entitlement] resource, otherwise returns an error.

pub async fn create_entitlement<'_>(
&'_ mut self,
request: impl IntoRequest<CreateEntitlementRequest>
) -> Result<Response<Operation>, Status>

Creates an entitlement for a customer.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT:

Missing or invalid required parameters in the request.

Cannot purchase an entitlement if there is already an entitlement for customer, for a SKU from the same product family.

INVALID_VALUE: Offer passed in isn't valid. Make sure OfferId is valid. If it is valid, then contact Google Channel support for further troubleshooting.

NOT_FOUND: If the customer or offer resource is not found for the reseller.

ALREADY_EXISTS: This failure can happen in the following cases:

If the SKU has been already purchased for the customer.

If the customer's primary email already exists. In this case retry after changing the customer's primary contact email.

CONDITION_NOT_MET or FAILED_PRECONDITION: This failure can happen in the following cases:

Purchasing a SKU that requires domain verification and the domain has not been verified.

Purchasing an Add-On SKU like Vault or Drive without purchasing the pre-requisite SKU, such as Google Workspace Business Starter.

Applicable only for developer accounts: reseller and resold domain. Must meet the following domain naming requirements:

Domain names must start with goog-test.

Resold domain names must include the reseller domain.

INTERNAL: Any non-user error related to a technical issue in the backend. Contact Cloud Channel Support in this case.

UNKNOWN: Any non-user error related to a technical issue in the backend. Contact Cloud Channel Support in this case.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn change_parameters<'_>(
&'_ mut self,
request: impl IntoRequest<ChangeParametersRequest>
) -> Result<Response<Operation>, Status>

Change parameters of the entitlement

An entitlement parameters update is a long-running operation and results in updates to the entitlement as a result of fulfillment.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request. For example, if the number of seats being changed to is greater than the allowed number of max seats for the resource. Or decreasing seats for a commitment based plan.

NOT_FOUND: Entitlement resource not found.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn change_renewal_settings<'_>(
&'_ mut self,
request: impl IntoRequest<ChangeRenewalSettingsRequest>
) -> Result<Response<Operation>, Status>

Updates the renewal settings for an existing customer entitlement.

An entitlement update is a long-running operation and results in updates to the entitlement as a result of fulfillment.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: Entitlement resource not found.

NOT_COMMITMENT_PLAN: Renewal Settings are only applicable for a commitment plan. Can't enable or disable renewal for non-commitment plans.

INTERNAL: Any non user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn change_offer<'_>(
&'_ mut self,
request: impl IntoRequest<ChangeOfferRequest>
) -> Result<Response<Operation>, Status>

Updates the Offer for an existing customer entitlement.

An entitlement update is a long-running operation and results in updates to the entitlement as a result of fulfillment.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: Offer or Entitlement resource not found.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn start_paid_service<'_>(
&'_ mut self,
request: impl IntoRequest<StartPaidServiceRequest>
) -> Result<Response<Operation>, Status>

Starts paid service for a trial entitlement.

Starts paid service for a trial entitlement immediately. This method is only applicable if a plan has already been set up for a trial entitlement but has some trial days remaining.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: Entitlement resource not found.

FAILED_PRECONDITION/NOT_IN_TRIAL: This method only works for entitlement on trial plans.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn suspend_entitlement<'_>(
&'_ mut self,
request: impl IntoRequest<SuspendEntitlementRequest>
) -> Result<Response<Operation>, Status>

Suspends a previously fulfilled entitlement. An entitlement suspension is a long-running operation.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: Entitlement resource not found.

NOT_ACTIVE: Entitlement is not active.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn cancel_entitlement<'_>(
&'_ mut self,
request: impl IntoRequest<CancelEntitlementRequest>
) -> Result<Response<Operation>, Status>

Cancels a previously fulfilled entitlement. An entitlement cancellation is a long-running operation.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller or if the reseller account making the request and reseller account being queried for are different.

FAILED_PRECONDITION: If there are any Google Cloud projects linked to the Google Cloud entitlement's Cloud Billing subaccount.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: Entitlement resource not found.

DELETION_TYPE_NOT_ALLOWED: Cancel is only allowed for Google Workspace add-ons or entitlements for Google Cloud's development platform.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The response will contain google.protobuf.Empty on success. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn activate_entitlement<'_>(
&'_ mut self,
request: impl IntoRequest<ActivateEntitlementRequest>
) -> Result<Response<Operation>, Status>

Activates a previously suspended entitlement. The entitlement must be in a suspended state for it to be activated. Entitlements suspended for pending ToS acceptance can't be activated using this method. An entitlement activation is a long-running operation and can result in updates to the state of the customer entitlement.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller or if the reseller account making the request and reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: Entitlement resource not found.

SUSPENSION_NOT_RESELLER_INITIATED: Can't activate an entitlement that is pending TOS acceptance. Only reseller initiated suspensions can be activated.

NOT_SUSPENDED: Can't activate entitlements that are already in ACTIVE state. Can only activate suspended entitlements.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn transfer_entitlements<'_>(
&'_ mut self,
request: impl IntoRequest<TransferEntitlementsRequest>
) -> Result<Response<Operation>, Status>

Transfers customer entitlements to new reseller.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: If the customer or offer resource is not found for the reseller.

ALREADY_EXISTS: If the SKU has been already transferred for the customer.

CONDITION_NOT_MET or FAILED_PRECONDITION: This failure can happen in the following cases:

Transferring a SKU that requires domain verification and the domain has not been verified.

Transferring an Add-On SKU like Vault or Drive without transferring the pre-requisite SKU, such as G Suite Basic

Applicable only for developer accounts: reseller and resold domain must follow the domain naming convention as follows:

Domain names must start with goog-test.

Resold domain names must include the reseller domain.

All transferring entitlements must be specified.

INTERNAL: Any non-user error related to a technical issue in the backend. Please contact Cloud Channel Support in this case.

UNKNOWN: Any non-user error related to a technical issue in the backend. Please contact Cloud Channel Support in this case.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn transfer_entitlements_to_google<'_>(
&'_ mut self,
request: impl IntoRequest<TransferEntitlementsToGoogleRequest>
) -> Result<Response<Operation>, Status>

Transfers customer entitlements from current reseller to Google.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: If the customer or offer resource is not found for the reseller.

ALREADY_EXISTS: If the SKU has been already transferred for the customer.

CONDITION_NOT_MET or FAILED_PRECONDITION: This failure can happen in the following cases:

Transferring a SKU that requires domain verification and the domain has not been verified.

Transferring an Add-On SKU like Vault or Drive without purchasing the pre-requisite SKU, such as G Suite Basic

Applicable only for developer accounts: reseller and resold domain must follow the domain naming convention as follows:

Domain names must start with goog-test.

Resold domain names must include the reseller domain.

INTERNAL: Any non-user error related to a technical issue in the backend. Please contact Cloud Channel Support in this case.

UNKNOWN: Any non-user error related to a technical issue in the backend. Please contact Cloud Channel Support in this case.

Return Value:
Long Running Operation ID.

To get the results of the operation, call the GetOperation method of CloudChannelOperationsService. The response will contain google.protobuf.Empty on success. The Operation metadata will contain an instance of [OperationMetadata][google.cloud.channel.v1.OperationMetadata].

pub async fn list_channel_partner_links<'_>(
&'_ mut self,
request: impl IntoRequest<ListChannelPartnerLinksRequest>
) -> Result<Response<ListChannelPartnerLinksResponse>, Status>

List [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink]s belonging to a distributor. To call this method, you must be a distributor.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

Return Value:
If successful, returns the list of [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resources for the distributor account, otherwise returns an error.

pub async fn get_channel_partner_link<'_>(
&'_ mut self,
request: impl IntoRequest<GetChannelPartnerLinkRequest>
) -> Result<Response<ChannelPartnerLink>, Status>

Returns a requested [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource. To call this method, you must be a distributor.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

NOT_FOUND: ChannelPartnerLink resource not found. Results due invalid channel partner link name.

Return Value:
[ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource if found, otherwise returns an error.

pub async fn create_channel_partner_link<'_>(
&'_ mut self,
request: impl IntoRequest<CreateChannelPartnerLinkRequest>
) -> Result<Response<ChannelPartnerLink>, Status>

Initiates a channel partner link between a distributor and a reseller or between resellers in an n-tier reseller channel. To accept the invite, the invited partner should follow the invite_link_uri provided in the response. If the link creation is accepted, a valid link is set up between the two involved parties. To call this method, you must be a distributor.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

ALREADY_EXISTS: If the ChannelPartnerLink sent in the request already exists.

NOT_FOUND: If no Cloud Identity customer exists for domain provided.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
Newly created [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource if successful, otherwise error is returned.

pub async fn update_channel_partner_link<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateChannelPartnerLinkRequest>
) -> Result<Response<ChannelPartnerLink>, Status>

Updates a channel partner link. A distributor calls this method to change a link's status. For example, suspend a partner link. To call this method, you must be a distributor.

Possible Error Codes:

PERMISSION_DENIED: If the reseller account making the request and the reseller account being queried for are different.

INVALID_ARGUMENT:

Missing or invalid required parameters in the request.

Updating link state from invited to active or suspended.

Sending reseller_cloud_identity_id, invite_url or name in update mask.

NOT_FOUND: ChannelPartnerLink resource not found.

INTERNAL: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

UNKNOWN: Any non-user error related to a technical issue in the backend. In this case, contact Cloud Channel support.

Return Value:
If successful, the updated [ChannelPartnerLink][google.cloud.channel.v1.ChannelPartnerLink] resource, otherwise returns an error.

pub async fn list_products<'_>(
&'_ mut self,
request: impl IntoRequest<ListProductsRequest>
) -> Result<Response<ListProductsResponse>, Status>

Lists the Products the reseller is authorized to sell.

Possible Error Codes:

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

pub async fn list_skus<'_>(
&'_ mut self,
request: impl IntoRequest<ListSkusRequest>
) -> Result<Response<ListSkusResponse>, Status>

Lists the SKUs for a product the reseller is authorized to sell.

Possible Error Codes:

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

pub async fn list_offers<'_>(
&'_ mut self,
request: impl IntoRequest<ListOffersRequest>
) -> Result<Response<ListOffersResponse>, Status>

Lists the Offers the reseller can sell.

Possible Error Codes:

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

pub async fn list_purchasable_skus<'_>(
&'_ mut self,
request: impl IntoRequest<ListPurchasableSkusRequest>
) -> Result<Response<ListPurchasableSkusResponse>, Status>

Lists the Purchasable SKUs for following cases:

SKUs that can be newly purchased for a customer

SKUs that can be upgraded/downgraded to, for an entitlement.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller

INVALID_ARGUMENT: Missing or invalid required parameters in the request.

pub async fn list_purchasable_offers<'_>(
&'_ mut self,
request: impl IntoRequest<ListPurchasableOffersRequest>
) -> Result<Response<ListPurchasableOffersResponse>, Status>

Lists the Purchasable Offers for the following cases:

Offers that can be newly purchased for a customer

Offers that can be changed to, for an entitlement.

Possible Error Codes:

PERMISSION_DENIED: If the customer doesn't belong to the reseller

INVALID_ARGUMENT: Missing or invalid required parameters in the request.