Struct gapi_grpc::google::iam::admin::v1::iam_client::IamClient

pub struct IamClient<T> { /* fields omitted */ }

Creates and manages service account objects.

Service account is an account that belongs to your project instead of to an individual end user. It is used to authenticate calls to a Google API.

To create a service account, specify the project_id and account_id for the account. The account_id is unique within the project, and used to generate the service account email address and a stable unique_id.

All other methods can identify accounts using the format projects/{PROJECT_ID}/serviceAccounts/{ACCOUNT}. Using - as a wildcard for the PROJECT_ID will infer the project from the account. The ACCOUNT value can be the email address or the unique_id of the service account.

Implementations

impl IamClient<Channel>

pub async fn connect<D>(dst: D) -> Result<Self, Error> where
D: TryInto<Endpoint>,
D::Error: Into<StdError>,

Attempt to create a new client by connecting to a given endpoint.

impl<T> IamClient<T> where
T: GrpcService<BoxBody>,
T::ResponseBody: Body + HttpBody + Send + 'static,
T::Error: Into<StdError>,
<T::ResponseBody as HttpBody>::Error: Into<StdError> + Send,

pub fn new(inner: T) -> Self

pub fn with_interceptor(inner: T, interceptor: impl Into<Interceptor>) -> Self

pub async fn list_service_accounts<'_>(
&'_ mut self,
request: impl IntoRequest<ListServiceAccountsRequest>
) -> Result<Response<ListServiceAccountsResponse>, Status>

Lists [ServiceAccounts][google.iam.admin.v1.ServiceAccount] for a project.

pub async fn get_service_account<'_>(
&'_ mut self,
request: impl IntoRequest<GetServiceAccountRequest>
) -> Result<Response<ServiceAccount>, Status>

Gets a [ServiceAccount][google.iam.admin.v1.ServiceAccount].

pub async fn create_service_account<'_>(
&'_ mut self,
request: impl IntoRequest<CreateServiceAccountRequest>
) -> Result<Response<ServiceAccount>, Status>

Creates a [ServiceAccount][google.iam.admin.v1.ServiceAccount] and returns it.

pub async fn update_service_account<'_>(
&'_ mut self,
request: impl IntoRequest<ServiceAccount>
) -> Result<Response<ServiceAccount>, Status>

Updates a [ServiceAccount][google.iam.admin.v1.ServiceAccount].

Currently, only the following fields are updatable: display_name and description.

pub async fn delete_service_account<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteServiceAccountRequest>
) -> Result<Response<()>, Status>

Deletes a [ServiceAccount][google.iam.admin.v1.ServiceAccount].

pub async fn list_service_account_keys<'_>(
&'_ mut self,
request: impl IntoRequest<ListServiceAccountKeysRequest>
) -> Result<Response<ListServiceAccountKeysResponse>, Status>

Lists [ServiceAccountKeys][google.iam.admin.v1.ServiceAccountKey].

pub async fn get_service_account_key<'_>(
&'_ mut self,
request: impl IntoRequest<GetServiceAccountKeyRequest>
) -> Result<Response<ServiceAccountKey>, Status>

Gets the [ServiceAccountKey][google.iam.admin.v1.ServiceAccountKey] by key id.

pub async fn create_service_account_key<'_>(
&'_ mut self,
request: impl IntoRequest<CreateServiceAccountKeyRequest>
) -> Result<Response<ServiceAccountKey>, Status>

Creates a [ServiceAccountKey][google.iam.admin.v1.ServiceAccountKey] and returns it.

pub async fn delete_service_account_key<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteServiceAccountKeyRequest>
) -> Result<Response<()>, Status>

Deletes a [ServiceAccountKey][google.iam.admin.v1.ServiceAccountKey].

pub async fn sign_blob<'_>(
&'_ mut self,
request: impl IntoRequest<SignBlobRequest>
) -> Result<Response<SignBlobResponse>, Status>

Signs a blob using a service account's system-managed private key.

pub async fn sign_jwt<'_>(
&'_ mut self,
request: impl IntoRequest<SignJwtRequest>
) -> Result<Response<SignJwtResponse>, Status>

Signs a JWT using a service account's system-managed private key.

If no expiry time (exp) is provided in the SignJwtRequest, IAM sets an an expiry time of one hour by default. If you request an expiry time of more than one hour, the request will fail.

pub async fn get_iam_policy<'_>(
&'_ mut self,
request: impl IntoRequest<GetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Returns the Cloud IAM access control policy for a [ServiceAccount][google.iam.admin.v1.ServiceAccount].

Note: Service accounts are both resources and identities. This method treats the service account as a resource. It returns the Cloud IAM policy that reflects what members have access to the service account.

This method does not return what resources the service account has access to. To see if a service account has access to a resource, call the getIamPolicy method on the target resource. For example, to view grants for a project, call the projects.getIamPolicy method.

pub async fn set_iam_policy<'_>(
&'_ mut self,
request: impl IntoRequest<SetIamPolicyRequest>
) -> Result<Response<Policy>, Status>

Sets the Cloud IAM access control policy for a [ServiceAccount][google.iam.admin.v1.ServiceAccount].

Note: Service accounts are both resources and identities. This method treats the service account as a resource. Use it to grant members access to the service account, such as when they need to impersonate it.

This method does not grant the service account access to other resources, such as projects. To grant a service account access to resources, include the service account in the Cloud IAM policy for the desired resource, then call the appropriate setIamPolicy method on the target resource. For example, to grant a service account access to a project, call the projects.setIamPolicy method.

pub async fn test_iam_permissions<'_>(
&'_ mut self,
request: impl IntoRequest<TestIamPermissionsRequest>
) -> Result<Response<TestIamPermissionsResponse>, Status>

Tests the specified permissions against the IAM access control policy for a [ServiceAccount][google.iam.admin.v1.ServiceAccount].

pub async fn query_grantable_roles<'_>(
&'_ mut self,
request: impl IntoRequest<QueryGrantableRolesRequest>
) -> Result<Response<QueryGrantableRolesResponse>, Status>

Queries roles that can be granted on a particular resource. A role is grantable if it can be used as the role in a binding for a policy for that resource.

pub async fn list_roles<'_>(
&'_ mut self,
request: impl IntoRequest<ListRolesRequest>
) -> Result<Response<ListRolesResponse>, Status>

Lists the Roles defined on a resource.

pub async fn get_role<'_>(
&'_ mut self,
request: impl IntoRequest<GetRoleRequest>
) -> Result<Response<Role>, Status>

Gets a Role definition.

pub async fn create_role<'_>(
&'_ mut self,
request: impl IntoRequest<CreateRoleRequest>
) -> Result<Response<Role>, Status>

Creates a new Role.

pub async fn update_role<'_>(
&'_ mut self,
request: impl IntoRequest<UpdateRoleRequest>
) -> Result<Response<Role>, Status>

Updates a Role definition.

pub async fn delete_role<'_>(
&'_ mut self,
request: impl IntoRequest<DeleteRoleRequest>
) -> Result<Response<Role>, Status>

Soft deletes a role. The role is suspended and cannot be used to create new IAM Policy Bindings. The Role will not be included in ListRoles() unless show_deleted is set in the ListRolesRequest. The Role contains the deleted boolean set. Existing Bindings remains, but are inactive. The Role can be undeleted within 7 days. After 7 days the Role is deleted and all Bindings associated with the role are removed.

pub async fn undelete_role<'_>(
&'_ mut self,
request: impl IntoRequest<UndeleteRoleRequest>
) -> Result<Response<Role>, Status>

Undelete a Role, bringing it back in its previous state.

pub async fn query_testable_permissions<'_>(
&'_ mut self,
request: impl IntoRequest<QueryTestablePermissionsRequest>
) -> Result<Response<QueryTestablePermissionsResponse>, Status>

Lists the permissions testable on a resource. A permission is testable if it can be tested for an identity on a resource.